% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-alt-text.R
\name{add_alt_text}
\alias{add_alt_text}
\title{Add alternative text to figures in the document via the `tagpdf` LaTeX
package}
\usage{
add_alt_text(tex_file = "hake-assessment.tex", alt_fig_text)
}
\arguments{
\item{tex_file}{The name of the TEX file}

\item{alt_fig_text}{A character vector of alternative text to insert}
}
\value{
Nothing, overwrites the `tex_file`
}
\description{
Add alternative text to figures in the document via the `tagpdf` LaTeX
package
}
\details{
Searches the given TEX file for the `includegraphics{}`,
`\\begin{caption}`, `\\end{caption}`, `\\begin{knitrout}`,
`\\end{knitrout}`, and `\\label{fig.*}` and manipulates the
structure of those commands with the figure tagging commands
injected. Alternative text comes from a data frame created
during the knitting process via `knitr` hooks that read in the
text from the `alt.text` knitr chunk tags in the figures. Those
have to be manually changed each year
}
