% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-mountains.R
\name{add_extra_ribbon_data}
\alias{add_extra_ribbon_data}
\title{Add extra data for decreasing-sized polygons which will be used to make
ribbons, one inside the other. See \code{\link[=plot_selex_mountains]{plot_selex_mountains()}}}
\usage{
add_extra_ribbon_data(d, prop_seq)
}
\arguments{
\item{d}{The data frame to add ribbon data to}

\item{prop_seq}{A sequence of decreasing proportion values which
will be applied to the height data iteratively to create smaller
polygon data}
}
\value{
A modified data frame with the same structure as \code{d}, but with
many more rows which contain the extra ribbon polygon data
}
\description{
Add extra data for decreasing-sized polygons which will be used to make
ribbons, one inside the other. See \code{\link[=plot_selex_mountains]{plot_selex_mountains()}}
}
