% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-add-major-ticks.R
\name{add_major_ticks}
\alias{add_major_ticks}
\title{Create a vector of labels to use when major and minor tick marks are
to be used in a plot, and the custom tick data frame to be used to draw
the lines outside the ggplot panel area using \code{\link[ggplot2:geom_linerange]{ggplot2::geom_linerange()}}}
\usage{
add_major_ticks(
  g,
  x_breaks = NULL,
  modulo = 5,
  prop = 0.8,
  major_tick_thickness = 0.5,
  ...
)
}
\arguments{
\item{g}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object}

\item{x_breaks}{A vector of values which are where all the x-axis ticks
are located on the plot \code{g}. If \code{NULL}, this will become a sequence from
the lower limit of the x-axis to the upper limit, by 1.}

\item{modulo}{The number to use as a modulus so that every n-th tick mark
is a major tick}

\item{prop}{A value between 0 and 1 to multiply the length of the major
tick lines by. This is used in case the full length ticks are overlapping
the tick labels}

\item{major_tick_thickness}{The thickness of the major tick marks}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object, the original plot \code{g} with x-axis
ticks modified to be minor and major ticks
}
\description{
Create a vector of labels to use when major and minor tick marks are
to be used in a plot, and the custom tick data frame to be used to draw
the lines outside the ggplot panel area using \code{\link[ggplot2:geom_linerange]{ggplot2::geom_linerange()}}
}
