% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-add-newlines.R
\name{add_newlines}
\alias{add_newlines}
\title{Replace '+' with a newline in the given string}
\usage{
add_newlines(x, ...)
}
\arguments{
\item{x}{A character string}

\item{...}{Absorbs other arguments not meant for this function}
}
\value{
A modified character string
}
\description{
Replace '+' with a newline in the given string
}
\details{
Used mainly for ggplot y axis labels, if they are too long
and get cut off
}
