% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addpoly}
\alias{addpoly}
\title{Add a polygon to a plot}
\usage{
addpoly(yrvec, lower, upper, color = 1, shade.col = NA)
}
\arguments{
\item{yrvec}{A vector of years}

\item{lower}{A vector of lower CI values}

\item{upper}{A vector of upper CI values}

\item{color}{The color to make the polygon lines}

\item{shade.col}{The shade color to fill in the polygon with}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Add a polygon to a plot
}
