% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_bubble_cohorts}
\alias{age_bubble_cohorts}
\title{A vector of years when larger-than-normal recruitment took place, used
only for plots showing age bubbles, for drawing the diagonal cohort lines}
\format{
A vector of numerical years
}
\usage{
age_bubble_cohorts
}
\description{
A vector of years when larger-than-normal recruitment took place, used
only for plots showing age bubbles, for drawing the diagonal cohort lines
}
\keyword{datasets}
