% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_diag_line_type}
\alias{age_diag_line_type}
\title{The default line type to use for diagonal cohort lines on age bubble plots}
\format{
A character string or numerical value representing an R line type
}
\usage{
age_diag_line_type
}
\description{
The default line type to use for diagonal cohort lines on age bubble plots
}
\keyword{datasets}
