% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_diag_line_width}
\alias{age_diag_line_width}
\title{The default line width to use for diagonal cohort lines on age bubble plots}
\format{
A numerical value
}
\usage{
age_diag_line_width
}
\description{
The default line width to use for diagonal cohort lines on age bubble plots
}
\keyword{datasets}
