% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_diag_linecolor}
\alias{age_diag_linecolor}
\title{The default color to use for diagonal cohort lines on age bubble plots}
\format{
A character string or numerical value representing an R color
}
\usage{
age_diag_linecolor
}
\description{
The default color to use for diagonal cohort lines on age bubble plots
}
\keyword{datasets}
