% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_fillcolor}
\alias{age_fillcolor}
\title{The default fill color to use for bubbles in age bubble plots}
\format{
A character string or numerical value representing an R color
}
\usage{
age_fillcolor
}
\description{
The default fill color to use for bubbles in age bubble plots
}
\keyword{datasets}
