% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-retros.R
\name{append_retros}
\alias{append_retros}
\title{Append retrospective runs to the already-built RDS file and re-save it}
\usage{
append_retros(rds_fn, retrospective_yrs = 1:10)
}
\arguments{
\item{rds_fn}{RDS filename}

\item{retrospective_yrs}{a vector of years to include in the loading}
}
\value{
Nothing
}
\description{
Append retrospective runs to the already-built RDS file and re-save it
}
