% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assess_changes_df}
\alias{assess_changes_df}
\title{A data frame containing information on changes in the assessment from
year to year}
\format{
A data frame with 16 rows and 8 variables:
\describe{
\item{Year}{Year of assessment}
\item{Framework}{Modeling framework used}
\item{Survey}{Was there a new year of survey data}
\item{Comp_Method}{Age-composition weighting method}
\item{Comp_Fishery}{Parameter estimate/value for the fishery}
\item{Comp_Survey}{Parameter estimate/value for the survey}
\item{MCMC}{Number of MCMC samples used for all results}
\item{Change}{Description of tchanges made in thee assessment year}
}
}
\usage{
assess_changes_df
}
\description{
A data frame containing information on changes in the assessment from
year to year
}
\keyword{datasets}
