% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assess_history_df}
\alias{assess_history_df}
\title{A data frame containing information on estimated biomass trajectories for
historical assessments from 1991 to present}
\format{
A data frame with 57 rows and 70 variables:
\describe{
\item{yr}{Assessment year}
\item{fishery_independent_cpue}{Toggle for if the index used was fishery-independent}
\item{author}{Assessment authors}
\item{model}{Type of model}
\item{value}{Biomass estimate specifics}
\item{model_type}{Type of model, eg. weight-at-age or growth}
\item{1960}{Biomass value for the year. There are additional columns, one for each year}
}
}
\usage{
assess_history_df
}
\description{
A data frame containing information on estimated biomass trajectories for
historical assessments from 1991 to present
}
\keyword{datasets}
