% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assess_history_probs_df}
\alias{assess_history_probs_df}
\title{A data frame containing information on probability of biomass decline
or of being below B40\% for hindcasts of the last 10 years}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{Year}{Year of probabilities}
\item{P_decline}{Probability of decline from previous year}
\item{P_below_B40}{Probability the biomass is below B40\%}
}
}
\usage{
assess_history_probs_df
}
\description{
A data frame containing information on probability of biomass decline
or of being below B40\% for hindcasts of the last 10 years
}
\keyword{datasets}
