% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-age.R
\name{atage_table}
\alias{atage_table}
\title{Makes a table of the estimated -at-age values for 5 different values:
Numbers-at-age, Catch-at-age, Biomass-at-age, Exploitation-at-age,
and Catch-at-age-biomass}
\usage{
atage_table(
  model,
  start_yr = NA,
  end_yr = NA,
  table_type = 1,
  digits = 0,
  csv_dir = "out-csv",
  xcaption = "default",
  xlabel = "default",
  font_size = 9,
  space_size = 10
)
}
\arguments{
\item{model}{A model in this project}

\item{start_yr}{Start year for the table}

\item{end_yr}{End year for the table}

\item{table_type}{1 = Numbers-at-age, 2 = Exploitation-rate-at-age, 3 = Catch-at-age-number
4 = Catch-at-age-biomass, 5 = Biomass-at-age}

\item{digits}{Number of decimal points}

\item{csv_dir}{Directory for CSV output}

\item{xcaption}{Table caption}

\item{xlabel}{The label used to reference the table in latex}

\item{font_size}{Size of the font for the table}

\item{space_size}{Size of the vertical spaces for the table}
}
\value{
An [xtable::xtable()]
}
\description{
Makes a table of the estimated -at-age values for 5 different values:
Numbers-at-age, Catch-at-age, Biomass-at-age, Exploitation-at-age,
and Catch-at-age-biomass
}
