% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{axis_label_color}
\alias{axis_label_color}
\title{The color to use by default for all axis titles}
\format{
A numerical value or character string representing an official
color in R
}
\usage{
axis_label_color
}
\description{
The color to use by default for all axis titles
}
\keyword{datasets}
