% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{axis_tick_font_size}
\alias{axis_tick_font_size}
\title{The size font to use by default for all axis tick labels}
\format{
A numerical value
}
\usage{
axis_tick_font_size
}
\description{
The size font to use by default for all axis tick labels
}
\keyword{datasets}
