% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-beta.R
\name{beta_get_shape_params}
\alias{beta_get_shape_params}
\title{Calculate the two shape parameters from the mean (\code{mu}) and \code{cv}
which is what is in the SS files.}
\usage{
beta_get_shape_params(mu, cv)
}
\arguments{
\item{mu}{The mean of the Beta distribution}

\item{cv}{The CV of the Beta distribution}
}
\value{
A vector of two values, the two shape parameters, also know as
\code{shape1} and \code{shape2} by the \code{stats::dbeta()} function
}
\description{
Calculate the two shape parameters from the mean (\code{mu}) and \code{cv}
which is what is in the SS files.
}
\details{
To make a plot of the Beta function based on values from SS3 which are
labelled \code{prior} and \code{pr_sd} in the control file, do this:
\code{s1 <- beta_get_shape_params(prior, pr_sd)[1]}
\code{s2 <- beta_get_shape_params(prior, pr_sd)[2]}
\verb{ggplot() +}
\verb{geom_function(fun = stats::dbeta,}
\verb{args = list(shape1 = s1, shape2 = s2, log = FALSE))}
}
