% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bind-cols-quiet.R
\name{bind_cols_quiet}
\alias{bind_cols_quiet}
\title{Wrapper for \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}} but silences the New names... messages
that appear when binding}
\usage{
bind_cols_quiet(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}}}
}
\value{
The output data frame from \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}}
}
\description{
Wrapper for \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}} but silences the New names... messages
that appear when binding
}
