% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-doc.R
\name{build_adnuts_doc}
\alias{build_adnuts_doc}
\title{Build the ADNUT diagnostics document entirely from within an R session}
\usage{
build_adnuts_doc(knit_only = FALSE, doc_name = "adnuts-diagnostics")
}
\arguments{
\item{knit_only}{Only knit the code, do not run latex engine}

\item{doc_name}{What to name the document (no extension)}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Build the ADNUT diagnostics document entirely from within an R session
}
\details{
Make sure you have created the .rds files by running \code{\link[=build_rds]{build_rds()}} in the appropriate manner.
Once you have done that and run this function once within an R session, it will be a little quicker since the RDS
file contents will have already been loaded into the R session. Only has the
PNG option since EPS files produce a huge final PDF; no need for bibtex
since no references. If those are needed then just adapt \code{build_doc} again.
}
