% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-doc.R
\name{build_doc}
\alias{build_doc}
\title{Build the Hake assessment document}
\usage{
build_doc(tag = FALSE, ...)
}
\arguments{
\item{tag}{Logical. If \code{TRUE}, add tagging commands to the
tex file so that a tagged, web-accessible document is built.
If \code{TRUE}, it takes longer for the \code{lualatex} step so typically this
is only use for the final build (or testing along the way)}
}
\value{
Nothing
}
\description{
Build the Hake assessment document
}
