% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-doc.R
\name{build_es}
\alias{build_es}
\title{Build the executive summary section only}
\usage{
build_es()
}
\arguments{
\item{doc_name}{What to name the document (no extension needed)}
}
\value{
[base::invisible()]
}
\description{
Build the executive summary section only
}
