% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.mcmc}
\alias{calc.mcmc}
\title{Return a list of mcmc calculations, e.g. quantiles for various values}
\usage{
calc.mcmc(
  mcmc,
  lower = 0.025,
  upper = 0.975,
  biomass.scale = 1e+06,
  recruitment.scale = 1e+06
)
}
\arguments{
\item{mcmc}{The output of the [r4ss::SS_getMCMC()] function as a data.frame}

\item{lower}{Lower quantile value}

\item{upper}{Upper quantile value}

\item{biomass.scale}{Scale the biomass by this amount. The default for
2021 and before was 2e6, but is not 1e6 due to changes in SS3 (hake Issue
 #866). Biomass will be shown in the millions of tonnes (and is females only).}

\item{recruitment.scale}{Scale the recruitment by this amount. The default is 1e6
because recruitment will be shown in millions of tonnes}
}
\value{
A named list of MCMC outputs
}
\description{
Return a list of mcmc calculations, e.g. quantiles for various values
}
