% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-fishery-ages.R
\name{calc_fishery_ages}
\alias{calc_fishery_ages}
\title{Load data and create age compositions for the fishery}
\usage{
calc_fishery_ages(weight_at_age)
}
\arguments{
\item{weight_at_age}{A data frame read in by \code{\link[r4ss:SS_readwtatage]{r4ss::SS_readwtatage()}} for
the current year's model.}
}
\description{
Load data and create age compositions for the fishery
}
\details{
TODO:
\enumerate{
\item Weight-at-age is summed over all ages 0--20 but ages are not available
for 16+?
}
}
\author{
Kelli F. Johnson
}
