% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-probs.R
\name{calc_historical_probs}
\alias{calc_historical_probs}
\title{Calculate probabilities from an MCMC model to compare with estimates
from past assessments}
\usage{
calc_historical_probs(model, start = 2012, end)
}
\arguments{
\item{model}{A model as loaded by [load_ss_files()]}

\item{start}{First assessment year to do comparisons}

\item{end}{Final assessment year to do comparisons}
}
\value{
Data frame with columns
* `year` - year (lower case to check when binding columnsin
  [combine_historical_probs()]
* `P_decline_curr` - current estimate (from `model`) of the probability that
  the spawning biomass declined from year to year + 1
* `P_below_B40_curr` - current estimate (from `model`) of the probability that
  the spawning biomass was below B_40 in year + 1
}
\description{
Calculate probabilities from an MCMC model to compare with estimates
from past assessments
}
\details{
Calculate a model's estimate of P(B_t+1 < B_t) and P(B_t+1 < B_40%)
to then compare with the estimates of those quantities in year t's
assessment, in [combine_historical_probs()].
}
