% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-retro-stats.R
\name{calc_retro_stats}
\alias{calc_retro_stats}
\title{Add a new year of biomass dispersion stats to the CSV file holding all
years and return the data frame}
\usage{
calc_retro_stats(
  model,
  yr = year(Sys.Date()),
  fn = here("data-tables", "assessment-history-dispersion.csv")
)
}
\arguments{
\item{model}{A model object}

\item{yr}{The year to calculate the statistics for}
}
\value{
The contents of the \verb{data-tables/"assessment-history-dispersion.csv}
file, after adding a new year if necessary
}
\description{
Add a new year of biomass dispersion stats to the CSV file holding all
years and return the data frame
}
\details{
Adds the current year to the CSV file if it is not present in it
}
