% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-risk.R
\name{calc_risk}
\alias{calc_risk}
\title{Calculate the probabilities of being under several reference points from
one forecast year to the next}
\usage{
calc_risk(model, ...)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A list of length 1 less than the number of forecast years. Each
element is a data.frame of catch levels holding the probabilities. For
example, list element 1 will hold the probabilities for each catch.level
of being under several reference points for the first two years in the
forecast_yrs vector. If forecast.outputs is NA, NA will be returned,
otherwise the risk.list will be returned
}
\description{
Calculate the probabilities of being under several reference points from
one forecast year to the next
}
