% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_sd_of_devs}
\alias{calc_sd_of_devs}
\title{Calculate the standard deviation of recruitment deviations in an MCMC
data frame}
\usage{
calc_sd_of_devs(posteriors, pattern = "Main")
}
\arguments{
\item{posteriors}{The MCMC data frame as generated by \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}}}

\item{pattern}{The regular expression used to match the columns}
}
\value{
A formatted string representing the standard deviation of the
recruitment deviations
}
\description{
Calculate the standard deviation of recruitment deviations in an MCMC
data frame
}
