% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-sum-rec-devs.R
\name{calc_sum_rec_devs}
\alias{calc_sum_rec_devs}
\title{Calculate the men, median, and range of the recruitment deviations in an
MCMC data frame}
\usage{
calc_sum_rec_devs(posteriors, pattern = "Main")
}
\arguments{
\item{posteriors}{The MCMC data frame as generated by \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}}}

\item{pattern}{The regular expression used to match the columns}
}
\value{
A list of the mean, median, and range of the recruitment deviations
}
\description{
Calculate the men, median, and range of the recruitment deviations in an
MCMC data frame
}
