% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-yoob.R
\name{calc_yoob}
\alias{calc_yoob}
\title{Calculate the data outside the range of the y limits and change the
\code{lo_col} and \code{hi_col} in the data frame to be equal to the limits for
those points below or above.  YOOB means Y-axis Out-Of-Bounds}
\usage{
calc_yoob(d, ylim, lo_col, med_col, hi_col, show_arrows = TRUE)
}
\arguments{
\item{d}{A data frame containing \code{lo_col}, \code{med_col}, and \code{hi_col}}

\item{ylim}{A vector of two values, a lower limit and an upper limits}

\item{lo_col}{The name of the column representing the lower values}

\item{med_col}{The name of the column representing the median values}

\item{hi_col}{The name of the column representing the upper values}
}
\value{
A list of length up to three,, containing the modified data
frame \code{d}, the data frame containing data below the lower \code{ylim} value
(\code{d_outside_lo}), and the data frame containing data above the higher
\code{ylim} value (\code{d_outside_hi}). If no data are below or above the \code{ylim}
values, \code{d_outside_lo} and/or \code{d_outside_hi} may not be present
}
\description{
Calculate the data outside the range of the y limits and change the
\code{lo_col} and \code{hi_col} in the data frame to be equal to the limits for
those points below or above.  YOOB means Y-axis Out-Of-Bounds
}
