% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ages_lst}
\alias{can_ages_lst}
\title{A data frame containing information on amount of sampling done by year}
\format{
A data frame with 48 rows and 11 variables:
\enumerate{
\item{Shoreside age comps}
{A dataframe with 28 rows and 15 columns}
\describe{
\item{1}{Proportion of age 1 for the given year (rowname)}
\item{2}{Proportion of age 2 for the given year (rowname)}
\item{3}{Proportion of age 3 for the given year (rowname)}
\item{4}{Proportion of age 4 for the given year (rowname)}
\item{5}{Proportion of age 5 for the given year (rowname)}
\item{6}{Proportion of age 6 for the given year (rowname)}
\item{7}{Proportion of age 7 for the given year (rowname)}
\item{8}{Proportion of age 8 for the given year (rowname)}
\item{9}{Proportion of age 9 for the given year (rowname)}
\item{10}{Proportion of age 10 for the given year (rowname)}
\item{11}{Proportion of age 11 for the given year (rowname)}
\item{12}{Proportion of age 12 for the given year (rowname)}
\item{13}{Proportion of age 13 for the given year (rowname)}
\item{14}{Proportion of age 14 for the given year (rowname)}
\item{15}{Proportion of age 15 for the given year (rowname)}
}
\item{Freezer trawler age comps}
{A dataframe with 17 rows and 15 columns}
\describe{
\item{1}{Proportion of age 1 for the given year (rowname)}
\item{2}{Proportion of age 2 for the given year (rowname)}
\item{3}{Proportion of age 3 for the given year (rowname)}
\item{4}{Proportion of age 4 for the given year (rowname)}
\item{5}{Proportion of age 5 for the given year (rowname)}
\item{6}{Proportion of age 6 for the given year (rowname)}
\item{7}{Proportion of age 7 for the given year (rowname)}
\item{8}{Proportion of age 8 for the given year (rowname)}
\item{9}{Proportion of age 9 for the given year (rowname)}
\item{10}{Proportion of age 10 for the given year (rowname)}
\item{11}{Proportion of age 11 for the given year (rowname)}
\item{12}{Proportion of age 12 for the given year (rowname)}
\item{13}{Proportion of age 13 for the given year (rowname)}
\item{14}{Proportion of age 14 for the given year (rowname)}
\item{15}{Proportion of age 15 for the given year (rowname)}
}
\item{Joint-venture age comps}
{A dataframe with 19 rows and 15 columns}
\describe{
\item{1}{Proportion of age 1 for the given year (rowname)}
\item{2}{Proportion of age 2 for the given year (rowname)}
\item{3}{Proportion of age 3 for the given year (rowname)}
\item{4}{Proportion of age 4 for the given year (rowname)}
\item{5}{Proportion of age 5 for the given year (rowname)}
\item{6}{Proportion of age 6 for the given year (rowname)}
\item{7}{Proportion of age 7 for the given year (rowname)}
\item{8}{Proportion of age 8 for the given year (rowname)}
\item{9}{Proportion of age 9 for the given year (rowname)}
\item{10}{Proportion of age 10 for the given year (rowname)}
\item{11}{Proportion of age 11 for the given year (rowname)}
\item{12}{Proportion of age 12 for the given year (rowname)}
\item{13}{Proportion of age 13 for the given year (rowname)}
\item{14}{Proportion of age 14 for the given year (rowname)}
\item{15}{Proportion of age 15 for the given year (rowname)}
}
\item{Shoreside number of trips sampled for age}
{A vector of number of samples by year}
\item{Freezer trawlers number of trips sampled for age}
{A vector of number of samples by year}
\item{Joint-venture number of trips sampled for age}
{A vector of number of samples by year}
}
}
\usage{
can_ages_lst
}
\description{
A data frame containing information on amount of sampling done by year
}
\keyword{datasets}
