% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_catch_by_year_fn}
\alias{can_catch_by_year_fn}
\title{The filename for the Canadian catch by year}
\format{
A character string
}
\usage{
can_catch_by_year_fn
}
\description{
The filename for the Canadian catch by year
}
\keyword{datasets}
