% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ft_age_df}
\alias{can_ft_age_df}
\title{A data frame containing information about the proportions  of fish
by age and year for the Freezer trawlers in Canada}
\format{
A data frame with 17 rows and 15 variables:
\describe{
\item{1}{Proportion of age 1 for the given year (rowname)}
\item{2}{Proportion of age 2 for the given year (rowname)}
\item{3}{Proportion of age 3 for the given year (rowname)}
\item{4}{Proportion of age 4 for the given year (rowname)}
\item{5}{Proportion of age 5 for the given year (rowname)}
\item{6}{Proportion of age 6 for the given year (rowname)}
\item{7}{Proportion of age 7 for the given year (rowname)}
\item{8}{Proportion of age 8 for the given year (rowname)}
\item{9}{Proportion of age 9 for the given year (rowname)}
\item{10}{Proportion of age 10 for the given year (rowname)}
\item{11}{Proportion of age 11 for the given year (rowname)}
\item{12}{Proportion of age 12 for the given year (rowname)}
\item{13}{Proportion of age 13 for the given year (rowname)}
\item{14}{Proportion of age 14 for the given year (rowname)}
\item{15}{Proportion of age 15 for the given year (rowname)}
}
}
\usage{
can_ft_age_df
}
\description{
A data frame containing information about the proportions  of fish
by age and year for the Freezer trawlers in Canada
}
\keyword{datasets}
