% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ft_catch_by_month_fn}
\alias{can_ft_catch_by_month_fn}
\title{The filename for the Freezer trawler catch by year and month data file}
\format{
A character string
}
\usage{
can_ft_catch_by_month_fn
}
\description{
The filename for the Freezer trawler catch by year and month data file
}
\keyword{datasets}
