% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ft_num_fish}
\alias{can_ft_num_fish}
\title{A data frame containing information about the number of fish sampled for
age by year for the Freezer trawlers in Canada}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{year}{Year of sample}
\item{num_fish}{The number of fish sampled for each year}
}
}
\usage{
can_ft_num_fish
}
\description{
A data frame containing information about the number of fish sampled for
age by year for the Freezer trawlers in Canada
}
\keyword{datasets}
