% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_raw_ft_age_counts_fn}
\alias{can_raw_ft_age_counts_fn}
\title{The filename for the Freezer trawler raw age counts by year data file}
\format{
A character string
}
\usage{
can_raw_ft_age_counts_fn
}
\description{
The filename for the Freezer trawler raw age counts by year data file
}
\keyword{datasets}
