% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_sample_data_rds_fn}
\alias{can_sample_data_rds_fn}
\title{The filename for the RDS file containing hake sample data as extracted
using \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}
\format{
A character string
}
\usage{
can_sample_data_rds_fn
}
\description{
The filename for the RDS file containing hake sample data as extracted
using \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}
}
\keyword{datasets}
