% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ss_age_props_fn}
\alias{can_ss_age_props_fn}
\title{The filename for the Shoreside age proportions by year data file}
\format{
A character string
}
\usage{
can_ss_age_props_fn
}
\description{
The filename for the Shoreside age proportions by year data file
}
\keyword{datasets}
