% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ss_catch_by_month_df}
\alias{can_ss_catch_by_month_df}
\title{A data frame containing catch information on Canadian Shoreside vessels
by month and year}
\format{
A data frame with 16 rows and 13 variables:
\describe{
\item{Year}{Year of samples}
\item{1}{January catch in tonnes}
\item{2}{February catch in tonnes}
\item{3}{March catch in tonnes}
\item{4}{April catch in tonnes}
\item{5}{May catch in tonnes}
\item{6}{June catch in tonnes}
\item{7}{July catch in tonnes}
\item{8}{August catch in tonnes}
\item{9}{September catch in tonnes}
\item{10}{October catch in tonnes}
\item{11}{November catch in tonnes}
\item{12}{December catch in tonnes}
}
}
\usage{
can_ss_catch_by_month_df
}
\description{
A data frame containing catch information on Canadian Shoreside vessels
by month and year
}
\keyword{datasets}
