% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_ss_gear_depth_df}
\alias{can_ss_gear_depth_df}
\title{A data frame containing information about the depths of the fishing gear
for catch by year for Canadian Shoreside vessels}
\format{
A data frame with 16 rows and 6 variables:
\describe{
\item{year}{Year of survey}
\item{lower95}{The lower 95\% percentile (2.5th)}
\item{lowerhinge}{The lower hinge value (bottom IQR)}
\item{median}{The median depth value}
\item{upperhinge}{The upper hinge value (upper IQR)}
\item{upper95}{The upper 95\% percentile (97.5th)}
}
}
\usage{
can_ss_gear_depth_df
}
\description{
A data frame containing information about the depths of the fishing gear
for catch by year for Canadian Shoreside vessels
}
\keyword{datasets}
