% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_waa_fn}
\alias{can_waa_fn}
\title{The filename for the Canadian weight-at-age CSV file}
\format{
A character string
}
\usage{
can_waa_fn
}
\description{
The filename for the Canadian weight-at-age CSV file
}
\keyword{datasets}
