% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-create-commercial-waa.R
\name{canada_create_commercial_waa}
\alias{canada_create_commercial_waa}
\title{create the weight-at-age CSV file for Canadian commercial data}
\usage{
canada_create_commercial_waa(
  d,
  weight_scale = 1000,
  num_sex = 2,
  ret_df = FALSE
)
}
\arguments{
\item{d}{A list as returned by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{weight_scale}{A value to divide the weights by}

\item{num_sex}{Either 1 or 2. If 1, then fill the column \code{Sex} with \code{U}
for every record. If 2, use the data to fill the \code{Sex} column with
\code{M}, \code{F}, or \code{U} depending on input values}

\item{ret_df}{Logical. If \code{TRUE}, return the weight-at-age data frame.
If \code{FALSE}, write the data frame to the Canadian weight-at-age CSV file}
}
\value{
a data frame with \code{Source} set to \code{CAN_freezer}, \code{CAN_shoreside},
\code{CAN_jv}, and \code{CAN_polish} for the four fishery types
}
\description{
create the weight-at-age CSV file for Canadian commercial data
}
\details{
No filtering is performed in this function.
Creates a CSV file containing the commercial weight-at-age
}
