% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-canada-create-fleet-month-catch-csv-files.R
\name{canada_create_fleet_month_catch_csv_files}
\alias{canada_create_fleet_month_catch_csv_files}
\title{Create CSV files with catch by year and month for Canadian fleets}
\usage{
canada_create_fleet_month_catch_csv_files(lst, catch_scale = 1000, digits = 3)
}
\arguments{
\item{lst}{A list of data frames as returned by
\code{\link[=canada_extract_fleet_catch]{canada_extract_fleet_catch()}}}

\item{catch_scale}{A value to divide the catch by prior to writing to
the file}

\item{digits}{The number of decimal points to report on the catches in
the csv files}
}
\value{
Nothing, creates three CSV files
}
\description{
Create CSV files with catch by year and month for Canadian fleets
}
