% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-canada-create-fleet-year-catch-csv-files.R
\name{canada_create_fleet_year_catch_csv_files}
\alias{canada_create_fleet_year_catch_csv_files}
\title{Calculate the catch by year and fishery for Canadian
fisheries}
\usage{
canada_create_fleet_year_catch_csv_files(
  lst,
  catch_scale = 1000,
  write_file = TRUE,
  digits = 2
)
}
\arguments{
\item{lst}{A list as returned by \code{\link[=canada_extract_fleet_catch]{canada_extract_fleet_catch()}}}

\item{catch_scale}{A value to divide the catch by prior to writing to
the file}

\item{write_file}{If \code{TRUE}, write the output to the data file given by
the package data variable \code{can_catch_by_year_fn}. If \code{FALSE}, return the
data frame}

\item{digits}{The number of decimal places to include in the table output}
}
\value{
If \code{write_file} is \code{TRUE}, nothing. If \code{write_file} is \code{FALSE},
a data frame containing the catch by year (rows) and fishery type (columns)
}
\description{
Calculate the catch by year and fishery for Canadian
fisheries
}
