% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-extract-depth-data-from-db.R
\name{canada_extract_depth_data_from_db}
\alias{canada_extract_depth_data_from_db}
\title{Extract the Canadian depth and fishing event data}
\usage{
canada_extract_depth_data_from_db(
  fn = file.path(can_sample_dr, can_depths_rds_fn),
  ret_df = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{ret_df}{Logical. If \code{TRUE}, return the data frame from the function
and do not write a csv file. If \code{FALSE}, write the csv file and return
nothing}

\item{overwrite}{Logical. Overwrite the RDS file for if it exists}

\item{dr}{The directory to save the depth RDS file in}
}
\value{
If \code{ret_df} is \code{TRUE}, return the depth data frame. If \code{ret_df}
is \code{FALSE}, return nothing, write an RDS file instead
}
\description{
Extract the Canadian depth and fishing event data from the \code{GFBIOSQL}
database. The data frame returned will contain the \code{fishing_event_id}
column along  with six depth columns.
}
