% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-extract-fleet-catch.R
\name{canada_extract_fleet_catch}
\alias{canada_extract_fleet_catch}
\title{Merge Canadian DMP and LOGS catch data and extract into 3 data frames,
for Freezer trawlers, Shoreside, and Joint-venture}
\usage{
canada_extract_fleet_catch(lst)
}
\arguments{
\item{lst}{a list of three data frames as returned by
\code{\link[=load_catch_data_canada]{load_catch_data_canada()}}}
}
\value{
A list of three data frames, one for each gear: Freezer trawlers,
Shoreside, and Joint Venture
}
\description{
Merge Canadian DMP and LOGS catch data and extract into 3 data frames,
for Freezer trawlers, Shoreside, and Joint-venture
}
