% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-extract-sample-data.R
\name{canada_extract_sample_data}
\alias{canada_extract_sample_data}
\title{Fetch the sample data from the GFBIOSQL database
The data will be filtered to only include hake major and minor areas}
\usage{
canada_extract_sample_data(
  dr = "/srv/hake/other/samples",
  overwrite = FALSE,
  species_code = 225
)
}
\arguments{
\item{dr}{The directory to save the samples RDS file in}

\item{overwrite}{Logical. Overwrite the RDS file for sample data if
it exists}

\item{species_code}{The species code from the GFBio database, or
species name if you're unsure}
}
\description{
Fetch the sample data from the GFBIOSQL database
The data will be filtered to only include hake major and minor areas
}
