% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-extract-samples-from-db.R
\name{canada_extract_samples_from_db}
\alias{canada_extract_samples_from_db}
\title{Extract the Canadian biological sample data}
\usage{
canada_extract_samples_from_db(
  dr = "/srv/hake/other/samples",
  ret_df = FALSE,
  overwrite = FALSE,
  unsorted_only = FALSE,
  species_code = 225
)
}
\arguments{
\item{dr}{The directory to save the samples RDS file in}

\item{ret_df}{Logical. If \code{TRUE}, return the data frame from the function
and do not write a csv file. If \code{FALSE}, write the csv file and return
nothing}

\item{overwrite}{Logical. Overwrite the RDS file for sample data if
it exists}

\item{unsorted_only}{Only include unsorted samples. For hake in Canada,
this needs to be \code{FALSE} or some years will show no ages at all.
See \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}} for additional information}

\item{species_code}{The DFO Groundfish database species code;
default: 225 (hake)}
}
\value{
If \code{ret_df} is \code{TRUE}, return the sample data frame. If \code{ret_df}
is \code{FALSE}, return nothing, write an RDS file instead
}
\description{
Extract the Canadian biological sample data from the \code{GFBIOSQL} database
The data frame returned will be raw and unfiltered
}
