% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-extract-sampling-by-month.R
\name{canada_extract_sampling_by_month}
\alias{canada_extract_sampling_by_month}
\title{Extract the number of age, length, and weight samples by fleet and
month for a given year}
\usage{
canada_extract_sampling_by_month(fleet_lst, yr)
}
\arguments{
\item{fleet_lst}{A list of fleet samples. This list's format must match
the output of \code{\link[=canada_get_fleet_samples]{canada_get_fleet_samples()}}}

\item{yr}{The year to extract by-month sampling summary for}
}
\value{
A list of the same length as \code{fleet_lst} of data frames, explained
in details above
}
\description{
Used to make data frames for tables that show these summary totals
}
\details{
Creates a data frame of \code{month}, \code{num_age}, \code{num_length}, and \code{num_weight}
for each data frame in the \code{fleet_lst} list.
}
