% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-get-age-proportions.R
\name{canada_get_age_proportions}
\alias{canada_get_age_proportions}
\title{Get the age data from the sample data}
\usage{
canada_get_age_proportions(
  d,
  min_date = as.Date("1972-01-01"),
  raw_proportions = FALSE,
  plus_grp = 15,
  lw_cutoff = 10,
  lw_tol = 0.1,
  lw_maxiter = 1000,
  weight_scale = 1000,
  by_month = FALSE
)
}
\arguments{
\item{d}{A data frame as returned by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{min_date}{Earliest date to include}

\item{raw_proportions}{Logical. If \code{TRUE}, return raw, unweighted age
proportions. If \code{FALSE}, return the age proportions weighted by sample
and catch weights}

\item{plus_grp}{Age plus group for maximum grouping}

\item{lw_cutoff}{How many length-weight records are required to estimate
a length/weight model}

\item{lw_tol}{See \code{\link[=fit_lw]{fit_lw()}}}

\item{lw_maxiter}{See \code{\link[=fit_lw]{fit_lw()}}}

\item{weight_scale}{A value to divide the weights by}

\item{by_month}{Logical. If \code{TRUE}, return a data frame with a
\code{month} column in addition to a \code{year} column}
}
\value{
Nothing
}
\description{
Get the age data from the sample data
}
