% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_get_commercial_wa.R
\name{canada_get_commercial_wa}
\alias{canada_get_commercial_wa}
\title{Get the weight-at-age data frame for commercial data.}
\usage{
canada_get_commercial_wa(d, weight_scale = 1000)
}
\arguments{
\item{d}{A list as returned by \code{\link[gfdata:get_commercial_catch]{gfdata::get_commercial_catch()}}}

\item{weight_scale}{A value to divide the weights by}
}
\value{
a data frame with \code{Source} set to \code{CAN_freezer}, \code{CAN_shoreside},
\code{CAN_jv}, and \code{CAN_polish} for the four fishery types
}
\description{
Get the weight-at-age data frame for commercial data.
}
\details{
No area filtering is preformed in this function
}
