% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-load-catch-data.R
\name{canada_load_catch_data}
\alias{canada_load_catch_data}
\title{Load Canadian catch data from files extracted using the FOS web
interface}
\usage{
canada_load_catch_data(
  dr = "/srv/hake/other/landings/canada",
  min_date = as.Date("2007-04-01")
)
}
\arguments{
\item{dr}{The directory name where the files reside}

\item{min_date}{Earliest date to include in the data returned}
}
\value{
A list of four named data frames, \code{dmp_df} for the outside DMP data,
\code{dmp_inside_df} for the inside DMP data,
\code{logs_df} for the LOGS data, and \code{logs_inside_df} for the area 4B data
which were not included. \code{logs_inside_df} is returned as a convenience
for you to see if any other data from area 4B should be included in the
assessment
}
\description{
Load Canadian catch data from files extracted using the FOS web
interface
}
\details{
The web interface is queried each year with the DMP landings from April 1,
2007 to present being in one file called \code{LandingsSpeciesDateDMP.csv} and
N files, 1 for each year or LOGS data (logbook discards) with names in
the form \code{LogCatchReportXXXX.csv} with \code{XXXX} being the four digit year
from 2007 to present. These must be separate files due to limitations of
the web software, it cannot process all the years in one query. One new
\code{LOGS} file is added each year and the \code{DMP} file is re-generated each year.
}
