% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-load-sample-data.R
\name{canada_load_sample_data}
\alias{canada_load_sample_data}
\title{Load the Canadian commercial sample data}
\usage{
canada_load_sample_data(dr = "/srv/hake/other/samples")
}
\arguments{
\item{dr}{The directory in which the sample RDS file resides}
}
\value{
A data frame containing filtered Canadian samples
}
\description{
Load the Canadian commercial sample data from an RDS file generated by
\code{\link[=canada_extract_samples_from_db]{canada_extract_samples_from_db()}}. The filename is stored in the package
data variable \link{can_sample_data_rds_fn}
}
\details{
The following filtering is performed in this function for all fleets:
\enumerate{
\item Areas - BC offshore areas + Strait of Juan de Fuca
\item Fishing gear - Midwater trawl only
}
}
