% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-sql-inject-fishery-filters.R
\name{canada_sql_inject_fishery_filters}
\alias{canada_sql_inject_fishery_filters}
\title{Inject SQL code for fishery categories and types based on the fleet type}
\usage{
canada_sql_inject_fishery_filters(sql, type = c("ft", "ss", "jv"))
}
\arguments{
\item{sql}{SQL code as a vector of character strings (from \code{\link[=readLines]{readLines()}})}

\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside,
or Joint Venture respectively}
}
\description{
Injects \verb{GFBIO VESSEL_REGISTRATION_NUMBER}s where the text
"-- inject vessel codes here" appears in the SQL code and \code{TRIP_CATEGORY}s
where "-- inject fishery categories here" appears
}
