% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caption_adjustments_fn}
\alias{caption_adjustments_fn}
\title{Name of file containing the caption adjustments for figures and tables}
\format{
A character string
}
\usage{
caption_adjustments_fn
}
\description{
Name of file containing the caption adjustments for figures and tables
}
\keyword{datasets}
