% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_targets_biomass_fn}
\alias{catch_targets_biomass_fn}
\title{The filename for the file containing coastwide catch targets (TACs),
realized catch and biomass estimates}
\format{
A character string
}
\usage{
catch_targets_biomass_fn
}
\description{
The filename for the file containing coastwide catch targets (TACs),
realized catch and biomass estimates
}
\keyword{datasets}
