% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_targets_df}
\alias{catch_targets_df}
\title{A data frame containing information on the actual catch, biomass estimate
and Total Allowable Catch by year}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{Year}{Year of catch/TAC}
\item{Realized catch}{Actual catch take for the year in tonnes}
\item{TAC}{Total allowable catch for the year}
\item{Biomass estimate}{The biomass estimate for the year from the current assessment}
\item{Depletion}{The depletion estimate for the year from the current assessment}
\item{Default HCR TAC}{The default harvest control rule TAC (F40\%)}
}
}
\usage{
catch_targets_df
}
\description{
A data frame containing information on the actual catch, biomass estimate
and Total Allowable Catch by year
}
\keyword{datasets}
