% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-change-mortality-.R
\name{change_mortality_ctl}
\alias{change_mortality_ctl}
\title{Change natural mortality in a control file to be age-specific}
\usage{
change_mortality_ctl(ctl, n = 6, regular_expression = "NatM", estimate = TRUE)
}
\arguments{
\item{ctl}{An input list from r4ss with the control file.}

\item{n}{An integer specifying the number of breakpoints.}

\item{regular_expression}{A regular expression to search for in the
parameter names.}

\item{estimate}{A logical specifying if the natural mortality base-line
parameters should be estimated or fixed.}
}
\description{
Change natural mortality in a control file to be age-specific
}
