% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-change-mortality-.R
\name{change_mortality_dat}
\alias{change_mortality_dat}
\title{Add natural mortality information in the data file to allow for deviations}
\usage{
change_mortality_dat(dat, n = 6, data, years)
}
\arguments{
\item{dat}{A data file list read in by r4ss.}

\item{n}{An integer specifying the number of deviations.}

\item{data}{A data frame of deviations.}

\item{years}{The years that those deviations pertain to.}
}
\description{
Add natural mortality information in the data file to allow for deviations
}
