% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check-catch-levels.R
\name{check_catch_levels}
\alias{check_catch_levels}
\title{Ensure that the \code{\link[=run_catch_levels]{run_catch_levels()}} function has been run successfully
by looking in to the catch levels directory and checking values in the
forecast files}
\usage{
check_catch_levels(
  model = NULL,
  model_path = NULL,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_path}{The directory the models are located in}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
Logical. \code{TRUE} if the catch levels appears to have been run
successfully and \code{FALSE} otherwise
}
\description{
Ensure that the \code{\link[=run_catch_levels]{run_catch_levels()}} function has been run successfully
by looking in to the catch levels directory and checking values in the
forecast files
}
