% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check-forecasts.R
\name{check_forecasts}
\alias{check_forecasts}
\title{Check that forecasts have been run}
\usage{
check_forecasts(
  model_path = NULL,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  ...
)
}
\arguments{
\item{model_path}{The directory the models are located in}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
Logical. \code{TRUE} if the catch levels appears to have been run
successfully and \code{FALSE} otherwise
}
\description{
Check that forecasts have been run
}
\details{
Ensure that the \code{\link[=run_forecasts]{run_forecasts()}} function has been run successfully
by looking in to the forecasts directory and checking values in the
forecast files
}
