% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean-all.R
\name{clean_all}
\alias{clean_all}
\title{Run \code{\link[=clean]{clean()}} on directories used for rendering the document and
beamer presentations}
\usage{
clean_all()
}
\value{
Nothing
}
\description{
Run \code{\link[=clean]{clean()}} on directories used for rendering the document and
beamer presentations
}
\details{
Typically this needs to be done before running the \code{\link[devtools:check]{devtools::check()}}
function to avoid those build files being analyzed
}
