% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cohort.catch}
\alias{cohort.catch}
\title{Calculate the total catch taken for a given cohort}
\usage{
cohort.catch(model, cohort, ages = 0:20, trim.end.year = NA)
}
\arguments{
\item{model}{The model to extract from}

\item{cohort}{The year the cohort was born}

\item{ages}{The ages to include in the summation calculation}

\item{trim.end.year}{Remove all years after this including this year}
}
\value{
The total catch for a given cohort
}
\description{
Calculate the total catch taken for a given cohort
}
