% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cohort-catch.R
\name{cohort_catch}
\alias{cohort_catch}
\title{Calculate the total catch taken for a given cohort}
\usage{
cohort_catch(model, cohort, ages = 0:20, trim_end_year = NA)
}
\arguments{
\item{model}{The model to extract from}

\item{cohort}{The year the cohort was born}

\item{ages}{The ages to include in the summation calculation}

\item{trim_end_year}{Remove all years after this including this year}
}
\value{
A vector of total catch at age for a given cohort
}
\description{
Calculate the total catch taken for a given cohort
}
