% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-probs.R
\name{combine_historical_probs}
\alias{combine_historical_probs}
\title{Calculate a model's probability of stock decline in a year and
combine with past assessment estimates}
\usage{
combine_historical_probs(
  model,
  fn = file.path(rootd_data, "assessment-history-probs.csv"),
  start = 2012,
  end,
  ...
)
}
\arguments{
\item{model}{A model as loaded by [load_ss_files()]}

\item{...}{Further arguments to pass to [calc_historical_probs()[]}

\item{file}{Filename (.csv) of historical values from past assessments}
}
\value{
Data frame with columns
* `year` - year of assessment
* `P_decline` - estimate from `year' assessment model of the probability
   that the spawning biomass will decline from `year` to `year` + 1 for a
   catch that turned out to be the catch in `year'
* `P_below_B40` - estimate from `year' assessment model of the probability
   that the spawning biomass will be below B_40 in `year`+1 for a catch that
  turned out to be the catch in `year'
* `P_decline_curr` - current estimate (from `model`) of the probability that
  the spawning biomass declined from year to year+1
* `P_below_B40_curr` - current estimate (from `model`) of the probability
   that the spawning biomass was below B_40 in year+1
}
\description{
Calculate a model's probability of stock decline in a year and
combine with past assessment estimates
}
