% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compute-hexagon-xy.R
\name{compute_hexagon_xy}
\alias{compute_hexagon_xy}
\title{Compute the coordinates for a hexagon to be shown in a spatial CPUE plot}
\usage{
compute_hexagon_xy(d, bin_width)
}
\arguments{
\item{d}{Data from \code{\link[gfdata:get_data]{gfdata::get_cpue_spatial()}},
\code{\link[gfdata:get_data]{gfdata::get_cpue_spatial_ll()}}, or \code{\link[gfdata:get_data]{gfdata::get_catch_spatial()}}}

\item{bin_width}{Bin width as defined in \code{\link[ggplot2:stat_summary_2d]{ggplot2::stat_summary_hex()}}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the columns 'hex_id', 'cpue$value,
X, and Y
}
\description{
Compute the coordinates for a hexagon to be shown in a spatial CPUE plot
}
