% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_fn}
\alias{control_fn}
\title{A character string containing the file name for the SS3 control file}
\format{
A character string
}
\usage{
control_fn
}
\description{
A character string containing the file name for the SS3 control file
}
\keyword{datasets}
