% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert-ctl-file-param-dfs.R
\name{convert_ctl_file_param_dfs}
\alias{convert_ctl_file_param_dfs}
\title{Convert a control file parameter data frame into a format usable in this
project}
\usage{
convert_ctl_file_param_dfs(ctl, df_nm)
}
\arguments{
\item{ctl}{The control file list as read in by \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}}}

\item{df_nm}{The name of the data frame found in the \code{ctl} list}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with modified format and containing only the
columns we need
}
\description{
Convert a control file parameter data frame into a format usable in this
project
}
