% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{copy_dirfiles}
\alias{copy_dirfiles}
\title{Call \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}} depending on the Operating System}
\usage{
copy_dirfiles(
  fromdir = getwd(),
  todir = NULL,
  files = c("starter.ss", "forecast.ss", "hake_control.ss", "hake_data.ss", "wtatage.ss"),
  overwrite = TRUE,
  ignore_missing_files = TRUE
)
}
\arguments{
\item{fromdir}{The directory to copy subdirectories and their \code{files} from}

\item{todir}{The directory to copy subdirectories and their \code{files} to}

\item{files}{A character vector of file names to include in the copy}

\item{overwrite}{If \code{todir} exists then overwrite its subdirs if they exist.
Everything else in \code{todir} not pertaining to the copy will remain after the
overwrite}

\item{ignore_missing_files}{If \code{FALSE} and any of \code{files} are missing in
subdirectories of \code{fromdir}, code execution will stop and an error will
be shown. If \code{TRUE}, missing files will be ignored}
}
\value{
Nothing
}
\description{
Call \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}} depending on the Operating System
}
\details{
Copy \code{files} and their containing directories from all subdirectories
of \code{fromdir} to \code{todir}. Only subdirs of \code{fromdir} are searched for
\code{files}, not \code{fromdir} itself, and there is no recursion.
}
