% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covar_fn}
\alias{covar_fn}
\title{A character string containing the name of the covariance output file in SS3}
\format{
A character string
}
\usage{
covar_fn
}
\description{
A character string containing the name of the covariance output file in SS3
}
\keyword{datasets}
