% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-data-hake.R
\name{create_data_hake}
\alias{create_data_hake}
\title{Create package data with one call}
\usage{
create_data_hake(var, val, ...)
}
\arguments{
\item{var}{The variable name}

\item{val}{The value for variable \code{var}}

\item{...}{Additional arguments for \code{\link[usethis:use_data]{usethis::use_data()}}}
}
\value{
Nothing, creates package data. See \code{\link[usethis:use_data]{usethis::use_data()}}
}
\description{
Create package data with one call
}
\details{
Usually you have to create a variable first, and then call
\code{\link[usethis:use_data]{usethis::use_data()}} with the variable name as an argument and the
package data will be created. This function just unifies that into one step
}
