% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-fontsize-str.R
\name{create_fontsize_str}
\alias{create_fontsize_str}
\title{Creates a LaTeX font size string which can be used in
table header cells to make \code{\link[kableExtra:kbl]{kableExtra::kbl()}} tables have headers with
a different font size than the table cells have}
\usage{
create_fontsize_str(
  header_font_size = 10,
  header_vert_spacing = NULL,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{header_font_size}{Size of font, in pt}

\item{header_vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A list of two character strings, one having double-backslashes
before the \code{fontsize} and \code{selectfont} commands and the other having 4
backslashes
}
\description{
Creates a LaTeX font size string which can be used in
table header cells to make \code{\link[kableExtra:kbl]{kableExtra::kbl()}} tables have headers with
a different font size than the table cells have
}
