% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-df-biomass.R
\name{create_group_df_biomass}
\alias{create_group_df_biomass}
\title{Create a data frame containing model output of biomass from a
model list which is in long format ready for [ggplot2::ggplot()]}
\usage{
create_group_df_biomass(
  model_lst = NULL,
  model_names = NULL,
  rel = FALSE,
  bo_yr = 4,
  ...
)
}
\arguments{
\item{model_lst}{A list of models, each created by [create_rds_file()]}

\item{model_names}{A vector of model names,the same length as `model_lst`}

\item{rel}{Logical. If `TRUE` return a list of one element, the relative
biomass (no B0 element)}

\item{bo_yr}{The number of years prior to the start of the time series
that you want to use as a plot point for B0 on biomass plots.}

\item{end_yrs}{A vector of the end years for each model.
If one value, it will apply to all models}
}
\value{
A list or one or two [tibble::tibble()]. If `rel` is `TRUE`, one
(biomass) If `rel` is `FALSE`, two (biomass and B0)
}
\description{
Create a data frame containing model output of biomass from a
model list which is in long format ready for [ggplot2::ggplot()]
}
