% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-df-index.R
\name{create_group_df_index}
\alias{create_group_df_index}
\title{Create a data frame containing model output index fits from a
model list which is in long format ready for [ggplot2::ggplot()]}
\usage{
create_group_df_index(
  model_lst = NULL,
  model_names = NULL,
  survey_type = c("age1", "age2")
)
}
\arguments{
\item{model_lst}{A list of models, each created by [create_rds_file()]}

\item{model_names}{A vector of model names,the same length as `model_lst`}

\item{type}{The type of survey, must be one of `age1` or `age2`.
`age2` means age 2+ acoustic survey and `age1``is the age 1 acoustic survey}
}
\value{
A list containing a [tibble::tibble()]
}
\description{
Create a data frame containing model output index fits from a
model list which is in long format ready for [ggplot2::ggplot()]
}
