% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-kn-files.R
\name{create_kn_files}
\alias{create_kn_files}
\title{Create the \code{nuisance} and \code{key} posteriors files}
\usage{
create_kn_files(model, posterior_regex, key_post_fn, nuisance_post_fn)
}
\arguments{
\item{model}{A model object representing the output from the SS model}

\item{posterior_regex}{A vector of key posterior names in regular
expression formats, which will be glued together with logical OR ('|') to
match the names}

\item{key_post_fn}{Filename for the key posterior file}

\item{nuisance_post_fn}{Filename for the nuisance posterior file}
}
\value{
Nothing
}
\description{
Create the \code{nuisance} and \code{key} posteriors files
}
