% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-labels-and-ticks.R
\name{create_labels_and_ticks}
\alias{create_labels_and_ticks}
\title{Create a vector of labels to use when major and minor tick marks are
to be used in a plot, and the custom tick data frame to be used to draw
the lines outside the ggplot panel area using \code{\link[ggplot2:geom_linerange]{ggplot2::geom_linerange()}}}
\usage{
create_labels_and_ticks(
  breaks,
  modulo,
  lower_lim = 0,
  mjr_tick_length = major_tick_length
)
}
\arguments{
\item{breaks}{A vector of values which are where the minor ticks will
be placed}

\item{modulo}{The number to use as a modulus so that every n-th tick mark
is a major tick}

\item{lower_lim}{The lower limit of the plot, usually zero}

\item{mjr_tick_length}{The length of the major tick mark, in centimeters}
}
\value{
A list of two elements, the first is a vector of labels to use
for the major ticks only. The minor tick locations do not have labels.
The second list element is a data frame of two columns, the breaks (values)
where the major ticks will occur and the ending position of the lines
used to mark the major ticks
}
\description{
Create a vector of labels to use when major and minor tick marks are
to be used in a plot, and the custom tick data frame to be used to draw
the lines outside the ggplot panel area using \code{\link[ggplot2:geom_linerange]{ggplot2::geom_linerange()}}
}
