% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-attach-retrospectives.R
\name{create_rds_attach_retrospectives}
\alias{create_rds_attach_retrospectives}
\title{Load retrospective RDS files and attach to as RDS file}
\usage{
create_rds_attach_retrospectives(
  model_path = NULL,
  verbose = TRUE,
  base_model_name = "Base model",
  ...
)
}
\arguments{
\item{model_path}{Directory name of model to be loaded}

\item{verbose}{Logical. If \code{TRUE}, write more output to the console}

\item{base_model_name}{The name to show in retrospective comparison plots
for the base model}

\item{...}{Passed to \code{\link[=check_catch_levels]{check_catch_levels()}}}
}
\value{
Nothing
}
\description{
Load retrospective RDS files and attach to as RDS file
}
\details{
Load retrospectives and attach it to an already-created RDS file as the list
element \code{retrospectives} and overwrite the RDS file
}
