% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-files-retro.R
\name{create_rds_files_retro}
\alias{create_rds_files_retro}
\title{Create the RDS files for all retrospective runs}
\usage{
create_rds_files_retro(model_path, verbose = TRUE, ...)
}
\arguments{
\item{model_path}{Directory name of model to be loaded}

\item{verbose}{Logical. If \code{TRUE}, write more output to the console}

\item{...}{Arguments to pass to \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Create the RDS files for all retrospective runs
}
