% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-natage.R
\name{create_rds_natage}
\alias{create_rds_natage}
\title{Create an rds file to hold the model's full MCMC samples for annual numbers at age.}
\usage{
create_rds_natage(
  model_path = NULL,
  verbose = TRUE,
  overwrite = FALSE,
  progress_n = 500,
  first = 0,
  ...
)
}
\arguments{
\item{model_path}{Directory name of model to be loaded}

\item{verbose}{Logical. If \code{TRUE}, write more output to the console}

\item{overwrite}{Logical. If \code{TRUE}, overwrite the RDS file if it exists}

\item{progress_n}{Report every time this many files are processed. Consider
how many posteriors there are, this should be a fairly large proportion of
that (around 1/8th) or there will be too much output and it will run slow}

\item{first}{Load this many of the files. If a non-positive number, load
them all. Used for debugging purposes to cut down the size of the
lists used}

\item{...}{Arguments to pass to \code{\link[=load_ss_files]{load_ss_files()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Create an rds file to hold the model's full MCMC samples for annual numbers at age.
}
\details{
Simplified from \code{\link[=create_rds_file]{create_rds_file()}} by Andy for highest density interval
manuscript, for which need to look at the full MCMC values for numbers at
age for each year. Save the rds file with name as used in
\code{\link[=create_rds_file]{create_rds_file()}} but with with \code{natage} appended.
This function does not load retrospectives and forecasts.
To do that you must run this first, then run other functions
to load those things and attach them to the model list
}
\examples{
\dontrun{
# create_rds_natage("/srv/hake/models/2024/02-version/01-base-models/01-base/")
}
}
