% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-version-dirs.R
\name{create_version_dirs}
\alias{create_version_dirs}
\title{Create set of directories for a new model version}
\usage{
create_version_dirs(
  year,
  dir_models = "/srv/hake/models",
  version = "01-version"
)
}
\arguments{
\item{year}{A four-digit integer specifying the assessment year. E.g., for
an assessment that uses data up to 2023 and was due in 2024, the
assessment year is 2024.}

\item{dir_models}{A file path where you want to create the year directory.
The default is the path that stores all of the model folders on hake-
precision, i.e., \code{"/srv/hake/models"}.}

\item{version}{A string that starts with a two-digit integer and is followed
by a dash and then the word "version".}
}
\value{
A vector of directory paths that were created is invisibly returned.
}
\description{
Create set of directories for a new model version
}
\examples{
\dontrun{
create_version_dirs(year = 2023, dir_models = getwd())
unlink("01-version", recursive = TRUE)
}
}
\author{
Kelli F. Johnson
}
