% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct}
\alias{ct}
\title{A data frame containing information on catch by year and fishery (fleet)}
\format{
A data frame with 57 rows and 22 variables:
\describe{
\item{Year}{Year of catch}
\item{U.S. Foreign}{Catch in tonnes}
\item{U.S. Joint-venture}{Catch in tonnes}
\item{U.S. Mothership}{Catch in tonnes}
\item{U.S. Catcher-processor}{Catch in tonnes}
\item{U.S. Shore-based}{Catch in tonnes}
\item{U.S. Research}{Catch in tonnes}
\item{Canada Foreign}{Catch in tonnes}
\item{Canada Joint-venture}{Catch in tonnes}
\item{Canada Shoreside}{Catch in tonnes}
\item{Canada Freezer-trawler}{Catch in tonnes}
\item{U.S. TAC}{Total Allowable catch in tonnes}
\item{Canada TAC}{Total Allowable catch in tonnes}
\item{Total TAC}{Total Allowable catch in tonnes}
\item{U.S. Total}{Total catch in tonnes for the U.S. each year}
\item{Canada Total}{Total catch in tonnes for Canada each year}
\item{us_prop}{U.S. proprtion of catch taken each year}
\item{can_prop}{Canadian proprtion of catch taken each year}
\item{us_attain}{U.S. proportion of U.S. TAC caught each year}
\item{can_attain}{Canadian proportion of Canadian TAC caught each year}
\item{tot_attain}{Proportion of total TAC caught each year}
}
}
\usage{
ct
}
\description{
A data frame containing information on catch by year and fishery (fleet)
}
\keyword{datasets}
