% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_levels_catch_tol}
\alias{ct_levels_catch_tol}
\title{The tolerance in tonnes. The iterations will stop if the difference
between the projected biomass between the first and second years is
less than this}
\format{
A single decimal number
}
\usage{
ct_levels_catch_tol
}
\description{
The tolerance in tonnes. The iterations will stop if the difference
between the projected biomass between the first and second years is
less than this
}
\keyword{datasets}
