% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_levels_max_iter}
\alias{ct_levels_max_iter}
\title{The maximum number of iterations to do. If this is reached, then no
catch value could be found within the tolerances above}
\format{
A single decimal number
}
\usage{
ct_levels_max_iter
}
\description{
The maximum number of iterations to do. If this is reached, then no
catch value could be found within the tolerances above
}
\keyword{datasets}
