% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_levels_path}
\alias{ct_levels_path}
\title{Directory name for storing the catch levels calculations.}
\format{
A character string
}
\usage{
ct_levels_path
}
\description{
Directory name for storing the catch levels calculations.
}
\keyword{datasets}
