% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_levels_spr_tol}
\alias{ct_levels_spr_tol}
\title{Custom catch levels calculations. The tolerance in the SPR away from
1 for the calculation of catch for SPR = 1}
\format{
A single decimal number
}
\usage{
ct_levels_spr_tol
}
\description{
Custom catch levels calculations. The tolerance in the SPR away from
1 for the calculation of catch for SPR = 1
}
\keyword{datasets}
