% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-curr-time-date.R
\name{curr_time_date}
\alias{curr_time_date}
\title{Create a time and date string for the title page of the doc}
\usage{
curr_time_date(testday = NULL)
}
\arguments{
\item{testday}{A test string in the format e.g. "May 02" or "July 1"}
}
\value{
A string with the nice date and time
}
\description{
Create a time and date string for the title page of the doc
}
