% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fn}
\alias{data_fn}
\title{A character string containing the file name for the SS3 data file}
\format{
A character string
}
\usage{
data_fn
}
\description{
A character string containing the file name for the SS3 data file
}
\keyword{datasets}
