% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ssnew_fn}
\alias{data_ssnew_fn}
\title{A character string containing the name of the file SS3 creates internally
as a copy of the \link{data_fn} input file}
\format{
A character string
}
\usage{
data_ssnew_fn
}
\description{
A character string containing the name of the file SS3 creates internally
as a copy of the \link{data_fn} input file
}
\keyword{datasets}
