% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tables_path}
\alias{data_tables_path}
\title{The directory name for the directory containing the data tables as
csv files}
\format{
A character string
}
\usage{
data_tables_path
}
\description{
The directory name for the directory containing the data tables as
csv files
}
\keyword{datasets}
