% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-decisions.R
\name{decision_table}
\alias{decision_table}
\title{Create the Executive Summary decision tables. `rows_to_label` and `row_labels`
need changing each year.}
\usage{
decision_table(
  model,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  type = "biomass",
  placement = "H",
  forecast_inds = seq_along(model$forecasts[[length(model$forecasts)]]),
  rows_to_show = NULL,
  rows_to_label = c("e", "f", "h", "j", "l", "m", "n", "o"),
  row_labels = list(c("10\\\\\%", "reduction"), c(assess_yr - 1, "catch"), c("10\\\\\%",
    "reduction"), c("10\\\\\%", "reduction"), c(assess_yr - 1, "TAC"), c("FI=",
    "100\\\\\%"), c("default", "HR"), c(paste0("C", assess_yr, "="), paste0("C",
    assess_yr + 1)))
)
}
\arguments{
\item{model}{An mcmc run, `model$forecasts` must be populated from [run_forecasts()]}

\item{xcaption}{Caption to appear in the calling document}

\item{xlabel}{The label used to reference the table in latex}

\item{font.size}{Size of the font for the table}

\item{space.size}{Size of the vertical spaces for the table}

\item{type}{Type to build. `biomass` or `spr`}

\item{placement}{Latex code for placement of table}

\item{forecast_inds}{Indices of forecast catch levels to show. By default all
catch levels will be shown}

\item{rows_to_label}{A vector of letters representing rows to add custom labels to}

\item{row_labels}{A list of length two vectors with text to place in the empty
two rows under the letter for the letters given in `rows_to_label`}
}
\value{
The latex code needed to build the table
}
\description{
Create the Executive Summary decision tables. `rows_to_label` and `row_labels`
need changing each year.
}
