% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_hr_path}
\alias{default_hr_path}
\title{Directory name for storing the default harvest rule catch catch levels}
\format{
A character string
}
\usage{
default_hr_path
}
\description{
Directory name for storing the default harvest rule catch catch levels
}
\keyword{datasets}
