% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-delete-files.R
\name{delete_dirs}
\alias{delete_dirs}
\title{Delete all directories and files of `sub_dir``}
\usage{
delete_dirs(models_dir, sub_dir = NULL)
}
\arguments{
\item{models_dir}{Directory name for all models location}

\item{sub_dir}{The subdirectory to delete recursively}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Delete all directories and files of `sub_dir``
}
