% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-delete-files.R
\name{delete_rds_files}
\alias{delete_rds_files}
\title{Delete all model rds files found in the subdirectories of the `models_dir`` directory}
\usage{
delete_rds_files(models_dir, dont_del = last_yr_base_model_dir_name)
}
\arguments{
\item{models_dir}{Directory in which the models reside}

\item{dont_del}{A vector of directory names to keep}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Delete all model rds files found in the subdirectories of the `models_dir`` directory
}
