% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{derposts_fn}
\alias{derposts_fn}
\title{A character string containing the file name for the SS3 derived posteriors
output file}
\format{
A character string
}
\usage{
derposts_fn
}
\description{
A character string containing the file name for the SS3 derived posteriors
output file
}
\keyword{datasets}
