% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doc_path}
\alias{doc_path}
\title{Directory name for the main assessment document.}
\format{
A character string
}
\usage{
doc_path
}
\description{
Directory name for the main assessment document.
}
\keyword{datasets}
