% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-draw-arrows-yoob.R
\name{draw_arrows_yoob}
\alias{draw_arrows_yoob}
\title{Draw arrows on the top and/or bottom of the plot pointing to where
out-of-bounds points are located. YOOB means Y-axis Out-Of-Bounds}
\usage{
draw_arrows_yoob(
  g,
  yoob,
  arrow_angle = 10,
  arrow_size_npc = 0.03,
  arrow_units = "npc"
)
}
\arguments{
\item{g}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object}

\item{yoob}{Anlist returned from the function \code{\link[=calc_yoob]{calc_yoob()}}}

\item{arrow_angle}{See \code{angle} parameter in  \code{\link[grid:arrow]{grid::arrow()}}}

\item{arrow_size_npc}{See \code{length} parameter in  \code{\link[grid:arrow]{grid::arrow()}}}

\item{arrow_units}{See \code{length} parameter in  \code{\link[grid:arrow]{grid::arrow()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Draw arrows on the top and/or bottom of the plot pointing to where
out-of-bounds points are located. YOOB means Y-axis Out-Of-Bounds
}
