% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enact-privacy-rule.R
\name{enact_privacy_rule}
\alias{enact_privacy_rule}
\title{Apply the confidentiality rule to spatial data. Copied from the \link{gfplot}
package}
\usage{
enact_privacy_rule(dat, bin_width, n_minimum_vessels, x_lim, y_lim)
}
\arguments{
\item{dat}{Data from \code{\link[gfdata:get_data]{gfdata::get_cpue_spatial()}},
\code{\link[gfdata:get_data]{gfdata::get_cpue_spatial_ll()}}, or \code{\link[gfdata:get_data]{gfdata::get_catch_spatial()}}}

\item{bin_width}{Bin width as defined in \code{\link[ggplot2:stat_summary_2d]{ggplot2::stat_summary_hex()}}}

\item{n_minimum_vessels}{The minimum number of vessels that must be present
in a hexagon for the hexagon to be shown}

\item{x_lim}{The length-2 vector representing the minimum and maximum
limits of the x-axis}

\item{y_lim}{The length-2 vector representing the minimum and maximum
limits of the y-axis}
}
\value{
A list containing 1) the data frame \code{dat} but with rows removed that
do not fulfill the privacy limitations, 2) The total number of fishing
events removed due to privacy limitations, and 3) The total number of
fishing events in \code{dat}
}
\description{
Apply the confidentiality rule to spatial data. Copied from the \link{gfplot}
package
}
\details{
Removes data which does not satisfy the confidentiality rules; the minimum
number of vessels with catch in a specific area
}
