% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export.depth}
\alias{export.depth}
\title{Take a data frame that has at least two columns called year and fdep,
convert to meters and calculate boxplot stats on each year write a
csv file to current directory}
\usage{
export.depth(x, fleet = "")
}
\arguments{
\item{x}{The data frame}

\item{fleet}{}
}
\value{
[base::invis]
}
\description{
Take a data frame that has at least two columns called year and fdep,
convert to meters and calculate boxplot stats on each year write a
csv file to current directory
}
\details{
Use spatial data frame (Canada) output to this function
}
\examples{
export.depth(d.ft, "freezer-trawlers")
export.depth(d.ss, "shoreside")
}
