% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-fleet-catch-canada.R
\name{extract_fleet_catch_canada}
\alias{extract_fleet_catch_canada}
\title{Merge Canadian DMP and LOGS catch data and extract into 3 data frames,
for Freezer trawlers, Shoreside, and Joint-venture}
\usage{
extract_fleet_catch_canada(lst)
}
\arguments{
\item{lst}{a list of three data frames as returned by
\code{\link[=load_catch_data_canada]{load_catch_data_canada()}}}
}
\value{
A data frame containing only data for the type given by \code{type}
}
\description{
Merge Canadian DMP and LOGS catch data and extract into 3 data frames,
for Freezer trawlers, Shoreside, and Joint-venture
}
