% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-inside-catch.R
\name{extract_inside_catch}
\alias{extract_inside_catch}
\title{Extract the inside (Strait of Georgia) hake catch by year}
\usage{
extract_inside_catch(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=canada_load_catch_data]{canada_load_catch_data()}}}
}
\value{
A list of three data frames, one for the landed, one for the
discarded, and one for the sum of the two
}
\description{
Extract the inside (Strait of Georgia) hake catch by year
}
