% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-label-from-figure-filename.R
\name{extract_label_from_figure_filename}
\alias{extract_label_from_figure_filename}
\title{Extract chunk label for the knitr chunk containing includegraphics and a
filename}
\usage{
extract_label_from_figure_filename(fn)
}
\arguments{
\item{fn}{The figure file name to match (no extension)}
}
\value{
The chunk label found for the chunk which includes the figure
file \code{fn}
}
\description{
Extract chunk label for the knitr chunk containing includegraphics and a
filename
}
\details{
\enumerate{
\item Starts by parsing the \verb{_bookdown.yml} file to extract the names of all
uncommented rmd files in the current build.
\item Searches all those files for any lines starting with (ref:variable).
Preceding spaces are ignored
\item Extracts the \code{chunkname} given the filename loaded within the chunk
that \code{chunkname} belongs to
}
}
