% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-mcmc-quant.R
\name{extract_mcmc_quant}
\alias{extract_mcmc_quant}
\title{Extract MCMC quantiles from models and return them
in a [ggplot2::ggplot()]-ready data frame}
\usage{
extract_mcmc_quant(
  model_lst,
  model_names,
  type,
  inc_model_year = FALSE,
  end_yrs = year(Sys.time())
)
}
\arguments{
\item{model_lst}{A list of models, each created by [create_rds_file()]}

\item{model_names}{A vector of model names,the same length as `models_lst`}

\item{type}{A name as found in an `mcmccalcs` object of a model object,
for example extraction of: `base_model$mcmccalcs$smed` would require
`type` = "smed"}

\item{end_yrs}{A vector of the end years for each model.
If one value, it will apply to all models}

\item{inc_models_yr}{Logical. If `TRUE`, include the `model` and `year`
columns in the data frame}
}
\value{
A [tibble::tibble()]
}
\description{
Extract MCMC quantiles from models and return them
in a [ggplot2::ggplot()]-ready data frame
}
