% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-ref-label-from-figure-filename.R
\name{extract_ref_label_from_figure_filename}
\alias{extract_ref_label_from_figure_filename}
\title{Extract ref label for the alt text of a knitr chunk containing
includegraphics and a filename}
\usage{
extract_ref_label_from_figure_filename(fn)
}
\arguments{
\item{fn}{The figure file name to match (no extension)}
}
\value{
The text found for the label given by \code{fn}
}
\description{
Extract ref label for the alt text of a knitr chunk containing
includegraphics and a filename
}
\details{
\enumerate{
\item Starts by parsing the \verb{_bookdown.yml} file to extract the names of all
uncommented rmd files in the current build.
\item Searches all those files for any lines starting with (ref:variable).
Preceding spaces are ignored
\item Extracts the \code{(ref:chunkname-alt)} given the filename loaded within
the chunk that \code{(ref:chunkname-alt)} belongs to
}
}
