% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-weight-at-age-vector-from-ss-inputs.R
\name{extract_weight_at_age_vector_from_ss_inputs}
\alias{extract_weight_at_age_vector_from_ss_inputs}
\title{Extract weight-at-age vector from SS3 input data}
\usage{
extract_weight_at_age_vector_from_ss_inputs(
  model,
  yrs = 2020:2024,
  yrs_mean = TRUE,
  flt = 1
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yrs}{A vector of years to use in the calculation}

\item{yrs_mean}{If \code{TRUE}, calculate the at-age means from the \code{yrs}
weight-at-age values}

\item{flt}{The fleet number 1 = fishery, 2 = survey}
}
\value{
A numeric named vector
}
\description{
Extract weight-at-age vector from SS3 input data
}
