% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{f}
\alias{f}
\title{Format x to have supplied number of decimal points}
\usage{
f(x, dec.points = 0)
}
\arguments{
\item{x}{The number}

\item{dec.points}{The number of decimal points to use}
}
\value{
A formatted string representing the number
}
\description{
Format x to have supplied number of decimal points
}
\details{
Make thousands separated by commas and the number of decimal
points given by \code{dec.points}
}
