% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-f-date.R
\name{f_date}
\alias{f_date}
\title{Format dates}
\usage{
f_date(x, format = c("\%m", "\%Y"), factor = FALSE)
}
\arguments{
\item{x}{A vector of dates with class \code{POSIXct}, \code{POSIXt}, or \code{Date}.}

\item{format}{A character value specifying the format you want
the output to be in. See \code{\link[base:strptime]{base::strptime()}} for options.}

\item{factor}{A logical specifying if you want a factor to be returned.
The factor is \emph{ALWAYS} limited to the levels that are available. The
default behavior is to \emph{NOT} return a factor. Thus, this must be invoked.}
}
\value{
A vector the same length as the input vector.
}
\description{
Format a date object the same as \code{\link[base:format]{base::format()}} but return a numerical
value when able instead of always returning a character string. Users
dictate what is returned by the same \code{format} argument that is used in
\code{\link[base:format]{base::format()}}.
}
\seealso{
\itemize{
\item \code{\link[base:format]{base::format()}}
\item \code{\link[base:factor]{base::factor()}}
\item \code{\link[base:droplevels]{base::droplevels()}}
\item \code{\link[base:strptime]{base::strptime()}}
}
}
\author{
Kelli F. Johnson
}
