% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{fetch_catch_levels}
\alias{fetch_catch_levels}
\title{Fetch catch levels from SS forecast files. This is used to retrieve ending values
after running the [run_catch_levels()] function for default HR, SPR 100, and stable catch}
\usage{
fetch_catch_levels(catch_levels_path, catch_levels = NULL, ...)
}
\arguments{
\item{catch_levels_path}{The path for the catch-levels output}

\item{catch_levels}{The catch levels list as defined in forecast-catch-levels.R}

\item{model}{The SS model output as loaded by [load_ss_files()]}

\item{spr_100_path}{The path for the spr-100 output}

\item{default_hr_path}{The path for the default hr output}

\item{stable_catch_path}{The path for the stable-catch output}
}
\value{
A list of 3-element lists of vectors of 3 catch levels corresponding to:
a) SPR-100%
b) Default harvest policy
c) Stable catch
Return object looks the same as the `catch_levels` object but with three more elements
}
\description{
Fetch catch levels from SS forecast files. This is used to retrieve ending values
after running the [run_catch_levels()] function for default HR, SPR 100, and stable catch
}
\details{
Assumes [run_catch_levels()] function has been run and the forecast files
are populated with 3 forecast years
}
