% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-forecasts-extra.R
\name{fetch_forecasts_extra}
\alias{fetch_forecasts_extra}
\title{Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}}
\usage{
fetch_forecasts_extra(
  forecasts_path,
  forecast_yrs,
  my_catch_levels = list(list(rep(470000, length(forecast_yrs)), "470,000 t",
    "16-470000"), list(rep(510000, length(forecast_yrs)), "510,000 t", "17-510000"),
    list(rep(6e+05, length(forecast_yrs)), "600,000 t", "18-600000"), list(rep(650000,
    length(forecast_yrs)), "650,000 t", "19-650000")),
  ...
)
}
\arguments{
\item{forecasts_path}{Path of the forecasts}

\item{forecast_yrs}{A vector of the forecast years}

\item{my_catch_levels}{The table of catch levels with same structure as
found in \code{forecast-catch-levels.R}}

\item{...}{}
}
\value{
A list of forecast outputs as read in by \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}}
}
\description{
Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}
}
\details{
If the forecasts directory does not exist or there is a problem
loading the forecasts, return \code{NA}
}
