% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-firstup.R
\name{firstup}
\alias{firstup}
\title{Convert the first letter of a string to uppercase}
\usage{
firstup(str)
}
\arguments{
\item{str}{The string}
}
\value{
The string \code{str} with the first letter in uppercase
}
\description{
Convert the first letter of a string to uppercase
}
