% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit-lw.R
\name{fit_lw}
\alias{fit_lw}
\title{Fit a length-weight model}
\usage{
fit_lw(d, tol = 0.1, maxiter = 1000)
}
\arguments{
\item{d}{Data frame containing the columns \code{length} and \code{weight}}

\item{tol}{See \code{\link[stats:nls]{stats::nls()}}}

\item{maxiter}{See \code{\link[stats:nls]{stats::nls()}}}
}
\value{
The \code{\link[stats:coef]{stats::coefficients()}} of the model fit
}
\description{
Fit a length-weight model
}
