% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{fix.posteriors}
\alias{fix.posteriors}
\title{Remove multiple header/datarows found in posteriors files to leave just one}
\usage{
fix.posteriors(dir)
}
\arguments{
\item{dir}{Directory where the files reside}
}
\value{
If the file contains only one header/datarow (i.e is of correct format)
then return a 1-row dataframe of the contents
}
\description{
Remove multiple header/datarows found in posteriors files to leave just one
}
