% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-posteriors.R
\name{fix_posteriors}
\alias{fix_posteriors}
\title{Remove multiple header/data rows found in posteriors files to leave just
one}
\usage{
fix_posteriors(fn)
}
\arguments{
\item{fn}{File name for the posterior file}
}
\value{
If the file contains only one header/data row (i.e is of correct
format) then return a 1-row data frame of the contents
}
\description{
Remove multiple header/data rows found in posteriors files to leave just
one
}
