% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fix-title-case-for-csl.R
\name{fix_title_case_for_csl}
\alias{fix_title_case_for_csl}
\title{Add an additional layer of curly braces around the titles found in the
supplied BIB file (BiBTeX format). This is necessary for CSL to write the
titles in the output document as-is. If it is not done, sentence-case is
implemented in CSL and your titles will be messed up.}
\usage{
fix_title_case_for_csl(
  bib_fn = here("doc/bib/refs.bib"),
  out_fn = here("doc/bib/refs_modified.bib")
)
}
\arguments{
\item{bib_fn}{The input BiBTeX file name}

\item{out_fn}{The output BiBTeX file name}
}
\value{
nothing
}
\description{
Add an additional layer of curly braces around the titles found in the
supplied BIB file (BiBTeX format). This is necessary for CSL to write the
titles in the output document as-is. If it is not done, sentence-case is
implemented in CSL and your titles will be messed up.
}
\details{
Only done once for a project. Added here for reference and to save the code
}
