% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_descriptions_fn}
\alias{forecast_descriptions_fn}
\title{Name of file containing the forecasting descriptions and settings}
\format{
A character string
}
\usage{
forecast_descriptions_fn
}
\description{
Name of file containing the forecasting descriptions and settings
}
\keyword{datasets}
