% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_fn}
\alias{forecast_fn}
\title{A character string containing the file name for the SS3 forecast file}
\format{
A character string
}
\usage{
forecast_fn
}
\description{
A character string containing the file name for the SS3 forecast file
}
\keyword{datasets}
