% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_probs}
\alias{forecast_probs}
\title{A vector of the probabilities to use for uncertainty calculations in the
forecasting part of the assessment document}
\format{
A vector of three values, for the lower credible interval limit,
a lower-middle value, the median, a middle-upper value, and the upper
credible interval limit
}
\usage{
forecast_probs
}
\description{
A vector of the probabilities to use for uncertainty calculations in the
forecasting part of the assessment document
}
\keyword{datasets}
