% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecasts_path}
\alias{forecasts_path}
\title{Directory name for storing the forecasting.}
\format{
A character string
}
\usage{
forecasts_path
}
\description{
Directory name for storing the forecasting.
}
\keyword{datasets}
