% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecasts_prepend}
\alias{forecasts_prepend}
\title{Name to be prepended to year for forecast directory names.}
\format{
A character string
}
\usage{
forecasts_prepend
}
\description{
Name to be prepended to year for forecast directory names.
}
\keyword{datasets}
