% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{further_tac_details_fn}
\alias{further_tac_details_fn}
\title{The filename for the file containing further TAC details such as U.S.
tribal allocations and Canadian carryover and Joint venture TAC}
\format{
A character string
}
\usage{
further_tac_details_fn
}
\description{
The filename for the file containing further TAC details such as U.S.
tribal allocations and Canadian carryover and Joint venture TAC
}
\keyword{datasets}
