% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{further_tac_df}
\alias{further_tac_df}
\title{A data frame containing information on extra TAC information}
\format{
A data frame with 9 rows and 12 variables:
\describe{
\item{Year}{Year of catch/TAC}
\item{us_tribal_quota}{U.S. tribal quota in tonnes}
\item{us_research_quota}{U.S. research quota in tonnes}
\item{us_nontribal_quota}{U.S. non-tribal quota in tonnes}
\item{us_tribal_quota_reallocated}{U.S. tribal quota reallocated in tonnes}
\item{us_tribal_reallocate_dates}{U.S. tribal quota reallocation dates}
\item{us_tribal_max_landed}{U.S. tribal maximum landed catch in tonnes}
\item{can_carried_over}{Canadian carry-over value in tonnes}
\item{us_shore_reallocated}{U.S. shore-based reallocation in tonnes}
\item{us_cp_reallocated}{U.S. catcher-processor reallocation in tonnes}
\item{us_ms_reallocated}{U.S. mothership reallocation in tonnes}
\item{can_jv_tac}{Canadian joint-venture allocation/reallocation in tonnes}
}
}
\usage{
further_tac_df
}
\description{
A data frame containing information on extra TAC information
}
\keyword{datasets}
