% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pointpath.R
\name{geom_pointpath}
\alias{geom_pointpath}
\title{Connect observations with lines and points}
\usage{
geom_pointpath(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}.}
}
\description{
\code{geom_pointpath()} connects the observations in the order in which they
appear in the data.
}
\section{Missing value handling}{

\code{geom_pointpath()} handle \code{NA} in the same way as \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
and \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}
}

