% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pointpath.R
\docType{data}
\name{GeomPointPath}
\alias{GeomPointPath}
\title{\code{ggproto} class for the \code{pointpath} Geom}
\format{
An object of class \code{GeomPointPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomPointPath
}
\description{
\code{ggproto} class for the \code{pointpath} Geom
}
\keyword{datasets}
