% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.align}
\alias{get.align}
\title{Create a character vector used in the align argument of the
\code{\link[xtable:xtable]{xtable::xtable()}} command}
\usage{
get.align(num, first.left = TRUE, just = "r")
}
\arguments{
\item{num}{The number of columns in the table}

\item{first.left}{Logical. Keep the first column left-justified. If
\code{FALSE}, it will be justified according to the \code{just} argument}

\item{just}{The justification to use for the columns,
i.e. "r", "l", or "c"}
}
\value{
A character vector used in the align argument of the \code{\link[xtable:xtable]{xtable::xtable()}} command
}
\description{
Create a character vector used in the align argument of the
\code{\link[xtable:xtable]{xtable::xtable()}} command
}
\details{
e.g. posterior output tables, reference point tables.
Most tables really
}
