% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.shade}
\alias{get.shade}
\title{Create an RGB string of the specified color and opacity}
\usage{
get.shade(color, opacity)
}
\arguments{
\item{color}{A vector of R color strings or numbers}

\item{opacity}{A number between 0 and 99}
}
\value{
An RGB string of the specified color and opacity
}
\description{
Create an RGB string of the specified color and opacity
}
\details{
Format of returned string is #RRGGBBAA where
RR = red, a 2-hexadecimal-digit string
GG = green, a 2-hexadecimal-digit string
BB = blue, a 2-hexadecimal-digit string
AA = opacity, 2-digit string
}
