% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_active_parameter_names}
\alias{get_active_parameter_names}
\title{Get a vector of the active parameter names for a model}
\usage{
get_active_parameter_names(model)
}
\arguments{
\item{model}{A model object as returned from [load.ss.files()]}
}
\value{
A vector of the active parameter names for a model
}
\description{
Get a vector of the active parameter names for a model
}
