% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-age-comp-limits.R
\name{get_age_comp_limits}
\alias{get_age_comp_limits}
\title{Get the start and end year of the age comp data, and maximum
proportion overall with its year and age}
\usage{
get_age_comp_limits(model, type = 1)
}
\arguments{
\item{model}{A model as returnded by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{type}{1 for Fishery and 2 for Survey}
}
\value{
A vector of 5 elements as described above
}
\description{
Get the start and end year of the age comp data, and maximum
proportion overall with its year and age
}
