% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-age-proportions-by-fleet.R
\name{get_age_proportions_by_fleet}
\alias{get_age_proportions_by_fleet}
\title{Calculates the age proportions by year and returns a list with one
data frame for each fishery type (Freezer trawlers, Shoreside, and
Joint venture)}
\usage{
get_age_proportions_by_fleet(d, raw_proportions = F)
}
\arguments{
\item{d}{A data frame as returned by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{raw_proportions}{Logical. If \code{TRUE}, return raw, unweighted age
proportions. If \code{FALSE}, return the age proportions weighted by sample
and catch weights}
}
\value{
A list of three data frames of age proportions by year
}
\description{
Calculates the age proportions by year and returns a list with one
data frame for each fishery type (Freezer trawlers, Shoreside, and
Joint venture)
}
