% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-assess-yr.R
\name{get_assess_yr}
\alias{get_assess_yr}
\title{Get the current assessment year}
\usage{
get_assess_yr(assess_yr_month_cutoff = 12)
}
\arguments{
\item{assess_yr_month_cutoff}{The month where 1 will be added to
the assessment year. So if this is 12 and the current date is in December,
the assessment year will be the next year, which starts in January}
}
\value{
A year
}
\description{
Get the current assessment year
}
\details{
There is a bash script that does this as well. See
\code{hake/bash-scripts/get-assess-year.sh}
}
