% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-quantities.R
\name{get_baa_mle}
\alias{get_baa_mle}
\title{Calculate Biomass-at-age for the years given for the given model}
\usage{
get_baa_mle(model, yrs = NULL, scale = 1000)
}
\arguments{
\item{model}{A model as loaded by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{yrs}{The years to return the values for}

\item{scale}{Scaling factor for the output values}
}
\value{
A data frame of the biomass-at-ages
}
\description{
Calculate Biomass-at-age for the years given for the given model
}
\details{
Uses numbers-at-age multiplied by weight-at-age. If any years
are missing the weight-at-age, the average will be used (Assumed to be
in the 1966 year slot for SS input)
}
\examples{
baa <- get_baa(base_model)
}
