% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-col-name-from-key-title.R
\name{get_col_name_from_key_title}
\alias{get_col_name_from_key_title}
\title{Get the name of a column in the input data frame that matches a regular
expression in the \code{key_posteriors} list, given a title as found in
\code{key_posteriors_titles}}
\usage{
get_col_name_from_key_title(df = NULL, pat = NULL)
}
\arguments{
\item{df}{The data frame with column names}

\item{pat}{The regular expression used to match a title in
\code{key_posteriors_titles}}
}
\value{
A string, which is the name of a column in the data frame \code{df}
}
\description{
Get the name of a column in the input data frame that matches a regular
expression in the \code{key_posteriors} list, given a title as found in
\code{key_posteriors_titles}
}
