% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-data-yr.R
\name{get_data_yr}
\alias{get_data_yr}
\title{Get the current data year}
\usage{
get_data_yr(data_yr_month_cutoff = 4)
}
\arguments{
\item{data_yr_month_cutoff}{The final month in the year that you would still
only want data from the previous year. Often times we might pull the data
in March but we do not actually want to include data from January,
February, and March of that year. The default is April, or 4, to be safe.
Thus, if you are working in June of 1999, the data year will be 1999 but if
you are working in February of 1999 the data year will be 1998 using the
default input.}
}
\value{
A four-digit integer is returned that specifies the year of interest.
}
\description{
Get the current data year
}
\author{
Kelli F. Johnson
}
