% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{get_depth_by_year}
\alias{get_depth_by_year}
\title{Get depth data by year and calculate boxplot stats on it.}
\usage{
get_depth_by_year(
  d,
  type = c("bottom", "gear"),
  yrs = NULL,
  min_depth_cutoff = 50/1.8288
)
}
\arguments{
\item{d}{Output from [load_spatial_catch_data().}

\item{type}{One of "bottom" or "gear" for depth type.}

\item{yrs}{A vector of years to include. If \code{NULL}, all years in the data
will be included.}

\item{min_depth_cutoff}{The depth for which to remove data. In fasthoms.
All data shallower than this will be removed.}
}
\value{
A tibble containing year and depth record stats.
}
\description{
Get depth data by year and calculate boxplot stats on it.
}
