% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-doc-type.R
\name{get_doc_type}
\alias{get_doc_type}
\title{Get the document type for the given RMD file}
\usage{
get_doc_type(fn)
}
\arguments{
\item{fn}{The RMD filename}
}
\value{
Either "beamer" or "doc". "doc" means the main assessment document
}
\description{
Get the document type for the given RMD file
}
