% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-key-nuisance-posts.R
\name{get_key_nuisance_posts}
\alias{get_key_nuisance_posts}
\title{Create the \code{nuisance} and \code{key} posteriors CSV files in the output
mcmc directory of the model}
\usage{
get_key_nuisance_posts(model)
}
\arguments{
\item{model}{A model object representing the output from the SS model}
}
\value{
Nothing, creates two CSV files
}
\description{
Create the \code{nuisance} and \code{key} posteriors CSV files in the output
mcmc directory of the model
}
