% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-obj-sizes.R
\name{get_obj_sizes}
\alias{get_obj_sizes}
\title{Get the object sizes in bytes for elements of the input list}
\usage{
get_obj_sizes(lst)
}
\arguments{
\item{lst}{The input list}
}
\value{
A data frame with one column if the elements have no names and two
columns if there are names
}
\description{
Get the object sizes in bytes for elements of the input list
}
