% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-os.R
\name{get_os}
\alias{get_os}
\title{Return the Operating System name in lower case}
\usage{
get_os()
}
\value{
A character string describing the OS, e.g. 'windows', 'linux',
or 'osx'
}
\description{
Return the Operating System name in lower case
}
