% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-param-est-comparison-df.R
\name{get_param_est_comparison_df}
\alias{get_param_est_comparison_df}
\title{Create the data frame that holds parameter estimates for several models
for comparison purposes in a table}
\usage{
get_param_est_comparison_df(
  models,
  model_nms,
  end_yr = models[[1]]$endyr,
  digits = 3,
  inc_loglike = TRUE,
  ...
)
}
\arguments{
\item{models}{A list of models which contain the MCMC output}

\item{model_nms}{A vector of names of the same length as the number of
models in the models list}

\item{end_yr}{The last year to include}

\item{digits}{The number of decimal points to include in the table}

\item{inc_loglike}{Logical. If \code{TRUE}, include the negative log-likelihood
values}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A data frame containing 1 column of parameter estimates for each
model in \code{models} list
}
\description{
Create the data frame that holds parameter estimates for several models
for comparison purposes in a table
}
