% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-post-cols.R
\name{get_post_cols}
\alias{get_post_cols}
\title{Extract time series columns for a given parameter and change the column
names to years. Optionally, divide all values in the resulting data frame
by some value}
\usage{
get_post_cols(df, pat, scale = 1, exact = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{pat}{A regular expression to be used to extract columns from \code{df}.
If the column names start with this, they will be extracted and renamed}

\item{scale}{A value to divide all resulting values in the table by}
}
\value{
A simple data frame with years as columns and values as elements
}
\description{
Extract time series columns for a given parameter and change the column
names to years. Optionally, divide all values in the resulting data frame
by some value
}
