% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-posterior-data.R
\name{get_posterior_data}
\alias{get_posterior_data}
\title{Get the posterior values for the given regular expressions of
parameter names}
\usage{
get_posterior_data(model, params_regex = NULL)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{params_regex}{A vector of regular expressions used to extract data
for parameter names}
}
\value{
A data frame of posterior vectors, one column for each of
the regular expressions in \code{params_regex}
}
\description{
Get the posterior values for the given regular expressions of
parameter names
}
\examples{
\dontrun{
get_posterior_data(base_model, "BH_steep")
get_posterior_data(base_model, "e")
get_posterior_data(base_model, "asdfg")
get_posterior_data(base_model, c("NatM", "SR_LN", "SR_BH_steep",
 "Q_extraSD"))
}
}
