% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-prior-string.R
\name{get_prior_string}
\alias{get_prior_string}
\title{Looks at the prior type p_type and phase, and if uniform will return
"Uniform"
If not uniform, it will parse the \code{vals} and build a string defining
the prior info.
If Fixed, it will return the initial value
If Lognormal, it will parse the \code{vals} and build a string defining the
prior info, with the exp function applied}
\usage{
get_prior_string(row, digits = 2)
}
\arguments{
\item{digits}{The number of decimal points to return}

\item{vals}{The parameter initial values and settings from the control file}
}
\value{
A string with the prior type and initial values in parentheses
}
\description{
Looks at the prior type p_type and phase, and if uniform will return
"Uniform"
If not uniform, it will parse the \code{vals} and build a string defining
the prior info.
If Fixed, it will return the initial value
If Lognormal, it will parse the \code{vals} and build a string defining the
prior info, with the exp function applied
}
