% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-quantities.R
\name{get_rec_ci_mle}
\alias{get_rec_ci_mle}
\title{Plot cohort recruitment MLE estimates with retrospectives}
\usage{
get_rec_ci_mle(model, end_yr = NULL, cohorts = NULL)
}
\arguments{
\item{model}{A model as loaded by \code{\link[=load_ss_files]{load_ss_files()}} with MLE retrospective
runs}

\item{end_yr}{End year for the plot}

\item{cohorts}{The cohort years to show}
}
\value{
A \link{ggplot2} object
}
\description{
Plot cohort recruitment MLE estimates with retrospectives
}
