% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-ss3-name.R
\name{get_ss3_exe_name}
\alias{get_ss3_exe_name}
\title{Get the name of the SS3 executable to use.}
\usage{
get_ss3_exe_name(ss_exe = NULL)
}
\arguments{
\item{ss_exe}{The name of executable to use or \code{NULL} to use the package
data variable \link{ss_executable}}
}
\value{
The name of the SS3 executable to use
error
}
\description{
Get the name of the SS3 executable to use.
}
\details{
Simple logic to either use:
\enumerate{
\item the one provided in the \code{ss_exe} argument, if it is not \code{NULL}
\item the one defined in \link{ss_executable} which is a package data variable
\item throw an error if both the above options are \code{NULL}
Note that this function does not check to see if the executable exists
or is on the 'PATH' or anything else, it is just a logic gate returning a
string
}
}
