% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-year-range.R
\name{get_year_range}
\alias{get_year_range}
\title{Get best start year and end year values given the data minimum and maximum
years and suggested start and end year values}
\usage{
get_year_range(start_yr = NA, end_yr = NA, min_yr = NA, max_yr = NA)
}
\arguments{
\item{start_yr}{The suggested start year}

\item{end_yr}{The suggested end year}

\item{min_yr}{The minimum year in the data}

\item{max_yr}{The maximum year in the data}
}
\value{
A named list of two, the start and end year that is best
}
\description{
Get best start year and end year values given the data minimum and maximum
years and suggested start and end year values
}
