% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-goback.R
\name{goback}
\alias{goback}
\title{Navigate back to the directory you were in before calling \code{\link[=gotest]{gotest()}}}
\usage{
goback()
}
\value{
Nothing
}
\description{
Navigate back to the directory you were in before calling \code{\link[=gotest]{gotest()}}
}
\details{
After running \code{\link[=gotest]{gotest()}}, you will be in a temporary directory,
and you'll be able to build a test document with arbitrary
knitr code chunks. Once you're done testing and checking your output,
run this function to return to the real document directory, and
to reset the \code{\link[here:here]{here::here()}} command so that it references the real
project again.

Assumes you're going back to a directory in the hake repository and that
the file \code{.here} exists in the hake repository root directory.
}
