% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gotest-goback.R
\name{gotest}
\alias{gotest}
\title{Create and change to a temporary directory and copy a skeleton
version of the document files for testing}
\usage{
gotest(copy_tmpdir = FALSE)
}
\arguments{
\item{copy_tmpdir}{Logical. If \code{TRUE}, copy the temporary path that
you were just placed in to the OS clipboard}
}
\value{
Nothing
}
\description{
Create and change to a temporary directory and copy a skeleton
version of the document files for testing
}
\details{
See the \emph{Debugging a figure or table, or anything else} section of the
README.md file for detailed explanation
}
