% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotest-beamer-get-src-dest-filenames.R
\name{gotest_beamer_get_src_dest_filenames}
\alias{gotest_beamer_get_src_dest_filenames}
\title{Create the lists of source and destination filenames needed for copying
for the beamer presentations testing}
\usage{
gotest_beamer_get_src_dest_filenames(bookdown_lst)
}
\value{
Nothing
}
\description{
Create the lists of source and destination filenames needed for copying
for the beamer presentations testing
}
\details{
Meant to be called by the wrapper function \code{\link[=gotest]{gotest()}}
}
