% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gotest-doc-get-src-dest-filenames.R
\name{gotest_doc_get_src_dest_filenames}
\alias{gotest_doc_get_src_dest_filenames}
\title{Create the lists of source and destination filenames needed for copying
for the main assessment document testing}
\usage{
gotest_doc_get_src_dest_filenames(bookdown_lst = NULL, figures_dir = NULL)
}
\value{
Nothing
}
\description{
Create the lists of source and destination filenames needed for copying
for the main assessment document testing
}
\details{
Meant to be called by the wrapper function \code{\link[=gotest]{gotest()}}
}
