% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-comparisons-nuts.R
\name{group_test_models}
\alias{group_test_models}
\title{Group the input list of models into a list of sub-lists of models which
are grouped by a common variable}
\usage{
group_test_models(lst, group_by = c("adapt_delta", "num_cores"))
}
\arguments{
\item{lst}{A list, as returned from \code{\link[=load_all_models_in_subdir]{load_all_models_in_subdir()}}}

\item{group_by}{A character string defining the variable to group by.
Must be one of "adapt_delta", "num_cores"}
}
\value{
A list of sub-lists, grouped by the unique value of the variable
defined by \code{group_by}
}
\description{
Group the input list of models into a list of sub-lists of models which
are grouped by a common variable
}
