% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-theme.R
\name{hake_theme}
\alias{hake_theme}
\title{Theme for the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plots in the document}
\usage{
hake_theme()
}
\description{
Theme for the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plots in the document
}
