% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-heatmap-add-extrap-yrs-wa.R
\name{heatmap_add_extrap_yrs_wa}
\alias{heatmap_add_extrap_yrs_wa}
\title{Add extrapolated years before and after the weight-at-age time
series data in \code{model$wtatage}}
\usage{
heatmap_add_extrap_yrs_wa(
  model = NULL,
  wa = NULL,
  pre_yrs = min(pull(dplyr::filter(model$wtatage, Yr > 0), Yr)):model$endyr,
  pre_func = mean,
  post_yrs = min(pull(dplyr::filter(model$wtatage, Yr > 0), Yr)):model$endyr,
  post_func = mean,
  pre_wa_vals = NULL,
  post_wa_vals = NULL,
  ...
)
}
\arguments{
\item{model}{An list of results read in from an SS model using
\code{\link[=load_ss_files]{load_ss_files()}}}

\item{wa}{The weight-at-age data frame found in \code{model$wtatage}}

\item{pre_yrs}{A vector of the years to use for the \code{pre_func}
function to fill in missing years before the weight-at-age data starts}

\item{pre_func}{The function to use on the data filtered by the years
given by \code{pre_yrs}}

\item{post_yrs}{A vector of the years to use for the \code{post_func}
function to fill in missing years after the weight-at-age data ends
(projection years)}

\item{post_func}{The function to use on the data filtered by the years
given by \code{post_yrs}}

\item{pre_wa_vals}{A vector of weight-at-age values to use instead of the
function \code{pre_func()} on the data found in the years defined by
\code{pre_yrs}. If this is not \code{NULL}, this vector will be used for all
years prior to the start of the weight-at-age data instead of the output
of \code{pre_func()}. Note this starts at age 0, so it may be one more
than you think}

\item{post_wa_vals}{A vector of weight-at-age values to use instead of the
function \code{post_func()} on the data found in the years defined by
\code{post_yrs}. If this is not \code{NULL}, this vector will be used for all
projection years instead of the output of \code{post_func()}  Note this
starts at age 0, so it may be one more than you think}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A data frame containing years (\code{yr}) column and columns for age,
represented as numbers (no text appended to them). The number of age
columns is the same as the number in \code{model$wtatage}
}
\description{
Add extrapolated years before and after the weight-at-age time
series data in \code{model$wtatage}
}
