% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-heatmap-calc-function.R
\name{heatmap_calc_function}
\alias{heatmap_calc_function}
\title{Calculate values for a given range of years, for ages which are
numerically-named columns in the data frame
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})}
\usage{
heatmap_calc_function(wa = NULL, func = mean, yr_col = "yr", ...)
}
\arguments{
\item{wa}{A weight-at-age data frame with a column \code{yr} for the years, and
numerically-named columns for ages}

\item{func}{A function to use on each age column to create a row of outputs}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A single row data frame containing columns for age,
represented as numbers. The number of age columns is the same as
the number in \code{wa}
}
\description{
Calculate values for a given range of years, for ages which are
numerically-named columns in the data frame
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})
}
