% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-extract-bf.R
\name{heatmap_extract_bf}
\alias{heatmap_extract_bf}
\title{Extract the boldface mask data frame for the heatmap plot
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})}
\usage{
heatmap_extract_bf(extrap_mask = NULL, fleet = NULL, wa = NULL, ...)
}
\arguments{
\item{extrap_mask}{A data frame where ages are columns (and start with
the letter 'a'). If the values are zero, the weight-at-age was
extrapolated/interpolated. If there is a value, the weight-at-age is data}

\item{fleet}{An integer value specifying which fleet you want plotted.
Fleet -2 will plot fecundity information.
Fleet -1 will plot population weight-at-age for the middle of the year.
Fleet 0 will plot population weight-at-age for the beginning of the year.
Positive values for fleet will link to a modeled fleet.}

\item{wa}{A weight-at-age data frame as created by \code{\link[=heatmap_extract_wa]{heatmap_extract_wa()}}}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A data frame containing years (\code{yr}) column and columns for age,
represented as numbers (no text appended to them). The number of age
columns is the same as the number in \code{model$wtatage}. Each cell is \code{TRUE}
or \code{FALSE} representing if text in that cell will be boldface or not.
The structure of the returned data frame will be identical to the
structure of \code{wa}
}
\description{
Extract the boldface mask data frame for the heatmap plot
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})
}
