% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-extract-wa.R
\name{heatmap_extract_wa}
\alias{heatmap_extract_wa}
\title{Extract the weight-at-age data frame for the heatmap plot
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})}
\usage{
heatmap_extract_wa(
  model = NULL,
  fleet = NULL,
  apply_pre_yrs = min(pull(filter(model$wtatage, Yr > 0), Yr)):model$endyr,
  apply_pre_func = mean,
  apply_post_yrs = min(pull(filter(model$wtatage, Yr > 0), Yr)):model$endyr,
  apply_post_func = mean,
  pre_wa_vals = NULL,
  post_wa_vals = NULL,
  ...
)
}
\arguments{
\item{model}{An list of results read in from an SS model using
\code{\link[=load_ss_files]{load_ss_files()}}}

\item{fleet}{An integer value specifying which fleet you want plotted.
Fleet -2 will plot fecundity information.
Fleet -1 will plot population weight-at-age for the middle of the year.
Fleet 0 will plot population weight-at-age for the beginning of the year.
Positive values for fleet will link to a modeled fleet.}

\item{apply_pre_yrs}{A vector of the years to use for the \code{apply_pre_func}
function to fill in missing years before the weight-at-age data starts}

\item{apply_pre_func}{The function to use on the data filtered by the years
given by \code{apply_pre_yrs}}

\item{apply_post_yrs}{A vector of the years to use for the \code{apply_post_func}
function to fill in missing years after the weight-at-age data ends
(projection years)}

\item{apply_post_func}{The function to use on the data filtered by the years
given by \code{apply_post_yrs}}

\item{pre_wa_vals}{A vector of weight-at-age values to use instead of the
function \code{apply_pre_func()} on the data found in the years defined by
\code{apply_pre_yrs}. If this is not \code{NULL}, this vector will be used for all
years prior to the start of the weight-at-age data instead of the output
of \code{apply_pre_func()}. Note this starts at age 0, so it may be one more
than you think}

\item{post_wa_vals}{A vector of weight-at-age values to use instead of the
function \code{apply_post_func()} on the data found in the years defined by
\code{apply_post_yrs}. If this is not \code{NULL}, this vector will be used for all
projection years instead of the output of \code{apply_post_func()}  Note this
starts at age 0, so it may be one more than you think}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A data frame containing years (\code{yr}) column and columns for age,
represented as numbers (no text appended to them). The number of age
columns is the same as the number in \code{model$wtatage}
}
\description{
Extract the weight-at-age data frame for the heatmap plot
(\code{\link[=plot_weight_at_age_heatmap]{plot_weight_at_age_heatmap()}})
}
