% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-insert-row.R
\name{insert_row}
\alias{insert_row}
\title{Insert a new row made up of the vector \code{row_vec} at row \code{row_ind} in}
\usage{
insert_row(d, row_vec, row_ind, ...)
}
\arguments{
\item{d}{A data frame}

\item{row_vec}{A vector which has the same number of elements as columns
in \code{d}}

\item{row_ind}{The row index to insert the new row at. Whatever row was
at that index before is pushed down, as are the ones following it}

\item{...}{Arguments passed to \code{\link[=vec2df]{vec2df()}}}
}
\value{
A data frame with the new row inserted
}
\description{
Insert a new row made up of the vector \code{row_vec} at row \code{row_ind} in
}
