% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-chunks.R
\name{interlace_chunks}
\alias{interlace_chunks}
\title{Interlace chunks of code. Opposite effect to \code{\link[=extract_chunks]{extract_chunks()}}}
\usage{
interlace_chunks(lst)
}
\arguments{
\item{lst}{A list, as returned by \code{\link[=extract_chunks]{extract_chunks()}}}
}
\value{
A vector, with the \code{between} chunks and \code{inbetween} chunks
interlaced
}
\description{
Interlace chunks of code. Opposite effect to \code{\link[=extract_chunks]{extract_chunks()}}
}
