% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_posteriors}
\alias{key_posteriors}
\title{A list of key posteriors for this assessment}
\format{
A list of regular expressions used to find the key posterior names
in MCMC output from SS3
}
\usage{
key_posteriors
}
\description{
A list of key posteriors for this assessment
}
\keyword{datasets}
