% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_posteriors_fn}
\alias{key_posteriors_fn}
\title{A character string containing the file name for the key posteriors CSV file}
\format{
A character string
}
\usage{
key_posteriors_fn
}
\description{
A character string containing the file name for the key posteriors CSV file
}
\keyword{datasets}
