% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_posteriors_titles}
\alias{key_posteriors_titles}
\title{A list of key posterior names for this assessment}
\format{
A list of names for key posteriors. These are the names that
will be visible n the assessment document
}
\usage{
key_posteriors_titles
}
\description{
A list of key posterior names for this assessment
}
\keyword{datasets}
