% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kriging_pars_df}
\alias{kriging_pars_df}
\title{A data frame containing information about the Kriging extrapolation
performed on the survey data}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{Year}{Year of data}
\item{SearchRadius}{The radius used in the kriging extrapolation}
\item{kmin}{Kringing parameter}
\item{kmax}{Kringing parameter}
}
}
\usage{
kriging_pars_df
}
\description{
A data frame containing information about the Kriging extrapolation
performed on the survey data
}
\keyword{datasets}
