% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label-parsed-space.R
\name{label_parsed_space}
\alias{label_parsed_space}
\title{Updates \code{\link[ggplot2:labellers]{ggplot2::label_parsed()}} to accommodate spaces}
\usage{
label_parsed_space(labels)
}
\arguments{
\item{labels}{Labels to use}
}
\description{
Updates \code{\link[ggplot2:labellers]{ggplot2::label_parsed()}} to accommodate spaces
}
