% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landings_tac_fn}
\alias{landings_tac_fn}
\title{The filename for the landings and TAC history for the US and Canada}
\format{
A character string
}
\usage{
landings_tac_fn
}
\description{
The filename for the landings and TAC history for the US and Canada
}
\keyword{datasets}
