% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_cohorts}
\alias{large_cohorts}
\title{A vector of years when larger-than-normal recruitment took place}
\format{
A vector of numerical years
}
\usage{
large_cohorts
}
\description{
A vector of years when larger-than-normal recruitment took place
}
\keyword{datasets}
