% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.amp}
\alias{latex.amp}
\title{Create a string with `n` ampersands separated by spaces}
\usage{
latex.amp(n = 1)
}
\arguments{
\item{n}{}
}
\value{
A string with `n` ampersands seperated by spaces
}
\description{
Create a string with `n` ampersands separated by spaces
}
\details{
The string will have one leading and one trailing space
}
