% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.bold}
\alias{latex.bold}
\title{Wrap the given text with the latex \\textbf{} macro around it}
\usage{
latex.bold(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The given text with the latex \\textbf{} macro around it
}
\description{
Wrap the given text with the latex \\textbf{} macro around it
}
