% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.cline}
\alias{latex.cline}
\title{Provide latex code to draw a horizontal line across the columns specified}
\usage{
latex.cline(cols)
}
\arguments{
\item{cols}{A string in this format: "1-3" which means the line should go
across columns 1 to 3}
}
\value{
A string of latex code to draw a horizontal line across the
columns specified
}
\description{
Provide latex code to draw a horizontal line across the columns specified
}
