% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.mlc}
\alias{latex.mlc}
\title{Returns a string which has been glued together using multi-line-cell
macro for latex}
\usage{
latex.mlc(latex.vec, make.bold = TRUE)
}
\arguments{
\item{latex.vec}{A vector of the strings to glue together}

\item{make.bold}{Logical. If TRUE, make the text bold by inserting
a \\textbf{} macro}
}
\value{
A string
}
\description{
Returns a string which has been glued together using multi-line-cell
macro for latex
}
