% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.mrow}
\alias{latex.mrow}
\title{Wrap the given text with the latex \\multirow{} macro around it}
\usage{
latex.mrow(nrow, just, txt)
}
\arguments{
\item{just}{Justification, e.g. "l", "c", or "r" for left, center, right}

\item{txt}{The text}

\item{ncol}{The number of columns}
}
\value{
The given text with the latex \\multirow{} macro around it
}
\description{
Wrap the given text with the latex \\multirow{} macro around it
}
