% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.paste}
\alias{latex.paste}
\title{Create a string comprised of each element in the vector \code{vec} with an
ampersand in between}
\usage{
latex.paste(vec)
}
\arguments{
\item{vec}{A vector of characters}
}
\value{
A string comprised of each element in the vector \code{vec} with an
ampersand in between
}
\description{
Create a string comprised of each element in the vector \code{vec} with an
ampersand in between
}
\details{
The string will have one leading and one trailing space
}
