% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.size.str}
\alias{latex.size.str}
\title{Creates a string which has the given font size and space size applied}
\usage{
latex.size.str(fnt.size, spc.size)
}
\arguments{
\item{fnt.size}{The font size}

\item{spc.size}{The space size (between text size)}
}
\value{
A string which has the given font size and space size applied
}
\description{
Creates a string which has the given font size and space size applied
}
