% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.subscr}
\alias{latex.subscr}
\title{Creates a latex string with \code{main.txt} subscripted by \code{subscr.txt}}
\usage{
latex.subscr(main.txt, subscr.txt)
}
\arguments{
\item{main.txt}{The main text to subscript}

\item{subscr.txt}{The subscript text}
}
\value{
A latex string with \code{main.txt} subscripted by \code{subscr.txt}
}
\description{
Creates a latex string with \code{main.txt} subscripted by \code{subscr.txt}
}
