% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex.supscr}
\alias{latex.supscr}
\title{Creates a latex string with `main.txt` superscripted by `supscr.txt`}
\usage{
latex.supscr(main.txt, supscr.txt)
}
\arguments{
\item{main.txt}{The main text to superscript}

\item{supscr.txt}{The superscript text}
}
\value{
A latex string with `main.txt` superscripted by `supscr.txt`
}
\description{
Creates a latex string with `main.txt` superscripted by `supscr.txt`
}
