% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latex-functions.R
\name{latex_amp}
\alias{latex_amp}
\title{Create a string with \code{n} ampersands separated by spaces}
\usage{
latex_amp(n = 1)
}
\arguments{
\item{n}{The number of space-ampersands to return}
}
\value{
A string with \code{n} ampersands separated by spaces
}
\description{
Create a string with \code{n} ampersands separated by spaces
}
\details{
The string will have one leading and one trailing space
}
