% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latex-functions.R
\name{latex_bold}
\alias{latex_bold}
\title{Wrap the given text with the latex \\textbf{} macro around it}
\usage{
latex_bold(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The given text with the latex \\textbf{} macro around it
}
\description{
Wrap the given text with the latex \\textbf{} macro around it
}
