% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex_cmidr}
\alias{latex_cmidr}
\title{Provide latex code to draw a horizontal line across the columns specified}
\usage{
latex_cmidr(cols, trim = "r")
}
\arguments{
\item{cols}{A string in this format: "1-3" which means the line should go
across columns 1 to 3}

\item{trim}{Can be l, r, or lr and tells it to trim the line a bit so that
if there are two lines they don't touch in the middle. See \code{booktabs}}
}
\value{
As string of latex code to draw a horizontal line across the columns specified
}
\description{
Provide latex code to draw a horizontal line across the columns specified
}
