% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex_continue}
\alias{latex_continue}
\title{Return the necessary latex to repeat longtable headers for continuing pages}
\usage{
latex_continue(n_col = 1, header = "Default")
}
\arguments{
\item{n_col}{The number of columns in the table}

\item{header}{The table header, needed for subsequent pages}
}
\value{
vector of strings needed to repeat the header of a longtable
and a footer which says 'Continued on next page ...'
}
\description{
Return the necessary latex to repeat longtable headers for continuing pages
}
