% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latex-functions.R
\name{latex_mcol}
\alias{latex_mcol}
\title{Wrap the given text with the latex \\multicolumnn{} macro around it}
\usage{
latex_mcol(ncol, just, txt)
}
\arguments{
\item{ncol}{The number of columns}

\item{just}{Justification, e.g. "l", "c", or "r" for left, center, right}

\item{txt}{The text}
}
\value{
The given text with the latex \\multicolumn{} macro around it
}
\description{
Wrap the given text with the latex \\multicolumnn{} macro around it
}
