% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex_mlc}
\alias{latex_mlc}
\title{Returns a string which has been glued together using multi-line-cell
macro for latex}
\usage{
latex_mlc(latex_vec, make_bold = TRUE)
}
\arguments{
\item{latex_vec}{A vector of the strings to glue together}

\item{make_bold}{Logical. If TRUE, make the text bold by inserting
a \\textbf{} macro}
}
\value{
A string
}
\description{
Returns a string which has been glued together using multi-line-cell
macro for latex
}
