% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex_mrow}
\alias{latex_mrow}
\title{Wrap the given text with the latex \\multirow{} macro}
\usage{
latex_mrow(nrow, just, txt)
}
\arguments{
\item{nrow}{The number of rows in the multirow cell}

\item{just}{Justification, e.g. "l", "c", or "r" for left, center, right}

\item{txt}{The text}
}
\value{
The given text wrapped in the latex \\multirow{} macro
}
\description{
Wrap the given text with the latex \\multirow{} macro
}
