% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latex-functions.R
\name{latex_size_str}
\alias{latex_size_str}
\title{Creates a string which has the given font size and space size applied}
\usage{
latex_size_str(fnt_size, spc_size)
}
\arguments{
\item{fnt.size}{The font size}

\item{spc.size}{The space size (between text size)}
}
\value{
A string which has the given font size and space size applied
}
\description{
Creates a string which has the given font size and space size applied
}
