% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latex-functions.R
\name{latex_subscr}
\alias{latex_subscr}
\title{Creates a latex string with \code{main_txt} subscripted by \code{subscr_txt}}
\usage{
latex_subscr(main_txt, subscr_txt)
}
\arguments{
\item{main_txt}{The main text to subscript}

\item{subscr_txt}{The subscript text}
}
\value{
A latex string with \code{main_txt} subscripted by \code{subscr_txt}
}
\description{
Creates a latex string with \code{main_txt} subscripted by \code{subscr_txt}
}
