% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latex_supscr}
\alias{latex_supscr}
\title{Creates a latex string with \code{main_txt} superscripted by \code{supscr_txt}}
\usage{
latex_supscr(main_txt, supscr_txt)
}
\arguments{
\item{main_txt}{The main text to superscript}

\item{supscr_txt}{The superscript text}
}
\value{
A latex string with \code{main_txt} superscripted by \code{supscr_txt}
}
\description{
Creates a latex string with \code{main_txt} superscripted by \code{supscr_txt}
}
