% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-linebreaker.R
\name{linebreaker}
\alias{linebreaker}
\title{Place a LaTeX \verb{\\\\makecell\{\}} macro around some text}
\usage{
linebreaker(x, align = c("l", "c", "r"), ...)
}
\arguments{
\item{x}{A vector of character strings}

\item{align}{Choose from "l", "c" or "r" for "left", "center", and "right"
respectively}

\item{...}{Arguments passed to \code{\link[kableExtra:linebreak]{kableExtra::linebreak()}}}
}
\value{
A vector of modified character strings or \code{NULL} if \code{x} is \code{NULL} or
\code{NA} if \code{x} is \code{NA}. Any empty strings will remain the same, without
wrapping text
}
\description{
Place a LaTeX \verb{\\\\makecell\{\}} macro around some text
}
\details{
This function is a wrapper around \code{\link[kableExtra:linebreak]{kableExtra::linebreak()}} which itself
does not wrap the makecell macro around strings without the linebreak
character. For table headers, this needs to be done to keep all headers
aligned the same way with respect to each other.
}
