% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-.R
\name{load_age_fishery}
\alias{load_age_fishery}
\title{Load data and create age compositions}
\usage{
load_age_fishery(dirdata = hakedata_wd())
}
\arguments{
\item{dirdata}{The directory with the data.}
}
\description{
Load data and create age compositions
}
