% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-comparisons-nuts.R
\name{load_all_models_in_subdir}
\alias{load_all_models_in_subdir}
\title{Load all models from a directory with subdirectories}
\usage{
load_all_models_in_subdir(
  num_samples = NULL,
  type_path = "/srv/hake/models/2024/02-version/05-test-models"
)
}
\arguments{
\item{num_samples}{The number of samples to use (extracted from the
filenames, so must be one of 500, 1000, 2500, 5000, 10000).}

\item{type_path}{The path where the subdirectories reside}
}
\value{
A list of lists
}
\description{
Load all models from a directory with subdirectories
}
\details{
Creates a list of lists of models, each model in the sublist has the same
adapt_delta value as the others in that list. The adapt_delta value is
extracted from the filename. The names of the list elements are the
adapt delta values as character strings
}
