% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-can-age-data.R
\name{load_can_age_data}
\alias{load_can_age_data}
\title{Load the Canadian age data from the CSV file which contains it}
\usage{
load_can_age_data(fn)
}
\arguments{
\item{fn}{The CSV filename}
}
\value{
A list of 6 data frames, one age comps and one number of samples
for each fleet: Freezer trawlers, shoreside, and JV
}
\description{
Load the Canadian age data from the CSV file which contains it
}
