% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ct-levels.R
\name{load_ct_levels}
\alias{load_ct_levels}
\title{Load catch levels from SS forecast files. This is used to retrieve ending
values after running the \code{\link[=run_ct_levels]{run_ct_levels()}} function for default HR, SPR 100,
and stable catch}
\usage{
load_ct_levels(model, ct_levels = NULL, ...)
}
\arguments{
\item{ct_levels}{The catch levels list as defined in
\code{forecast-catch-levels.R}}

\item{...}{Absorbs arguments intended for other functions}

\item{ct_levels_path}{The path for the catch-levels output}
}
\value{
A list of 3-element lists of vectors of 3 catch levels
corresponding to:
a) SPR-100\%
b) Default harvest policy
c) Stable catch
Return object looks the same as the \code{ct_levels} object but with three more
elements
}
\description{
Load catch levels from SS forecast files. This is used to retrieve ending
values after running the \code{\link[=run_ct_levels]{run_ct_levels()}} function for default HR, SPR 100,
and stable catch
}
\details{
Assumes \code{\link[=run_ct_levels]{run_ct_levels()}} function has been run and the forecast
files are populated with 3 forecast years.
}
