% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-age-comps.R
\name{load_extra_mcmc_age_comps}
\alias{load_extra_mcmc_age_comps}
\title{Load the age composition estimates from the extra MCMC output}
\usage{
load_extra_mcmc_age_comps(
  compreps,
  verbose = TRUE,
  start_yr = NULL,
  end_yr = NULL,
  progress_n,
  ret_quants = TRUE,
  ...
)
}
\arguments{
\item{compreps}{A list of vectors representing one element per
compreport file}

\item{verbose}{Logical. If \code{TRUE}, show messages}

\item{start_yr}{Filter years earlier than this out of the results. If
\code{NULL}, no filtering will occur for the start year}

\item{end_yr}{Filter years greater than this out of the results. If
\code{NULL}, no filtering will occur for the end year}

\item{progress_n}{Report every time this many list items are processed.
Consider how many posteriors there are, this should be a fairly large
proportion of that (around 1/8th) or there will be too much output and
it will run really slow}

\item{ret_quants}{Logical. If \code{TRUE}, return the quantiles of the data.
If \code{FALSE}, return all posteriors (the \code{iter} column will be present in
this case)}

\item{...}{Arguments passed to \code{\link[=load_extra_mcmc_get_chunk]{load_extra_mcmc_get_chunk()}}}
}
\value{
A data frame with columns fleet, yr, iter (posterior number) and
a column for each age in the output. If \code{ret_quants} is \code{TRUE}, the data
frame will be in the long format and have columns fleet, yr, age, and
one column for each of the quantiles found in the \link{probs} variable
(normally 2.5\%, 50\%, and 97.5\%)
}
\description{
Load the age composition estimates from the extra MCMC output
}
