% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-atage.R
\name{load_extra_mcmc_atage}
\alias{load_extra_mcmc_atage}
\title{Load at-age data chunks from a list of report file output,
and calculate the medians at-age by year}
\usage{
load_extra_mcmc_atage(
  reps,
  verbose = TRUE,
  txt = "output",
  scale = 1,
  start_yr = NULL,
  end_yr = NULL,
  ...
)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{verbose}{Logical. Show messages}

\item{txt}{Text to write to the screen while processing}

\item{scale}{Value to divide the values by in the output table}

\item{start_yr}{Filter years earlier than this out of the results. If
\code{NULL}, no filtering will occur for the start year}

\item{end_yr}{Filter years greater than this out of the results. If
\code{NULL}, no filtering will occur for the end year}

\item{...}{Absorbs arguments meant for other functions}

\item{probs}{The quantile values to use on the MCMC posterior data}

\item{progress_n}{Report every time this many list items are processed.
Consider how many posteriors there are, this should be a fairly large
proportion of that (around 1/8th) or there will be too much output and
it will run really slow}

\item{beg_pat}{A regular expression that marks the beginning of the chunk
of data}

\item{end_pat}{A regular expression that marks the end of the chunk of data}

\item{beg_off}{Offset from the beginning marker to the start of the data}

\item{end_off}{Offset from the beginning marker to the end of the data.
Should be a negative number}
}
\value{
A list with one element being the data frame of values by year with
every posterior included (\code{atage}) and the other being a much smaller data
frame containing the median across posteriors by year for the values (\code{med})
}
\description{
Load at-age data chunks from a list of report file output,
and calculate the medians at-age by year
}
