% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-get-chunk.R
\name{load_extra_mcmc_get_chunk}
\alias{load_extra_mcmc_get_chunk}
\title{Extracts a vector of lines of data or output (a chunk) from each
element of a list of identically-formatted vectors, typically read
in from files with identical formats (SS3 Report files)}
\usage{
load_extra_mcmc_get_chunk(lst, beg_pat, end_pat, ...)
}
\arguments{
\item{lst}{A list of identically-formatted vectors, typically from a
call to \code{\link[=readLines]{readLines()}} where each file is identically-formatted}

\item{beg_pat}{A regular expression that marks the beginning of the chunk
of data}

\item{end_pat}{A regular expression that marks the end of the chunk of data}
}
\value{
A list of vectors of output or data, each list element will have
the exact same number of lines and the same structure within those lines
}
\description{
Extracts a vector of lines of data or output (a chunk) from each
element of a list of identically-formatted vectors, typically read
in from files with identical formats (SS3 Report files)
}
