% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-likelihoods.R
\name{load_extra_mcmc_likelihoods}
\alias{load_extra_mcmc_likelihoods}
\title{Load the likelihoods as defined in Report_mce_xxxx.sso files}
\usage{
load_extra_mcmc_likelihoods(reps, verbose = TRUE, ...)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{verbose}{Logical. Show messages}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A list of outputs
}
\description{
Load the likelihoods as defined in Report_mce_xxxx.sso files
}
