% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-m.R
\name{load_extra_mcmc_m}
\alias{load_extra_mcmc_m}
\title{Load the natural mortality outputs from the extra MCMC output}
\usage{
load_extra_mcmc_m(reps, probs, verbose = TRUE, progress_n, ...)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{probs}{The quantile values to use on the MCMC posterior data}

\item{verbose}{Logical. Show messages}

\item{progress_n}{Report every time this many list items are processed.
Consider how many posteriors there are, this should be a fairly large
proportion of that (around 1/8th) or there will be too much output and
it will run really slow}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A list of outputs
}
\description{
Load the natural mortality outputs from the extra MCMC output
}
