% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-recr-cohorts.R
\name{load_extra_mcmc_recr_cohorts}
\alias{load_extra_mcmc_recr_cohorts}
\title{Load the recruitment outputs from the extra MCMC output for certain cohorts}
\usage{
load_extra_mcmc_recr_cohorts(reps, cohorts = NULL, verbose = TRUE, ...)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{cohorts}{A vector of years of birth for cohorts to extract
recruitment quantiles for}

\item{verbose}{Logical. Show messages}

\item{...}{Absorbs arguments meant for other functions}

\item{progress_n}{Report every time this many list items are processed.
Consider how many posteriors there are, this should be a fairly large
proportion of that (around 1/8th) or there will be too much output and
it will run really slow}
}
\value{
A list of outputs
}
\description{
Load the recruitment outputs from the extra MCMC output for certain cohorts
}
