% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-recr-devs.R
\name{load_extra_mcmc_recr_devs}
\alias{load_extra_mcmc_recr_devs}
\title{Load the inital numbers-at-age as defined in Report_mce_xxxx.sso files}
\usage{
load_extra_mcmc_recr_devs(reps, start_yr, end_yr, verbose = TRUE, ...)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{verbose}{Logical. Show messages}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A list of outputs
}
\description{
Load the inital numbers-at-age as defined in Report_mce_xxxx.sso files
}
