% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-mcmc-vals.R
\name{load_mcmc_vals}
\alias{load_mcmc_vals}
\title{Load some values from the MCMC output and store in a list}
\usage{
load_mcmc_vals(model, assess_yr, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{model}{A model list as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{assess_yr}{The current assessment year}

\item{probs}{A vector of 3 values, the lower CI, median, and upper CI}
}
\value{
A list of values extracted from the MCMC data frame of a model
}
\description{
Load some values from the MCMC output and store in a list
}
