% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parameter-priors.R
\name{load_parameter_priors}
\alias{load_parameter_priors}
\title{Load values (priors) from the \code{parameters} data frame in the model output}
\usage{
load_parameter_priors(model)
}
\arguments{
\item{model}{A model list as created by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A list of values extracted from the \code{parameter} data frame
}
\description{
Load values (priors) from the \code{parameters} data frame in the model output
}
