% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-retrospectives.R
\name{load_retrosepctives}
\alias{load_retrosepctives}
\title{Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt}
\usage{
load_retrosepctives(retro_path, retro_yrs, ...)
}
\arguments{
\item{retro_path}{The path in which the retrospective directories reside}

\item{retro_yrs}{A vector of the years to run retrospectives for}

\item{...}{Arguments passed to \code{\link[=load_ss_files]{load_ss_files()}} and \code{\link[=load_extra_mcmc]{load_extra_mcmc()}}}
}
\value{
The list of retrospective outputs
}
\description{
Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt
}
