% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-retrospectives.R
\name{load_retrospectives}
\alias{load_retrospectives}
\title{Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt}
\usage{
load_retrospectives(retro_path, retro_yrs, ...)
}
\arguments{
\item{retro_path}{The path in which the retrospective directories reside}

\item{retrospective_yrs}{A vector of years for the retrospectives}

\item{end_yr}{The end year of the model. Retrospectives will have
their \code{endyr} element set to this minus the numbers of years of the
retorspective}
}
\value{
The list of retrospective outputs
}
\description{
Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt
}
