% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ss-files.R
\name{load_ss_files}
\alias{load_ss_files}
\title{Load all the SS files for output and input, and return the model object}
\usage{
load_ss_files(
  model_path = NA,
  key_posts = c("NatM", "SR_LN", "SR_BH_steep", "Q_extraSD", "ln.EffN_mult._1",
    "ln.EffN_mult._2"),
  posts_fn = "posteriors.sso",
  derposts_fn = "derived_posteriors.sso",
  key_posts_fn = "keyposteriors.csv",
  nuisance_posts_fn = "nuisanceposteriors.csv",
  printstats = FALSE,
  ...
)
}
\arguments{
\item{model_path}{The directory that the model is in}

\item{key_posts}{Vector of key posteriors used to create key posteriors file}

\item{posts_fn}{Posteriors output filename}

\item{key_posts_fn}{Name of the key posteriors filename}

\item{nuisance_posts_fn}{Name of the nuisance posteriors filename}

\item{printstats}{Print info on each model loaded via \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A model object representing the output from the SS model
}
\description{
Load all the SS files for output and input, and return the model object
}
\details{
If MCMC directory is present, load that and perform calculations
for mcmc parameters.
}
