% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_alpha}
\alias{main_alpha}
\title{The transparency to use by default for all fills in the plots in the
document}
\format{
A numerical value between 0 and 1
}
\usage{
main_alpha
}
\description{
The transparency to use by default for all fills in the plots in the
document
}
\keyword{datasets}
