% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_fill}
\alias{main_fill}
\title{The color to use by default for all fills in the plots in the document}
\format{
A character string or numerical value representing an official
color in R
}
\usage{
main_fill
}
\description{
The color to use by default for all fills in the plots in the document
}
\keyword{datasets}
