% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_tick_length}
\alias{major_tick_length}
\title{The length of the major tick marks in plots in cm}
\format{
A numerical value
}
\usage{
major_tick_length
}
\description{
The length of the major tick marks in plots in cm
}
\keyword{datasets}
