% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-comp-forecast.R
\name{make.age.comp.forecast.plot}
\alias{make.age.comp.forecast.plot}
\title{Make a plot shoeing the forecast age ompositions for the next year}
\usage{
make.age.comp.forecast.plot(model)
}
\arguments{
\item{model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
a base R plot
}
\description{
Make a plot shoeing the forecast age ompositions for the next year
}
