% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-data.R
\name{make.age.comp.plot}
\alias{make.age.comp.plot}
\title{A figure for age-composition data with years (x axis) and ages (y axis)}
\usage{
make.age.comp.plot(
  d,
  yrs = (year(now()) - 6):(year(now()) - 1),
  title = NULL,
  title.cex = 1,
  bg.col = "blue",
  ...
)
}
\arguments{
\item{d}{The data, with years in the rows and ages in the columns.}

\item{yrs}{A vector of years that you want to include. This vector will be used to subset the
data in \code{d} based on rownames.}

\item{title}{The title to pass to \link{plotBubbles} \code{main} parameter.}

\item{title.cex}{A numeric value specifying the size of the text for the title.}

\item{bg.col}{A color for the \code{fill} of the bubbles.}

\item{...}{Arguments to pass to \link{plotBubbles} such as \code{hide0}.}
}
\description{
A bubble plot of age compositions by year that
facilitates tracking cohorts through time.
The function is a wrapper for \link{plotBubbles} that subsets the
data and sets the color scheme.
}
