% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-assessment-history.R
\name{make.assessment.history.disp.plot}
\alias{make.assessment.history.disp.plot}
\title{Make an interquartile dispersion plot fir historic assessments}
\usage{
make.assessment.history.disp.plot(base, assessment.history.disp)
}
\arguments{
\item{base}{This year's base model}

\item{assessment.history.disp}{A data frame read in from the assessment history SSBdispersion file}
}
\value{
A base R plot
}
\description{
Make an interquartile dispersion plot fir historic assessments
}
