% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-assessment-history.R
\name{make.assessment.history.plot}
\alias{make.assessment.history.plot}
\title{Plot the history of assessments spawning biomass estimated trajectory}
\usage{
make.assessment.history.plot(
  base,
  assessment.history,
  par.mar = c(3.1, 3.1, 1.1, 6.1),
  cex.legend = 0.5
)
}
\arguments{
\item{base}{This year's base model}

\item{assessment.history}{A data frame read in from the assessment
history file}

\item{par.mar}{par(mar = c(x, x, x, x)) values; default is for doc}

\item{cex.legend}{The text size for the legend}
}
\description{
Plot the history of assessments spawning biomass estimated trajectory
}
\details{
The csv file should be set up with blanks for SSB in the most
recent year
}
