% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-timeseries.R
\name{make.ci.posterior.table}
\alias{make.ci.posterior.table}
\title{Make a latex table of credible intervals with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.}
\usage{
make.ci.posterior.table(
  model,
  start_yr,
  end_yr,
  weight.factor = 1000,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  digits = 1,
  lower_col = NULL,
  upper_col = NULL
)
}
\arguments{
\item{model}{A model list as output by [load_models()]}

\item{start_yr}{Start year in the table}

\item{end_yr}{End year in the table}

\item{weight.factor}{Factor to divide biomass values by and multiply the recruitment by}

\item{xcaption}{Caption for the latex table}

\item{xlabel}{Label for the latex table}

\item{font.size}{Size of font}

\item{space.size}{Vertical spacing between text in table}

\item{digits}{Number of decimal points to round to}

\item{lower_col}{Character string name of the column containing the lower credible interval values}

\item{upper_col}{Character string name of the column containing the upper credible interval values}
}
\value{
An [xtable::xtable()]
}
\description{
Make a latex table of credible intervals with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.
}
