% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-data.R
\name{make.cumulative.catch.plot}
\alias{make.cumulative.catch.plot}
\title{Create a plot of cumulative catch}
\usage{
make.cumulative.catch.plot(
  d,
  yrs = (year(now()) - 4):(year(now()) - 1),
  title = NULL,
  scale = 1000,
  title.cex = 1,
  cex.axis = 1,
  leg.cex = 1
)
}
\arguments{
\item{d}{The data as found in the catch-by-month CSV files}

\item{yrs}{A vector of years to include}

\item{title}{The plot title}

\item{scale}{A value to divide the catch by}

\item{title.cex}{The font size for the title}

\item{cex.axis}{The font size for the axis titles}

\item{leg.cex}{The font size for the legend}
}
\value{
A base R plot
}
\description{
Create a plot of cumulative catch
}
