% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-probs.R
\name{make.historical.probs.plot}
\alias{make.historical.probs.plot}
\title{Plots to compare historical and current probabilities}
\usage{
make.historical.probs.plot(
  model,
  type = "decline",
  end = assess_yr - 1,
  xLim = NULL,
  add.50 = TRUE,
  add.50.col = "darkgrey",
  one.year = 2019,
  add.projs = FALSE,
  num.projs = catch.levels.num,
  colors = c("red", "blue"),
  pch = c(16, 17),
  lwd.val = 1,
  legend.cex = 1,
  legend.loc = "bottomright",
  main.title = NULL,
  legend.text = c("From year t's assessment", "From current base model"),
  add.to.plot = FALSE,
  ...
)
}
\arguments{
\item{model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{"decline" to show probs of spawning biomass declining in year
after historical assessment year, "decline.one.year" to show that for just
one year (to explain in a talk), "bforty" to show prob of being below
\verb{B_40\\\%} in the year after ther historical assessment year}

\item{end}{final year for calculations}

\item{xLim}{range of x (years) axis}

\item{add.50}{Whether to add horizontal line at 50\%}

\item{add.50.col}{Colour for 50\% line}

\item{one.year}{A single year to plot (may automatically work for more
years)}

\item{add.projs}{Whether to add future projections from current base model}

\item{num.projs}{Number of projection catch levels to show}

\item{colors}{The colors to use for the lines in the plot}

\item{pch}{change legend call also if change this}

\item{lwd.val}{Line width}

\item{legend.cex}{Legend font size}

\item{legend.loc}{Legend location}

\item{main.title}{Main title text}

\item{legend.text}{Legent title text}

\item{add.to.plot}{whether to add to an existing plot}

\item{...}{Arguments passed to \code{\link[=combine_historical_probs]{combine_historical_probs()}}}
}
\value{
A base R plot
}
\description{
Plot either:
(a) base model's estimate of P(B_t+1 < B_t) and compare with the estimate of
that from the year t assessment (using the known catch in year t)
(b) base model's estimate of P(B_t+1 < B_40\%) and compare with the estimates
of that from the year t assessment (using the known catch in year
t).
Included in management presentation.
}
