% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-age.R
\name{make.input.age.data.table}
\alias{make.input.age.data.table}
\title{Make an at-age table showing percentage of each age by year}
\usage{
make.input.age.data.table(
  model,
  fleet = 1,
  start_yr,
  end_yr,
  csv.dir = "out-csv",
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "htbp",
  decimals = 2
)
}
\arguments{
\item{model}{The model to make the table for}

\item{fleet}{The fishing fleet (1 = commercial, 2 = survey)}

\item{start_yr}{First year in table}

\item{end_yr}{Last year in table}

\item{csv.dir}{Directory for output of the table in csv format}

\item{xcaption}{Caption for the table}

\item{xlabel}{Latex label to use in the main document for this table}

\item{font.size}{Point size of font}

\item{space.size}{Vertical space between rows of table}

\item{placement}{Latex table placement string}

\item{decimals}{Number if decimal points to show}
}
\value{
an \code{\link[xtable:xtable]{xtable::xtable()}}
}
\description{
Make an at-age table showing percentage of each age by year
}
