% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-maturity.R
\name{make.maturity.samples.table}
\alias{make.maturity.samples.table}
\title{Make the ovary collection table}
\usage{
make.maturity.samples.table(
  ovary_samples_df,
  xcaption = "default",
  xlabel = "default",
  font.size = 10,
  space.size = 10
)
}
\arguments{
\item{ovary_samples_df}{The values as read in from the ovary samples CSV file ("ovary-samples.csv")
using \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{xcaption}{The caption}

\item{xlabel}{The latex label to use}

\item{font.size}{Size of the font for the table}

\item{space.size}{Size of the vertical spaces for the table}
}
\value{
An \link{xtable} object
}
\description{
Make the ovary collection table
}
\details{
The header in the file is used as the header in the table. If you wish to place a newline
in any header label, insert a '\n' in the label
}
