% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.mcmc.catchability.plot}
\alias{make.mcmc.catchability.plot}
\title{Make a bar plot of the catchability parameter for a survey}
\usage{
make.mcmc.catchability.plot(
  model,
  model2 = NULL,
  type = "age2plus",
  hist_color = "grey60",
  hist_alpha = 0.5,
  med_color = "royalblue",
  model2_med_color = "red",
  model2_mle_color = "red"
)
}
\arguments{
\item{model}{The model to plot bars for}

\item{model2}{A model to draw a median line for}

\item{type}{Either "age2plus" or "age1" for those two surveys}

\item{hist_color}{Bar color}

\item{hist_alpha}{Bar transparency}

\item{med_color}{Median line color}

\item{model2_med_color}{Model 2 median line color}

\item{model2_mle_color}{Model 2 median line transparency}
}
\description{
Make a bar plot of the catchability parameter for a survey
}
