% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.mcmc.diag.pairs.plot}
\alias{make.mcmc.diag.pairs.plot}
\title{Plot pairs for MCMC posteriors}
\usage{
make.mcmc.diag.pairs.plot(
  model,
  inc.key.params = TRUE,
  key_posteriors_regex = NULL,
  key_posteriors_names = NULL,
  recr = NULL,
  bratio = NULL,
  forecatch = NULL
)
}
\arguments{
\item{model}{A model object as output by [load_ss_models()]}

\item{inc.key.params}{If TRUE, use the arguments `key_posteriors_regex` and `key_posteriors_names`}

\item{key_posteriors_regex}{A vector of regular experessions represting key posteriors}

\item{key_posteriors_names}{A vector of names to show for the key posteriors}

\item{recr}{A vector of recruitment deviation years to include if `inc.key.params` is FALSE}

\item{bratio}{A vector of Bratio years to include if `inc.key.params` is FALSE}

\item{forecatch}{A vector of forecast catch years to include if `inc.key.params` is FALSE}
}
\value{
A pairs plot
}
\description{
Plot pairs for MCMC posteriors
}
