% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-timeseries.R
\name{make.median.posterior.table}
\alias{make.median.posterior.table}
\title{Make a latex table of medians with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.}
\usage{
make.median.posterior.table(
  model,
  start_yr = NULL,
  end_yr = NULL,
  weight.factor = 1000,
  csv.dir = NULL,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  digits = 1
)
}
\arguments{
\item{model}{A model list as output by \code{\link[=load_models]{load_models()}}}

\item{start_yr}{Start year in the table}

\item{end_yr}{End year in the table}

\item{weight.factor}{Factor to divide biomass values by and multiply the recruitment by}

\item{csv.dir}{Directory for outputting the CSV version of the table}

\item{xcaption}{Caption for the latex table}

\item{xlabel}{Label for the latex table}

\item{font.size}{Size of font}

\item{space.size}{Vertical spacing between text in table}

\item{digits}{Number of decimal points to round to}
}
\value{
An \code{\link[xtable:xtable]{xtable::xtable()}}
}
\description{
Make a latex table of medians with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.
}
