% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-risk.R
\name{make.risk.table}
\alias{make.risk.table}
\title{Creates LaTeX code to make a probability of risk table with various probabilities
of things happening with the stock}
\usage{
make.risk.table(
  model,
  forecast.yrs,
  index = 1,
  digits = 0,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  let_col_width_cm = 0.2,
  catch_col_width_cm = 1.6,
  col_width_cm = 1.6,
  placement = "H"
)
}
\arguments{
\item{model}{A model from this project}

\item{forecast.yrs}{A vector of forecast years}

\item{index}{Index for which forecast year data to use. e.g. 1 = second forecast year
compared to the first. If there were N forecast years, this can be from 1 to N-1.}

\item{digits}{Number of decimal places to round to for values in table}

\item{xcaption}{Caption to appear in the calling document}

\item{xlabel}{Label used to reference the table in LaTeX}

\item{font.size}{Point size of font in table}

\item{space.size}{Vertical space between rows}

\item{let_col_width_cm}{Column width in cm for the letters column}

\item{catch_col_width_cm}{Column width in cm for the catch column}

\item{col_width_cm}{Column width in cm for all other columns}

\item{placement}{LaTeX code for placement of the table, e.g. "H" or "tbp"}
}
\value{
LaTeX code to render the table
}
\description{
Creates LaTeX code to make a probability of risk table with various probabilities
of things happening with the stock
}
