% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{make.survey.age1.plot}
\alias{make.survey.age1.plot}
\title{Make a plot of the age-1 index overlaid with model fit to that index}
\usage{
make.survey.age1.plot(model, legendloc = "topleft", ylim = c(0.015, 50))
}
\arguments{
\item{model}{A model as output by [load_models()].}

\item{legendloc}{See [legend()].}

\item{ylim}{The limits of the y axis as a range.}
}
\value{
A plot of the age-1 index fit
}
\description{
Make a plot of the age-1 index overlaid with model fit to that index
}
