% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{make.survey.history.table}
\alias{make.survey.history.table}
\title{An xtable of the history of the acoustic survey}
\usage{
make.survey.history.table(
  dat,
  CPUE,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10
)
}
\arguments{
\item{dat}{A data frame containing the survey history.}

\item{CPUE}{A data frame containing the obs and se_log by year and index.
Observations are in 1000s but will be displayed as billions.}

\item{digits}{Number of decimal places.}

\item{xcaption}{Caption to appear in the calling document.}

\item{xlabel}{The label used to reference the table in latex.}

\item{font.size}{Size of the font for the table.}

\item{space.size}{Size of the vertical spaces for the table.}
}
\description{
An xtable of the history of the acoustic survey
}
\details{
The vessel names need to be fixed. They are seperated by spaces, and may
or may not have dashes in their names. The dashes will be replaced with
spaces, and the spaces will be replaced by newlines in the output
}
