% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-comps.R
\name{make_age_comp_bubble_plot}
\alias{make_age_comp_bubble_plot}
\title{Make an age composition bubble plot}
\usage{
make_age_comp_bubble_plot(model, type = c("fishery", "survey"), ...)
}
\arguments{
\item{model}{A model object as returned from \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{Either \code{fishery} or \code{survey}}

\item{...}{Additional parameters passed to \code{\link[=plot_bubbles]{plot_bubbles()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make an age composition bubble plot
}
