% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{make_catch_table_latex}
\alias{make_catch_table_latex}
\title{Make summary of catch for given country (US, Canada, or both).
Default is a longtable for main-tables.rnw, use tabular
for Executive Summary.}
\usage{
make_catch_table_latex(
  ct,
  country = 3,
  start_yr,
  end_yr,
  xcaption = "default",
  xlabel = "default",
  font_size = 9,
  space_size = 10,
  table_placement = "H",
  tabular_environment = "longtable",
  hline_after = NULL
)
}
\arguments{
\item{ct}{Output from \code{\link[=load_catches]{load_catches()}}}

\item{country}{1 for Canada, 2 for US, anything else for both}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{xcaption}{Latex caption}

\item{xlabel}{Latex label}

\item{font_size}{Font size}

\item{space_size}{Between-letter space size}

\item{table_placement}{Latex for where to put table (H, tbp, etc.)}

\item{tabular_environment}{longtable or tabular}

\item{hline_after}{Default NULL is for longtable, gets changed for tabluar}
}
\value{
An \code{\link[xtable:xtable]{xtable::xtable()}}
}
\description{
Make summary of catch for given country (US, Canada, or both).
Default is a longtable for main-tables.rnw, use tabular
for Executive Summary.
}
