% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make_key_posteriors_mcmc_priors_vs_posts_plot}
\alias{make_key_posteriors_mcmc_priors_vs_posts_plot}
\title{Make a grid of posterior plots with priors and initial values}
\usage{
make_key_posteriors_mcmc_priors_vs_posts_plot(
  model,
  posterior_regex,
  titles = NULL,
  x_range = c("posterior", "prior"),
  ...
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
[load_ss_files()].}

\item{posterior_regex}{A vector of regular expressions that can be matched
to parameter names. Use [get_active_parameter_names()] to see all active
parameter names. Order the vector in the same order as you want the
figures to be plotted but make sure you also have `titles` ordered
appropriately as well otherwise the names will not line up.}

\item{titles}{A vector of titles for the plots. The titles can be used as is
or they can be transformed via a labeller passed to
[ggplot2::facet_wrap()] via `...`. The default value of `NULL` leads to
each panel being labelled with a lower-case letter enclosed in open
brackets, e.g., `(a)`. If you truly do not want a label, just pass the
argument a vector of empty strings, e.g., `c("", "", ...)`, with the same
length as the result of parameters that are found via `posterior_regex`.}

\item{x_range}{A string specifying the method to limit the range of each
x axis in the panels. Typically, all geoms used in the figure dictate the
range. But, with the use of [ggh4x::facetted_pos_scales()] we can specify
which data set should specify the range limits. Setting this argument to
`"posterior"`, instead of the default `"prior"`, will limit the ranges to
only the realized values in the posterior. This can be helpful when the
prior is quite vague and the posterior only covers a small range of the
parameter space.}

\item{...}{Parameters to be passed to [ggplot2::facet_wrap()]. For example,
`labeller = label_parsed_space`, which is available in this package, will
remove the spaces from the strings and implement
[ggplot2::label_parsed()]. And, `ncol` and `nrow` can be used to direct
the output or you can let {ggplot2} figure it out.}
}
\value{
A {ggplot2} object.
}
\description{
Make a grid of posterior plots with priors and initial values
}
\examples{
make_key_posteriors_mcmc_priors_vs_posts_plot(
  base_model,
  key_posteriors,
  titles = key_posteriors_titles,
  labeller = label_parsed_space,
  nrow = 3,
  ncol = 3
)
}
