% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{make_landings_tac_table}
\alias{make_landings_tac_table}
\title{Make a summary table of landings and TACS}
\usage{
make_landings_tac_table(
  d,
  start_yr,
  end_yr,
  xcaption = "default",
  xlabel = "default",
  font_size = 9,
  space_size = 10,
  placement = "H"
)
}
\arguments{
\item{d}{Data frame as output from [load_catches()]}

\item{start_yr}{Start year for the table}

\item{end_yr}{End year for the table. If past data range, the last year in `d` will be used}

\item{xcaption}{Caption for the table}

\item{xlabel}{Latex label to use}

\item{placement}{Latex table placement character}

\item{font.size}{Size of the font}

\item{space.size}{Size of spaces between letters}
}
\value{
An [xtable::xtable()]
}
\description{
Make a summary table of landings and TACS
}
