% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-make-major-tick-labels.R
\name{make_major_tick_labels}
\alias{make_major_tick_labels}
\title{Create a vector of x-axis tick labels by only populating the labels
found at a given modulo, i.e. every nth tick mark will have a label}
\usage{
make_major_tick_labels(x_breaks, modulo = 5)
}
\arguments{
\item{x_breaks}{A vector of the x-axis tick locations that appear on the
plot}

\item{modulo}{The value to use in the modulo operation to determine
which tick marks will have a label and which won't}
}
\value{
A vector of labels
}
\description{
Create a vector of x-axis tick labels by only populating the labels
found at a given modulo, i.e. every nth tick mark will have a label
}
