% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-comps.R
\name{make_numbers_at_age_plot}
\alias{make_numbers_at_age_plot}
\title{Plot the numbers-at-age bubble plot with mean age line}
\usage{
make_numbers_at_age_plot(model, scale = 1000, ...)
}
\arguments{
\item{model}{A model object as output from \code{\link[=create_rds_file]{create_rds_file()}}}

\item{scale}{Number to scale the numbers-at-age by so the legend values are nicer}

\item{...}{Additional parameters passed to \code{\link[=plot_bubbles]{plot_bubbles()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the numbers-at-age bubble plot with mean age line
}
