% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-small-rds.R
\name{make_small_rds}
\alias{make_small_rds}
\title{Make RDS files smaller by removing unnecessary objects}
\usage{
make_small_rds(drs)
}
\arguments{
\item{drs}{Output from \code{\link[=set_dirs]{set_dirs()}}}
}
\description{
Make RDS files smaller by removing unnecessary objects
}
