% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-makebox.R
\name{makebox}
\alias{makebox}
\title{Make a box plot from a data frame of quantiles}
\usage{
makebox(
  data,
  col = c("gray", "color"),
  lty = 1,
  labels = c("Year", "Depth (m)"),
  ...
)
}
\arguments{
\item{data}{A data frame of quantiles by year. The function assumes that
the depth data are in meters.}

\item{col}{The default is to use gray for all boxes, but one can change
to "color" which will use custom colors per year that are generated by
\code{\link[=plot_color]{plot_color()}}.}

\item{lty}{To change the line type, where the default is \code{1} and leads
to solid lines, where \code{2} would produce dashed lines, etc. These lines
are both for the whiskers and the median that cuts the center of the box}

\item{labels}{Character values for the horizontal and vertical axes labels}

\item{...}{Items passed to \code{\link[graphics:bxp]{graphics::bxp()}}.}
}
\value{
A base R plot
}
\description{
Make a box plot from a data frame of quantiles
}
\details{
Use \code{\link[graphics:bxp]{graphics::bxp()}}, which is normally called by \code{\link[graphics:boxplot]{graphics::boxplot()}}
to create a boxplot. Given the data are provided directly in \code{data}
there is no need to use the wrapper and \code{\link[graphics:bxp]{graphics::bxp()}} is called
directly.
}
