% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_at_age}
\alias{maturity_at_age}
\title{A data frame of empirical weight at age for all fisheries and all surveys}
\format{
A data frame with many rows and eight columns:
\describe{
\item{source}{The fishery or survey the data came from}
\item{weight}{The fish weight ()}
\item{sex}{The sex of the fish, i.e., F, M, or U}
\item{age}{The age in years of the fish}
\item{month}{The month the fish was collected}
\item{year}{The year the fish was collected}
\item{data_type}{If the fish was collected on a fishing or survey vessel}
\item{country}{The country the fish was collected in}
}

Proportion mature at each age

A vector of maturity values from the maturity ogive. The length of the vector
is the same as the number of ages in the model, not the number of ages in the
data.

\subsection{\code{maturity_at_age}}{

A vector with 20 observations, one for each modelled age starting with age
zero.
}
}
\source{
Melissa Head
}
\usage{
maturity_at_age
}
\description{
A data frame of empirical weight at age for all fisheries and all surveys
}
\details{
The data are stored in a data object that can be updated/built by running
the script in \code{data-raw/maturity_at_age.R}.
}
\keyword{datasets}
