% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_estimates_df}
\alias{maturity_estimates_df}
\title{A data frame containing estimates of maturity at age}
\format{
A data frame with 4 variables:
\describe{
\item{age}{Age (years) of fish}
\item{year}{A four digit integer representing the year}
\item{p_mature}{The probability a fish is mature}
\item{model}{A string defining which model the probability came from}
}
}
\usage{
maturity_estimates_df
}
\description{
A data frame containing estimates of maturity at age
}
\keyword{datasets}
