% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_samples_df}
\alias{maturity_samples_df}
\title{A data frame containing information on ovary sampling}
\format{
A data frame with 2636 rows and 33 variables:
\describe{
\item{sampling_platform}{Group that performed the sampling}
\item{yr}{Year of sample}
\item{primary_reader}{Initials of the primary sample-taker}
\item{haul_date}{Date of the haul (tow) that the sample was from}
\item{month}{Month of the sample}
\item{rounded_latitude}{Latitude of sample, rounded to degree}
\item{n_or_s_of_34_44}{Sample was North or South of 34.44 degrees North}
\item{season}{Season of sample}
\item{visual_maturity_code}{Maturity code for sample taken by visual means only}
\item{weight_kg}{Weight of fish}
\item{length_cm}{Length in centimeters of fish}
\item{age}{Age of fish or NA if not aged}
\item{certainty}{Certain or uncertain of maturity assignment}
\item{o_1}{Oocyte markers}
\item{o_2}{Oocyte markers}
\item{o_4_1}{Oocyte markers}
\item{o_4_2}{Oocyte markers}
\item{o_5}{Oocyte markers}
\item{o_6}{Oocyte markers}
\item{o_7}{Oocyte markers}
\item{o_8}{Oocyte markers}
\item{o_9}{Oocyte markers}
\item{highest_matstage}{Highest maturity stage assigned}
\item{biological_maturity}{0 for not mature and 1 for mature}
\item{prop_atresia}{Proportion for atresia (closedness)}
\item{functional_maturity}{0 for not mature and 1 for mature}
\item{spent}{Toggle for spent ovary}
\item{batch_spent}{0 for not spent, 1 for spent}
\item{spawning}{Toggle for spawning fish or non-spawning fish}
\item{pofs_present}{0 for not present, 1 for present}
\item{histology_notes}{Notes taken during histilogical analysis}
\item{data_notes}{Notes taken during sample analysis}
}
}
\usage{
maturity_samples_df
}
\description{
A data frame containing information on ovary sampling
}
\keyword{datasets}
