% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_path}
\alias{mcmc_path}
\title{Directory name for the mcmc output for models}
\format{
A character string
}
\usage{
mcmc_path
}
\description{
Directory name for the mcmc output for models
}
\keyword{datasets}
