% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minor_tick_length}
\alias{minor_tick_length}
\title{The length of the minor tick marks in plots in cm}
\format{
A numerical value
}
\usage{
minor_tick_length
}
\description{
The length of the minor tick marks in plots in cm
}
\keyword{datasets}
