% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_output_log_fn}
\alias{model_output_log_fn}
\title{A character string containing the file name for the model output logfile}
\format{
A character string
}
\usage{
model_output_log_fn
}
\description{
A character string containing the file name for the model output logfile
}
\keyword{datasets}
