% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-modify-starter-mcmc-type.R
\name{modify_starter_mcmc_type}
\alias{modify_starter_mcmc_type}
\title{Change the MCMC setting in an SS starter file so the extra MCMC files are
not created}
\usage{
modify_starter_mcmc_type(path, value)
}
\arguments{
\item{path}{The path where \code{starter.ss} resides}

\item{value}{Value to change the MCMC setting to. 1 = Do not create extra
MCMC files, 2 or 3 - create extra MCMC files}
}
\value{
Nothing
}
\description{
Change the MCMC setting in an SS starter file so the extra MCMC files are
not created
}
