% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mutsel.R
\name{mutsel}
\alias{mutsel}
\title{Link together \code{\link[dplyr:mutate]{dplyr::mutate()}} and \code{\link[dplyr:select]{dplyr::select()}}}
\usage{
mutsel(d, ...)
}
\arguments{
\item{d}{A data frame}

\item{...}{The column names to mutate/select}
}
\value{
The modified data frame
}
\description{
Link together \code{\link[dplyr:mutate]{dplyr::mutate()}} and \code{\link[dplyr:select]{dplyr::select()}}
}
\details{
This function allows you to \code{\link[dplyr:mutate]{dplyr::mutate()}} to create/modify columns,
and then \code{\link[dplyr:select]{dplyr::select()}} them so that the newly mutated columns are in
the order supplied to the function. Other remaining columns will follow in
the same order they appear in the original data frame \code{d}.
}
