% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuisance_posteriors_fn}
\alias{nuisance_posteriors_fn}
\title{A character string containing the file name for the nuisance posteriors
CSV file}
\format{
A character string
}
\usage{
nuisance_posteriors_fn
}
\description{
A character string containing the file name for the nuisance posteriors
CSV file
}
\keyword{datasets}
