% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-number-to-word.R
\name{number_to_word}
\alias{number_to_word}
\title{Change a number into an English word}
\usage{
number_to_word(x = NA, th = FALSE, cap_first = FALSE)
}
\arguments{
\item{x}{The number to convert}

\item{th}{Logical. If TRUE the \code{th} versions will be returned,
e.g. 4 = fourth}

\item{cap_first}{Logical. Capitalize the first letter of the returned
string?}
}
\value{
The English word representing the number
}
\description{
Change a number into an English word
}
\details{
See
https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r
See Function by John Fox found here:
http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
}
\examples{
number_to_word(c(1000,2,3,10000001), th = TRUE, cap_first = TRUE)
}
