% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{object_placement_fn}
\alias{object_placement_fn}
\title{Name of file containing the object placement settings for figures and
tables (LaTeX placement values such as H, !b, tbh, etc.)}
\format{
A character string
}
\usage{
object_placement_fn
}
\description{
Name of file containing the object placement settings for figures and
tables (LaTeX placement values such as H, !b, tbh, etc.)
}
\keyword{datasets}
