% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordered_decision_table_paths}
\alias{ordered_decision_table_paths}
\title{An ordered vector of the paths used in the decision table for those catch
streams with a calculated value. Other catch streams have a numbered path}
\format{
A character vector
}
\usage{
ordered_decision_table_paths
}
\description{
An ordered vector of the paths used in the decision table for those catch
streams with a calculated value. Other catch streams have a numbered path
}
\keyword{datasets}
