% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_csv_path}
\alias{out_csv_path}
\title{Directory name for outputting the CSV versions of at-age tables in the
document}
\format{
A character string
}
\usage{
out_csv_path
}
\description{
Directory name for outputting the CSV versions of at-age tables in the
document
}
\keyword{datasets}
