% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_permissions}
\alias{output_permissions}
\title{The Linux file permissions value to use for fs::file_chmod()}
\format{
A character string containing 3 octal digits (0-7). e.g. "777"
for full permissions for everybody
}
\usage{
output_permissions
}
\description{
The Linux file permissions value to use for fs::file_chmod()
}
\keyword{datasets}
