% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovary_samples_df}
\alias{ovary_samples_df}
\title{A data frame containing information on ovary sampling}
\format{
A data frame with 12 rows and 8 variables:
\describe{
\item{Year}{Year of samples}
\item{NWFSC Trawl Survey}{Number of ovary samples}
\item{CAN Acoustic Survey/Research (Summer)}{Number of ovary samples}
\item{U.S. Acoustic Survey/Research (Summer)}{Number of ovary samples}
\item{U.S. Acoustic Survey/Research (Winter)}{Number of ovary samples}
\item{U.S. At-Sea Hake Observer Program (Spring)}{Number of ovary samples}
\item{U.S. At-Sea Hake Observer Program (Fall)}{Number of ovary samples}
\item{OR Dept. Fish & Wildlife}{Number of ovary samples}
}
}
\usage{
ovary_samples_df
}
\description{
A data frame containing information on ovary sampling
}
\keyword{datasets}
