% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pacea-save.R
\name{pacea_save}
\alias{pacea_save}
\title{Save some results as \code{.rds} files to then be imported into the pacea R
package.}
\usage{
pacea_save(
  model_lst = list(base_model),
  model_names = list(base_model_name),
  dir_to_save = NULL
)
}
\arguments{
\item{model_lst}{List of models, but only works for one, just change if want
to save results from a different model. Code keep the list format to
remain similar to other \code{hake} package functions.}

\item{model_names}{List of model names, again only one will work.}

\item{dir_to_save}{Directory to put the resulting \code{.rds} files into, gets
created if doesn't exist. The default (gets built if \code{NULL}) is somewhat
tailored for Andy's machine, so check the function code and set the value
if needed. You'll have to manually include \code{assess_yr}.}
}
\value{
Nothing, but saves the \code{.rds} files in the appropriate directory.
}
\description{
Need to first have results loaded in, so just run this after building the
document (function does use \code{assess_yr} which will have been defined
already). Function creates a directory and places the files in there.
See pacea/data-raw/groundfish/hake.R for next steps. Some results were
automatically already saved for the assessment (e.g. recruitment), but
some had to be calculated here (e.g. recruitment scaled by R0).
}
\author{
Andrew Edwards
}
