% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pad_num}
\alias{pad_num}
\title{Pad the beginning of a number with zeroes}
\usage{
pad_num(num, digits = 1)
}
\arguments{
\item{num}{A vector of the numbers to pad}

\item{digits}{The number of characters that the resulting strings
should have}
}
\value{
A vector of strings of the padded numbers
}
\description{
Pad the beginning of a number with zeroes
}
