% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pad-weight-at-age.R
\name{pad_weight_at_age}
\alias{pad_weight_at_age}
\title{Fill in necessary years of weight-at-age file}
\usage{
pad_weight_at_age(
  data,
  n_forecast_years = 5,
  n_years_used_for_forecast = 5,
  year_global_average = -1940,
  ages = 0:20
)
}
\arguments{
\item{data}{A data frame of weight-at-age data in long form where the
weight-at-age information is stored in a column called \code{pred_weight},
the four-digit year is in a column called \code{year}, and integer ages are
in a column called \code{age}.}

\item{n_forecast_years}{An integer specifying the number of forecast years
you want to extend the time series by.}

\item{n_years_used_for_forecast}{An integer specifying the number of years
of data that you want to average over for the forecast period.}

\item{year_global_average}{A four digit integer, typically \code{-1940}, that is
used to delineate the equilibrium year, all years after this and prior to
the first year of data will be filled in by the information stored in this
row.}

\item{ages}{A vector of integers specifying the ages you want in your file.
If there are no ages in the data to represent the older ages present in
vector, then the information for the oldest age will be repeated for any
given year.}
}
\value{
A data frame of weight-at-age information is returned that can immediately
be passed to \code{\link[r4ss:SS_writewtatage]{r4ss::SS_writewtatage()}}.
}
\description{
Fill in necessary years of weight-at-age file
}
\details{
Note that this function can also be used to pad maturity information
such that it has the same format as the weight-at-age information allowing
for the multiplication of maturity by weight at age to get fecundity.
}
\author{
Kelli F. Johnson
}
