% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{panel.letter}
\alias{panel.letter}
\title{Add a letter to the plot}
\usage{
panel.letter(letter)
}
\arguments{
\item{letter}{The letter to add}
}
\value{
[base::invisible()]
}
\description{
Add a letter to the plot
}
