% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_fn}
\alias{par_fn}
\title{A character string containing the file name for the SS3 par file
(parameter output)}
\format{
A character string
}
\usage{
par_fn
}
\description{
A character string containing the file name for the SS3 par file
(parameter output)
}
\keyword{datasets}
