% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot-catch-by-month.R
\name{plot_age_by_month}
\alias{plot_age_by_month}
\title{Plot the Canadian age by month and fleet}
\usage{
plot_age_by_month(
  fleet_lst,
  x_lim = c(1, 15),
  x_breaks = seq(x_lim[1], x_lim[2], 2),
  y_lim = c(0, 0.55),
  yr = 2025
)
}
\arguments{
\item{x_lim}{Limits for the x-axis}

\item{x_breaks}{Breaks for the x-axis}

\item{y_lim}{Limits for the y-axis}

\item{yr}{Year of data to plot}
}
\value{
A \code{\link[=ggplot2:ggplot]{ggplot2:ggplot()}} object
}
\description{
Plot the Canadian age by month and fleet
}
