% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-bubbles-data.R
\name{plot_age_comp_bubbles_data}
\alias{plot_age_comp_bubbles_data}
\title{Create a multi-panel plot of age proportion bubble plots, one for each
sector of the fishery}
\usage{
plot_age_comp_bubbles_data(
  age_data_lst,
  nms_vec,
  years = c(year(now()) - 6, year(now()) - 1),
  yrs_not_used_in_model = NULL,
  yrs_not_used_in_model_color = "red",
  yrs_not_used_in_model_width = 1
)
}
\arguments{
\item{age_data_lst}{A list of age proportion data frames}

\item{nms_vec}{A vector of names for the sectors, must be the same length
as \code{age_data-lst}}

\item{years}{A vector of two values re0-presenting the minimum and maximum
year values to plot}

\item{yrs_not_used_in_model}{A vector of years to draw a box around
signifying they were not used in the current assessment model}

\item{yrs_not_used_in_model_color}{If \code{yrs_not_used_in_model} is not \code{NULL},
use this color for the box}

\item{yrs_not_used_in_model_width}{If \code{yrs_not_used_in_model} is not \code{NULL},
use this for the width of the box outline}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a multi-panel plot of age proportion bubble plots, one for each
sector of the fishery
}
