% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-comparison-bubbles.R
\name{plot_age_comp_comparison_bubbles}
\alias{plot_age_comp_comparison_bubbles}
\title{Create a plot of age proportions as bubbles with fishery and survey values
overlaid in one plot}
\usage{
plot_age_comp_comparison_bubbles(
  model,
  alpha = 0.3,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color
)
}
\arguments{
\item{model}{A model object}

\item{alpha}{Transparanecy of the bubbles}

\item{ax_title_font_size}{The font size for the axis titles and strip
titles (sector names)}

\item{ax_tick_font_size}{The foint size for the axis tick labels
(years and ages)}

\item{ax_label_color}{The color for the axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of age proportions as bubbles with fishery and survey values
overlaid in one plot
}
