% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-fit.R
\name{plot_age_comp_fit}
\alias{plot_age_comp_fit}
\title{Create a set of barplots by year with the bars being observed
age comps and errorbars being the MCMC fits with 95\% credible
interval}
\usage{
plot_age_comp_fit(
  model,
  type = c("fishery", "survey"),
  ages = NULL,
  n_col = 4,
  x_breaks = seq(2, max(ages), by = 2),
  y_breaks = c(0, 0.2, 0.4, 0.6),
  axis_title_font_size = 14,
  axis_tick_font_size = 12,
  axis_label_color = "black",
  label_loc = c(ages[length(ages)] - 2, 0.45),
  label_font_size = 5,
  point_size = 1.5,
  whisker_width = 0.15
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{The type of fits/observations to plot. Must be either
\code{fishery} or \code{survey}}

\item{ages}{A vector of ages to show on the x-axis}

\item{n_col}{The number of columns for the facets (years)}

\item{x_breaks}{The x-axis breaks passed to \code{\link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete()}}}

\item{y_breaks}{The y-axis breaks passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}

\item{axis_title_font_size}{The font size for the axis titles}

\item{axis_tick_font_size}{The font size for the axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}

\item{label_loc}{The x-y location of the labels in each panel (vector of 2)}

\item{label_font_size}{The label font size}

\item{point_size}{The point size. Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}ends of
the errorbar}

\item{whisker_width}{The width (size) of the top and bottom bars on the}
}
\description{
Create a set of barplots by year with the bars being observed
age comps and errorbars being the MCMC fits with 95\% credible
interval
}
