% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-assessment-history-biomass.R
\name{plot_assessment_history_biomass}
\alias{plot_assessment_history_biomass}
\title{Title}
\usage{
plot_assessment_history_biomass(
  base_model,
  history_df,
  x_labs_mod = 10,
  ylim = c(0, 7),
  y_breaks = ylim[1]:ylim[2],
  y_major_lab_adj = 45,
  tick_prop = 1,
  vjust_x_labels = -0.25,
  leg_pos = c(0.75, 0.8),
  leg_ncol = 3,
  leg_font_size = 8
)
}
\arguments{
\item{base_model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{history_df}{A data frame, previously read in from the file
\code{assessment-history.csv}}

\item{x_labs_mod}{Value for major X-axis tick marks. Every Nth tick
will be longer and have a label. The first and last will be shown
regardless of what this number is}

\item{ylim}{A vector of two representing the minimum and maximum values to
show on the y-axis}

\item{y_breaks}{A vector of the values to show on the y-axis}

\item{y_major_lab_adj}{A value to move the major tick abels closer/further
away from the major tick marks. Reducing this moves them closer, increasing
it moves them further away}

\item{leg_pos}{The position of the legend inside the plot. If \code{NULL},
\code{NA}, or \code{none}, the legend will not be shown}

\item{leg_ncol}{The number of columns to show in the legend}

\item{leg_font_size}{The legend font size}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Title
}
