% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-assessment-history-disp.R
\name{plot_assessment_history_disp}
\alias{plot_assessment_history_disp}
\title{Create Interquartile range or quartile coefficient of dispersion plots
for the assessment history spawning biomass estimates}
\usage{
plot_assessment_history_disp(
  base_model,
  history_disp_df,
  type = c("inter", "coeff"),
  y_lim = c(0, 2000),
  x_breaks = seq(2012, 2022, 2),
  point_shape = 1,
  point_size = 2,
  line_width = 1
)
}
\arguments{
\item{base_model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{history_disp_df}{A data frame, previously read in from the file
\code{assessment-history-probs.csv}}

\item{type}{One of 'inter' or 'coeff' for Inter-quartile range or
Quartile coefficient of dispersion respectively}

\item{x_breaks}{A vector of the values to show on the x-axis}

\item{point_shape}{A numeric point shape code}

\item{point_size}{Size of the points}

\item{line_width}{Width of the line}

\item{ylim}{A vector of two representing the minimum and maximum values to
show on the y-axis}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create Interquartile range or quartile coefficient of dispersion plots
for the assessment history spawning biomass estimates
}
