% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot-assessment-timeline.R
\name{plot_assessment_timeline}
\alias{plot_assessment_timeline}
\title{Create a plot showing the relative numbers of days we have had to
produce the assessment each year since 2017}
\usage{
plot_assessment_timeline(
  yrs = c(2017, 2018, 2019, 2020, 2021, 2022, 2023),
  days = c(17, 26, 23, 25, 23, 18, 36)
)
}
\arguments{
\item{yrs}{A vector of years to plot}

\item{days}{A vector of days required each year to produce the
assessment. Must be the same length as \code{yrs}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot showing the relative numbers of days we have had to
produce the assessment each year since 2017
}
