% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-autocor.R
\name{plot_autocor}
\alias{plot_autocor}
\title{Create an autocorelation plot for a particular parameter}
\usage{
plot_autocor(
  model,
  post_regex,
  lag_max = 20,
  y_lim = c(-1, 1),
  bar_width = 0.5
)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{post_regex}{A regular expression that matches one parameter name
and nothing else}

\item{lag_max}{See \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create an autocorelation plot for a particular parameter
}
