% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R, R/plot-cohort-catch.R,
%   R/plot-depl-fore-comparison.R, R/plot-exploitation-fraction.R,
%   R/plot-fishing-intensity.R, R/plot-fore-compare.R, R/plot-recdevs.R,
%   R/plot-recruitment.R, R/plot-rel-biomass.R, R/plot-survey-index-fits.R
\name{plot_biomass}
\alias{plot_biomass}
\alias{plot_cohort_catch}
\alias{plot_depl_fore_comparison}
\alias{plot_exploitation_fraction}
\alias{plot_fishing_intensity}
\alias{plot_fore_compare}
\alias{plot_recdevs}
\alias{plot_recruitment}
\alias{plot_rel_biomass}
\alias{plot_survey_index_fits}
\title{Plot biomass from MCMC output for one or more models}
\usage{
plot_biomass(
  model_lst = NULL,
  model_names,
  xlim = c(1964, year(Sys.time())),
  x_breaks = 1966:year(Sys.time()),
  x_labs_mod = 5,
  x_expansion = 3,
  ylim = c(0, 4.5),
  y_breaks = seq(ylim[1], ylim[2], by = 0.5),
  alpha = 0.1,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  point_size = 2.5,
  point_shape = 16,
  line_width = 1,
  clip_cover = 2,
  single_line_color = "black",
  single_ribbon_color = "blue",
  ribbon_line_type = "dotted",
  rev_colors = FALSE,
  wrap_y_label = FALSE,
  d_obj = NULL
)

plot_cohort_catch(
  model,
  cohorts = c(1999, 2010, 2014, 2016, 2020),
  ages = 0:20,
  y_breaks = seq(0, 1250, 250),
  leg_pos = c(0.85, 0.5),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  line_width = 1,
  show_arrowheads = TRUE,
  arrow_size = 0.03
)

plot_depl_fore_comparison(
  model,
  fore_inds = c(1, 2, 6, 12, 14),
  xlim = c(max(forecast_yrs) - 16, max(forecast_yrs)),
  x_breaks = xlim[1]:xlim[length(xlim)],
  ylim = c(0, 3.5),
  y_breaks = c(0, 0.1, 0.4, 0.5, 1, seq(1.5, ylim[2], by = 0.5)),
  y_labels = expression("0", "0.1B"[0], "0.4B"[0], "0.5", "B"[0], "1.5", "2", "2.5", "3",
    "3.5"),
  y_colors = c("black", "red", "green", "black", "blue", rep("black", length(seq(1.5,
    ylim[2], by = 0.5)))),
  alpha = 0.2,
  leg_pos = c(0.15, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  forecast_yrs
)

plot_exploitation_fraction(
  model,
  xlim = c(1964, year(Sys.time()) - 1),
  x_breaks = 1966:year(Sys.time()),
  x_labs_mod = 5,
  x_expansion = 1,
  ylim = c(0, 0.25),
  y_breaks = seq(ylim[1], ylim[2], by = 0.05),
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  point_size = 1,
  point_shape = 16,
  line_width = 0.5,
  line_color = "blue"
)

plot_fishing_intensity(
  model,
  xlim = c(1964, year(Sys.time()) - 1),
  x_breaks = 1966:year(Sys.time()),
  x_labs_mod = 5,
  x_expansion = 1,
  ylim = c(0, 1.4),
  y_breaks = seq(ylim[1], ylim[2], by = 0.2),
  axis_title_font_size = 12,
  axis_tick_font_size = 11,
  point_size = 1,
  point_shape = 16,
  line_width = 0.5,
  line_color = "blue"
)

plot_fore_compare(
  model,
  forecast_yrs,
  fore_yr,
  colors = c("black", "blue", "green", "orange", "red", "tan"),
  shapes = c(16, 17, 17, 17, 15, 18),
  leg_font_size = 8,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  leg_pos = NULL,
  leg_ncol = 1,
  remove_x_val = NULL,
  show_50_line = TRUE,
  short = FALSE
)

plot_recdevs(
  model_lst = NULL,
  model_names,
  xlim = c(1946, year(Sys.time())),
  x_breaks = xlim[1]:xlim[2],
  x_labs_mod = 5,
  x_expansion = 3,
  ylim = c(-5, 5),
  y_breaks = seq(ylim[1], ylim[2], by = 1),
  y_labels = expression("-5", "-4", "-3", "-2", "-1", "0", "1", "2", "3", "4", "5"),
  y_colors = c("black", "black", "black", "black", "black", "blue", "black", "black",
    "black", "black", "black"),
  alpha = 0.1,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  point_size = 1.5,
  line_width = 0.5,
  clip_cover = 2,
  single_point_color = "black",
  single_line_color = "black",
  crossbar_width = 0,
  dodge_val = 0.5,
  rev_colors = FALSE,
  d_obj = NULL
)

plot_recruitment(
  model_lst = NULL,
  model_names,
  inc_means = FALSE,
  xlim = c(1966, year(Sys.time())),
  x_breaks = xlim[1]:xlim[2],
  x_labs_mod = 5,
  x_expansion = 3,
  ylim = c(0, 40),
  y_breaks = seq(ylim[1], ylim[2], by = 10),
  y_labels = y_breaks,
  y_colors = rep("black", length(y_breaks)),
  alpha = 0.2,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  point_size = 1.5,
  line_width = 0.5,
  clip_cover = 2,
  single_point_color = "black",
  single_line_color = "black",
  crossbar_width = 0,
  dodge_val = 0.5,
  rev_colors = FALSE,
  d_obj = NULL
)

plot_rel_biomass(
  model_lst = NULL,
  model_names,
  xlim = c(1966, year(Sys.time())),
  x_breaks = xlim[1]:xlim[2],
  x_labs_mod = 5,
  x_expansion = 3,
  ylim = c(0, 3.5),
  alpha = 0.1,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  minor_tick_length = 0.1,
  point_size = 2,
  point_shape = 16,
  line_width = 1,
  clip_cover = 2,
  single_line_color = "black",
  single_ribbon_color = "blue",
  ribbon_line_type = "dotted",
  rev_colors = FALSE,
  wrap_y_label = FALSE,
  d_obj = NULL
)

plot_survey_index_fits(
  model_lst = NULL,
  model_names,
  survey_type = c("age1", "age2"),
  xlim = c(1995, 2021),
  x_breaks = xlim[1]:xlim[2],
  x_labs_mod = 5,
  x_expansion = 3,
  ylim = c(0, 3),
  y_breaks = seq(ylim[1], ylim[2], by = 0.5),
  alpha = 0.1,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  point_size = 1.5,
  line_width = 0.5,
  clip_cover = 2,
  rev_colors = FALSE,
  dodge_val = 0.5,
  d_obj = NULL
)
}
\arguments{
\item{model_lst}{A list of models, each created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_names}{A vector of model names,the same length as \code{models_lst}}

\item{xlim}{The year limits to plot}

\item{x_breaks}{The year value tick marks to show for the x axis}

\item{x_labs_mod}{Value for major X-axis tick marks. Every Nth tick
will be longer and have a label. The first and last will be shown
regardless of what this number is}

\item{x_expansion}{Amount to expand the x axis. See the \code{expand} argument in
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}

\item{ylim}{The depletion limits to plot}

\item{y_breaks}{The depletion value tick marks to show for the y axis}

\item{alpha}{The transparency for all ribbons}

\item{leg_pos}{The position of the legend inside the plot. If \code{NULL},
\code{NA}, or \code{none}, the legend will not be shown}

\item{leg_ncol}{The number of columns to show in the legend}

\item{leg_font_size}{The legend font size}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{point_size}{Size of all points shown in plot}

\item{point_shape}{The R shape number for the points}

\item{line_width}{Width of all lines on the plot}

\item{clip_cover}{There is a white rectangle drawn on top of the plot
to cover any of the plot that made it outside the plot area. \code{clip} has to
be set to \code{off} for the major x-axis tick marks to work, So, this is required.
If you make the plot in a grid, the rectangle may overwrite some of the plot
above it, and this number will have to be changed through trial and error
until you cannot see the white rectangle anymore.}

\item{single_line_color}{Line color for the case where there is only
one model to plot}

\item{single_ribbon_color}{The ribbon color if there is only a single model}

\item{ribbon_line_type}{Linetype for ribbon edges; use 0 for no line.}

\item{rev_colors}{Logical. If \code{TRUE}, reverse the order of the colors
in the plot. Only applies if more than one model is plotted}

\item{wrap_y_label}{Logical. If \code{TRUE}, adds a newline to the y axis
label so that it doesn't get cut off}

\item{d_obj}{If not \code{NULL} this is a list which has been
pre-processed to contain all models in a format that is ready to plot.
Essentially the first steps of this function have been replicated
outside the function (The code inside the \verb{if(is.null(d_obj))})
is done to stop the Rmd process from taking forever}

\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{cohorts}{A vector of cohorts to make lines for}

\item{ages}{A vector of ages to show}

\item{show_arrowheads}{Logical. If \code{TRUE} show arrowheads on the ends of
the lines}

\item{arrow_size}{size of the arrow heads, \code{npc} units. See \code{\link[grid:unit]{grid::unit()}}
For description of \code{npc}}

\item{fore_inds}{Indices of the \code{model$catch_levels} list to include as
forecast catch streams in the plot}

\item{y_labels}{Labels for the tick marks on the Y axis}

\item{y_colors}{Colors for the tick labels on the Y axis}

\item{forecast_yrs}{A vector of the forecast years}

\item{fore_yr}{The forecast year to use in the plot. Must be in
\code{forecast_yrs}}

\item{colors}{Colors to use for the lines in the plot}

\item{shapes}{Shapes to use for the points in the plot. Must be the same
length as \code{colors}}

\item{remove_x_val}{A vector of values to remove labels for on the x axis.
Plot first leaving this as \code{NULL}, then add values to this that overlap
others}

\item{show_50_line}{Logical. If \code{TRUE}, draw a horizontal line at 50\\\%}

\item{short}{Logical. If \code{TRUE}, draw only 3 lines (decline, under B40,
under B10)}

\item{single_point_color}{Color to use for the points if there is only
one model}

\item{crossbar_width}{The width of the crossbar ends in
\code{\link[ggplot2:geom_linerange]{ggplot2::geom_errorbar()}}}

\item{dodge_val}{The amount to separate lines between unique models
multiple model plots}

\item{inc_means}{Logical. If \code{TRUE} include the mean values in the plot}

\item{minor_tick_length}{The length of the small x-axis ticks}

\item{survey_type}{Either \code{age1} or \code{age2}}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot biomass from MCMC output for one or more models

Create a plot of cohorts by age as a set of lines

Plot the relative spawning biomass with several forecast trajectories

Create a plot of exploitation fraction

Create a plot of fishing intensity

Title

Plot recruitment deviations from MCMC output for one or more models

Plot recruitment from MCMC output for one or more models

Plot relative biomass from MCMC output for one or more models

Plot survey index fits from MCMC output for one or more models,
for one of the acoustic surveys
}
