% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass-fished-unfished.R
\name{plot_biomass_fished_unfished}
\alias{plot_biomass_fished_unfished}
\title{Plot the fished biomass vs. the unfished biomass for a model}
\usage{
plot_biomass_fished_unfished(
  model,
  model_nms = c("Base model - Fished", "Base model - Unfished"),
  rel = FALSE,
  ...
)
}
\arguments{
\item{model}{A model to plot}

\item{model_nms}{A vector of length 2 containing the names to use in the
legend for the fished/unfished trajectories}

\item{rel}{Logical. If \code{TRUE}, plot the biomass relative to dynamic B0
biomass}

\item{...}{Arguments passed to \code{\link[=plot_biomass]{plot_biomass()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the fished biomass vs. the unfished biomass for a model
}
