% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch-by-month.R
\name{plot_catch_by_month}
\alias{plot_catch_by_month}
\title{Create a plot of cumulative catches, with one panel for each fishery}
\usage{
plot_catch_by_month(
  catch_lst,
  names_lst,
  yrs = c((year(now()) - 4), (year(now()) - 1)),
  scale = 1000,
  type = c("catch", "proportion", "cumulative", "quota"),
  quota_lst = NULL,
  disclaimer_text = NULL,
  y_breaks = NULL,
  leg_pos = c(0.1, 0.7),
  leg_font_size = 8,
  line_width = ts_linewidth,
  line_gap = ts_linegap,
  point_shape = ts_pointshape,
  point_size = ts_pointsize,
  point_stroke = ts_pointstroke,
  title_font_size = axis_title_font_size,
  ax_title_x = axis_title_font_size,
  ax_title_y = axis_title_font_size,
  ax_text_x = axis_tick_font_size,
  ax_text_y = axis_tick_font_size
)
}
\arguments{
\item{catch_lst}{A list of data tables with either three columns: month,
year, and catch or 13 columns: Year, and one column each for the 12 months}

\item{names_lst}{A list of names for the panel titles in the plot. Must be
the same length as the \code{catch_lst} list}

\item{yrs}{A vector of two years representing the maximum and minimum years
to include in the data shown in the plot}

\item{scale}{A number to divide the catch by}

\item{type}{The type of plot, one of \code{catch}, \code{proportion}, \code{cumulative},
or \code{quota}}

\item{quota_lst}{A list the same length as \code{catch_lst} and \code{names_lst}
containing data frames which have two columns each, one named \code{Year}
and one representing the sector quota for that year for which the name
does not matter}

\item{disclaimer_text}{Text to show in the empty panel at the bottom
right of the plot grid. If \code{NULL}, Nothing will be shown there}

\item{y_breaks}{A vector of y-axis values to show. The viewable part of
the y-axis will also be set to the minimum to maximum of these values}

\item{leg_pos}{A two-element vector describing the placement of the
legend inside the plotting area where the x,y values are both between 0
and 1 or "none" for no legend}

\item{leg_font_size}{The font size for the legend items}

\item{line_width}{The width of the lines}

\item{line_gap}{The gap between the points and lines (blank spacing)}

\item{point_shape}{The point shape type}

\item{point_size}{The size of the points}

\item{point_stroke}{The stroke value for the points}

\item{title_font_size}{Size of the title text for each panel}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of cumulative catches, with one panel for each fishery
}
