% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catchability-density.R
\name{plot_catchability_density}
\alias{plot_catchability_density}
\title{Plot the density histogram for the posteriors for catchability for
either the acoustic age-2+ survey or the age-1 index}
\usage{
plot_catchability_density(
  model,
  model2 = NULL,
  type = c("age2", "age1"),
  num_bins = 30,
  x_breaks = seq(0, 1.4, 0.1),
  y_breaks = seq(0, 3, 0.5),
  bar_outline_color = "black",
  bar_fill = main_fill,
  bar_alpha = main_alpha,
  line_colors = c("green", "red"),
  line_types = c("solid", "solid"),
  line_widths = c(1, 1)
)
}
\arguments{
\item{model}{The base model}

\item{model2}{The second model, typically last year's base model}

\item{type}{One of "age2" or "age1}

\item{num_bins}{The number of bins to use for the histogram created using
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}

\item{x_breaks}{A vector of values to show on the x-axis}

\item{y_breaks}{A vector of values to show on the y-axis}

\item{bar_outline_color}{The color of the outline of each bar in the histogram}

\item{bar_fill}{The color of the fill for each bar in the histogram}

\item{bar_alpha}{The amount of transparency for each bar in the histogram}

\item{line_colors}{A vector of two line colors, one for the base model
(\code{model}) median catchability and the second for last year's model
(\code{model2})median catchability}

\item{line_types}{A vector of two line types for the two median lines}

\item{line_widths}{A vector of two line widths for the two median lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the density histogram for the posteriors for catchability for
either the acoustic age-2+ survey or the age-1 index
}
