% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchvmonthbyyear.R
\name{plot_catchvmonthbyyear}
\alias{plot_catchvmonthbyyear}
\title{Plot catch versus month by year}
\usage{
plot_catchvmonthbyyear(
  data,
  file = NULL,
  title,
  Yrs,
  quotas = NULL,
  leg.cex = 1.2,
  divisor = 1000
)
}
\arguments{
\item{file}{A path to a png file where you want the plot to be saved to. The
default is \code{NULL}, which leads to the figure being returned rather than
saved to a file.}

\item{title}{todo: document}

\item{Yrs}{A vector of years that you want to plot.}

\item{quotas}{A named vector of quotas for the fleet of interest. The years
in \code{Yrs} must be present in the vector.}

\item{leg.cex}{The font size for the legend.}

\item{divisor}{A numeric value used to determine if you want the figure to
be in metric tons or in kg (i.e., 1000 mt).}

\item{A}{data frame of landings by month. For example,
\code{us_ms_catch_by_month_df}.}
}
\description{
Plot catch versus month by year
}
