% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colour.R
\name{plot_colour}
\alias{plot_colour}
\title{Colours for the Hake Assessment}
\usage{
plot_colour(n = 10)
}
\arguments{
\item{n}{The default number of colours to generate is 10.}
}
\value{
A vector of colours
}
\description{
Generate the colour scheme for the hake assessment
such that as more years of historical data are added on
that the more recent years are still the same colour.
For example, the most recent year of data will always be
black and the second most recent year will always be red.
}
\examples{
n <- 15
plot_colour(n)

}
\author{
Kelli F. Johnson
}
