% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-comparisons-nuts.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Create an n-panel plot of the biomass, relative biomass, or recruitment
for a set of models comparing a value}
\usage{
plot_comparison(
  lst = NULL,
  plot_func = NULL,
  label_text = "Variable x",
  panel_axis_titles_size = 11,
  panel_axis_text_size = 8,
  scale = 1,
  digits = 0,
  ...
)
}
\arguments{
\item{lst}{A list, as returned from \code{\link[=group_test_models]{group_test_models()}}}

\item{plot_func}{The plotting function to use. Typically \code{\link[=plot_biomass]{plot_biomass()}},
\code{\link[=plot_rel_biomass]{plot_rel_biomass()}}, or \code{\link[=plot_recruitment]{plot_recruitment()}}}

\item{label_text}{The text to paste on the label that is shown on
each panel}

\item{panel_axis_titles_size}{Font size for the axis labels}

\item{panel_axis_text_size}{Font size for the axis tick labels}

\item{scale}{Number to divide the label by (used to divide adapt
delta values by 100 so they are less than 1)}

\item{digits}{The number of decimal places to make the label numbers}

\item{...}{Arguments to pass to the plotting function defined by \code{plot_func}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create an n-panel plot of the biomass, relative biomass, or recruitment
for a set of models comparing a value
}
