% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cumulative-catch.R
\name{plot_cumulative_catches}
\alias{plot_cumulative_catches}
\title{Create a plot of cumulative catches, with one panel for each fishery}
\usage{
plot_cumulative_catches(
  catch_lst,
  names_lst,
  yrs = c((year(now()) - 4), (year(now()) - 1)),
  scale = 1000,
  type = c("default", "proportion", "cumulative"),
  leg_pos = c(0.1, 0.7),
  leg_font_size = 8,
  line_width = ts_linewidth,
  line_gap = ts_linegap,
  point_shape = ts_pointshape,
  point_size = ts_pointsize,
  point_stroke = ts_pointstrok
)
}
\arguments{
\item{catch_lst}{A list of data tables with either three columns: month,
year, and catch or 13 columns: Year, and one column each for the 12 months}

\item{names_lst}{A list of names for the panel titles in the plot. Must be
the same length as the \code{catch_lst} list}

\item{yrs}{A vector of two years representing the maximum and minimum years
to include in the data shown in the plot}

\item{scale}{A number to divide the catch by}

\item{type}{The type of plot, one of default (no change to data),
proportion, or cumulative}

\item{leg_pos}{A two-element vector describing the placement of the
legend inside the plotting area where the x,y values are both between 0
and 1 or "none" for no legend}

\item{leg_font_size}{The font size for the legend items}

\item{line_width}{The width of the lines}

\item{line_gap}{The gap between the points and lines (blank spacing)}

\item{point_shape}{The point shape type}

\item{point_size}{The size of the points}

\item{point_stroke}{The stroke value for the points}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of cumulative catches, with one panel for each fishery
}
