% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{plot_depth}
\alias{plot_depth}
\title{Create a boxplot of depths for a given data frame}
\usage{
plot_depth(
  depth_df,
  yrs,
  axis_tick_font_size = 11,
  title_text = NULL,
  line_col = "grey",
  box_width = 0.75,
  ...
)
}
\arguments{
\item{depth_df}{A data frame containing boxplot input, see files like
\verb{depth-us-*.csv}}

\item{yrs}{A vector of the years to show in the plot}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a boxplot of depths for a given data frame
}
