% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth-2-panel.R
\name{plot_depth_2_panel}
\alias{plot_depth_2_panel}
\title{Create a 2-panel plot with fishing depth on the left and bottom depth on
the right}
\usage{
plot_depth_2_panel(fishing_depth_df, bottom_depth_df, label_size = 14, ...)
}
\arguments{
\item{fishing_depth_df}{A data frame containing boxplot input, see files like
\verb{depth-us-*.csv}}

\item{bottom_depth_df}{A data frame containing boxplot input, see files like
\verb{depth-us-*.csv}}

\item{label_size}{The font size for the two axis labels,
\code{Year} and \code{Depth (m)}}

\item{...}{Arguments passed to \code{\link[=plot_depth]{plot_depth()}}}
}
\value{
A \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}} object
}
\description{
Create a 2-panel plot with fishing depth on the left and bottom depth on
the right
}
