% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth-4-panel.R
\name{plot_depth_4_panel}
\alias{plot_depth_4_panel}
\title{Create a 4-panel plot with fishing depth on the left and bottom depth on
the right, with one fleet on top and another below}
\usage{
plot_depth_4_panel(
  fishing_depth_df1,
  bottom_depth_df1,
  fishing_depth_df2,
  bottom_depth_df2,
  fleet_1_name = "",
  fleet_2_name = "",
  label_size = 14,
  ...
)
}
\arguments{
\item{fishing_depth_df1}{A data frame containing boxplot input, see files
like \verb{depth-us-*.csv}}

\item{bottom_depth_df1}{A data frame containing boxplot input, see files
like \verb{depth-us-*.csv}}

\item{fishing_depth_df2}{A data frame containing boxplot input, see files
like \verb{depth-us-*.csv}}

\item{bottom_depth_df2}{A data frame containing boxplot input, see files
like \verb{depth-us-*.csv}}

\item{label_size}{The font size for the two axis labels,
\code{Year} and \code{Depth (m)}}

\item{...}{Arguments passed to \code{\link[=plot_depth]{plot_depth()}}}
}
\value{
A \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}} object
}
\description{
Create a 4-panel plot with fishing depth on the left and bottom depth on
the right, with one fleet on top and another below
}
