% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fecundity.R
\name{plot_fecundity}
\alias{plot_fecundity}
\title{Create a aplot of fecundity-at-age and mean-weight-at-age}
\usage{
plot_fecundity(
  model,
  yrs = NULL,
  line_colors = c("blue", "red"),
  line_alpha = 0.5,
  leg_pos = c(0.68, 0.13),
  leg_ncol = 1,
  leg_font_size = 14,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yrs}{A vector of the years to use in the calculations}

\item{leg_pos}{The position of the legend inside the plot. If \code{NULL},
\code{NA}, or \code{none}, the legend will not be shown}

\item{leg_ncol}{The number of columns to show in the legend}

\item{leg_font_size}{The legend font size}

\item{...}{Arguments passed to \code{\link[=plot_maturity]{plot_maturity()}}}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a aplot of fecundity-at-age and mean-weight-at-age
}
