% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fore-age-comps.R
\name{plot_fore_age_comps}
\alias{plot_fore_age_comps}
\title{Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)}
\usage{
plot_fore_age_comps(
  model,
  x_lim = c(1, 15),
  x_breaks = seq(x_lim[1], x_lim[2], 2),
  bar_alpha = 0.3
)
}
\arguments{
\item{model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{x_lim}{A vector of 2 representing the minimum and maximum x-axis
values to plot}

\item{bar_alpha}{The transparency of the bars}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} object
}
\description{
Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)
}
