% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-hcr.R
\name{plot_hcr}
\alias{plot_hcr}
\title{Plot the effects of the harvest control rule on the MCMC output of a model.}
\usage{
plot_hcr(
  model,
  year,
  lrp = 0.1,
  urp = 0.4,
  background_alpha = 0.1,
  show_labels = TRUE,
  point_size = 0.25,
  arrowhead_crit_x = 0.05,
  arrowhead_crit_y = 0.1,
  arrowhead_caut_x = 0.3,
  arrowhead_caut_y = 0.5,
  arrowhead_healthy_x = 0.7,
  arrowhead_healthy_y = 0.8,
  label_crit_x = 1.5,
  label_crit_y = 0.1,
  label_caut_x = 1.3,
  label_caut_y = 0.1,
  label_healthy_x = 0.8,
  label_healthy_y = 0.1,
  arrow_obj = arrow(length = unit(0.5, "cm"), angle = 20, type = "closed")
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{year}{The year to extract Bratio for from the SS3 MCMC output matrix}

\item{lrp}{Limit Reference Point}

\item{urp}{upper reference point, also called USR or Upper Stock Reference}

\item{background_alpha}{The opacity of the colored areas representing the
zones}

\item{show_labels}{If \code{TRUE}, show the labels specifying the number of
posteriors in each zone}

\item{point_size}{The size of the points shown. Smaller points will allow you
to see individual points, larger ones make it seem like a solid line}

\item{arrowhead_crit_x}{The x value for the head of the arrow for the
critical zone text label}

\item{arrowhead_crit_y}{The y value for the head of the arrow for the
critical zone text label}

\item{arrowhead_caut_x}{The x value for the head of the arrow for the
cautious zone text label}

\item{arrowhead_caut_y}{The y value for the head of the arrow for the
cautious zone text label}

\item{arrowhead_healthy_x}{The x value for the head of the arrow for the
healthy zone text label}

\item{arrowhead_healthy_y}{The y value for the head of the arrow for the
healthy zone text label}

\item{label_crit_x}{The x value for the label location for the
critical zone text label}

\item{label_crit_y}{The y value for the label location for the
critical zone text label}

\item{label_caut_x}{The x value for the label location for the
cautious zone text label}

\item{label_caut_y}{The y value for the label location for the
cautious zone text label}

\item{label_healthy_x}{The x value for the label location for the
healthy zone text label}

\item{label_healthy_y}{The y value for the label location for the
healthy zone text label}

\item{arrow_obj}{Definition of the arrow shape and size for the text label
arrows}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the effects of the harvest control rule on the MCMC output of a model.
}
\details{
Plots relative spawning biomass for a given \code{year} on the x-axis by default
HCR relative fishing intensity on the y-axis. Extracts all MCMC posteriors
from SS3 output, specifically from the column called \verb{Bratio_<year>} and uses
those as the basis for the calculation to make th plot.
}
