% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap-sample-size-weight-at-age.R
\name{plot_heatmap_sample_size_weight_at_age}
\alias{plot_heatmap_sample_size_weight_at_age}
\title{Create a weight-at-age heatmap including extrapolated years}
\usage{
plot_heatmap_sample_size_weight_at_age(
  model,
  sample_size_df = NULL,
  fleet = 1,
  proj_line_color = "royalblue",
  proj_line_width = 1,
  proj_line_yr = NULL,
  yrs = NULL,
  cell_font_size = 4,
  sum_col_fill_color = "white",
  ...
)
}
\arguments{
\item{model}{An list of results read in from an SS model using
\code{\link[=load_ss_files]{load_ss_files()}}}

\item{sample_size_df}{A data frame where ages are columns (and start with
the letter 'a'). If the values are zero, the weight-at-age was
extrapolated/interpolated. If there is a value, the weight-at-age is data}

\item{fleet}{An integer value specifying which fleet you want plotted.
Fleet -2 will plot fecundity information.
Fleet -1 will plot population weight-at-age for the middle of the year.
Fleet 0 will plot population weight-at-age for the beginning of the year.
Positive values for fleet will link to a modeled fleet.}

\item{proj_line_color}{Line color to separate projection years.}

\item{proj_line_width}{Line width to separate projection years.}

\item{yrs}{A vector of the years to include in the table. If \code{NULL}, all
years will be included}

\item{cell_font_size}{Font size of the values printed in each cell of
the table}

\item{sum_col_fill_color}{The fill color for the row sum column}

\item{...}{Arguments passed to `\code{\link[=heatmap_add_extrap_yrs_wa]{heatmap_add_extrap_yrs_wa()}},
\code{\link[=heatmap_get_wa_ggplot_vals]{heatmap_get_wa_ggplot_vals()}}, and \code{\link[=heatmap_set_colors]{heatmap_set_colors()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a weight-at-age heatmap including extrapolated years
}
\details{
Produce a matrix of colors and values of weight-at-age information
by year (y-axis) and age (x-axis). The plot uses output from a
Stock Synthesis model.

Ages zero through the maximum age modeled in the data are shown
for each year plotted. This maximum-age shown is found by removing ages
that are the duplicate of the previous age, no user input is needed.

Users have many options to control what years are included in the plot.
The beginning years of data can be removed using \code{first_yr}.
Users do not have control over what are noted as projection years, which
are all years after the end of the data included in the assessment model
up to last year included in the weight-at-age matrix.
These projection years are noted using a horizontal line but if the
demarcation is not wanted, then they can change the line width to zero.
Finally, users can calculate the colors using a range of data specified
with \code{first_yr} and the last year of the data in the model. Then,
the resulting plot can be truncated to a specified year range using
\code{print_yrs}. The truncation is helpful to facilitate plots that fit
on a single page or on a slide with readable values. Using this truncation,
the colors would show trends across all plots made rather than just the
truncated data.
}
