% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-param-hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Make a histogram and density plot of posterior samples for a given parameter and model (or set of models)}
\usage{
plot_hist(
  model = list(base_model),
  model.name = base_model_name,
  param_name = "SR_LN(R0)",
  legend.pos = "top",
  xlabel = expression(ln(R[0]))
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=load_ss_files]{load_ss_files()}} or a list of models (see examples)}

\item{model.name}{model name or set of model names}

\item{param_name}{The SS name for a given parameter}

\item{legend.pos}{Position for the legend, e.g., top or none for no legend}

\item{xlabel}{The label used for the x-axis}

\item{vert.line}{Location on x-axis of vertical line, used to show single estimate with distribution}
}
\value{
Nothing
}
\description{
Make a histogram and density plot of posterior samples for a given parameter and model (or set of models)
}
\examples{
plot_hist(model=list(base_model), model.name=base_model_name,param_name = "Late_RecrDev_2019",xlabel = "RecDev_2019",legend.pos="none")
plot_hist(model=c(list(base_model), list(sens_models_7),sens_models_2),model.name=c(base_model_name,"rwMH",sens_model_dir_names_2))
plot_hist(model=list(base_model), model.name=base_model_name,param_name = "Recr_2019",xlabel = "Rec_2019",legend.pos="none",vert.line=16149000)

}
