% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-historical-probs-retro.R
\name{plot_historical_probs_retro}
\alias{plot_historical_probs_retro}
\title{Make individual plots (or one overlaid) of historical probability calculations
that uses retrospective runs (given we have them). Calls
\code{make.historical.probs.plot()} for each retrospective}
\usage{
plot_historical_probs_retro(
  model,
  hist_probs,
  type = "decline",
  xLim = c(2012, assess_yr),
  make.one.figure = TRUE,
  cols = c("lightgrey", "darkgreen", "lightgreen", "brown", "pink", "orange", "gold",
    "darkgrey", "violet", "black"),
  time.sleep = 0,
  omit.current = TRUE,
  select.retros = NULL,
  lwd.val.for.retros = 1,
  legend.text.model = "From current base model",
  ...
)
}
\arguments{
\item{model}{Model with retrospectives}

\item{hist_probs}{A data frame containing the data found in
\code{assessment-history-probs.csv}}

\item{type}{One of \code{decline} or \code{bforty} as needed for
\code{\link[=make.historical.probs.plot]{make.historical.probs.plot()}}}

\item{xLim}{Range of x (years) axis}

\item{make.one.figure}{Whether to plot everything on one figure or new
figure for each retro.}

\item{cols}{Vector colors to use}

\item{time.sleep}{Seconds to pause between plots}

\item{omit.current}{Omit current base model plot when doing individual plots
(e.g. for panel plot don't need it, probably already shown it)}

\item{select.retros}{Vector of retrospective runs to use, where 1:4 would be
the earliest four retrospective runs (for not doing all plots to be able to
spread them out over multiple pages)}

\item{lwd.val.for.retros}{lwd for retro lines, default of 1 is for one
figure}

\item{legend.text.model}{text to describe \code{model} in the legend}

\item{...}{Arguments to pass to \code{\link[=make.historical.probs.plot]{make.historical.probs.plot()}}}
}
\description{
Make individual plots (or one overlaid) of historical probability calculations
that uses retrospective runs (given we have them). Calls
\code{make.historical.probs.plot()} for each retrospective
}
