% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-intensity-illustration.R
\name{plot_intensity_illustration}
\alias{plot_intensity_illustration}
\title{Create a plot illustrating the fishing intensity as a function of SPR}
\usage{
plot_intensity_illustration(target = 0.4)
}
\arguments{
\item{target}{The target value for SPR as a proportion between 0 and 1}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot illustrating the fishing intensity as a function of SPR
}
