% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-management-catch-vs-tac.R
\name{plot_management_catch_vs_tac}
\alias{plot_management_catch_vs_tac}
\title{Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch}
\usage{
plot_management_catch_vs_tac(
  d,
  curr_assess_biomass = NULL,
  line_type = "solid",
  line_width = 1,
  line_alpha = 0.5,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  point_size = 3
)
}
\arguments{
\item{d}{The data as read in using \code{\link[readr:read_delim]{readr::read_csv()}} from the file
"catch-targets.csv"}

\item{curr_assess_biomass}{Current year's assessment-estimated biomass
(or any value). If \code{NULL} it will not be shown}

\item{line_type}{Which linetype to use for connecting the
values with vertical lines}

\item{line_width}{Width of the connecting lines}

\item{line_alpha}{Which alpha level (0-1) to use for connecting
the values with vertical lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch
}
