% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-management-catch-vs-tac-1-to-1.R
\name{plot_management_catch_vs_tac_1_to_1}
\alias{plot_management_catch_vs_tac_1_to_1}
\title{Plot the TAC vs realized catch on a 1-to-1 plot}
\usage{
plot_management_catch_vs_tac_1_to_1(
  d,
  top_yrs = NULL,
  x_lim = c(0, 1000),
  y_lim = c(0, 800),
  leg_xmin = 20,
  leg_xmax = 620,
  leg_ymin = 700,
  leg_ymax = 780
)
}
\arguments{
\item{d}{The data as read in using \code{\link[readr:read_delim]{readr::read_csv()}} from the file
"catch-targets.csv"}

\item{xlim}{See \code{\link[base:plot]{base::plot()}}}

\item{ylim}{See \code{\link[base:plot]{base::plot()}}}
}
\description{
Plot the TAC vs realized catch on a 1-to-1 plot
}
