% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-management-catch-vs-tac-diffs.R
\name{plot_management_catch_vs_tac_diffs}
\alias{plot_management_catch_vs_tac_diffs}
\title{Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch}
\usage{
plot_management_catch_vs_tac_diffs(
  inc_biomass_ests = FALSE,
  line_type = "solid",
  line_width = 1,
  line_alpha = 0.5,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  point_size = 3,
  font_size = 10,
  right_cols_cm = 1.8,
  y_breaks = seq(0, 1000, 100),
  x_labs_mod = 2,
  ...
)
}
\arguments{
\item{inc_biomass_ests}{If \code{TRUE}, include a line representing the estimated
biomass (medians)}

\item{line_type}{Which linetype to use for connecting the
values with vertical lines}

\item{line_width}{Width of the connecting lines}

\item{line_alpha}{Which alpha level (0-1) to use for connecting
the values with vertical lines}

\item{leg_pos}{The legend position as x-y coordinates (vector of 2)}

\item{leg_ncol}{The number of columns to split the legend into}

\item{leg_font_size}{The legend font size}

\item{point_size}{The point size}

\item{y_breaks}{A vector of values to show on the y-axis}

\item{x_labs_mod}{Show a year label every Nth tick, This is N. Default 2}

\item{...}{Arguments passed to \code{\link[=setup_data_frame_for_past_management_plots]{setup_data_frame_for_past_management_plots()}}
Includes the catch/tac data frame, model, and arguments to include biomass
estimates and diffs (differences between subsequent year proportions)}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch
}
