% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-maturity-ogives.R
\name{plot_maturity_ogives}
\alias{plot_maturity_ogives}
\title{Create a plot of the maturity ogives used in the assessment. The plot
shows two of them; above and below the line if latitude at Point
Conception (34.44° latitude)}
\usage{
plot_maturity_ogives(
  model,
  d,
  colors = c("blue", "red"),
  point_alpha = 0.3,
  line_alpha = 0.3,
  leg_pos = c(0.85, 0.15),
  leg_ncol = 1,
  leg_font_size = 16,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color,
  ret_df = FALSE
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{d}{The data frame which is read in from}

\item{colors}{A vector of two colors for North and South of 34 degrees
respectively}

\item{point_alpha}{Transparency of the bubbles}

\item{line_alpha}{Transparency of the lines
\code{inst/extdata/data/hake-maturity-data.csv}}

\item{leg_pos}{The position of the legend inside the plot. If \code{NULL},
\code{NA}, or \code{none}, the legend will not be shown}

\item{leg_ncol}{The number of columns to show in the legend}

\item{leg_font_size}{The legend font size}

\item{ax_title_font_size}{Size of the font for the X and Y axis labels}

\item{ax_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{ax_label_color}{Color of the font for the X and Y axis tick and
title labels}

\item{ret_df}{Logical. If \code{TRUE}, don't plot, instead return the
calculation of \code{d} which may be needed as part of the fecundity calculation
elsewhere}

\item{alpha}{The transparency for all ribbons}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of the maturity ogives used in the assessment. The plot
shows two of them; above and below the line if latitude at Point
Conception (34.44° latitude)
}
