% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mcmc-autocor.R
\name{plot_mcmc_autocor}
\alias{plot_mcmc_autocor}
\title{Plot the autocorrelation from a pre-calculated data frame}
\usage{
plot_mcmc_autocor(
  d,
  col_nm,
  x_lim = NULL,
  y_lim = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  x_brk = 0.1,
  y_brk = 20,
  ...
)
}
\arguments{
\item{d}{A dta frame containing the column with name \code{col_nm} which
contains the autocorrelation calculations as done in
\code{\link[=calc_mcmc_param_stats]{calc_mcmc_param_stats()}}. These are calculated using \code{\link[stats:acf]{stats::acf()}}}

\item{col_nm}{The name of the column in \code{d}containing the
autocorrelation values}

\item{x_lim}{A vector of two values. The min and max to show on the
x-axis. If \code{NULL}, values will be calculated from the input values}

\item{y_lim}{A vector of two values. The min and max to show on the
y-axis. If \code{NULL}, values will be calculated from the input values}

\item{x_breaks}{A vector of values to show on the x-axis. If \code{NULL},
values will be calculated from the input values}

\item{y_breaks}{A vector of values to show on the y-axis. If \code{NULL},
values will be calculated from the input values}

\item{x_brk}{A value to use to create the \code{x_breaks} values if it is
\code{NULL}. This value is used as the \code{by} argument like this:
\code{seq(xlim[1], xlim[2], by = x_brk)}. This is also used as the \code{binwidth}
in the \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} call}

\item{y_brk}{A value to use to create the \code{y_breaks} values if it is
\code{NULL}. This value is used as the \code{by} argument like this:
\code{seq(ylim[1], ylim[2], by = y_brk)}}

\item{...}{Additional arguments to pass to the plotting function
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the autocorrelation from a pre-calculated data frame
}
