% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mcmc-diagnostics-all-params.R
\name{plot_mcmc_diagnostics_all_params}
\alias{plot_mcmc_diagnostics_all_params}
\title{Creates the plot showing the 4 MCMC diagnostics for  all parameters
combined}
\usage{
plot_mcmc_diagnostics_all_params(
  model,
  diag_fill = "royalblue",
  diag_alpha = 0.3,
  diag_outline_color = "black",
  ...
)
}
\arguments{
\item{model}{A model object as output by \code{\link[=create_rds_file]{create_rds_file()}}.}

\item{diag_fill}{The fill color for the bars in the plots}

\item{diag_alpha}{The transparency for the bars in the plots}

\item{diag_outline_color}{The bar outline color for the bars in the plots}

\item{...}{Arguments passed to \code{\link[=calc_mcmc_param_stats]{calc_mcmc_param_stats()}} and
\code{\link[=plot_mcmc_histogram]{plot_mcmc_histogram()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Creates the plot showing the 4 MCMC diagnostics for  all parameters
combined
}
