% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mcmc-histogram.R
\name{plot_mcmc_histogram}
\alias{plot_mcmc_histogram}
\title{Plot a histogram for a given column in a data frame}
\usage{
plot_mcmc_histogram(
  d,
  col_nm,
  x_lab = col_nm,
  y_lab = "Count",
  barplot = FALSE,
  lvls = NULL,
  scale_effn = 1,
  show_ro = FALSE,
  show_bar_values = TRUE,
  ro_arrow_length = 20,
  ro_text_nudge = 0,
  ro_text_size = 5,
  bar_label_limit = 10,
  bar_text_color = "navyblue",
  bar_text_size = 2,
  x_lim = NULL,
  y_lim = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  x_brk = 0.1,
  y_brk = 20,
  ...
)
}
\arguments{
\item{d}{A dta frame containing the column with name \code{col_nm} which
contains the values to use in the plot}

\item{col_nm}{The name of the column in \code{d}containing the
values to use in the plot}

\item{x_lab}{The label to show on the x-axis}

\item{y_lab}{The label to show on the y-axis}

\item{barplot}{Logical. If \code{TRUE}, return a \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}, which
is used for character columns (discrete data). If \code{FALSE}, return a
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} which is for numerical data (continuous data)}

\item{lvls}{A vector of levels that should appear in the data if
\code{barplot} is \code{TRUE}. If \code{NULL}, the data will be plotted as-is}

\item{scale_effn}{}

\item{show_ro}{Logical. If \code{TRUE}, make a label with arrow showing which
bin the R0 parameter is in}

\item{ro_arrow_length}{The length in count units (y-axis units) of the
arrow pointing to the bin that the R0 parameter is in. Only used if
\code{show_ro} is \code{TRUE}}

\item{ro_text_nudge}{Amount to nudge to the left(negative) or right
(positive) the R0 location label}

\item{ro_text_size}{The font size for the R0 labels}

\item{bar_label_limit}{The number in a bin (count - y-axis), above which
the label of the count for the bar is in white in the middle of the bar
vertically. Below this value, the bar value will be above the bar. To
have all value labels above the bars, set this to a value larger than the
highest count bin}

\item{bar_text_color}{The color of the count labels for each bar. Only
used if \code{show_bar_values} is \code{TRUE}}

\item{bar_text_size}{The font size of the count labels for each bar. Only
used if \code{show_bar_values} is \code{TRUE}}

\item{x_lim}{A vector of two values. The min and max to show on the
x-axis. If \code{NULL}, values will be calculated from the input values}

\item{y_lim}{A vector of two values. The min and max to show on the
y-axis. If \code{NULL}, values will be calculated from the input values}

\item{x_breaks}{A vector of values to show on the x-axis. If \code{NULL},
values will be calculated from the input values}

\item{y_breaks}{A vector of values to show on the y-axis. If \code{NULL},
values will be calculated from the input values}

\item{x_brk}{A value to use to create the \code{x_breaks} values if it is
\code{NULL}. This value is used as the \code{by} argument like this:
\code{seq(xlim[1], xlim[2], by = x_brk)}. This is also used as the \code{binwidth}
in the \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} call}

\item{y_brk}{A value to use to create the \code{y_breaks} values if it is
\code{NULL}. This value is used as the \code{by} argument like this:
\code{seq(ylim[1], ylim[2], by = y_brk)}}

\item{...}{Additional arguments to pass to the plotting functions
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} and \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a histogram for a given column in a data frame
}
