% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-param-stats.R
\name{plot_mcmc_param_stats}
\alias{plot_mcmc_param_stats}
\title{Plot Autocorrelation, Effective sample size, Geweke statistic, and
Heidelberger & Welch statistic histograms. This is a re-coded version of
https://github.com/r4ss/r4ss/blob/bioscale/R/mcmc.nuisance.R}
\usage{
plot_mcmc_param_stats(
  model,
  col = get.shade(color = "blue", opacity = 30),
  effn_labels = FALSE,
  ylim_mult = 1,
  show_ro_loc = FALSE,
  ro_yloc = 50,
  ro_yloc_arr_start = 60,
  ro_yloc_arr_end = 30
)
}
\arguments{
\item{model}{A model as output by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{col}{A color to shade the bars}

\item{effn_labels}{If TRUE, add labels to the top of the bars on the effective sample size plot}

\item{ylim_mult}{Multiply the maximum ylim value by this (to show labels on bars better if \code{effn_labels} is \code{TRUE})}

\item{show_ro_loc}{If TRUE, place the text "R0" over the bar in which the R0 parameter falls}

\item{ro_yloc}{If \code{show_ro_loc} is \code{TRUE}, use this value as a y value for the text}

\item{ro_yloc_arr_start}{If \code{show_ro_loc} is \code{TRUE}, use this value as a y value for the start of the arrow
pointing to the bar where R0 is}

\item{ro_yloc_arr_end}{If \code{show_ro_loc} is \code{TRUE}, use this value as a y value for the end of the arrow
pointing to the bar where R0 is}
}
\value{
A \code{\link[=barplot]{barplot()}}
}
\description{
Plot Autocorrelation, Effective sample size, Geweke statistic, and
Heidelberger & Welch statistic histograms. This is a re-coded version of
https://github.com/r4ss/r4ss/blob/bioscale/R/mcmc.nuisance.R
}
