% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-param-stats.R
\name{plot_mcmc_param_stats_makecaption}
\alias{plot_mcmc_param_stats_makecaption}
\title{Make the caption for \link{plot_mcmc_param_stats}}
\usage{
plot_mcmc_param_stats_makecaption(modelname)
}
\arguments{
\item{modelname}{A character string to represent the model name that
will be pasted into the caption.}
}
\description{
Make the caption for \link{plot_mcmc_param_stats}
}
