% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-overview-map.R
\name{plot_overview_map}
\alias{plot_overview_map}
\title{Create a map of the west coast of North America showing places of interest}
\usage{
plot_overview_map(
  ports_df,
  states_df,
  crs = 4326,
  extents = tibble(lon = c(-140, -113), lat = c(33, 58)),
  label_size = 4
)
}
\arguments{
\item{ports_df}{A data frame as read in by \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{states_df}{A data frame as read in by \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{crs}{Coordinate Reference System (CRS) number. Default is 4326 which
is WGS84: See \href{https://epsg.org/home.html}{Epsg.org} for details. Click
\verb{Text search} tab and look at the \code{code} column for valid crs numbers}

\item{extents}{The extents of the map as a two-column data frame with
columns \code{latitude} and \code{longitude}. Change this to make the map larger
or smaller, i.e. to zoom map in/out}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a map of the west coast of North America showing places of interest
}
