% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Create a pairs plot showing correlations among key parameters or
recruitment deviations against the R0 parameter}
\usage{
plot_pairs(
  model,
  type = c("key", "devs"),
  key_posteriors = NULL,
  key_posteriors_titles = NULL,
  recr = NULL,
  bratio = NULL,
  forecatch = NULL,
  point_color = "royalblue",
  point_alpha = 0.3,
  point_size = 0.05,
  lm_color = "black",
  lm_linetype = "solid",
  lm_linewidth = 1,
  strip_fill_color = "transparent",
  strip_fill_alpha = 0.3
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
\code{\link[=create_rds_file]{create_rds_file()}}.}

\item{type}{Either "key" or "devs", which describe the type of pairs plot
to create. If this is "key", all parameters listed in the \code{key_posteriors}
argument will be plotted as well as the objective function and recruitment
parameters for years listed in \code{recr}. If this is "devs", The R0 parameter
will be plotted against the recruitment deviation parameters for the years
listed in \code{recr}}

\item{key_posteriors}{A list of key posterior names (as regular
expressions)}

\item{key_posteriors_titles}{A list of key posterior titles (pretty
names which are meant to be shown on plots)}

\item{recr}{A vector of the recruitment parameter years to include.}

\item{bratio}{A vector of years to plot the relative spawning biomass for}

\item{forecatch}{A vector of years to plot the default harvest catch for.
There are forecasts, hence the name}

\item{point_color}{The color of the scatter plot points}

\item{point_alpha}{The transparency of the scatter plot points}

\item{point_size}{Th size of the scatter plot points}

\item{lm_color}{The color of the regression line found on the scatter plot
panels}

\item{lm_linetype}{The type of the regression line found on the scatter
plot panels}

\item{lm_linewidth}{The width of the regression line found on the scatter
plot panels}

\item{strip_fill_color}{The color to fill in the strip labels
(facet labels)}

\item{strip_fill_alpha}{The transparency for the strip labels
(facet labels)}
}
\value{
A \code{\link[GGally:ggmatrix]{GGally::ggmatrix()}} object
}
\description{
Create a pairs plot showing correlations among key parameters or
recruitment deviations against the R0 parameter
}
