% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-param-density.R
\name{plot_param_density}
\alias{plot_param_density}
\title{Create a plot showing the density of the posterior for the catch forecast
for a given year}
\usage{
plot_param_density(
  model,
  param_nm,
  prob_vec = probs,
  scale = 1000,
  x_lab = "",
  y_lab = "Density",
  x_lim = c(0, 3000),
  y_lim = c(0, 1e-04),
  expand_vec = c(0, 0),
  x_breaks = seq(x_lim[1], x_lim[2], by = (x_lim[2] - x_lim[1])/6),
  fill_color = ts_single_model_ribbon_fill,
  fill_alpha_tail = 0.3,
  fill_alpha_main = 0.7,
  show_y_labels = TRUE,
  show_med_label = TRUE,
  med_label_units = "kt",
  med_label_digits = 3,
  arrow_obj = arrow(length = unit(0.5, "cm"), angle = 20, type = "closed"),
  label_x_offset = 500,
  label_y_offset = 1e-04,
  ...
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
\code{\link[=create_rds_file]{create_rds_file()}}}

\item{param_nm}{A parameter name present in \code{names(model$mcmc)}}

\item{prob_vec}{A vector of probabilities to use in the \code{quantile()}
calculation}

\item{scale}{A value to divide the parameter values by}

\item{x_lab}{The x-axis label}

\item{y_lab}{The y-axis label. Only shown if \code{show_y_labels} is \code{TRUE}}

\item{x_lim}{A vector of two values. The x-axis minimum and maximum values}

\item{expand_vec}{A vector to pass to the \code{expand} argument of the
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} function. Default is \code{c(0, 0)}}

\item{fill_color}{The color used to fill under the density curve}

\item{fill_alpha_tail}{The transparency used for the color of the tails}

\item{fill_alpha_main}{The transparency used for the color of the main part
of the curve (-1SD - 1SD)}

\item{show_y_labels}{Logical. If \code{TRUE}, show the y-axis label , ticks,
and tick labels}

\item{show_med_label}{Logical. If \code{TRUE}, show the median label with arrow}

\item{med_label_units}{Units (as text) to place in the label showing the
median. Only used if \code{show_med_label} is \code{TRUE}}

\item{label_x_offset}{An x-axis offset to move the median label to,
Positive means move right. Only used if \code{show_med_label} is \code{TRUE}}

\item{label_y_offset}{A y-axis offset to move the median label to,
Positive means move up. Only used if \code{show_med_label} is `TRUE}

\item{...}{Arguments to pass to \code{\link[=theme_axis_text]{theme_axis_text()}}}

\item{show_med_digits}{Number of decimal places to show in the median label.
Only used if \code{show_med_label} is \code{TRUE}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot showing the density of the posterior for the catch forecast
for a given year
}
