% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-phase.R
\name{plot_phase}
\alias{plot_phase}
\title{Create plot of the relative fishing intensity in year t-1 against relative
spawning biomass in year t, as a historical look at the fishery for the
MCMC given by model}
\usage{
plot_phase(
  model,
  start_yr,
  end_yr,
  x_lim = c(0, 1.4),
  y_lim = c(0, 1.2),
  init_lbl_x_off = 0,
  init_lbl_y_off = 0,
  final_lbl_x_off = 0,
  final_lbl_y_off = 0,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{start_yr}{The minimum year in the plot}

\item{end_yr}{The maximum year in the plot}

\item{x_lim}{Vector of two values for minimum and maximum x-axis values}

\item{y_lim}{Vector of two values for minimum and maximum y-axis values}

\item{init_lbl_x_off}{Initial year label x offset in units of axis}

\item{init_lbl_y_off}{Initial year label y offset in units of axis}

\item{final_lbl_x_off}{Final year label x offset in units of axis}

\item{final_lbl_y_off}{Final year label y offset in units of axis}

\item{ax_title_font_size}{Size of the font for the X and Y axis labels}

\item{ax_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{ax_label_color}{Color of the font for the X and Y axis tick and
title labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create plot of the relative fishing intensity in year t-1 against relative
spawning biomass in year t, as a historical look at the fishery for the
MCMC given by model
}
