% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rawmeasure.R
\name{plot_raw_age}
\alias{plot_raw_age}
\title{Workup At-Sea and PacFIN lengths and weights}
\usage{
plot_raw_age(
  atsea.ages = get_local("nages.Rdat"),
  page = get_local("page.Rdat"),
  years = as.numeric(format(Sys.time(), "\%Y")) - 0:4,
  xlim_weight = c(0, 3),
  xlim_length = c(0, 70)
)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is \code{NULL},
where the object will be read from a saved object within a directory called
extractedData.}

\item{years}{A vector of years that you want plotted. The default is to plot
the most recent five years.}

\item{ncatch}{An R object of NORPAC catches.
The default value is loaded from the disk using \code{\link[=get_local]{get_local()}}.
Otherwise, the R object is typically stored in the \code{hakedata} environment.}
}
\value{
Four figures are saved to the disk. Figures are of length and weight
distributions for the at-sea and shoreside sectors.
}
\description{
Workup At-Sea and PacFIN lengths and weights
}
\author{
Kelli F. Johnson
}
