% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-retro-recruits.R
\name{plot_retro_recruits}
\alias{plot_retro_recruits}
\title{Make squid plot of retrospectives of recruitment deviations.}
\usage{
plot_retro_recruits(
  retroSummary,
  endyrvec,
  cohorts,
  ylim = NULL,
  uncertainty = FALSE,
  labels = c("Recruitment deviation", "Recruitment (billions)",
    "relative to recent estimate", "Age"),
  main = "Retrospective analysis of recruitment deviations",
  mcmcVec = FALSE,
  devs = TRUE,
  relative = FALSE,
  labelyears = TRUE,
  legend = FALSE,
  leg.ncols = 4
)
}
\arguments{
\item{retroSummary}{List object created by \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}}}

\item{endyrvec}{Vector of years representing the final year of values to
show for each model}

\item{cohorts}{Which cohorts to show in plot}

\item{ylim}{Limits of y-axis}

\item{uncertainty}{Logical. If \code{TRUE}, show uncertainty intervals around lines}

\item{labels}{Vector of plot labels}

\item{main}{Title for plot}

\item{mcmcVec}{Either vector of \code{TRUE}/\code{FALSE} values indicating which models
use MCMC.  Or single value applied to all models}

\item{devs}{Plot deviations instead of absolute recruitment values?}

\item{relative}{Show deviations relative to most recent estimate or relative
to 0}

\item{labelyears}{Label cohorts with text at the end of each line?}

\item{legend}{Add a legend showing which color goes with which line (as
alternative to \code{labelyears}).}

\item{leg.ncols}{Number of columns for the legend.}
}
\description{
Make squid plot of retrospectives of recruitment deviations.
}
\details{
Based on \code{\link[r4ss:SSplotRetroRecruits]{r4ss::SSplotRetroRecruits()}}
Inspired by Jim Ianelli and named by Sean Cox, the squid plot is a way to
examine retrospective patterns in estimation of recruitment deviations.
}
