% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_retro_yrs}
\alias{plot_retro_yrs}
\title{The years previous to the end of the time series to include in plotting.
This value must be a subset of \code{retrospective_yrs}}
\format{
A vector of numbers

A logical value
}
\usage{
plot_retro_yrs

plot_retro_yrs
}
\description{
The years previous to the end of the time series to include in plotting.
This value must be a subset of \code{retrospective_yrs}

A toggle for showing the output from the SS model runs on the console
}
\keyword{datasets}
