% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-running-quants.R
\name{plot_running_quants}
\alias{plot_running_quants}
\title{Create a plot with running quantiles across the posterior for a parameter}
\usage{
plot_running_quants(
  model,
  post_regex,
  probs = c(0.025, 0.5, 0.975),
  rib_alpha = 0.2,
  y_lim = NULL
)
}
\arguments{
\item{model}{A model list, created by [create_rds_file()]}

\item{post_regex}{A regular expression that matches one parameter name
and nothing else}

\item{probs}{A vector of 3 values for the lower, median, and upper
quantiles}

\item{rib_alpha}{The alpha value for the credible interval ribbon}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Create a plot with running quantiles across the posterior for a parameter
}
