% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-mountains.R
\name{plot_selex_mountains}
\alias{plot_selex_mountains}
\title{Create a plot of median selectivity as a set of overlapping "mountains"
which are shaded top-down for visual appeal}
\usage{
plot_selex_mountains(
  model,
  yrs = NULL,
  ages = 1:8,
  scale = 1,
  fill_num_colors = 20,
  fill_col_upper = "darkblue",
  fill_col_lower = "lightblue"
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
\code{\link[=create_rds_file]{create_rds_file()}}.}

\item{yrs}{A vector of years to include in the plot. If \code{NULL}, all years
found in the data will be included}

\item{ages}{A vector of ages to include in the plot}

\item{scale}{A scaling factor to increase or decrease the height of the
selectivity plots}

\item{fill_num_colors}{The number of colors for the gradient of colors
between \code{fill_col_upper} and \code{fill_col_lower}}

\item{fill_col_upper}{The top color for the gradient color scheme}

\item{fill_col_lower}{The bottom color for the gradient color scheme}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of median selectivity as a set of overlapping "mountains"
which are shaded top-down for visual appeal
}
