% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-posteriors.R
\name{plot_selex_posteriors}
\alias{plot_selex_posteriors}
\title{Plot the selectivity parameter estimates from all posteriors
as many lines with an overlaid errorbar for each age}
\usage{
plot_selex_posteriors(
  model = NULL,
  type = c("fishery", "survey"),
  probs = c(0.025, 0.975),
  age_range = NULL,
  line_color = "black",
  line_thickness = 0.05,
  line_color_unc = "red",
  line_thickness_unc = 1,
  point_size_unc = 3
)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{One of 1 or 2 where 1 = Fishery and 2 = Acoustic age 2+ survey}

\item{probs}{A vector of two quantiles to include for the uncertainty}

\item{age_range}{A vector of two for the min and max ages to plot. If \code{NULL},
all ages in the data will be plotted}

\item{line_color}{Color of the posterior lines}

\item{line_thickness}{Thickness of the posterior lines}

\item{line_color_unc}{Color of the uncertainty lines and points}

\item{line_thickness_unc}{Thickness of the uncertainty lines}

\item{point_size_unc}{Size of the points}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the selectivity parameter estimates from all posteriors
as many lines with an overlaid errorbar for each age
}
