% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-uncertainty.R
\name{plot_selex_uncertainty}
\alias{plot_selex_uncertainty}
\title{Create a plot of time-varying selectivity panels}
\usage{
plot_selex_uncertainty(
  model,
  yr_lim = c(1990, last_data_yr),
  ages = 1:8,
  n_col = 1,
  rev = FALSE,
  point_size = 1,
  point_fatten = 2,
  border_width = 2,
  border_color = "black",
  border_linetype = "solid",
  show_panel_borders = FALSE,
  label_loc = c(1, 0.75),
  label_font_size = 4,
  pad_top = TRUE,
  pad_bottom = TRUE
)
}
\arguments{
\item{model}{A model list, as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yr_lim}{A vector of two values representing the minimum and
maximum years to plot panels for}

\item{ages}{A vector of ages to include}

\item{n_col}{The number of columns to hold panels}

\item{rev}{Logical. If \code{TRUE}, reverse the order of the years}

\item{point_size}{The point size}

\item{point_fatten}{The fatness of the point. Needed when using
\code{\link[ggplot2:geom_linerange]{ggplot2::geom_errorbar()}}}

\item{border_width}{Thickness of the border line. If \code{NULL} no border will
be shown}

\item{border_color}{Color of the column borders}

\item{border_linetype}{Line type of the column borders}

\item{show_panel_borders}{Logical. If \code{TRUE}, show borders around the panels}

\item{label_loc}{A vector of two (x, y) describing the label location inside
the panels}

\item{label_font_size}{The labels font size}

\item{pad_top}{Add blank cells across the top as "padding"}

\item{pad_bottom}{Add blank cells across the bottom as "padding"}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of time-varying selectivity panels
}
\details{
The panels show the medians and the credible interval as
calculated in \code{\link[=load_extra_mcmc]{load_extra_mcmc()}}.
}
