% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spatial-cpue.R
\name{plot_spatial_cpue}
\alias{plot_spatial_cpue}
\title{Create a map of the west coast of North America showing places of interest}
\usage{
plot_spatial_cpue(
  d,
  crs_ll = 4326,
  crs_utm = 32609,
  bathy_resolution = 1,
  x_lim = c(-131.25, -124.5),
  y_lim = c(48, 52.5),
  bin_width = 7,
  n_minimum_vessels = 3,
  bath_contours = c(100, 200, 500),
  hexagon_border_color = "black",
  hexagon_border_thickness = 0.5
)
}
\arguments{
\item{d}{Data from \code{\link[gfdata:get_data]{gfdata::get_cpue_spatial()}},
\code{\link[gfdata:get_data]{gfdata::get_cpue_spatial_ll()}}, or \code{\link[gfdata:get_data]{gfdata::get_catch_spatial()}}}

\item{crs_ll}{Coordinate Reference System (CRS) number for a
latitude/longitude-based projection. This could be NAD83 or WGS84 or
others. The default is 4326 which is WGS84: See
\href{https://epsg.org/home.html}{Epsg.org} for details. Click \verb{Text search}
tab and look at the \code{code} column for valid crs numbers}

\item{crs_utm}{Coordinate Reference System (CRS) number for a
Easting/Northing-based projection. This could be UTM9 or UTM10 or
others. There is a different numbre for each UTM zone. The default is
32609 which is UTM zone 9 (west coast Canada). See
\href{https://epsg.org/home.html}{Epsg.org} for details. Click \verb{Text search}
tab and look at the \code{code} column for valid crs numbers}

\item{bathy_resolution}{The resolution for the bathymetry in minutes.
See \code{\link[marmap:getNOAA.bathy]{marmap::getNOAA.bathy()}}}

\item{x_lim}{The length-2 vector representing the minimum and maximum
limits of the x-axis in degrees of longitude}

\item{y_lim}{The length-2 vector representing the minimum and maximum
limits of the y-axis in degrees of latitude}

\item{bin_width}{Bin width as defined in \code{\link[ggplot2:stat_summary_2d]{ggplot2::stat_summary_hex()}}}

\item{n_minimum_vessels}{The minimum number of vessels that must be present
in a hexagon for the hexagon to be shown}

\item{bath_contours}{A vector of contours to show contour lines for.
These can be either positive or negative; they will all be made negative
in the function}

\item{hexagon_border_color}{The color of the hexagon border lines}

\item{hexagon_border_thickness}{The thickness of the hexagon border lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a map of the west coast of North America showing places of interest
}
