% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spr-illustration.R
\name{plot_spr_illustration}
\alias{plot_spr_illustration}
\title{Create a three-panel plot illustrating how the SPR is calculated.}
\usage{
plot_spr_illustration(
  model,
  yrs = NULL,
  color = "royalblue",
  show_legend = TRUE
)
}
\arguments{
\item{model}{A model object, , created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yrs}{A vector of years to include in the calculation. If \code{NULL},
the years found in the weight-at-age data frame (\code{model$wtatage})
will be used}

\item{color}{The color to use for the bars}

\item{show_legend}{Logical. Whether or not to show the text 'legend'}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a three-panel plot illustrating how the SPR is calculated.
}
