% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-squid.R
\name{plot_squid}
\alias{plot_squid}
\title{Make the retrospective recruitment deviations plot (AKA 'squid plot')}
\usage{
plot_squid(model, subplot = 1, cohorts, plot_mcmc = TRUE, getdevs = TRUE)
}
\arguments{
\item{model}{Model with retrospectives}

\item{subplot}{1 = Recruitment deviations, 2 = Rec dev strength relative to most recent estimate}

\item{cohorts}{Vector of cohort years to plot (for labels)}

\item{plot_mcmc}{If \code{TRUE} plot the MCMC version based on medians. If \code{FALSE} plot MLE version.}

\item{getdevs}{A logical value specifying if the plot should be of
recruitment deviations, which is the default. If \code{FALSE}, then
the squid plot will be made using absolute recruitment instead of
deviations.}
}
\description{
Make the retrospective recruitment deviations plot (AKA 'squid plot')
}
