% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survey-biomass.R
\name{plot_survey_biomass}
\alias{plot_survey_biomass}
\title{Create a plot of survey biomass as points with errorbars}
\usage{
plot_survey_biomass(model, index = c("age1", "age2"), y_lim = c(0, 3))
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{index}{The index type to plot}

\item{y_lim}{A vector of two for the minimum and maximum values
for the y-axis on the plot}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of survey biomass as points with errorbars
}
\details{
Extra SD is shown for the 2009 "squid year"
}
