% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Theme For \code{\link[ggplot2]{theme}}}
\usage{
plot_theme()
}
\value{
A \code{\link[ggplot2]{theme}} object with specific
instructions on how to create plots for this repository.
}
\description{
Create a special theme for hake plots when plotting using the
\code{\link[ggplot2]{ggplot}} functions.
}
\author{
Kelli F. Johnson
}
