% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weight-at-age.R, R/plot_weight_at_age.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Create a plot of weight-at-age lines by age and year}
\usage{
plot_weight_at_age(data, maxage)

plot_weight_at_age(data, maxage)
}
\arguments{
\item{data}{A data frame of weight-at-age values.}

\item{maxage}{The maximum age in the modelled data.}

\item{wa}{A data frame of weight-at-age with a \code{yr} column and
numerically-named columns for ages}

\item{ages}{A vector of ages to plot. If \code{NULL}, all ages in the \code{wa}
data frame will be plotted}

\item{bold_ages}{A vector of ages to make bold in the plot. The lines for
these ages will be thicker and the labels for them will be bold as well}

\item{cols}{A vector of colors to base the color ramp on. See
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}

\item{tick_prop}{A value that the length of the major tick marks are
multiplied by. This proportion must be set by trial and error. Make sure
to change \code{vjust_x_labels} so the labels are not overlapping the lines or
are too far away from the lines}

\item{vjust_x_labels}{Adjustment to move the x-axis tick labels and label
up or down. Negative numbers move down}

\item{remove_yr_labels}{A vector of years to remove the ,labels for in
case they are overlapping}

\item{x_lim}{The min and max limits as a vector for the x-axis}

\item{y_lim}{The min and max limits as a vector for the y-axis}

\item{y_lab_by}{The amount between each y-axis label}

\item{x_labs_mod}{How many years between year labels on the x-axis}

\item{age_label_font_size}{The font size for the labels pointing to the age
lines}

\item{age_label_side}{Where to place the age labels next to the lines.
One of "none", "both", "left", or "right"}

\item{x_labels_mod}{Show every nth year where n is this number}

\item{...}{Arguments passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object

A {ggplot2} object.
}
\description{
Create a plot of weight-at-age lines by age and year

Plot weight-at-age data
}
\author{
Kelli F. Johnson
}
