% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weight-at-age-heatmap.R
\name{plot_weight_at_age_heatmap}
\alias{plot_weight_at_age_heatmap}
\title{Create a weight-at-age heatmap including extrapolated years}
\usage{
plot_weight_at_age_heatmap(
  model,
  fleet = 1,
  proj_line_color = "royalblue",
  proj_line_width = 1,
  proj_line_yr = NULL,
  yrs = NULL,
  extrap_mask = NULL,
  font_size = 4,
  axis_font_size = 10,
  samplesize = FALSE,
  print_yrs = NULL,
  colour = c("all", "age", "both"),
  ...
)
}
\arguments{
\item{model}{An list of results read in from an SS model using
\code{\link[=load_ss_files]{load_ss_files()}}}

\item{fleet}{An integer value specifying which fleet you want plotted.
Fleet -2 will plot fecundity information.
Fleet -1 will plot population weight-at-age for the middle of the year.
Fleet 0 will plot population weight-at-age for the beginning of the year.
Positive values for fleet will link to a modeled fleet.}

\item{proj_line_color}{Line color to separate projection years.}

\item{proj_line_width}{Line width to separate projection years.}

\item{yrs}{A vector of the years to include in the table. If \code{NULL}, all
years will be included}

\item{extrap_mask}{A data frame where ages are columns (and start with
the letter 'a'). If the values are zero, the weight-at-age was
extrapolated/interpolated. If there is a value, the weight-at-age is data}

\item{font_size}{Font size of the values printed in each box.}

\item{axis_font_size}{Font size for axis labels.}

\item{samplesize}{A logical value specifying if the heat map should be of
input sample size used to generate the data rather than the weights-at-age.}

\item{print_yrs}{A vector of years that will be included in the output,
but users should note that all data from \code{first_yr} to the terminal
year of the data in the model will be used to calculate means and colors.
This parameter facilitates splitting the plot into two separate plots when
there are a number of years.}

\item{colour}{A character string of \code{both}, \code{all}, or \code{age}. If the
default of \code{both}, then the colors of the boxes will be based on all of
the data and the transparency of the colors will strictly be a function
of each individual age in turn. That is, the darkness of the colors
used across years for age zero specifies which years have the largest
age-0 fish. \code{all} specifies colors based on the min and max values of
all ages without transparency. \code{age} colors the min and max color
specific for each age without transparency}

\item{longterm_mean_ages}{A vector of mean weight-at-age values
per ages zero to the maximum age of fish in the data.
If \code{NULL}, the first year will be assumed to be the mean  of all years.}

\item{start_yr}{Start year}

\item{end_yr}{End year}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a weight-at-age heatmap including extrapolated years
}
\details{
Produce a matrix of colors and values of weight-at-age information
by year (y-axis) and age (x-axis). The plot uses output from a
Stock Synthesis model.

Ages zero through the maximum age modeled in the data are shown
for each year plotted. This maximum-age shown is found by removing ages
that are the duplicate of the previous age, no user input is needed.

Users have many options to control what years are included in the plot.
The beginning years of data can be removed using \code{first_yr}.
Users do not have control over what are noted as projection years, which
are all years after the end of the data included in the assessment model
up to last year included in the weight-at-age matrix.
These projection years are noted using a horizontal line but if the
demarcation is not wanted, then they can change the line width to zero.
Finally, users can calculate the colors using a range of data specified
with \code{first_yr} and the last year of the data in the model. Then,
the resulting plot can be truncated to a specified year range using
\code{print_yrs}. The truncation is helpful to facilitate plots that fit
on a single page or on a slide with readable values. Using this truncation,
the colors would show trends across all plots made rather than just the
truncated data.
}
