% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weights-lengths.R
\name{plot_weights_lengths}
\alias{plot_weights_lengths}
\title{Plot some years of weight of length histograms}
\usage{
plot_weights_lengths(
  df,
  type = c("length", "weight"),
  include_sex = TRUE,
  min_total = 20,
  bin_width = 0.075,
  x_breaks = pretty(d[[bin_col]], 4L),
  label_font_size = 2.25,
  fill_col = c(F = rgb(1, 0, 0, 0.3), M = rgb(0, 0, 1, 0.3)),
  line_col = c(F = rgb(1, 0, 0, 0.3), M = rgb(0, 0, 1, 0.3)),
  unsexed_fill_col = rgb(0, 0, 1, 0.3),
  unsexed_col = "black",
  unit_str = ifelse(type == "length", "cm", "kg")
)
}
\arguments{
\item{df}{A data frame as returned by \code{\link[gfplot:tidy_comps]{gfplot::tidy_lengths_raw()}}}

\item{type}{Either "length" or "weight"}

\item{include_sex}{Logical. If \code{TRUE}, include the sex data in the plot,
with two bar colors}

\item{min_total}{Minimum number of fish for a given survey and year needed
before a histogram is shown.}

\item{bin_width}{The width of the bars in units of the plot}

\item{x_breaks}{A vector of values to show on the x-axis as tick labels}

\item{label_font_size}{The font size for the sample size labels}

\item{fill_col}{A named vector of two colors for Female and male bars.
The names are "F" and "M" respectively.}

\item{line_col}{A named vector of two colors for Female and male bar
outlines. The names are "F" and "M" respectively.}

\item{unsexed_fill_col}{The color to use for non-sex-specific bar fill}

\item{unsexed_col}{Thew color to use for the bar border for unsexed bars}

\item{unit_str}{A string representing the units, e.g. 'g' or 'kg'}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot some years of weight of length histograms
}
