% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotBars.fn}
\alias{plotBars.fn}
\title{Plot bars}
\usage{
plotBars.fn(
  x,
  y,
  scale = 1,
  gap = 0,
  add = FALSE,
  ciCol = "black",
  ciLty = 1,
  ciLwd = 1,
  ...
)
}
\arguments{
\item{x}{The x axis values (i.e., years)}

\item{y}{A data frame with columns:
value: estimate (point) to plot
lo: lower CI
hi: higher CI}

\item{scale}{Amount to divide y-axis values by}

\item{gap}{Gap size for the segments}

\item{add}{Logical. If \code{TRUE}, add to the current plot, if \code{FALSE},
create a new plot}

\item{ciCol}{Credible interval color}

\item{ciLty}{Credible interval line type}

\item{ciLwd}{Credible interval width}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Plot bars
}
