% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ports_df}
\alias{ports_df}
\title{A data frame containing information about the ports used in making the
overview map at the beginning of the figures section}
\format{
A data frame with 28 rows and 6 variables:
\describe{
\item{pos}{Position (side) to place label on}
\item{lat}{Latitude to place label}
\item{lon}{Longitude to place label}
\item{north_south}{Grouping variable for shiftin base vertical line for placement}
\item{name}{Name of the port}
}
}
\usage{
ports_df
}
\description{
A data frame containing information about the ports used in making the
overview map at the beginning of the figures section
}
\keyword{datasets}
