% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-appendix-counter-resets.R
\name{post_process_add_appendix_counter_resets}
\alias{post_process_add_appendix_counter_resets}
\title{Add appendix counter rests LaTeX code for equations, figures, and tables
so that each lettered appendix starts at 1 again for those things}
\usage{
post_process_add_appendix_counter_resets(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add appendix counter rests LaTeX code for equations, figures, and tables
so that each lettered appendix starts at 1 again for those things
}
