% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-counters.R
\name{post_process_add_counters}
\alias{post_process_add_counters}
\title{Add LaTeX counter code so that the Executive summary has letters for
tables and figures, and the rest of the document has numbers}
\usage{
post_process_add_counters(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add LaTeX counter code so that the Executive summary has letters for
tables and figures, and the rest of the document has numbers
}
