% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-disclaimer-presentations.R
\name{post_process_add_disclaimer_presentations}
\alias{post_process_add_disclaimer_presentations}
\title{Add the disclaimer to the title slide for the presentations}
\usage{
post_process_add_disclaimer_presentations(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Arguments passed to all the post-processing functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add the disclaimer to the title slide for the presentations
}
