% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-headers-footers.R
\name{post_process_add_headers_footers}
\alias{post_process_add_headers_footers}
\title{Add left and right footer text to the pages}
\usage{
post_process_add_headers_footers(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add left and right footer text to the pages
}
