% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-horiz-lines-decision-table.R
\name{post_process_add_horiz_lines_decision_table}
\alias{post_process_add_horiz_lines_decision_table}
\title{Adds two horizontal lines to the decision tables}
\usage{
post_process_add_horiz_lines_decision_table(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Adds two horizontal lines to the decision tables
}
