% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-picture-to-title-page.R
\name{post_process_add_picture_to_title_page}
\alias{post_process_add_picture_to_title_page}
\title{Insert an image on the title page}
\usage{
post_process_add_picture_to_title_page(
  x,
  title_page_image = NULL,
  title_page_image_width_cm = NULL,
  figures_dir = NULL,
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{title_page_image}{The file to include as the title page image}

\item{title_page_image_width_cm}{The width in cm to use for the picture}

\item{figures_dir}{The subdirectory of the "doc" directory containing
images that have previously been made such as pictures and logos, and all
other figures made outside the scope of this project}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Insert an image on the title page
}
