% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-start-appendices-code.R
\name{post_process_add_start_appendices_code}
\alias{post_process_add_start_appendices_code}
\title{Add start appendices LaTeX code}
\usage{
post_process_add_start_appendices_code(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add start appendices LaTeX code
}
\details{
Add LaTeX code for the start of the appendices section. This is needed
to tell the LaTeX compiler to start numbering the sections using letters
and set u-p tables, figures, and equations to be of the format A.2 for
example. It also adds TOC information so this has to come before the TOC
post processing step
}
