% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-beamer-add-extra-headers-etc.R
\name{post_process_beamer_add_extra_headers_etc}
\alias{post_process_beamer_add_extra_headers_etc}
\title{Add extra headers needed for tables, colors in tables, special
cell formats and page numbering}
\usage{
post_process_beamer_add_extra_headers_etc(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add extra headers needed for tables, colors in tables, special
cell formats and page numbering
}
