% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-beamer-insert-title-page-logos.R
\name{post_process_beamer_insert_title_page_logos}
\alias{post_process_beamer_insert_title_page_logos}
\title{Insert the NOAA and DFO logos and a hake picture on the title page of a
beamer presentation}
\usage{
post_process_beamer_insert_title_page_logos(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Arguments passed to all the post-processing functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Insert the NOAA and DFO logos and a hake picture on the title page of a
beamer presentation
}
