% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-beamer-make-title-page-plain.R
\name{post_process_beamer_make_title_page_plain}
\alias{post_process_beamer_make_title_page_plain}
\title{Remove theme shadow and navigation bar from the title page}
\usage{
post_process_beamer_make_title_page_plain(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Arguments passed to all the post-processing functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Remove theme shadow and navigation bar from the title page
}
