% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-beamer-old.R
\name{post_process_beamer_old}
\alias{post_process_beamer_old}
\title{Post-process Beamer Rmarkdown TeX files}
\usage{
post_process_beamer_old(tex_fn = "beamer-hake-data.tex")
}
\arguments{
\item{tex_fn}{The TeX file name}
}
\value{
Nothing, overwrites the input file \code{tex_fn}
}
\description{
Post-process Beamer Rmarkdown TeX files
}
\details{
Inserts the logos and picture on the title slide, and makes the disclaimer
text tiny so it fits on the slide
Creates a backup of the \code{tex_fn} file before modifying it. The backup file
is called 'backup.tex',
Must run \verb{lualatex beamer-hake-data.tex;lualatex beamer-hake-data.tex} in
a command-line terminal once done to re-create the PDF.
}
