% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-convert-to-article.R
\name{post_process_convert_to_article}
\alias{post_process_convert_to_article}
\title{Replace chapter with section for all sections, and prepend a "sub" to
each section, subsection etc to lower each by one level}
\usage{
post_process_convert_to_article(x)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}
}
\value{
The modified Tex code, as a vector
}
\description{
Replace chapter with section for all sections, and prepend a "sub" to
each section, subsection etc to lower each by one level
}
