% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-error-check.R
\name{post_process_error_check}
\alias{post_process_error_check}
\title{Check the input LaTex for errors or omissions before post processing occurs}
\usage{
post_process_error_check(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
`\code{\link[=invisible]{invisible()}}
}
\description{
Check the input LaTex for errors or omissions before post processing occurs
}
