% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-extract-chunks.R
\name{post_process_extract_chunks}
\alias{post_process_extract_chunks}
\title{Extract chunks of lines from a vector of lines}
\usage{
post_process_extract_chunks(x, beg_inds, end_inds)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by}

\item{beg_inds}{A vector of beginning indices for the chunks to
extract. Must be the same length as \code{end_inds}}

\item{end_inds}{A vector of ending indices for the chunks to
extract. Must be the same length as \code{beg_inds}}
}
\value{
A list of three elements, the first being the list of vectors
which are 'in-between' the \code{beg_inds} and \code{end_inds} values. The second
is a list of vectors which are 'between' the \code{beg_inds} and \code{end_inds}
values. the third is \code{first} which is explained above in \verb{@details}
}
\description{
Extract chunks of lines from a vector of lines
}
\details{
Creates two lists of slices of the input vector, the first list will
contain the slices between \code{beg_inds} and \code{end_inds}. The second
list will contain the slices NOT between \code{beg_inds} and \code{end_inds}.
If the first element of \code{beg_ind} is 1, then the return flag \code{first} will
be set to \code{TRUE}. If the first element of \code{beg_ind} is not a 1, the
return flag \code{first} will be set to \code{FALSE}. If the difference between
an \code{end_inds} value and the corresponding next \code{beg_inds} value, a
line with an empty string will be placed between them so that the
interlacing between the two lists remains perfect, so they 'zipper'
together 1 after the other.
}
\examples{
library(hake)
k <- extract_chunks(letters, c(2, 10), c(9, 12))
}
