% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-fix-landscape-issues.R
\name{post_process_fix_landscape_issues}
\alias{post_process_fix_landscape_issues}
\title{Remove landscape/portrait code between tables/figures which are sequential.
If this is not done, there will be a blank portrait page between
sequential landscape pages}
\usage{
post_process_fix_landscape_issues(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Remove landscape/portrait code between tables/figures which are sequential.
If this is not done, there will be a blank portrait page between
sequential landscape pages
}
