% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-fix-tildes.R
\name{post_process_fix_tildes}
\alias{post_process_fix_tildes}
\title{Change tildes (~) so that they will make it through the Pandoc conversion
to be processed in latex as intended}
\usage{
post_process_fix_tildes(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Change tildes (~) so that they will make it through the Pandoc conversion
to be processed in latex as intended
}
