% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-interlace-chunks.R
\name{post_process_interlace_chunks}
\alias{post_process_interlace_chunks}
\title{Interlace chunks of code that have been extracted by \code{\link[=extract_chunks]{extract_chunks()}}}
\usage{
post_process_interlace_chunks(lst)
}
\arguments{
\item{lst}{A list, as returned by \code{\link[=extract_chunks]{extract_chunks()}}}
}
\value{
A vector, with the \code{between} chunks and \code{inbetween} chunks
interlaced
}
\description{
Interlace chunks of code that have been extracted by \code{\link[=extract_chunks]{extract_chunks()}}
}
