% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-landscape-figures.R
\name{post_process_landscape_figures}
\alias{post_process_landscape_figures}
\title{Post-process landscape figures in TEX code}
\usage{
post_process_landscape_figures(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Post-process landscape figures in TEX code
}
\details{
Searches the TeX code in \code{x} for the keywords \verb{BEGIN LANDSCAPE} and
\verb{END LANDSCAPE} (all in caps) that surround either a knitr chunk that
produces a figure or a Bookdown image import statement with this format:
\verb{![Caption](path/to/filename)}
}
