% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-landscape-tables.R
\name{post_process_landscape_tables}
\alias{post_process_landscape_tables}
\title{Post-process landscape tables in TEX code}
\usage{
post_process_landscape_tables(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Post-process landscape tables in TEX code
}
