% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-longtables.R
\name{post_process_longtables}
\alias{post_process_longtables}
\title{Post-process LaTeX longtables in TEX code}
\usage{
post_process_longtables(x)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}
}
\value{
The modified Tex code, as a vector
}
\description{
Post-process LaTeX longtables in TEX code
}
\details{
Injects some LaTeX code which adds 'Continued on next page ...' and
'... Continued from previous page' to the longtables that split pages.
Removes the caption for the subsequent pages, but leaves the header row
}
