% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-modify-font-info.R
\name{post_process_modify_font_info}
\alias{post_process_modify_font_info}
\title{Modify the font type and size in the LaTeX code}
\usage{
post_process_modify_font_info(x, font = NULL, font_size_pt = NULL, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{font}{The name of the font, assumed to be present on the machine}

\item{font_size_pt}{The font size to use, in pt. Some fonts do not support
all sizes so this may have no effect depending on font}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify the font type and size in the LaTeX code
}
