% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-move-table-captions.R
\name{post_process_move_table_captions}
\alias{post_process_move_table_captions}
\title{Move certain table captions, usually to left-justify. There is no way
to center kable tables and left-justify the captions to the left side
of the page}
\usage{
post_process_move_table_captions(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Move certain table captions, usually to left-justify. There is no way
to center kable tables and left-justify the captions to the left side
of the page
}
