% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-object-placement.R
\name{post_process_object_placement}
\alias{post_process_object_placement}
\title{Modify the tex code \code{x} by changing the table or figure label placement
code for the \code{knitr_label} to the value of \code{place}}
\usage{
post_process_object_placement(x, row_df, row_num, num_search_lines = 5, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by}

\item{row_df}{A data frame which is a single row, and has the columns
\code{type}, \code{label}, \code{file_name}, and \code{placement}}

\item{row_num}{Which row this is from the object placement file}

\item{...}{Absorbs other arguments not meant for this function}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify the tex code \code{x} by changing the table or figure label placement
code for the \code{knitr_label} to the value of \code{place}
}
