% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-remove-title-page-number.R
\name{post_process_remove_title_page_number}
\alias{post_process_remove_title_page_number}
\title{Removes the page number from the title page}
\usage{
post_process_remove_title_page_number(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by}

\item{...}{Absorbs other arguments not meant for this function}
}
\value{
The modified Tex code, as a vector
}
\description{
Removes the page number from the title page
}
