% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-repair-unnumbered-section-links.R
\name{post_process_repair_unnumbered_section_links}
\alias{post_process_repair_unnumbered_section_links}
\title{Add phantomsection tag to sections that are unnumbered to force the
bookmark links to work right for them}
\usage{
post_process_repair_unnumbered_section_links(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add phantomsection tag to sections that are unnumbered to force the
bookmark links to work right for them
}
