% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-section-headers.R
\name{post_process_section_headers}
\alias{post_process_section_headers}
\title{Modify TeX code to capitalize and/or center section or subsection headers}
\usage{
post_process_section_headers(tex_vec)
}
\arguments{
\item{tex_vec}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify TeX code to capitalize and/or center section or subsection headers
}
