% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-set-object-placement.R
\name{post_process_set_object_placement}
\alias{post_process_set_object_placement}
\title{Inject LaTeX float placement codes for figure and tables, based on their
knitr chunk label names}
\usage{
post_process_set_object_placement(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Inject LaTeX float placement codes for figure and tables, based on their
knitr chunk label names
}
