% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-set-tab-fig-placement.R
\name{post_process_set_tab_fig_placement}
\alias{post_process_set_tab_fig_placement}
\title{Modify the tex code \code{x} by changing the table or figure label placement
code for the \code{knitr_label} to the value of \code{place}}
\usage{
post_process_set_tab_fig_placement(
  x,
  type = c("table", "figure"),
  knitr_label = NULL,
  file_name = NULL,
  place,
  n_lines = 5
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by}

\item{type}{One of \code{table} or \code{figure}
\code{\link[=readLines]{readLines()}}}

\item{knitr_label}{The knitr chunk label for the chunk which creates the
figure. If \code{NULL}, \code{file_name} will be used}

\item{file_name}{The file name for an image file inserted using rmarkdown
code. If \code{NULL}, \code{knitr_label} will be used}

\item{place}{One of the following LaTeX placement strings (or a
combination of them):
\itemize{
\item h: Place the float here, i.e., approximately at the same point it occurs
in the source text.
\item t: Position at the top of the page.
\item b: Position at the bottom of the page.
\item p: Put on a special page for floats only.
\item !: Override internal parameters LaTeX uses for determining “good” float
positions.
\item H: Place the float at precisely the location in the LaTeX code. This
requires the float package
}}

\item{figs_dir}{The name of the directory in which the knitr-built
figures reside. Used for matching figure import locations inside the tex
file}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify the tex code \code{x} by changing the table or figure label placement
code for the \code{knitr_label} to the value of \code{place}
}
