% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-subtract-section-space.R
\name{post_process_subtract_section_space}
\alias{post_process_subtract_section_space}
\title{Add LaTeX vertical space code before all sections with a \code{hyperref}
declared before it}
\usage{
post_process_subtract_section_space(x)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}
}
\value{
The modified Tex code, as a vector
}
\description{
Add LaTeX vertical space code before all sections with a \code{hyperref}
declared before it
}
