% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-table-of-contents.R
\name{post_process_table_of_contents}
\alias{post_process_table_of_contents}
\title{Insert the table of contents into TeX code}
\usage{
post_process_table_of_contents(
  x,
  section_indent_inch = 0,
  section_num_width_inch = 0.2,
  subsection_indent_inch = 0.3,
  subsection_num_width_inch = 0.35,
  subsubsection_indent_inch = 0.5,
  subsubsection_num_width_inch = 0.45,
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{section_indent_inch}{Indent value in inches for section headers in
the TOC}

\item{section_num_width_inch}{Width of section numbers in inches in
the TOC}

\item{subsection_indent_inch}{Indent value in inches for subsection headers
in the TOC}

\item{subsection_num_width_inch}{Width of subsection numbers in inches in
the TOC}

\item{subsubsection_indent_inch}{Indent value in inches for subsubsection
headers in the TOC}

\item{subsubsection_num_width_inch}{Width of subsubsection numbers in inches
in the TOC}

\item{...}{Absorb arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Insert the table of contents into TeX code
}
