% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posts_fn}
\alias{posts_fn}
\title{A character string containing the file name for the SS3 posteriors
output file}
\format{
A character string
}
\usage{
posts_fn
}
\description{
A character string containing the file name for the SS3 posteriors
output file
}
\keyword{datasets}
