% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probs}
\alias{probs}
\title{A vector of the probabilities to use for uncertainty calculations in the
assessment document}
\format{
A vector of three values, for the lower credible interval limit,
the median, and the upper credible interval limit
}
\usage{
probs
}
\description{
A vector of the probabilities to use for uncertainty calculations in the
assessment document
}
\keyword{datasets}
