% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_age-.R
\name{process_age_shore}
\alias{process_age_shore}
\title{Workup shoreside composition data}
\usage{
process_age_shore(page = get_local("page.Rdat"), ages = 1:15)
}
\arguments{
\item{page}{PacFIN biological data from the comprehensive database.}

\item{ages}{Vector of ages to keep.}
}
\value{
todo: document the return
}
\description{
Provide composition data from the U.S. Shoreside
hake fishery. Written by Allan Hicks in 2016 and
revised in 2017 by Ian G. Taylor.
}
\author{
Kelli F. Johnson
}
