% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_catch-.R
\name{process_catch_norpac}
\alias{process_catch_norpac}
\title{Process NORPAC catches}
\usage{
process_catch_norpac(
  ncatch = get_local(file = "ncatch.Rdat"),
  nyears = 5,
  savedir = hakedata_wd()
)
}
\arguments{
\item{ncatch}{A data frame with catch information extracted from the NORPAC
database. The default is to read the \code{.Rdat} file from the disk using
\code{\link[=get_local]{get_local()}}. Else, users can just pass the R object itself, which would
typically be \code{hakedata[["ncatch"]]}.}

\item{nyears}{The number of years for plotting.}
}
\value{
Figures and csv files are saved to the disk regarding
catch rates and depth (fishing and bottom depths).

In the \code{Catches} directory, the following csv files are saved:
\itemize{
\item depth-us-atsea-bottom.csv
\item depth-us-atsea-fishing.csv
\item us-cp-catch-by-month.csv
\item us-ms-catch-by-month.csv
\item us-cp-catch-rate-by-month.csv
\item us-ms-catch-rate-by-month.csv
}
}
\description{
Summarize and plot NORPAC catches for the at-sea hake fishery. Fleet types
are assigned to create summaries by fleet and year.
}
\section{Tribal catches}{
The column \code{CDQ_CODE} contains strings starting with an upper-case followed
by and the last two digits of the the CDQ code of the vessel. The code was
created in 1995 and is used to track tribal catches in the hake fishery. The
following codes exist:
\itemize{
\item C51 -- Aleutian Pribilof Island Community Development Association
\item C52 -- Bristol Bay Economic Development Corporation
\item C53 -- Central Bering Sea Fishermen's Association
\item C54 -- Coastal Villages Fishing Cooperative
\item C55 -- Norton Sound Economic Development Corporation
\item C56 -- Yukon Delta Fisheries Development Association
\item R10 -- Research Permit - Alaska Fisheries Science Center Catch Estimation
\item M01 -- Makah Tribe Whiting Association
\item R11 -- Research Permit - WA Sea Grant bird/longline interaction
\item R12 -- Research Permit - WA Sea Grant bird/longline interaction
}
}

\section{Raw data}{
Raw data consists of sampled and unsampled hauls with the total weight
including bycatch weight. For sampled hauls, \code{EXTRAPOLATED_WEIGHT} is the
weight of the catch only. For unsampled hauls that have no knowledge of the
amount of bycatch included in the haul, \code{OFFICIAL_TOTAL_CATCH} must be used
along with a estimated bycatch rate to calculate \code{EXTRAPOLATED_WEIGHT}.
Monthly fleet-specific bycatch rates are estimated from the sampled hauls and
multiplied by the \code{OFFICIAL_TOTAL_CATCH} to get an estimate of the amount of
bycatch that should be subtracted from \code{OFFICIAL_TOTAL_CATCH} to get just the
weight of Pacific Hake.
The raw column names are as follows \code{colnames(ncatch)}:
\itemize{
\item "HAULJOIN"
\item "CRUISE"
\item "PERMIT": "a unique vessel identifier"
\item "VESSEL"
\item "VESSEL_TYPE"
\item "HAUL_DATE"
\item "HAUL"
\item "DEPLOYMENT_DATE"
\item "RETRIEVAL_DATE"
\item "HRS"
\item "CDQ_CODE"
\item "OFFICIAL_TOTAL_CATCH"
\item "HAUL_SAMPLED_BY"
\item "LATDD_START"
\item "LONDD_START"
\item "LATDD_END"
\item "LONDD_END"
\item "FISHING_DEPTH_FATHOMS"
\item "BOTTOM_DEPTH_FATHOMS"
\item "CATCHER_BOAT_ADFG": "unique vessel identifier for catcher vessels
delivering to motherships"
\item "SPECIES"
\item "EXTRAPOLATED_WEIGHT"
}
}

\section{Definitions}{
\enumerate{
\item \code{OFFICIAL_TOTAL_CATCH} is measured in metric tons (mt) and includes both
retained and discarded species. Thus, this measurement sums the targeted
species, prohibited species, and non-allocated species.
\item \code{EXTRAPOLATED_WEIGHT} is measured in kilograms (kg) and includes just the
weight of the relevant \code{SPECIES} included in that row of data, e.g., if
\verb{SPECIES == }r species_norpac``, then \code{EXTRAPOLATED_WEIGHT} is the kg of
Pacific Hake in that haul.
}
}

\section{Assumptions}{
\itemize{
\item sampled hauls have an entry in the \code{SPECIES} column, i.e., hake are
denoted with 206
\item an unsampled haul is represented by one unique record because there no
additional lines of catch for the bycatch species given it was not sampled
}
}

\author{
Kelli F. Johnson
}
