% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weight_at_age-.R
\name{process_weight_at_age_survey}
\alias{process_weight_at_age_survey}
\title{Get the weight-at-age data from the NWFSC server}
\usage{
process_weight_at_age_survey(savedir = hakedata_wd())
}
\arguments{
\item{savedir}{A string specifying the path of interest.}
}
\description{
Get the weight-at-age data from the NWFSC server
}
\author{
Kelli F. Johnson
}
