% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weight_at_age-.R
\name{process_weight_at_age_us}
\alias{process_weight_at_age_us}
\title{Get the weight-at-age data from the commercial fisheries}
\usage{
process_weight_at_age_us(savedir = hakedata_wd())
}
\arguments{
\item{savedir}{A string specifying the path of interest.}
}
\description{
After reading in the saved \code{.Rdat} files from each database, the data
is massaged and anything from 2008 or newer is changed.
}
\author{
Kelli F. Johnson
}
