% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psv_fn}
\alias{psv_fn}
\title{A character string containing the file name for the SS3 psv file
(binary ADMB output needed for the mceval step)}
\format{
A character string
}
\usage{
psv_fn
}
\description{
A character string containing the file name for the SS3 psv file
(binary ADMB output needed for the mceval step)
}
\keyword{datasets}
