% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pull-survey-.R
\name{pull_survey_weight_at_age}
\alias{pull_survey_weight_at_age}
\title{Pull survey weight-at-age data from the NWFSC Network}
\usage{
pull_survey_weight_at_age(dir = "\\\\\\\\nwcfile\\\\fram\\\\Survey.Acoustics")
}
\arguments{
\item{dir}{The directory of interest for the NWFSC Network Acoustic Survey
data. The default is \code{"\\\\\\\\nwcfile\\\\fram\\\\Survey.Acoustics"}.}
}
\description{
Bring both survey years and research years together but as of right now the
only research years available are 2016, 2017, and 2018 because the other
years do not have files on the network that I can grab. Alicia is working on
getting everything into a database but that work is going slow.
}
\details{
This function takes a while because it uses \code{\link[fs:dir_ls]{fs::dir_ls()}} to find the
appropriate file names when looking for the research-year data.
}
