% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_surveyageproportions.R
\name{pull_surveyageproportions}
\alias{pull_surveyageproportions}
\title{Calculate age proportions for the acoustic age-2+ survey}
\usage{
pull_surveyageproportions(year = hake::get_data_yr(), round = options()$digits)
}
\arguments{
\item{year}{An integer specifying the year of interest.
The default uses \code{\link[=get_data_yr]{get_data_yr()}} to get the current year for the data.}

\item{round}{An integer specifying how many digits should be printed.
The default is to print the options for \code{digits} stored in your R session.}
}
\description{
Calculate age proportions for the acoustic age-2+ survey based on
the estimated \strong{abundance} at age extrapolated to the KS stratifications.
This information is produced by the acoustic team and stored on the
network in the \code{kriged_len_age_abundance_table} file.
The original code to generate these proportions was provided by
Allan Hicks and can be found on GitHub
\url{https://github.com/iantaylor-NOAA/hake-data/blob/master/Rcode/AcousticData2.R}.
Particularly lines 170-174 that explains the file location.
The actual code to compute the compositions is pretty trivial,
the difficulty is knowing what file to use.
As mentioned above, abundance is key here rather than biomass.
}
\examples{
proportions <- pull_surveyageproportions(
  year = 2021, round = 2
)
# Values to paste into .dat file
glue::glue_collapse(proportions, sep = " ")
}
