% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quotas}
\alias{quotas}
\title{Quota information for U.S. sectors of Pacific Hake fishery}
\format{
\subsection{\code{who}}{

A data frame with 5 rows and many columns:
\describe{
\item{Fleet}{Names of sectors that receive allocation}
\item{2008,...}{Year-specific allocations}
...
}
}
}
\usage{
quotas
}
\description{
Quota information for U.S. sectors of Pacific Hake fishery
}
\keyword{datasets}
