% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rbeta_ab}
\alias{rbeta_ab}
\alias{dbeta_ab}
\title{The Beta distribution}
\usage{
rbeta_ab(n, prior, sd, min, max)

dbeta_ab(x, prior, sd, min, max)
}
\arguments{
\item{n}{The number of random values you want in the returned vector.}

\item{prior, sd}{The mean and standard deviation of the prior as input into
Stock Synthesis.}

\item{min, max}{The minimum and maximum bound on the parameter as input into
Stock Synthesis.}

\item{x}{Vector of quantiles.}

\item{force}{A logical specifying if the returned values should be forced to
the bounds defined by \verb{[min,max]} or if they should just be soft bounds
used to characterize the shape of the distribution but not actually
limiting, i.e., force the likelihood to be zero for values beyond the
bounds.}
}
\value{
A vector of doubles, where the length of the vector depends on the
length of \code{x} value passed to \code{n}.
\code{dbeta_ab} returns density estimates for each input value and \code{rbeta_ab}
provides random deviates from the Beta distribution.
}
\description{
The Beta distribution
}
\details{
Random generation for the Beta distribution with parameters based on the
mean and standard deviation as well as the ranges as implemented in
CASAL. This formulation is also available on \href{https://en.wikipedia.org/wiki/Beta_distribution#Mean_and_variance}{Wikipedia}.
This function is a direct copy of the code available from \href{https://github.com/chantelwetzel-noaa/XSSS/blob/4dce917ee06d52416b58dd9440796ba1f7357943/R/rbeta_ab_fn.R}{Wetzel and Punt}.
}
\examples{
# A histogram of the 2023 prior used for the base model
hist(
  rbeta_ab(prior = 0.777, sd = 0.113, min = 0.2, max = 1.0),
  freq = FALSE,
  breaks = seq(0.2, 1.0, 0.01)
)
dist <- dbeta_ab(
  x = seq(0.2, 1.0, length = 1e6),
  prior = 0.777,
  sd = 0.113,
  min = 0.20,
  max = 1.0
)
lines(
  x = seq(0.2, 1.0, length = 1e6),
# rescale y-values by mean height * width
  y = exp(-dist) / ((1.0 - 0.2) * mean(exp(dist))),
  type = "l",
  col = 2,
  lwd = 3
)
}
