% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-bookdown-file.R
\name{read_bookdown_file}
\alias{read_bookdown_file}
\title{Read a _bookdown.yml file into a reasonable format which is a list
containing the book filename, the RMD filenames, and the variable
\code{delete_merged_file}}
\usage{
read_bookdown_file(fn = "./_bookdown.yml")
}
\arguments{
\item{fn}{The YML filename}
}
\value{
The list created. Can be \code{NULL} if the regular expressions fail
to match
}
\description{
Read a _bookdown.yml file into a reasonable format which is a list
containing the book filename, the RMD filenames, and the variable
\code{delete_merged_file}
}
