% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recruit_dev_start_yr}
\alias{recruit_dev_start_yr}
\title{The year in which the recruitment deviations start in the model}
\format{
A single numerical value, a year
}
\usage{
recruit_dev_start_yr
}
\description{
The year in which the recruitment deviations start in the model
}
\keyword{datasets}
