% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_bo_linetype}
\alias{refpt_bo_linetype}
\title{The default line type to use for the horizontal line representing B0}
\format{
A character string or numerical value representing an R line type
}
\usage{
refpt_bo_linetype
}
\description{
The default line type to use for the horizontal line representing B0
}
\keyword{datasets}
