% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_bo_linewidth}
\alias{refpt_bo_linewidth}
\title{The default line width to use for the horizontal line representing B0}
\format{
A numerical value
}
\usage{
refpt_bo_linewidth
}
\description{
The default line width to use for the horizontal line representing B0
}
\keyword{datasets}
