% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_lrp_linecolor}
\alias{refpt_lrp_linecolor}
\title{The color of the horizontal line representing the limit reference point}
\format{
A character string or numerical value representing an R color
}
\usage{
refpt_lrp_linecolor
}
\description{
The color of the horizontal line representing the limit reference point
}
\keyword{datasets}
