% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_lrp_linetype}
\alias{refpt_lrp_linetype}
\title{The default line type to use for the horizontal line representing the
limit reference point}
\format{
A character string or numerical value representing an R line type
}
\usage{
refpt_lrp_linetype
}
\description{
The default line type to use for the horizontal line representing the
limit reference point
}
\keyword{datasets}
