% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_lrp_linewidth}
\alias{refpt_lrp_linewidth}
\title{The default line width to use for the horizontal line representing the
limit reference point}
\format{
A numerical value
}
\usage{
refpt_lrp_linewidth
}
\description{
The default line width to use for the horizontal line representing the
limit reference point
}
\keyword{datasets}
