% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_usr_linecolor}
\alias{refpt_usr_linecolor}
\title{The color of the horizontal line representing the upper stock reference}
\format{
A character string or numerical value representing an R color
}
\usage{
refpt_usr_linecolor
}
\description{
The color of the horizontal line representing the upper stock reference
}
\keyword{datasets}
