% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_usr_linetype}
\alias{refpt_usr_linetype}
\title{The default line type to use for the horizontal line representing the
upper stock reference}
\format{
A character string or numerical value representing an R line type
}
\usage{
refpt_usr_linetype
}
\description{
The default line type to use for the horizontal line representing the
upper stock reference
}
\keyword{datasets}
