% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refpt_usr_linewidth}
\alias{refpt_usr_linewidth}
\title{The default line width to use for the horizontal line representing the
upper stock reference}
\format{
A numerical value
}
\usage{
refpt_usr_linewidth
}
\description{
The default line width to use for the horizontal line representing the
upper stock reference
}
\keyword{datasets}
