% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regex_eml_recdevs}
\alias{regex_eml_recdevs}
\title{Regular expression for the "early", "main", and "late" recruitment
deviation parameters (Early_RecrDev, etc.)in SS3 MCMC output. Used
in load_mcmc_vals.R}
\format{
A character string
}
\usage{
regex_eml_recdevs
}
\description{
Regular expression for the "early", "main", and "late" recruitment
deviation parameters (Early_RecrDev, etc.)in SS3 MCMC output. Used
in load_mcmc_vals.R
}
\keyword{datasets}
