% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regex_extra_mcmc_compreport}
\alias{regex_extra_mcmc_compreport}
\title{Regular expression for the CompReport files in the extra MCMC runs.
Used in load_extra_mcmc.R}
\format{
A character string
}
\usage{
regex_extra_mcmc_compreport
}
\description{
Regular expression for the CompReport files in the extra MCMC runs.
Used in load_extra_mcmc.R
}
\keyword{datasets}
