% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regex_main_recdevs}
\alias{regex_main_recdevs}
\title{Regular expression for the "main" recruitment deviation
parameters in SS3 MCMC output. Used in load_mcmc_vals.R}
\format{
A character string
}
\usage{
regex_main_recdevs
}
\description{
Regular expression for the "main" recruitment deviation
parameters in SS3 MCMC output. Used in load_mcmc_vals.R
}
\keyword{datasets}
