% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regex_recdev_early}
\alias{regex_recdev_early}
\title{Regular expression for the "early initial" (Early_InitAge) recruitment
deviation parameters in SS3 MCMC output}
\format{
A character string
}
\usage{
regex_recdev_early
}
\description{
Regular expression for the "early initial" (Early_InitAge) recruitment
deviation parameters in SS3 MCMC output
}
\keyword{datasets}
