% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-forecast-outputs.R
\name{remove_forecast_outputs}
\alias{remove_forecast_outputs}
\title{Remove the \code{outputs} list item from the \code{catch_levels} lists in each year
in the \code{forecasts} list}
\usage{
remove_forecast_outputs(forecasts)
}
\arguments{
\item{forecasts}{A list of years of forecasts, each of which is a list of
catch levels, each of which is a list of various values}
}
\value{
The forecasts list, with the \code{outputs} list items removed
}
\description{
Remove the \code{outputs} list item from the \code{catch_levels} lists in each year
in the \code{forecasts} list
}
\details{
The \code{outputs} list item for each catch level/year combination is set to
\code{NULL} which effectively removes it from the list. This is done because this
data frame can be over 87 million data values (if there are 8,000
posteriors and 726 columns as there are for the hake model)
}
