% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render the Hake assessment document PDF file}
\usage{
render(input = "000-launcher.rmd", output_dir = ".", tex_fn = "hake.tex", ...)
}
\arguments{
\item{input}{The name of the launching \code{rmd} file. This should be the
same as the first \code{rmd} file named in the \verb{_bookdown.yml} file.
See \code{input} argument for \code{\link[bookdown:render_book]{bookdown::render_book()}}}

\item{output_dir}{Which directory to build the PDF in.
See \code{output_dir} argument for \code{\link[bookdown:render_book]{bookdown::render_book()}}}

\item{tex_fn}{The filename of the TeX file generated by
\code{\link[bookdown:render_book]{bookdown::render_book()}} to send to the post processing function.
See \code{fn} argument for \code{\link[=post_process]{post_process()}}}

\item{...}{Arguments to pass to the \code{\link[bookdown:render_book]{bookdown::render_book()}} and
\code{\link[=post_process]{post_process()}} functions}
}
\value{
Nothing
}
\description{
Render the Hake assessment document PDF file
}
