% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pointpath.R
\name{repair_segment_arrow}
\alias{repair_segment_arrow}
\title{Copied from ggplot2 source code file geom-path.R}
\usage{
repair_segment_arrow(arrow, group)
}
\arguments{
\item{arrow}{See \code{\link[ggplot2:geom_path]{ggplot2::geom_path()}}}

\item{group}{See \code{\link[ggplot2:geom_path]{ggplot2::geom_path()}}}
}
\value{
Repaired \code{arrow}
}
\description{
Copied from ggplot2 source code file geom-path.R
}
