% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-replace-dotted-paths.R
\name{replace_dotted_paths}
\alias{replace_dotted_paths}
\title{Replace dotted paths in the vector with their correct full paths with
respect to the current directory and prepend the regular (non-dotted
paths) with the current directory path.}
\usage{
replace_dotted_paths(x)
}
\arguments{
\item{x}{A vector of filenames, with or without leading dotted paths
(\verb{../../}) or (\verb{../../../../../}) or to any level}
}
\value{
The same vector of filenames with the dotted paths replaced by
full paths and the non-dotted paths prepended by the current directory
path
}
\description{
Replace dotted paths in the vector with their correct full paths with
respect to the current directory and prepend the regular (non-dotted
paths) with the current directory path.
}
