% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report_fn}
\alias{report_fn}
\title{A character string containing the file name for the SS3 Report file
(MLE output)}
\format{
A character string
}
\usage{
report_fn
}
\description{
A character string containing the file name for the SS3 Report file
(MLE output)
}
\keyword{datasets}
