% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retropectives_path}
\alias{retropectives_path}
\title{Directory name for storing the retrospectives.}
\format{
A character string
}
\usage{
retropectives_path
}
\description{
Directory name for storing the retrospectives.
}
\keyword{datasets}
