% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retrospective_yrs}
\alias{retrospective_yrs}
\title{The years previous to the end of the time series to strip off for
retrospective analysis}
\format{
A vector of numbers
}
\usage{
retrospective_yrs
}
\description{
The years previous to the end of the time series to strip off for
retrospective analysis
}
\keyword{datasets}
