% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retrospectives_prepend}
\alias{retrospectives_prepend}
\title{Name to be prepended to year for retrospective directory names.}
\format{
A character string
}
\usage{
retrospectives_prepend
}
\description{
Name to be prepended to year for retrospective directory names.
}
\keyword{datasets}
