% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rich.colors.short}
\alias{rich.colors.short}
\title{Create rich colors as RGB strings}
\usage{
rich.colors.short(n, alpha = 1)
}
\arguments{
\item{n}{The number of colors}

\item{alpha}{The transparency for all colors}
}
\value{
A vector of RGB strings representing rich colors
}
\description{
Create rich colors as RGB strings
}
\examples{
rich.colors.short(10)
}
