% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rich-colors-short.R
\name{rich_colors_short}
\alias{rich_colors_short}
\title{Create rich colors as RGB strings}
\usage{
rich_colors_short(n, alpha = 1)
}
\arguments{
\item{n}{The number of colors}

\item{alpha}{The transparency for all colors}
}
\value{
A vector of RGB strings representing rich colors
}
\description{
Create rich colors as RGB strings
}
