% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run catch-levels, forecasting, and retrospectives in one function}
\usage{
run(
  model = NULL,
  model_path = NULL,
  run_catch_levels = FALSE,
  run_forecasts = FALSE,
  run_retrospectives = FALSE,
  ...
)
}
\arguments{
\item{model_path}{The directory the models are located in}

\item{run_catch_levels}{Logical. If \code{TRUE} run catch levels estimation}

\item{run_forecasts}{Logical. If \code{TRUE} run forecasting}

\item{run_retrospectives}{Logical. If \code{TRUE} run retrospectives}

\item{...}{Arguments passed to the \code{\link[=run_ct_levels]{run_ct_levels()}}, \code{\link[=run_forecasts]{run_forecasts()}},
and \code{\link[=run_retrospectives]{run_retrospectives()}} functions}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Run catch-levels, forecasting, and retrospectives in one function
}
\details{
This is a wrapper function for calling \code{\link[=run_ct_levels]{run_ct_levels()}},
\code{\link[=run_forecasts]{run_forecasts()}}, and \code{\link[=run_retrospectives]{run_retrospectives()}} functions.
}
