% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-adnuts.R
\name{run_adnuts}
\alias{run_adnuts}
\title{Run the adnuts MCMC for the model found in the directory given}
\usage{
run_adnuts(
  path,
  num_chains = NULL,
  seed = 42,
  num_samples = 8000,
  num_warmup_samples = 250,
  adapt_delta = 0.95,
  run_extra_mcmc = FALSE,
  hess_step = TRUE,
  fn_exe = ifelse(exists("ss_executable"), ss_executable, "ss3"),
  overwrite = FALSE,
  fn_logfile = "model_output.log",
  input_files = c(ifelse(exists("starter_file_name"), starter_file_name, "starter.ss"),
    ifelse(exists("starter_file_name"), forecast_file_name, "forecast.ss"),
    ifelse(exists("weight_at_age_file_name"), weight_at_age_file_name, "wtatage.ss"),
    ifelse(exists("control_file_name"), control_file_name, "hake_control.ss"),
    ifelse(exists("data_file_name"), data_file_name, "hake_data.ss"))
)
}
\arguments{
\item{path}{Directory where the model files reside}

\item{num_chains}{The number of chains to run in parallel. If `NULL`,
1 less than the number of cores on the machine will be used
1 less than the number of cores on the machine will be used}

\item{seed}{The random seed used to draw the random seeds for each chain}

\item{num_samples}{The number of samples to output}

\item{num_warmup_samples}{The warmup samples (equivalent of burnin)}

\item{adapt_delta}{The target acceptance rate. See [adnuts::sample_nuts()]}

\item{run_extra_mcmc}{If `TRUE`, run SS extra mcmc option which outputs
files into the `sso` subdirectory. If `FALSE`, those files will not be
created and the `posteriors.sso` and `dervied_posteriors.sso` files
will be in the running directory}

\item{fn_exe}{The name of the executable which was built using ADMB}

\item{overwrite}{Logical. If `TRUE`, don't ask user if they want to
overwrite if the directory already exists, just do it}

\item{input_files}{The input files for SS}

\item{check_issues}{Run [adnuts::launch_shinyadmb()] after initial short
run to discover issues. This will stop the function before the main
iterations are done, so you will have to re-run the function again
with this set to `FALSE`.}
}
\value{
Nothing
}
\description{
Run the adnuts MCMC for the model found in the directory given
}
\details{
`path` is the directory in which the MLE will be run, a subdirectory of
this called `mcmc` is where the MCMC will be run using the `NUTS`
algorithm. Inside the `mcmc` directory, several temporary subdirectories
will be created, one for each MCMC chain labeled `chain_*`, AkA CPU
number used in the parallel execution. These will disappear once the
run has completed and the output has been merged.
}
