% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{run_catch_levels}
\alias{run_catch_levels}
\title{A wrapper to run the catch levels determination routines}
\usage{
run_catch_levels(model_path, catch_levels_path = catch_levels_path, ...)
}
\arguments{
\item{model_path}{The model directory name}

\item{...}{Passed to \code{\link[=run_catch_levels_default_hr]{run_catch_levels_default_hr()}}, \code{\link[=run_catch_levels_spr_100]{run_catch_levels_spr_100()}},
and \code{\link[=run_catch_levels_stable_catch]{run_catch_levels_stable_catch()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
A wrapper to run the catch levels determination routines
}
