% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{run_catch_levels_default_hr}
\alias{run_catch_levels_default_hr}
\title{Run the model iteratively reducing the difference between the first and second year projections to
find a stable catch within the the given tolerance}
\usage{
run_catch_levels_default_hr(model, default_hr_path, forecast_yrs, ...)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{...}{}
}
\description{
Run the model iteratively reducing the difference between the first and second year projections to
find a stable catch within the the given tolerance
}
