% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{run_catch_levels_spr_100}
\alias{run_catch_levels_spr_100}
\title{Run the model iteratively zoning in on a catch value that reduces the SPR to 1, within the tolerance given
(\code{catch_levels_spr_tol})}
\usage{
run_catch_levels_spr_100(
  model,
  spr_100_path,
  forecast_yrs,
  catch_levels_spr_tol,
  catch_levels_catch_tol,
  catch_levels_max_iter,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{spr_100_path}{Path where the SPR model files are found}

\item{forecast_yrs}{A vector if the years to forecast for}

\item{catch_levels_spr_tol}{The tolerance to be within 1 for the SPR}

\item{catch_levels_catch_tol}{Catch tolerance. If the upper and lower catch in the algorithm are within this,
it is assumed that the SPR is close enough}

\item{catch_levels_max_iter}{The maximum number of iterations to do}

\item{...}{Not used}
}
\description{
Run the model iteratively zoning in on a catch value that reduces the SPR to 1, within the tolerance given
(\code{catch_levels_spr_tol})
}
