% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{run_catch_levels_stable_catch}
\alias{run_catch_levels_stable_catch}
\title{Run the model iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance}
\usage{
run_catch_levels_stable_catch(
  model,
  stable_catch_path,
  forecast_yrs,
  catch_levels_catch_tol,
  catch_levels_max_iter,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{stable_catch_path}{The path of the stable catch scenario}

\item{forecast_yrs}{A vector of forecast years}

\item{catch_levels_catch_tol}{The tolerance for stopping based on catch
difference}

\item{catch_levels_max_iter}{The maximum number of iterations}

\item{...}{Absorbs other arguments intended for other functions}
}
\description{
Run the model iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance
}
