% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels.R
\name{run_ct_levels}
\alias{run_ct_levels}
\title{A wrapper to run the catch levels determination routines}
\usage{
run_ct_levels(model_path, ct_levels_path = ct_levels_path, ...)
}
\arguments{
\item{model_path}{The model directory name}

\item{ct_levels_path}{The catch levels list, which is a list of lists
of length 3}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
A wrapper to run the catch levels determination routines
}
