% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels-spr-100.R
\name{run_ct_levels_spr_100}
\alias{run_ct_levels_spr_100}
\title{Run the model iteratively zoning in on a catch value that reduces the
SPR to 1, within the tolerance given (\code{ct_levels_spr_tol})}
\usage{
run_ct_levels_spr_100(
  model,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  ss_exe = ss_executable,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{forecast_yrs}{A vector if the years to forecast for}

\item{ss_exe}{The name of the SS3 executable. If run standalone,
this will be \link{ss_executable}. If run from the context of of the \link{bookdown}
document, this will be set as a YAML key/tag}

\item{...}{Not used}
}
\value{
Nothing, the 'forecast.ss' file will have the catch numbers at the
end of the file under 'Catch_or_F'
}
\description{
Run the model iteratively zoning in on a catch value that reduces the
SPR to 1, within the tolerance given (\code{ct_levels_spr_tol})
}
