% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-extra-forecasts.R
\name{run_extra_forecasts}
\alias{run_extra_forecasts}
\title{Run extra forecasting for the model supplied}
\usage{
run_extra_forecasts(
  model_path,
  forecast_yrs,
  catch_levels = list(list(rep(470000, length(forecast_yrs)), "470,000 t", "16-470000"),
    list(rep(510000, length(forecast_yrs)), "510,000 t", "17-510000"), list(rep(6e+05,
    length(forecast_yrs)), "600,000 t", "18-600000"), list(rep(650000,
    length(forecast_yrs)), "650,000 t", "19-650000")),
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=load_ss_files]{load_ss_files()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Run extra forecasting for the model supplied
}
\details{
You must modify the catch-levels list inside this function prior
to running. Start withthe directory number after what is already
in the base model \code{forecasts} directory.

If there is no mcmc component to the model, an error will be
given and the program will be stopped
}
