% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-forecasts-retrospectives.R
\name{run_forecasts_retrospectives}
\alias{run_forecasts_retrospectives}
\title{Run catch-levels, forecasting, and retrospectives in one function}
\usage{
run_forecasts_retrospectives(
  model = NULL,
  model_path = NULL,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  run_catch_levels = FALSE,
  run_forecasts = FALSE,
  run_retrospectives = FALSE,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_path}{The directory the models are located in}

\item{forecast_yrs}{A vector of forecast years}

\item{run_catch_levels}{Logical. If \code{TRUE} run catch levels estimation.
If this is \code{FALSE} and \code{run_forecasts} is \code{TRUE} but the directory
that is supposed to hold the catch-levels output (\link{ct_levels_path}) is
missing, then the catch levels will run anyway, because those value
are needed for forecasting}

\item{run_forecasts}{Logical. If \code{TRUE} run forecasting}

\item{run_retrospectives}{Logical. If \code{TRUE} run retrospectives}

\item{...}{Arguments passed to the \code{\link[=run_ct_levels]{run_ct_levels()}}, \code{\link[=run_forecasts]{run_forecasts()}},
and \code{\link[=run_retrospectives]{run_retrospectives()}} functions}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Run catch-levels, forecasting, and retrospectives in one function
}
\details{
This is a wrapper function for calling \code{\link[=run_ct_levels]{run_ct_levels()}},
\code{\link[=run_forecasts]{run_forecasts()}}, and \code{\link[=run_retrospectives]{run_retrospectives()}} functions.
}
