% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-retrospectives.R
\name{run_retrospectives}
\alias{run_retrospectives}
\title{Runs retrospectives for the given model and for the vector of years given}
\usage{
run_retrospectives(
  model_path,
  retrospective_yrs = NA,
  remove_blocks = FALSE,
  retro_mcmc = TRUE,
  run_extra_mcmc = TRUE,
  fn_exe = ifelse(exists("ss_executable"), ss_executable, "ss3"),
  ...
)
}
\arguments{
\item{model_path}{The path of the model run}

\item{retrospective_yrs}{The years (e.g. 1:6) to run so each of these numbers means that many
years of data removed from the model}

\item{remove_blocks}{If `TRUE`, remove block designs from control file prior to running}

\item{retro_mcmc}{If `TRUE`, run the ADNUTS MCMC in the *mcmc* subdirectory for each
retrospective in addition to the MLE run}

\item{...}{Arguments passed to [load_ss_files()]}

\item{num_samples}{Same as in [run_adnuts()]}

\item{num_warmup_samples}{Same as in [run_adnuts()]}

\item{num_chains}{Same as in [run_adnuts()]}
}
\value{
[base::invisible()]
}
\description{
Runs retrospectives for the given model and for the vector of years given
}
\details{
This will create a *retrospectives* directory in the same directory as the model resides,
create a directory for each retrospective year, copy all model files into each directory,
run the retrospectives, and make a list of the [r4ss::SS_output()] call to each
Warning - This function will completely delete all previous retrospectives that have been run without notice.
}
