% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampling_history_df}
\alias{sampling_history_df}
\title{A data frame containing information on amount of sampling done by year}
\format{
A data frame with 48 rows and 11 variables:
\describe{
\item{Year}{Year of samples}
\item{U.S. Foreign (hauls)}{Number of samples}
\item{U.S. Joint-venture (hauls)}{Number of samples}
\item{U.S. Mothership (hauls)}{Number of samples}
\item{U.S. Combined Mothership Catcher-processor (hauls)}{Number of samples}
\item{U.S. Catcher-processor (hauls)}{Number of samples}
\item{U.S. Shore-based (trips)}{Number of samples}
\item{Canada Foreign (hauls)}{Number of samples}
\item{Canada Joint-venture (hauls)}{Number of samples}
\item{Canada Shoreside (trips)}{Number of samples}
\item{Canada Freezer trawlers (hauls)}{Number of samples}
}
}
\usage{
sampling_history_df
}
\description{
A data frame containing information on amount of sampling done by year
}
\keyword{datasets}
