% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-set-figure-placement.R
\name{set_figure_placement}
\alias{set_figure_placement}
\title{Modify the tex code \code{x} by changing the figure label placement code for the
knitr_label\verb{to the value of}place`}
\usage{
set_figure_placement(x, figs_dir, knitr_label, place)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{figs_dir}{The name of the directory in which the knitr-built
figures reside. Used for matching figure import locations inside the tex
file}

\item{knitr_label}{The knitr chunk label for the chunk which creates the
figure}

\item{place}{One of the following LaTeX placement strings (or a
combination of them):
\itemize{
\item h: Place the float here, i.e., approximately at the same point it occurs
in the source text.
\item t: Position at the top of the page.
\item b: Position at the bottom of the page.
\item p: Put on a special page for floats only.
\item !: Override internal parameters LaTeX uses for determining “good” float
positions.
\item H: Place the float at precisely the location in the LaTeX code. This
requires the float package
}}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify the tex code \code{x} by changing the figure label placement code for the
knitr_label\verb{to the value of}place`
}
