% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-data-frame-for-past-management-plots.R
\name{setup_data_frame_for_past_management_plots}
\alias{setup_data_frame_for_past_management_plots}
\title{Create a data frame of ggplot-format containing a time series of the
management-chosen TAC, assessment estimated TAC, and realized catch, and if
requested the estimated spawning and total biomasses}
\usage{
setup_data_frame_for_past_management_plots(
  model,
  d,
  inc_biomass_ests = FALSE,
  diffs = FALSE,
  yrs = NA,
  ret_tbl = FALSE,
  font_size = 10,
  right_cols_cm = 1.8,
  ...
)
}
\arguments{
\item{d}{The data as read in using \code{\link[readr:read_delim]{readr::read_csv()}} from the file
"catch-targets.csv"}

\item{inc_biomass_ests}{If \code{TRUE}, include a line representing the estimated
biomass (medians)}

\item{diffs}{If \code{TRUE}, show proportion change from year to year, normalized
by the maximum value in \code{yrs}}

\item{yrs}{A vector of years to show on the plot. If \code{NA}, the \code{Year}
column in \code{d} will be used}

\item{ret_tbl}{Logical. If \code{TRUE}, return a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} containing
the outputs instead of the data frame}

\item{font_size}{The table data and header font size in points. Only used
if \code{ret_tbl} is \code{TRUE}}

\item{right_cols_cm}{The number of centimeters wide to make all of the
rightmost columns (all the value columns). Only used if \code{ret_tbl} is \code{TRUE}}

\item{...}{Absorbs arguments not expected to be passed to this function}
}
\value{
A \link{data.frame}
}
\description{
Create a data frame of ggplot-format containing a time series of the
management-chosen TAC, assessment estimated TAC, and realized catch, and if
requested the estimated spawning and total biomasses
}
\details{
if \code{diffs} is \code{TRUE} then proportions from year to year based on 1
being set to the maximum value in each time series will be returned instead
of actual values
}
