% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{show_ss_output}
\alias{show_ss_output}
\title{A character string containing the name of the variable flag for showing
SS3 output messages or not}
\format{
A character string
}
\usage{
show_ss_output
}
\description{
A character string containing the name of the variable flag for showing
SS3 output messages or not
}
\keyword{datasets}
