% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split.prior.info}
\alias{split.prior.info}
\title{Extract priors information from `prior.str``}
\usage{
\method{split}{prior.info}(prior.str, dec.points = 1, first.to.lower = FALSE)
}
\arguments{
\item{prior.str}{A string with the format `Lognormal(2.0,1.01)`}

\item{dec.points}{The number of decimal points to use}

\item{first.to.lower}{Make the first letter of the prior name lower case}
}
\value{
A vector of length 3 with the following format:
`c("Lognormal", 2.0, 1.01)`
}
\description{
Extract priors information from `prior.str``
}
