% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-split-prior-info.R
\name{split_prior_info}
\alias{split_prior_info}
\title{Extract priors information from a string with a particular format into
a vector}
\usage{
split_prior_info(prior_str, dec_points = 1, first_to_lower = FALSE)
}
\arguments{
\item{prior_str}{A string with the format \code{Lognormal(2.0,1.01)}}

\item{dec_points}{The number of decimal points to use}

\item{first_to_lower}{Make the first letter of the prior name lower case}
}
\value{
A vector of length 3 with the following format:
\code{c("Lognormal", 2.0, 1.01)}
}
\description{
Extract priors information from a string with a particular format into
a vector
}
