% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spr_100_path}
\alias{spr_100_path}
\title{Directory name for storing the SPR 100 catch levels}
\format{
A character string
}
\usage{
spr_100_path
}
\description{
Directory name for storing the SPR 100 catch levels
}
\keyword{datasets}
