% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sql_path}
\alias{sql_path}
\title{Directory name for any SQL code files}
\format{
A character string
}
\usage{
sql_path
}
\description{
Directory name for any SQL code files
}
\keyword{datasets}
