% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_executable}
\alias{ss_executable}
\title{A character string containing the name of the SS3 executable
(without extension)}
\format{
A character string
}
\usage{
ss_executable
}
\description{
A character string containing the name of the SS3 executable
(without extension)
}
\keyword{datasets}
