% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_input_files}
\alias{ss_input_files}
\title{A vector of character strings containing the file names needed for inputs
for SS3}
\format{
A character string
}
\usage{
ss_input_files
}
\description{
A vector of character strings containing the file names needed for inputs
for SS3
}
\keyword{datasets}
