% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sso_path}
\alias{sso_path}
\title{Directory name for the mcmc output when MCMC_output_detail in SS3 is 2
which means one report file and CompReport file for each posterior}
\format{
A character string
}
\usage{
sso_path
}
\description{
Directory name for the mcmc output when MCMC_output_detail in SS3 is 2
which means one report file and CompReport file for each posterior
}
\keyword{datasets}
