% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stable_catch_path}
\alias{stable_catch_path}
\title{Directory name for storing the stable catch catch levels}
\format{
A character string
}
\usage{
stable_catch_path
}
\description{
Directory name for storing the stable catch catch levels
}
\keyword{datasets}
