% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{start_yr}
\alias{start_yr}
\title{The starting year for the model estimates}
\format{
A single numerical value, a year
}
\usage{
start_yr
}
\description{
The starting year for the model estimates
}
\keyword{datasets}
