% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{start_yr_age_comps}
\alias{start_yr_age_comps}
\title{The starting year for the age composition data in the model}
\format{
A single numerical value, a year
}
\usage{
start_yr_age_comps
}
\description{
The starting year for the age composition data in the model
}
\keyword{datasets}
