% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starter_fn}
\alias{starter_fn}
\title{A character string containing the file name for the SS3 starter file}
\format{
A character string
}
\usage{
starter_fn
}
\description{
A character string containing the file name for the SS3 starter file
}
\keyword{datasets}
