% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_df}
\alias{states_df}
\title{A data frame containing information about the states and provinces used
in making the overview map at the beginning of the figures section}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{lat}{Latitude to place label}
\item{lon}{Longitude to place label}
\item{name}{Name of the port}
}
}
\usage{
states_df
}
\description{
A data frame containing information about the states and provinces used
in making the overview map at the beginning of the figures section
}
\keyword{datasets}
