% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop the program without issuing an error message}
\usage{
stop_quietly(...)
}
\value{
Nothing
}
\description{
Stop the program without issuing an error message
}
