% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str-comma-sep.R
\name{str_comma_sep}
\alias{str_comma_sep}
\title{Create a string in a comma-separated format from a vector}
\usage{
str_comma_sep(v, and_on_last = TRUE)
}
\arguments{
\item{v}{The vector of values}
}
\value{
A single string
}
\description{
Create a string in a comma-separated format from a vector
}
