% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip.columns}
\alias{strip.columns}
\title{Remove values from a named vector based on names}
\usage{
strip.columns(vec, names)
}
\arguments{
\item{vec}{A named vector}

\item{names}{A vector of names to remove from the vector `vec`}
}
\value{
The vector `vec` with items removed
}
\description{
Remove values from a named vector based on names
}
