% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surv_yrs}
\alias{surv_yrs}
\title{A vector of years in which the acoustic survey took place and there
are data in the model for}
\format{
A single numerical value, a year
}
\usage{
surv_yrs
}
\description{
A vector of years in which the acoustic survey took place and there
are data in the model for
}
\keyword{datasets}
