% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_end_yr}
\alias{survey_end_yr}
\title{The last year for the acoustic survey in the model}
\format{
A single numerical value, a year
}
\usage{
survey_end_yr
}
\description{
The last year for the acoustic survey in the model
}
\keyword{datasets}
