% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_start_yr}
\alias{survey_start_yr}
\title{The starting year for the acoustic survey in the model}
\format{
A single numerical value, a year
}
\usage{
survey_start_yr
}
\description{
The starting year for the acoustic survey in the model
}
\keyword{datasets}
