% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-system.R
\name{system_}
\alias{system_}
\title{Call \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}} depending on the Operating System}
\usage{
system_(...)
}
\arguments{
\item{...}{Arguments to pass to either \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}}}
}
\value{
The output from the command function called
}
\description{
Call \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}} depending on the Operating System
}
