% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-age-data.R
\name{table_age_data}
\alias{table_age_data}
\title{Creates a table containing single-fleet age proportions which are
input into the model}
\usage{
table_age_data(
  fleet = c("can_ft", "can_ss", "can_jv", "us_cp", "us_ms", "us_sb"),
  start_yr,
  end_yr,
  digits = 2,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{fleet}{Use codes of \code{can_ft} = Canada Freezer trawlers, \code{can_ss} =
Canada Shoreside, \code{can_jv} = Canada Joint venture, \code{us_cp} = U.S. Catcher-
processor, \code{us_ms} = U.S. Mothership, or \code{us_sb} = U.S. Shoreside}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Creates a table containing single-fleet age proportions which are
input into the model
}
