% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assessment-changes.R
\name{table_assessment_changes}
\alias{table_assessment_changes}
\title{Title}
\usage{
table_assessment_changes(
  d,
  start_yr,
  end_yr,
  digits = 3,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{d}{A data frame containing the ovary sampling information  as read
in from the file \code{ovary-samples.csv}}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{digits}{Number of decimal points to show in values in the table}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Title
}
