% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-biomass.R, R/table-param-est-bounds.R
\name{table_biomass}
\alias{table_biomass}
\alias{table_param_est_bounds}
\title{Create a table of spawning biomass and relative spawning biomass
medians and quantiles}
\usage{
table_biomass(
  model,
  start_yr,
  end_yr,
  digits = 1,
  scale = 1000,
  font_size = 8,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)

table_param_est_bounds(
  model,
  start_rec_dev_yr,
  end_rec_dev_yr,
  section_sep_lines = FALSE,
  digits = 2,
  font_size = 8,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ret_df = FALSE,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{digits}{Number of decimal places to round to}

\item{scale}{A scale factor to divide the biomass by}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{start_rec_dev_yr}{First year of estimated recruitment devs}

\item{end_rec_dev_yr}{Last year of estimated recruitment devs}

\item{section_sep_lines}{Logical. If \code{TRUE}, place a line under the
sections in each cell as a way to separate them vertically from the
years above and below}

\item{ret_df}{If \code{TRUE}, return a data frame of the results, if \code{FALSE},
return a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} table}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object

Either a \code{\link[tibble:tibble]{tibble::tibble()}} or \code{\link[kableExtra:kbl]{kableExtra::kbl()}} object
}
\description{
Create a table of spawning biomass and relative spawning biomass
medians and quantiles

Creates a table summarizing lead parameters
}
