% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-can-age-data.R
\name{table_can_age_data}
\alias{table_can_age_data}
\title{Creates a table containing the U.S. age data which is input into the model}
\usage{
table_can_age_data(
  lst,
  d_num_aged,
  fleet = c("ft", "ss", "jv"),
  start_yr,
  end_yr,
  digits = 2,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{lst}{A list of data frames of the age data as read in from the files
\verb{us-*-age-data.csv} found in the data directory. The list has one data frame
per fleet}

\item{d_num_aged}{A data frame of number of fish aged as read in from
the files \verb{can-*-num-fish-aged.csv} found in the data directory}

\item{fleet}{Use codes of \code{ft} = Freezer trawlers, \code{ss} = Shoreside, and
\code{jv} = Joint venture}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Creates a table containing the U.S. age data which is input into the model
}
