% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of total catch}
\usage{
table_catch(
  ct,
  country = c("both", "can", "us"),
  start_yr,
  end_yr,
  inc_foreign_jv = FALSE,
  col_width = 0.35,
  font_size = 9,
  ...
)
}
\arguments{
\item{ct}{Output from \code{\link[=load_catches]{load_catches()}}}

\item{country}{One of "both", "can", or "us"}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{col_width}{Width in inches for all the columns}

\item{font_size}{The table data and geader font size in points}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of total catch
}
