% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-cohort.R
\name{table_cohort}
\alias{table_cohort}
\title{Create a table of cohort biomass and natural mortality as they pass
through their life from year to year}
\usage{
table_cohort(
  model,
  cohorts,
  cohort_italics = TRUE,
  cohort_bold = TRUE,
  cohort_underline = TRUE,
  cohort_line_above = TRUE,
  cohort_line_below = TRUE,
  reverse_cohorts = FALSE,
  digits = 1,
  csv_dir = here::here("doc", out_csv_path),
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{cohorts}{A vector of cohorts (years of birth) to use}

\item{cohort_italics}{Logical. If \code{TRUE}, make the cohort header lines
italicized}

\item{cohort_bold}{Logical. If \code{TRUE}, make the cohort header lines
boldface}

\item{cohort_underline}{Logical. If \code{TRUE}, make the cohort header lines
underlined}

\item{cohort_line_above}{Logical. If \code{TRUE}, place a horizontal line above
cohort header lines}

\item{cohort_line_below}{Logical. If \code{TRUE}, place a horizontal line below
cohort header lines}

\item{reverse_cohorts}{Logical. If \code{TRUE}, show the cohorts in the table in
descending order, with the most recent cohort at the top of the table}

\item{digits}{Number of decimal points to show in values in the table}

\item{csv_dir}{Directory for CSV output}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of cohort biomass and natural mortality as they pass
through their life from year to year
}
