% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-decision-original.R
\name{table_decision_original}
\alias{table_decision_original}
\title{Create the Executive Summary decision tables. \code{rows_to_label} and
\code{row_labels} need changing each year.}
\usage{
table_decision_original(
  model,
  font_size = 9,
  space_size = 10,
  type = "biomass",
  placement = "H",
  forecast_inds = seq_along(model$forecasts[[length(model$forecasts)]]),
  rows_to_show = NULL,
  rows_to_label = c("e", "f", "h", "j", "l", "m", "n", "o"),
  row_labels = list(c("10\\\\\%", "reduction"), c(assess_yr - 1, "catch"), c("10\\\\\%",
    "reduction"), c("10\\\\\%", "reduction"), c(assess_yr - 1, "TAC"), c("FI=",
    "100\\\\\%"), c("default", "HR"), c(paste0("C", assess_yr, "="), paste0("C",
    assess_yr + 1))),
  ...
)
}
\arguments{
\item{model}{An mcmc run, \code{model$forecasts} must be populated from
\code{\link[=run_forecasts]{run_forecasts()}}}

\item{font_size}{Size of the font for the table}

\item{space_size}{Size of the vertical spaces for the table}

\item{type}{Type to build. \code{biomass} or \code{spr}}

\item{placement}{Latex code for placement of table}

\item{forecast_inds}{Indices of forecast catch levels to show. By
default all catch levels will be shown}

\item{rows_to_show}{A vector of letters of rows to show. If \code{NULL}, show
all rows}

\item{rows_to_label}{A vector of letters representing rows to add
custom labels to}

\item{row_labels}{A list of length two vectors with text to place in the
empty}

\item{...}{arguments passed to \code{\link[xtable:xtable]{xtable::xtable()}}}
}
\value{
The latex code needed to build the table
}
\description{
Create the Executive Summary decision tables. \code{rows_to_label} and
\code{row_labels} need changing each year.
}
