% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-fishing-intensity.R
\name{table_fishing_intensity}
\alias{table_fishing_intensity}
\title{Create a table of fishing intensity and exploitation fraction medians
and quantiles}
\usage{
table_fishing_intensity(
  model,
  start_yr,
  end_yr,
  digits = 1,
  font_size = 8,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{digits}{Number of decimal places to round recruitment values to}

\item{font_size}{The table data and header font size in points}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of fishing intensity and exploitation fraction medians
and quantiles
}
