% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-fore-age-props.R
\name{table_fore_age_props}
\alias{table_fore_age_props}
\title{Creates a table showing the forecast age comp percentages by number and
weight for select cohorts}
\usage{
table_fore_age_props(
  props_bynum,
  props_bywt,
  ages_filt = c(2:8, 10, 14),
  font_size = 8
)
}
\arguments{
\item{props_bynum}{A named vector of percentages by number to report. The
names must be the ages}

\item{props_bywt}{A named vector of percentages by weight to report. The
names must be the ages}

\item{ages_filt}{A vector of the ages to show in the table}

\item{font_size}{The font size for the text in the table}
}
\value{
A \code{\link[kableExtra:kbl]{kableExtra::kbl()}} table
}
\description{
Creates a table showing the forecast age comp percentages by number and
weight for select cohorts
}
