% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-input-age-data.R
\name{table_input_age_data}
\alias{table_input_age_data}
\title{Creates a table containing the age data which are input into the model}
\usage{
table_input_age_data(
  model,
  fleet = c("fishery", "survey"),
  start_yr,
  end_yr,
  digits = 2,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{fleet}{Use codes of \code{fishery} = Commercial fishery and \code{survey} =
Acoustic age 2+ survey}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Creates a table containing the age data which are input into the model
}
