% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-landings-tac.R
\name{table_landings_tac}
\alias{table_landings_tac}
\title{Create a summary table of landings and TACS}
\usage{
table_landings_tac(d, start_yr, end_yr, font_size = 8, ...)
}
\arguments{
\item{d}{Data frame as in the file \code{landing-tac-histroy.csv}}

\item{start_yr}{Start year for the table}

\item{end_yr}{End year for the table. If past data range, the last year
in \code{d} will be used}

\item{font_size}{The table data and header font size in points}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a summary table of landings and TACS
}
