% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-median-posteriors.R
\name{table_median_posteriors}
\alias{table_median_posteriors}
\title{Make a latex table of medians with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.}
\usage{
table_median_posteriors(
  model,
  start_yr = NULL,
  end_yr = NULL,
  digits = 1,
  scale = 1000,
  cohort_italics = TRUE,
  cohort_bold = TRUE,
  cohort_line_above = TRUE,
  cohort_line_below = TRUE,
  reverse_cohorts = FALSE,
  csv_dir = here::here("doc", "out-csv"),
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{digits}{Number of decimal points to show in values in the table}

\item{cohort_italics}{Logical. If \code{TRUE}, make the cohort header lines
italicized}

\item{cohort_bold}{Logical. If \code{TRUE}, make the cohort header lines
boldface}

\item{cohort_line_above}{Logical. If \code{TRUE}, place a horizontal line above
cohort header lines}

\item{cohort_line_below}{Logical. If \code{TRUE}, place a horizontal line below
cohort header lines}

\item{reverse_cohorts}{Logical. If \code{TRUE}, show the cohorts in the table in
descending order, with the most recent cohort at the top of the table}

\item{csv_dir}{Directory for CSV output}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{cohorts}{A vector of cohorts (years of birth) to use}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Make a latex table of medians with columns for female spawning biomass,
relative spawning biomass, total biomass, age 2+ biomass, age-0 recruits, relative
fishing intensity, and exploitation fraction. Also output a CSV file with the table contents.
}
