% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est.R
\name{table_param_est}
\alias{table_param_est}
\title{Returns an \code{\link[xtable:xtable]{xtable::xtable()}} in the proper format for the MCMC posterior
parameter estimates for all the models in the list, one for each column}
\usage{
table_param_est(
  models,
  model_nms,
  end_yr,
  section_italics = TRUE,
  section_bold = TRUE,
  section_underline = TRUE,
  section_line_above = FALSE,
  section_line_below = TRUE,
  digits = 3,
  rec_yrs = NA,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{models}{A list of models which contain the MCMC output}

\item{model_nms}{A vector of names of the same length as the number of
models in the models list}

\item{end_yr}{The last year to include}

\item{section_italics}{Logical. If \code{TRUE}, make the section header lines
italicized}

\item{section_bold}{Logical. If \code{TRUE}, make the section header lines
boldface}

\item{section_underline}{Logical. If \code{TRUE}, make the section header lines
underlined}

\item{section_line_above}{Logical. If \code{TRUE}, place a horizontal line above
section header lines}

\item{section_line_below}{Logical. If \code{TRUE}, place a horizontal line below
section header lines}

\item{digits}{Number of decimal points for the estimates}

\item{rec_yrs}{A vector of integers supplying the years for which you
want estimates of recruitment}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Returns an \code{\link[xtable:xtable]{xtable::xtable()}} in the proper format for the MCMC posterior
parameter estimates for all the models in the list, one for each column
}
