% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est-bounds.R
\name{table_param_est_bounds}
\alias{table_param_est_bounds}
\title{Creates a table summarizing lead parameters}
\usage{
table_param_est_bounds(
  model,
  start.rec.dev.yr,
  end.rec.dev.yr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  return.xtable = TRUE
)
}
\arguments{
\item{model}{The model to use for the table}

\item{start.rec.dev.yr}{First year of estimated recruitment devs}

\item{end.rec.dev.yr}{Last year of estimated recruitment devs}

\item{digits}{Number of decimal points}

\item{xcaption}{Caption for the table}

\item{xlabel}{Latex label reference in the main doc}

\item{font.size}{Point size of font}

\item{space.size}{Vertical size between rows in the table}

\item{return.xtable}{If \code{TRUE}, return the latex-formatted table code. If \code{FALSE}
return the data frame instead}
}
\value{
Either a data frame or \code{\link[xtable:xtable]{xtable::xtable()}} object
}
\description{
Creates a table summarizing lead parameters
}
