% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est-last-year-hardcoded.R
\name{table_param_est_last_year_hardcoded}
\alias{table_param_est_last_year_hardcoded}
\title{Returns a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} containing a comparison of parameter
estimates for 1 model compared to hard coded values for the base
model in 2023. This function is meant to be destroyed after the
2024 assessment season. It is used for table 26 only because we
could not locate the output files to reproduce the table.}
\usage{
table_param_est_last_year_hardcoded(
  d = NULL,
  model = NULL,
  model_nms = NULL,
  show_loglike = FALSE,
  section_row_inds = c(1, 11, 20, 21 + length(large_cohorts)),
  section_row_headers = c("Parameters", "Derived Quantities",
    paste0("Reference Points based on ", fspr_40_bold_for_latex_table),
    "Negative log likelihoods"),
  section_italics = TRUE,
  section_bold = TRUE,
  section_underline = TRUE,
  section_line_above = FALSE,
  section_line_below = TRUE,
  digits = 3,
  right_cols_cm = 1.8,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  longtable = FALSE,
  caption = NULL,
  ...
)
}
\arguments{
\item{model}{A list of models which contain the MCMC output}

\item{model_nms}{A vector of names of the same length as the number of
models in the models list}

\item{show_loglike}{Logical. If \code{TRUE} show the negative log likelihood
values in the table}

\item{section_italics}{Logical. If \code{TRUE}, make the section header lines
italicized}

\item{section_bold}{Logical. If \code{TRUE}, make the section header lines
boldface}

\item{section_underline}{Logical. If \code{TRUE}, make the section header lines
underlined}

\item{section_line_above}{Logical. If \code{TRUE}, place a horizontal line above
section header lines}

\item{section_line_below}{Logical. If \code{TRUE}, place a horizontal line below
section header lines}

\item{digits}{Number of decimal points for the estimates}

\item{right_cols_cm}{The number of centimeters wide to make all of the
rightmost columns (all the value columns)}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{longtable}{Logical. Passed to \code{\link[kableExtra:kbl]{kableExtra::kbl()}}}

\item{caption}{The table caption, or \code{NULL} for none. Passed to
\code{\link[kableExtra:kbl]{kableExtra::kbl()}}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Returns a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} containing a comparison of parameter
estimates for 1 model compared to hard coded values for the base
model in 2023. This function is meant to be destroyed after the
2024 assessment season. It is used for table 26 only because we
could not locate the output files to reproduce the table.
}
