% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-parameters-long-summary.R
\name{table_parameters_long_summary}
\alias{table_parameters_long_summary}
\title{Title}
\usage{
table_parameters_long_summary(
  model,
  posterior_regex = key_posteriors,
  digits = 4,
  font_size = 12
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{posterior_regex}{A list of regular expressions to match key
posteriors. See the package data item \link{key_posteriors}}

\item{digits}{The number of decimal places to report in the table}

\item{font_size}{The size of the font for all text in the table}
}
\value{
A \code{\link[kableExtra:kbl]{kableExtra::kbl()}} table object
}
\description{
Title
}
