% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-reference-points.R
\name{table_reference_points}
\alias{table_reference_points}
\title{Creates a table containing reference point estimates}
\usage{
table_reference_points(model, font_size = 10, ...)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{font_size}{Size of the font in points for the table data}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
a \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Creates a table containing reference point estimates
}
