% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-risk.R
\name{table_risk}
\alias{table_risk}
\title{Creates LaTeX code to make a probability of risk table with various
probabilities of things happening with the stock}
\usage{
table_risk(model, forecast_yrs, index = 1, digits = 0, font_size = 9, ...)
}
\arguments{
\item{model}{A model from this project}

\item{forecast_yrs}{A vector of forecast years}

\item{index}{Index for which forecast year data to use. e.g. 1 = second
forecast year compared to the first. If there were N forecast years, this
can be from 1 to N-1}

\item{digits}{The number of decimal places to print in the table output}
}
\value{
LaTeX code to render the table
}
\description{
Creates LaTeX code to make a probability of risk table with various
probabilities of things happening with the stock
}
