% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sampling-by-month.R
\name{table_sampling_by_month}
\alias{table_sampling_by_month}
\title{Extract the number of samples by fleet and month for a given year}
\usage{
table_sampling_by_month(
  d,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  right_cols_cm = 1.5,
  ...
)
}
\arguments{
\item{d}{A data frame, as one of the list elements of the list output
by \code{\link[=canada_extract_sampling_by_month]{canada_extract_sampling_by_month()}}}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{right_cols_cm}{The number of centimeters wide to make all of the
rightmost columns (all the value columns)}

\item{...}{}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A list of the same length as \code{fleet_lst} of data frames, explained
in details above
}
\description{
Extract the number of samples by fleet and month for a given year
}
