% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sampling-history.R
\name{table_sampling_history}
\alias{table_sampling_history}
\title{Title}
\usage{
table_sampling_history(
  d,
  start_yr,
  end_yr,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{d}{Data frame as in the file \code{fishery-sampling-histroy.csv}}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Title
}
