% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-survey-by-country.R
\name{table_survey_by_country}
\alias{table_survey_by_country}
\title{Create a table of values of the age 2+ survey biomass estimates by country}
\usage{
table_survey_by_country(
  d,
  digits = 3,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ret_df = FALSE,
  ...
)
}
\arguments{
\item{d}{A data frame containing the survey history as read in from
the file \code{survey-by-country.csv}}

\item{digits}{Number of decimal places}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{ret_df}{Logical. If \code{TRUE}, return a data frame of the table, If
\code{FALSE}, return a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} object}

\item{...}{Arguments passed to \code{\link[kableExtra:kbl]{kableExtra::kbl()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of values of the age 2+ survey biomass estimates by country
}
