% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-survey-history.R
\name{table_survey_history}
\alias{table_survey_history}
\title{Create a table of the acoustic survey history}
\usage{
table_survey_history(
  model,
  d,
  digits = 3,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{d}{A data frame containing the survey history as read in from
the file \code{survey-history.csv}}

\item{digits}{Number of decimal places}

\item{xcaption}{Caption to appear in the calling document}

\item{xlabel}{The label used to reference the table in latex}

\item{font.size}{Size of the font for the table}

\item{space.size}{Size of the vertical spaces for the table}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of the acoustic survey history
}
\details{
The vessel names need to be fixed. They are seperated by spaces, and may
or may not have dashes in their names. The dashes will be replaced with
spaces, and the spaces will be replaced by newlines in the output
}
