% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-theme-axis-text.R
\name{theme_axis_text}
\alias{theme_axis_text}
\title{Add a theme element to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to standardize the
tick and titla labels on a plot}
\usage{
theme_axis_text(
  g,
  face = "plain",
  angle_title_x = 0,
  angle_title_y = 90,
  vjust_tick_x = -1,
  vjust_title_x = -1,
  vjust_tick_y = 0.5,
  vjust_title_y = 0,
  hjust_tick_x = 0.5,
  hjust_title_x = 0.5,
  hjust_tick_y = 0.5,
  hjust_title_y = 0.5,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color
)
}
\arguments{
\item{g}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object}

\item{face}{Font face. one of "plain", "bold", or "italics"}

\item{angle_title_x}{The angle to set the x-axis title to}

\item{angle_title_y}{The angle to set the y-axis title to}

\item{vjust_tick_x}{The amount of vertical adjustment to the x-axis tick
labels}

\item{vjust_title_x}{The amount of vertical adjustment to the x-axis title}

\item{vjust_tick_y}{The amount of vertical adjustment to the x-axis tick
labels}

\item{vjust_title_y}{The amount of vertical adjustment to the y-axis title}

\item{hjust_tick_x}{The amount of horizontal adjustment to the x-axis tick
labels}

\item{hjust_title_x}{The amount of horizontal adjustment to the x-axis title}

\item{hjust_tick_y}{The amount of horizontal adjustment to the x-axis tick
labels}

\item{hjust_title_y}{The amount of horizontal adjustment to the y-axis title}

\item{ax_title_font_size}{Size of the font for the x and y-axis labels}

\item{ax_tick_font_size}{Size of the font for the x and y-axis tick labels}

\item{ax_label_color}{Color of the font for the x and y-axis tick and
title labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Add a theme element to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to standardize the
tick and titla labels on a plot
}
