% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{top.coh}
\alias{top.coh}
\title{Create text describing the top \code{num.cohorts} cohorts by year and
percentage as a sentence}
\usage{
top.coh(
  model = NULL,
  yr = NA,
  num.cohorts = 3,
  decimals = 0,
  cap = TRUE,
  spec.yr = NA,
  use.catage = FALSE,
  fleet = 1
)
}
\arguments{
\item{model}{The model as returned from \code{\link[=load_ss_files]{load_ss_files()}}}

\item{yr}{The year the cohort was born}

\item{num.cohorts}{The number of cohorts to include in the sentence}

\item{decimals}{The number of decimal points to use}

\item{cap}{Logical. Capitalize the first word in the sentence?}

\item{spec.yr}{If supplied, the percentage of catch that this cohort made
to the \code{yr} catch will be returned}

\item{use.catage}{If TRUE, use the \verb{model$catage`` object which are the estimates. If }FALSE\verb{, use the }model$dat$agecomp` object which are the
input data}

\item{fleet}{A integer value allowing the selection of a given fleet, where
for Pacific hake, \code{fleet = 1}, the default, selects the fishery data.}
}
\value{
Text describing the top \code{num.cohorts} cohorts by year and
percentage as a sentence
}
\description{
Create text describing the top \code{num.cohorts} cohorts by year and
percentage as a sentence
}
\details{
top.coh(base_model, 2018, 2) produces:
"The 2018 cohort was the largest (29\\\%), followed by the 2010
cohort (27\\\%)"
}
