% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-top-coh.R
\name{top_coh}
\alias{top_coh}
\title{Create text describing the top \code{num_cohorts} cohorts by year and
percentage as a sentence}
\usage{
top_coh(
  model = NULL,
  yr = year(Sys.Date()),
  num_cohorts = 3,
  decimals = 0,
  cap = TRUE,
  use_catage = FALSE,
  fleet = 1,
  ret_cohort = NA
)
}
\arguments{
\item{model}{The model as returned from \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yr}{the year to calculate the cohort catch percentages for}

\item{num_cohorts}{The top \code{num_cohorts} cohorts will be mentioned in
the sentence}

\item{decimals}{The number of decimal points to use}

\item{cap}{Logical. Capitalize the first word in the sentence?}

\item{use_catage}{Logical. If \code{TRUE}, use the
\code{model$extra_mcmc$catage_median} object which are the estimates. If
\code{FALSE}, use the \code{model$dat$agecomp} object which are the input data}

\item{fleet}{A integer value allowing the selection of a given fleet, where
\code{fleet = 1}, the default, selects the fishery data. This is only used for
the case where \code{use_catage} is \code{FALSE}. If \code{use_catage} is \code{TRUE}, \code{fleet}
will be ignored and fishery data are used}

\item{ret_cohort}{If not \code{NA}, filter out this cohort from the result and
return the proportion for it for year \code{yr}}
}
\value{
Text describing the top \code{num_cohorts} cohorts by year and
percentage as a sentence
}
\description{
Create text describing the top \code{num_cohorts} cohorts by year and
percentage as a sentence
}
\details{
top_coh(base_model, 2020, 2022, 2) produces:
"The 2020 cohort was the largest (33\\\%), followed by the 2016 cohort
(24\\\%)"
}
