% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_linegap}
\alias{ts_linegap}
\title{The gap between the lines connecting the median points in the time
series plots}
\format{
A numerical value
}
\usage{
ts_linegap
}
\description{
The gap between the lines connecting the median points in the time
series plots
}
\keyword{datasets}
