% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_linewidth}
\alias{ts_linewidth}
\title{The line width to use for the line connecting the median points in the
time series plots}
\format{
A numerical value
}
\usage{
ts_linewidth
}
\description{
The line width to use for the line connecting the median points in the
time series plots
}
\keyword{datasets}
