% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_pointshape}
\alias{ts_pointshape}
\title{The point shape to use for the medians in the time series plots}
\format{
A numerical value representing the point shape (R standard point
types)
}
\usage{
ts_pointshape
}
\description{
The point shape to use for the medians in the time series plots
}
\keyword{datasets}
