% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_pointsize}
\alias{ts_pointsize}
\title{The point size to use for the medians in the time series plots}
\format{
A numerical value
}
\usage{
ts_pointsize
}
\description{
The point size to use for the medians in the time series plots
}
\keyword{datasets}
