% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_pointstroke}
\alias{ts_pointstroke}
\title{The point stroke (boldness) to use for the points representing the median
points in the time series plots}
\format{
A numerical value
}
\usage{
ts_pointstroke
}
\description{
The point stroke (boldness) to use for the points representing the median
points in the time series plots
}
\keyword{datasets}
