% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_ribbon_alpha}
\alias{ts_ribbon_alpha}
\title{The transparency to use by default to fill time series CI ribbons in the
plots in the document}
\format{
A numerical value between 0 and 1
}
\usage{
ts_ribbon_alpha
}
\description{
The transparency to use by default to fill time series CI ribbons in the
plots in the document
}
\keyword{datasets}
