% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_ribbon_fill}
\alias{ts_ribbon_fill}
\title{The color to use by default to fill time series CI ribbons in the plots
in the document}
\format{
A character string or numerical value representing an official
color in R
}
\usage{
ts_ribbon_fill
}
\description{
The color to use by default to fill time series CI ribbons in the plots
in the document
}
\keyword{datasets}
