% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_ribbon_linetype}
\alias{ts_ribbon_linetype}
\title{The line type for the edges of the CI ribbon for time series CI ribbons}
\format{
A numerical value or character string representing line type for
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\usage{
ts_ribbon_linetype
}
\description{
The line type for the edges of the CI ribbon for time series CI ribbons
}
\keyword{datasets}
