% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_linetype}
\alias{ts_single_model_linetype}
\title{The line type to use for the line connecting the median points in the
time series plots where there is only one model plotted}
\format{
A numerical value
}
\usage{
ts_single_model_linetype
}
\description{
The line type to use for the line connecting the median points in the
time series plots where there is only one model plotted
}
\keyword{datasets}
