% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_linewidth}
\alias{ts_single_model_linewidth}
\title{The line width to use for the line connecting the median points in the
time series plots where there is only one model plotted}
\format{
A numerical value
}
\usage{
ts_single_model_linewidth
}
\description{
The line width to use for the line connecting the median points in the
time series plots where there is only one model plotted
}
\keyword{datasets}
