% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_pointcolor}
\alias{ts_single_model_pointcolor}
\title{The point color to use for the median points in the time series plots,
when there is only one model plotted}
\format{
A numerical value or R color string
}
\usage{
ts_single_model_pointcolor
}
\description{
The point color to use for the median points in the time series plots,
when there is only one model plotted
}
\keyword{datasets}
