% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_pointstroke}
\alias{ts_single_model_pointstroke}
\title{The point stroke (boldness) to use for the points representing the median
points in the time series plots, when there is only one model plotted}
\format{
A numerical value
}
\usage{
ts_single_model_pointstroke
}
\description{
The point stroke (boldness) to use for the points representing the median
points in the time series plots, when there is only one model plotted
}
\keyword{datasets}
