% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_ribbon_fill}
\alias{ts_single_model_ribbon_fill}
\title{The color to use to fill time series CI ribbons for plots where there
is only one model plotted}
\format{
A character string or numerical value representing an R color
}
\usage{
ts_single_model_ribbon_fill
}
\description{
The color to use to fill time series CI ribbons for plots where there
is only one model plotted
}
\keyword{datasets}
