% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_single_model_ribbon_linetype}
\alias{ts_single_model_ribbon_linetype}
\title{The line type to use for the borders on the CI ribbons for plots where
there is only one model plotted}
\format{
A character string or numerical value representing a line type
}
\usage{
ts_single_model_ribbon_linetype
}
\description{
The line type to use for the borders on the CI ribbons for plots where
there is only one model plotted
}
\keyword{datasets}
