% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unfished_eq_yr}
\alias{unfished_eq_yr}
\title{The year in which unfished equilibrium is set}
\format{
A single numerical value, a year
}
\usage{
unfished_eq_yr
}
\description{
The year in which unfished equilibrium is set
}
\keyword{datasets}
