% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-update-ss3-ageing-error.R
\name{update_ss3_ageing_error}
\alias{update_ss3_ageing_error}
\title{Insert ageing error matrix if needed}
\usage{
update_ss3_ageing_error(data_list)
}
\arguments{
\item{data_list}{A list read in by \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.}
}
\value{
A list with the same structure as the list that was input.
}
\description{
Determine if an additional ageing error matrix is needed and if so, then
add it to the list.
}
\author{
Kelli F. Johnson
}
