% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-update-ss3-catch.R
\name{update_ss3_catch}
\alias{update_ss3_catch}
\title{Update the catches in a Stock Synthesis file}
\usage{
update_ss3_catch(file_ss3)
}
\arguments{
\item{file_ss3}{A file path to the Stock Synthesis data file that you want to
integrate catches into.}
}
\value{
A list containing the Stock Synthesis data file is returned.
}
\description{
Update the catches in a Stock Synthesis file
}
\details{
Information from \code{landings-tac-history.csv} is summed and integrated into
the catch portion of a Stock Synthesis data file. The terminal year of
data becomes the same as the terminal year of the catch time series.
}
\author{
Kelli F. Johnson
}
