% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_cp_age_props_fn}
\alias{us_cp_age_props_fn}
\title{The filename for the Catcher-processor fleet age proportions by year
data file}
\format{
A character string
}
\usage{
us_cp_age_props_fn
}
\description{
The filename for the Catcher-processor fleet age proportions by year
data file
}
\keyword{datasets}
