% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-us-extract-sampling-by-month.r
\name{us_extract_sampling_by_month}
\alias{us_extract_sampling_by_month}
\title{Extract the number of ages, lengths, and weights month for year by month}
\usage{
us_extract_sampling_by_month(colnames, data, yr = last_data_yr)
}
\arguments{
\item{colnames}{A vector of column names for ages, lengths, and weights.}

\item{data}{A data frame of biological samples.}

\item{yr}{The year to extract by-month sampling summary for}
}
\value{
A data frame with four columns, \code{month}, \code{num_age}, \code{num_length},
and \code{num_weight}.
}
\description{
Used to make data frames for tables that show these summary totals.
}
