% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_ms_prop_tac}
\alias{us_ms_prop_tac}
\title{The proportion of the U.S. TAC that the Mothership sector gets each year}
\format{
A numerical value less than one
}
\usage{
us_ms_prop_tac
}
\description{
The proportion of the U.S. TAC that the Mothership sector gets each year
}
\keyword{datasets}
