% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_ti_ct_by_month_df}
\alias{us_ti_ct_by_month_df}
\title{A data frame containing catch information on U.S. tribal vessels
by month and year}
\format{
A data frame with 88 rows and 3 variables:
\describe{2
\item{month}{Month of catch}
\item{year}{Year of catch}
\item{catch}{Catch in tonnes}
}
}
\usage{
us_ti_ct_by_month_df
}
\description{
A data frame containing catch information on U.S. tribal vessels
by month and year
}
\keyword{datasets}
