% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usa-pull-data.R
\name{usa_pull_data}
\alias{usa_pull_data}
\title{Pull data from U.S. databases for Pacific Hake}
\usage{
usa_pull_data(
  password_file,
  start_year = 1900,
  end_year = get_data_yr(),
  save_dir = fs::path(here::here(), "data-tables")
)
}
\arguments{
\item{password_file}{A file path specifying where to find the passwords. The
path can be full or relative to your current working directory. The text
file must contain two lines, where the first line is the NORPAC password
and the second line is the PacFIN password. These passwords should not be
surrounded with quotes. If a file name is not provided, which is the
default behavior, then the user will be prompted for their passwords. This
also happens if the file cannot be found given the path provided. These
passwords are needed because the databases store confidential information.}

\item{start_year}{An integer value specifying the start year of data to
include in the extraction. The default is 1900 to include all the data
possible, including years where only a foreign fleet fished.}

\item{end_year}{An integer value specifying the final year of data to include
in the extraction. The default uses \code{\link[=get_data_yr]{get_data_yr()}} to determine the
terminal year.}

\item{save_dir}{A string specifying the full path to a directory where you
want files saved. The default is to save them in a folder found using
\code{fs::path(here::here(), "data-tables")()}, which returns a path specific to the personal preferences
of JTC members.}
}
\value{
An environment or \code{list} with several objects pulled from the desired
databases are returned invisibly. Additionally, several \code{.Rdat} files are
saved to the disk in \code{extractedData}, which is created inside of \code{save_dir}.
The following files are saved:
\itemize{
\item nages.Rdat
\item nweight.Rdat
\item ncatch.Rdat
\item page.Rdat
\item pcatch.Rdat
\item pspec.Rdat
}
}
\description{
Extraction of
\itemize{
\item catch data,
\item weight measurements,
\item length measurements,
\item age reads, and
\item management quantities
from NORPAC and PacFIN databases for the Pacific Hake assessment.
}
}
\details{
\code{usa_pull_data()} includes several steps, outlined below, and leads to many
files being saved to your computer. Originally, files were \strong{NOT}
overwritten. In 2022, Kelli F. Johnson changed the code to not save previous
downloads to save on disk space, reduce the storing of redundant information,
and simplify the code base that needed maintenance. The code steps through
the following steps:
\itemize{
\item Assign usernames and passwords, where the latter are read in from the
specified file or obtained from the user.
\item Find the folder where the data should be saved, typically using
\code{fs::path(here::here(), "data-tables")()}.
\item Extract data from NORPAC.
\itemize{
\item Catch data
\item Weight and age data
\item Squash table of ages (that also includes lengths)
\item Foreign ages
\item Species list
}
\item Extract data from PacFIN.
\itemize{
\item Catch data
\item Age, length, and weight data from bds table
\item At-sea data
\item Management quantities such as allowable catch limits (ACL) for the
Pacific Fisheries Management Council, i.e., U.S. quantities of interest.
}
\item Save each object to the disk in the "extractedData" directory, which is
created in the directory specified by save_dir if it does not already
exist.
}
}
\examples{
\dontrun{
# An environment with objects is returned
dataenv <- usa_pull_data()
# Access individual objects using `get()`
head(get("ncatch", envir = dataenv))
# Access individual objects pretending the environment is a list
dataenv[["ncatch"]][1:5, ]
}

}
\seealso{
\itemize{
\item \code{\link[=get_data_yr]{get_data_yr()}}
}
}
\author{
Kelli F. Johnson
}
