% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usa_query_database.R
\name{usa_query_database}
\alias{usa_query_database}
\title{Wrapper to make a query from a sql file and get the data from a database using RODBC}
\usage{
usa_query_database(
  queryFilename,
  db,
  uid,
  pw = "",
  sp = " ",
  start = " ",
  end = " ",
  asis = FALSE
)
}
\arguments{
\item{queryFilename}{The filename of the sql query to read in.}

\item{db}{The name of your database connection.}

\item{uid}{Your user ID.}

\item{pw}{The password for the database connection. If omitted, RODBC will
prompt you for a password and will mask it. If entered, it will be visible
and unsecure.}

\item{sp}{The species scientific name to extract from the database.}

\item{start}{The start year.}

\item{end}{The end year.}

\item{ais}{A logical, specifying whether or not to convert columns, as in
\code{\link[utils]{read.table}}.}
}
\description{
Wrapper to make a query from a sql file and get the data from a database using RODBC
}
\details{
Calls \code{RODBC} functions. The function first opens an ODBC
connection and, using sql, extracts data from the \code{db}.
}
\seealso{
\code{\link[RODBC]{sqlQuery}}
}
\author{
Kelli F. Johnson
}
