% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-update-catch-sums.R
\name{utils_update_catch_sums}
\alias{utils_update_catch_sums}
\title{Updates the most recent catch time values stored in data-tables/<>.csv}
\usage{
utils_update_catch_sums(
  dir_data = here::here("data-tables"),
  digits = 5,
  values_prev_assess = c(NA, NA, NA)
)
}
\arguments{
\item{dir_data}{A path leading to data-tables where the csv files for the
stock assessment are stored.}

\item{digits}{An integer specifying how many significant digits you want
to save in the stored files. The default is 5.}

\item{values_prev_assess}{A vector of three values,
\itemize{
\item estimate of spawning biomass for terminal year in 1000s mt,
\item estimate of depletion in percent with one decimal place,
\item estimate of TAC for the first year of the forecast period in mt
rounded to the nearest whole number.
}}
}
\value{
The following two files are updated in the data-tables:
\itemize{
\item landings-tac-history.csv
\item catch-targets-biomass.csv
}
}
\description{
Read monthly catch time series for each fleet and report their summary
values in the appropriate csv files stored in data-tables.
}
\author{
Kelli F. Johnson
}
