% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vec2df}
\alias{vec2df}
\title{Create a one-row data.frame from a vector of values}
\usage{
vec2df(vec, nms = NULL)
}
\arguments{
\item{vec}{A vector}

\item{nms}{Optional column names for the new data frame}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row
}
\description{
Create a one-row data.frame from a vector of values
}
\details{
This function should not be used for massive amounts of conversions in
a loop as it has a bad time complexity
}
