% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-size-at-age.R
\name{weight.at.age.heatmap}
\alias{weight.at.age.heatmap}
\title{Weight-At-Age Heatmap Including Extrapolated Years}
\usage{
weight.at.age.heatmap(
  model,
  fleet = 1,
  proj.line.color = "royalblue",
  proj.line.width = 1,
  proj.line.yr = NULL,
  first.year = 1975,
  start_yr = 1966,
  end_yr = assess_yr - 1,
  extrap.mask = NULL,
  longterm.mean.ages = NULL,
  font.size = 4,
  axis.font.size = 10,
  samplesize = FALSE,
  print.years = NULL,
  colour = c("all", "age", "both")
)
}
\arguments{
\item{model}{An list of results read in from an SS model using
\code{\link{load_ss_files}}.}

\item{fleet}{An integer value specifying which fleet you want plotted.
Fleet -2 will plot fecundity information.
Fleet -1 will plot population weight-at-age for the middle of the year.
Fleet 0 will plot population weight-at-age for the beginning of the year.
Positive values for fleet will link to a modeled fleet.}

\item{proj.line.color}{Line color to separate projection years.}

\item{proj.line.width}{Line width to separate projection years.}

\item{proj.line.yr}{Year to separate projection years from the data period,
where if \code{NULL}, then the last year of the data in the assessment will
be used to specify when the data end and projections begin.}

\item{first.year}{The first year to plot because users may want to truncate
the data to only years for which were informed by data and not all years
in the weight-at-age file.}

\item{extrap.mask}{A list of vectors of which years are extrapolations so
that they can be colored accordingly. 1's means they are extrapolations
and 0's means they are not.}

\item{longterm.mean.ages}{A vector of mean weight-at-age values
per ages zero to the maximum age of fish in the data.
If \code{NULL}, the first year will be assumed to be the mean  of all years.}

\item{font.size}{Font size of the values printed in each box.}

\item{axis.font.size}{Font size for axis labels.}

\item{samplesize}{A logical value specifying if the heat map should be of
input sample size used to generate the data rather than the weights-at-age.}

\item{print.years}{A vector of years that will be included in the output,
but users should note that all data from \code{first.year} to the terminal
year of the data in the model will be used to calculate means and colors.
This parameter facilitates splitting the plot into two separate plots when
there are a number of years.}

\item{colour}{A character string of \code{"both"}, \code{"all"},
or \code{"age"}. If the default of \code{"both"}, then the colors of the
boxes will be based on all of the data and the transparency of the colors
will strictly be a function of each individual age in turn. That is, the
darkness of the colors used across years for age zero specifies which years
have the largest age-0 fish. \code{"all"} specifies colors based on the min
and max values of all ages without transparency. \code{"age"} colors the min
and max color specific for each age without transparency.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Produce a matrix of colors and values of weight-at-age information
by year (y-axis) and age (x-axis). The plot uses output from a
Stock Synthesis model.
}
\details{
Ages zero through the maximum age modeled in the data are shown
for each year plotted. This maximum-age shown is found by removing ages
that are the duplicate of the previous age, no user input is needed.

Users have many options to control what years are included in the plot.
The beginning years of data can be removed using \code{first.year}.
Users do not have control over what are noted as projection years, which
are all years after the end of the data included in the assessment model
up to last year included in the weight-at-age matrix.
These projection years are noted using a horizontal line but if the
demarcation is not wanted, then they can change the line width to zero.
Finally, users can calculate the colors using a range of data specified
with \code{first.year} and the last year of the data in the model. Then,
the resulting plot can be truncated to a specified year range using
\code{print.years}. The truncation is helpful to facilitate plots that fit
on a single page or on a slide with readable values. Using this truncation,
the colours would show trends across all plots made rather than just the
truncated data.
}
