% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_age_sample_sizes_df}
\alias{weight_age_sample_sizes_df}
\title{A data frame containing information on weight-at-age}
\format{
A data frame with 49 rows and 22 variables:
\describe{
\item{Yr}{Year of samples}
\item{seas}{Season code (SS3)}
\item{gender}{Gender code (SS3)}
\item{GP}{Unknown}
\item{bseas}{Unknown}
\item{Fleet}{Fishery fleet (SS3)}
\item{a0}{Number of age 0 fish sampled}
\item{a1}{Number of age 1 fish sampled}
\item{a2}{Number of age 2 fish sampled}
\item{a3}{Number of age 3 fish sampled}
\item{a4}{Number of age 4 fish sampled}
\item{a5}{Number of age 5 fish sampled}
\item{a6}{Number of age 6 fish sampled}
\item{a7}{Number of age 7 fish sampled}
\item{a8}{Number of age 8 fish sampled}
\item{a9}{Number of age 9 fish sampled}
\item{a10}{Number of age 10 fish sampled}
\item{a11}{Number of age 11 fish sampled}
\item{a12}{Number of age 12 fish sampled}
\item{a13}{Number of age 13 fish sampled}
\item{a14}{Number of age 14 fish sampled}
\item{a15}{Number of age 15 fish sampled}
}
}
\usage{
weight_age_sample_sizes_df
}
\description{
A data frame containing information on weight-at-age
}
\keyword{datasets}
