% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_at_age_estimates_df}
\alias{weight_at_age_estimates_df}
\title{A data frame containing information on the maturity estimates by year and
model type. The data are coded as proportion mature by age for a given year
and model.}
\format{
A data frame with 3 variables:
\describe{
\item{year}{Year sample was taken}
\item{age}{Age of fish}
\item{p_weight}{Predicted weight of fish in kg, I think}
}
}
\usage{
weight_at_age_estimates_df
}
\description{
A data frame containing information on the maturity estimates by year and
model type. The data are coded as proportion mature by age for a given year
and model.
}
\keyword{datasets}
