% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_at_age_fn}
\alias{weight_at_age_fn}
\title{A character string containing the file name for the SS3 weight-at-age file}
\format{
A character string
}
\usage{
weight_at_age_fn
}
\description{
A character string containing the file name for the SS3 weight-at-age file
}
\keyword{datasets}
