% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{weight_at_age_outlier}
\alias{weight_at_age_outlier}
\alias{plot_weight_at_age_outlier}
\title{Identify weight-at-age outliers}
\usage{
weight_at_age_outlier(data, filter = TRUE, drop = TRUE)

plot_weight_at_age_outlier(data, ...)
}
\arguments{
\item{data}{A data frame of weight-at-age values.}

\item{filter}{A logical specifying if the outliers should be removed or not.
The default is to remove them.}

\item{drop}{A logical specifying if the new column called \code{outlier} should
be removed from the data frame prior to returning it.}
}
\value{
A data frame with the same columns or one more column, i.e.,
\code{outlier}, than the input data frame.
}
\description{
Get weight-at-age data and identify any outliers using three screening
methods.
\itemize{
\item Data in September of 2003 are screened for fish less than 0.52 kg and
greater than 45 cm in length.
\item Data are screen for a weight-length relationship beyond that deemed to be
normal, i.e., 20e-6\emph{length^3 and 2e-6}length^3.
\item Identify fish with a recorded age of greater than or equal to 99, which
is sometimes used for data that should be flagged or not used instead of
an \code{NA} value.
These protocols started in 2011.
}
}
\seealso{
Other weight-at-age: 
\code{\link{weight_at_age_read}()}

Other weight-at-age: 
\code{\link{weight_at_age_read}()}
}
\author{
Kelli F. Johnson
}
\concept{weight-at-age}
