% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{weight_at_age_read}
\alias{weight_at_age_read}
\title{Read weight-at-age files from the disk and filter appropriately}
\usage{
weight_at_age_read(file)
}
\arguments{
\item{file}{The file path, full or relative, to the \code{.csv} file you are
importing. The file should have the following seven columns: Source,
Weight_kg, Sex, Age_yrs, Length_cm, Month, and Year.}
}
\value{
A data frame of weight-at-age information.
The data frame includes the following columns:
\enumerate{
\item Source
\item Weight_kg
\item Sex
\item Age_yrs
\item Length_cm
\item Month
\item Year
\item OutlierL
}.
}
\description{
Read the file using the appropriate method based on the file extension and
filter the data such that data are not duplicated. For example, Canadian data
are in the cumulative file but a better version was provided by Chris Grandin
in 2022 as a csv file.
}
\seealso{
Other weight-at-age: 
\code{\link{weight_at_age_outlier}()}
}
\author{
Ian G. Taylor
}
\concept{weight-at-age}
