% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{weight_at_age_wide}
\alias{weight_at_age_wide}
\title{Turn a long data frame of weight-at-age data into a wide data frame by age}
\usage{
weight_at_age_wide(dat, value = c("weight", "length", "count"), maxage = 15)
}
\arguments{
\item{dat}{A data frame created from \code{\link[=weight_at_age_read]{weight_at_age_read()}} or one that has the
following column names:
\itemize{
\item Source,
\item Weight_kg,
\item Sex,
\item Age_yrs,
\item Length_cm,
\item Month, and
\item Year.
}}

\item{value}{A character value specifying which type of data you are
interested in. The options include \code{"weight"}, \code{"length"}, and \code{"count"},
where the latter will supply the sample size.}

\item{maxage}{The age of the plus-group bin. The default is 15 years old.
This age and all older ages will be combined into a single age bin..}
}
\value{
A data frame with the first six columns pertaining to metadata,
i.e., #year, seas, sex, GP, bseas, and fleet, and additional columns
pertaining to each age starting with age zero up to the maximum age
supplied with the argument \code{maxage}. Ages for which there were no
samples are filled with \code{NA}. If \code{value = "count"}, then
the data frame will return sample sizes for each age rather than mean
weight- or length-at-age.
}
\description{
Take the long-format, which is indicative of raw data, weight-at-age data
and make a wide data frame based on ages such that the data frame can be
written for Stock Synthesis.
}
\seealso{
\code{\link[=weight_at_age_read]{weight_at_age_read()}}
}
\author{
Ian G. Taylor
}
