% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workspace.R
\name{workspace}
\alias{workspace}
\title{Create and change to a temporary directory where the loading files have
been copied to}
\usage{
workspace()
}
\value{
Nothing
}
\description{
Create and change to a temporary directory where the loading files have
been copied to
}
\details{
Run \code{bookdown::render_book("000-launcher.rmd", output_dir = ".")} once
in the temporary directory to load everything necessary. Once that has
completed you can run any table or figure code in the document for
debugging purposes. You can also see what a plot or table will look like
in the doc by adding it to the blank "006-test.rmd" file and then running
bookdown as above.
}
