% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{write_wtatage_file}
\alias{write_wtatage_file}
\title{Write weight-at-age file}
\usage{
write_wtatage_file(
  file = paste0("wtatage_", format(Sys.time(), "\%Y"), "created_", format(Sys.time(),
    "\%d-\%b-\%Y_\%H.\%M"), ".ss"),
  data,
  maturity
)
}
\arguments{
\item{file}{A filename that you want to save the information as. The default is to have
an extension of \code{.ss} such that the file can be used for Stock Synthesis.
The file path can either be relative or absolute.}

\item{data}{Weight at age matrix.}

\item{maturity}{A vector of maturity at age.}
}
\description{
Write weight-at-age file
}
