% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml-check-margin-format.R
\name{yaml_check_margin_format}
\alias{yaml_check_margin_format}
\title{Check the format of the margin command entered into the YAML header.}
\usage{
yaml_check_margin_format(fig_margins)
}
\arguments{
\item{fig_margins}{The string containing the command to verify}
}
\value{
\code{\link[=invisible]{invisible()}} if format is correct, or stops if the format is
incorrect
}
\description{
Check the format of the margin command entered into the YAML header.
}
