% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-add-alt-text.R
\name{add_alt_text}
\alias{add_alt_text}
\title{Add alternative text to figures in the document via the \code{tagpdf} LaTeX
package}
\usage{
add_alt_text(x)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}
}
\value{
The modified Tex code, as a vector
}
\description{
Add alternative text to figures in the document via the \code{tagpdf} LaTeX
package
}
\details{
Searches the given TEX file for the \verb{includegraphics\{\}},
finds the correct alternative text for those calls by searching through
the project RMD files for the matching \code{(ref:variable-alt)} definitions
and injects the text and the \code{tagpdf} structural LaTeX lines for
web accessibility
}
