% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-rds.R
\name{build_rds}
\alias{build_rds}
\title{Build RDS file for a list of models by running forecasts, retrospectives, and extra-mcmc
where applicable}
\usage{
build_rds(model_dirs = model_list, build_file = TRUE, ...)
}
\arguments{
\item{model_dirs}{A vector of model directory names (not full paths)}

\item{build_file}{If \code{TRUE}, create the RDS file. If \code{FALSE}, do not create it (used for running retros, etc)}

\item{...}{Arguments passed to \code{\link[=run]{run()}} and \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
Nothing
}
\description{
Build RDS file for a list of models by running forecasts, retrospectives, and extra-mcmc
where applicable
}
