% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-fishery-ages.R
\name{calc_fishery_ages}
\alias{calc_fishery_ages}
\title{Load data and create age compositions for the fishery}
\usage{
calc_fishery_ages(weight_at_age)
}
\arguments{
\item{weight_at_age}{A data frame read in by \code{\link[r4ss:SS_readwtatage]{r4ss::SS_readwtatage()}} for
the current year's model.}
}
\value{
A data frame is returned that is appropriate for using in a Stock Synthesis
data file for the age-composition section.
}
\description{
Uses loaded age data from each fleet to create the age-composition data for
the Stock Synthesis file after weighting by catch and weight-at-age. The
weight-at-age information is passed as a function argument and subset for
fleet 1 and the catch data is available internally in the package.
}
\details{
TODO:
\enumerate{
\item Weight-at-age is summed over all ages 0--20 but ages are not available
for 16+?
}
}
\author{
Kelli F. Johnson
}
