% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-mcmc.R
\name{calc_mcmc}
\alias{calc_mcmc}
\title{Return a list of mcmc calculations, e.g. quantiles for various values}
\usage{
calc_mcmc(mcmc, biomass_scale = 1e+06, recr_scale = 1e+06, ...)
}
\arguments{
\item{mcmc}{The output of the \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}} function as a data.frame}

\item{biomass_scale}{A scale factor to divide biomass values by}

\item{recr_scale}{A scale factor to divide recruitment values by}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A named list of MCMC outputs
}
\description{
Return a list of mcmc calculations, e.g. quantiles for various values
}
\details{
The values calculated in this function are used in the tables and plots in
the hake document. If a new value is needed, add that calculation here
and re-run the \code{\link[=create_rds_file]{create_rds_file()}} function to re-build the RDS file
for the model.
}
