% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-get-commercial-wa.R
\name{canada_get_commercial_wa}
\alias{canada_get_commercial_wa}
\title{Get the weight-at-age data frame for commercial data and}
\usage{
canada_get_commercial_wa(d, weight_scale = 1000, num_sex = 1, ret_df = FALSE)
}
\arguments{
\item{d}{A list as returned by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{weight_scale}{A value to divide the weights by}

\item{num_sex}{Either 1 or 2. If 1, then fill the column \code{Sex} with \code{U}
for every record. If 2, use the data to fill the \code{Sex} column with
\code{M}, \code{F}, or \code{NA} depending on input values}

\item{ret_df}{Logical. If \code{TRUE}, return the weight-at-age data frame.
If \code{FALSE}, write the data frame to the Canadian weight-at-age CSV file}
}
\value{
a data frame with \code{Source} set to \code{CAN_freezer}, \code{CAN_shoreside},
\code{CAN_jv}, and \code{CAN_polish} for the four fishery types
}
\description{
Get the weight-at-age data frame for commercial data and
}
\details{
No filtering is performed in this function.
Creates a CSV file containing the commercial weight-at-age
}
