% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-canada-load-sample-data.R
\name{canada_load_sample_data}
\alias{canada_load_sample_data}
\title{Load the commercial sample data from a file generated
by \code{\link[=canada_extract_sample_data]{canada_extract_sample_data()}}}
\usage{
canada_load_sample_data(dr = "/srv/hake/other/samples")
}
\arguments{
\item{dr}{The directory in which the sample RDS file resides}
}
\value{
Nothing, writes an RDS file
}
\description{
Load the commercial sample data from a file generated
by \code{\link[=canada_extract_sample_data]{canada_extract_sample_data()}}
}
\details{
Filtering is performed in this function for
\enumerate{
\item Areas - BC offshore areas + Strait of Juan de Fuca
\item Fishing gear - Midwater trawl only
}
}
