% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{clean}
\alias{clean}
\title{Clean the doc directory of all build output including knitr cache and
knitr figures directories}
\usage{
clean(
  knitr_figures_dir = here::here("doc/knitr-figs"),
  knitr_cache_dir = here::here("doc/knitr-cache"),
  out_csv_dir = here::here("doc/out-csv")
)
}
\arguments{
\item{knitr_figures_dir}{Directory where4 the knitr-generated
figures reside}

\item{knitr_cache_dir}{Directory where the knitr cached chunk
databases reside}

\item{out_csv_dir}{Directory where the table outputs reside}
}
\value{
Nothing
}
\description{
Clean the doc directory of all build output including knitr cache and
knitr figures directories
}
\details{
This function can be called from anywhere in the hake project and will
clean the "doc" directory, not the local directory. If you are testing
with \code{gotest()} and are in a temporary directory, the cleaning will be
applied there instead. The cleaning takes place on whatever directory
is at \code{here::here("doc")}
}
