% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{clean}
\alias{clean}
\title{Clean (delete) all build output including knitr cache and knitr figures
directories from the current directory}
\usage{
clean(..., knitr_figures_dir = "knitr-figs", knitr_cache_dir = "knitr-cache")
}
\arguments{
\item{...}{Optional quoted or unquoted names of chunks that should be
'cleaned'.The chunk name(s) will be used to delete only the objects
relating to them from the \code{knitr_figures_dir} and \code{knitr-cache}.as well
as all the LaTeX and intermediate RMD and MD files}

\item{knitr_figures_dir}{Directory where the knitr-generated
figures reside}

\item{knitr_cache_dir}{Directory where the knitr cached chunk
databases reside}
}
\value{
Nothing
}
\description{
Clean (delete) all build output including knitr cache and knitr figures
directories from the current directory
}
\details{
This function will clean the current working directory. It works from
within the main \code{doc} directory as well as in all the beamer presentation
directories. If you are testing with \code{gotest()} and are in a temporary
directory, in the subdirectory \code{doc}, the cleaning will be applied there.
}
