% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-cohort.R
\name{cohort_table}
\alias{cohort_table}
\title{Make a table of values for cohorts as they pass through their life}
\usage{
cohort_table(
  model,
  end_yr,
  cohorts,
  csv_dir = "out-csv",
  xcaption = "default",
  xlabel = "default",
  digits = 1,
  font_size = 9,
  space_size = 10
)
}
\arguments{
\item{model}{A model as output by \code{\link[=load_models]{load_models()}}}

\item{end_yr}{The last year to use}

\item{cohorts}{A vector of cohorts (years) to use}
}
\value{
An \code{\link[xtable:xtable]{xtable::xtable()}}
}
\description{
Make a table of values for cohorts as they pass through their life
}
