% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-combine-historical-probs.R
\name{combine_historical_probs}
\alias{combine_historical_probs}
\title{Calculate a model's probability of stock decline in a year and
combine with past assessment estimates}
\usage{
combine_historical_probs(model, start = 2012, end, ...)
}
\arguments{
\item{model}{A model as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{start}{First assessment year to do comparisons}

\item{end}{Last assessment year to do comparisons}

\item{...}{Further arguments to pass to \code{\link[=calc_historical_probs]{calc_historical_probs()}}}
}
\value{
Data frame with columns
\itemize{
\item \code{Year} - year of assessment
\item \code{P_decline} - estimate from \verb{year' assessment model of the probability that the spawning biomass will decline from }year\code{to}year\verb{+ 1 for a catch that turned out to be the catch in}year'
\item \code{P_below_B40} - estimate from \verb{year' assessment model of the probability that the spawning biomass will be below B_40 in }year\verb{+1 for a catch that turned out to be the catch in }year'
\item \code{P_decline_curr} - current estimate (from \code{model}) of the probability that
the spawning biomass declined from year to year+1
\item \code{P_below_B40_curr} - current estimate (from \code{model}) of the probability
that the spawning biomass was below B_40 in year+1
}
}
\description{
Calculate a model's probability of stock decline in a year and
combine with past assessment estimates
}
