% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{copy_dirfiles}
\alias{copy_dirfiles}
\title{Call [shell()] or [system()] depending on the Operating System}
\usage{
copy_dirfiles(
  fromdir = getwd(),
  todir = NULL,
  files = c("starter.ss", "forecast.ss", "hake_control.ss", "hake_data.ss", "wtatage.ss"),
  overwrite = TRUE,
  ignore_missing_files = TRUE
)
}
\arguments{
\item{fromdir}{The directory to copy subdirectories and their `files` from}

\item{todir}{The directory to copy subdirectories and their `files` to}

\item{files}{A character vector of file names to include in the copy}

\item{overwrite}{If `todir` exists then overwrite its subdirs if they exist.
Everything else in `todir` not pertaining to the copy will remain after the
overwrite}

\item{ignore_missing_files}{If `FALSE` and any of `files` are missing in
subdirectories of `fromdir`, code execution will stop and an error will
be shown. If `TRUE`, missing files will be ignored}
}
\value{
Nothing
}
\description{
Call [shell()] or [system()] depending on the Operating System
}
\details{
Copy `files` and their containing directories from all subdirectories
of `fromdir` to `todir`. Only subdirs of `fromdir` are searched for
`files`, not `fromdir` itself, and there is no recursion.
}
