% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-df-biomass.R
\name{create_group_df_biomass}
\alias{create_group_df_biomass}
\title{Create a data frame containing model output of biomass from a
model list which is in long format ready for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
\usage{
create_group_df_biomass(
  model_lst = NULL,
  model_names = NULL,
  rel = FALSE,
  bo_yr = 4,
  ...
)
}
\arguments{
\item{model_lst}{A list of models, each created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_names}{A vector of model names,the same length as \code{model_lst}}

\item{rel}{Logical. If \code{TRUE} return a list of one element, the relative
biomass (no B0 element)}

\item{bo_yr}{The number of years prior to the start of the time series
that you want to use as a plot point for B0 on biomass plots.}

\item{...}{Arguments passed to \code{\link[=extract_mcmc_quant]{extract_mcmc_quant()}}}
}
\value{
A list or one or two \code{\link[tibble:tibble]{tibble::tibble()}}. If \code{rel} is \code{TRUE}, one
(biomass) If \code{rel} is \code{FALSE}, two (biomass and B0)
}
\description{
Create a data frame containing model output of biomass from a
model list which is in long format ready for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
