% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-df-index.R
\name{create_group_df_index}
\alias{create_group_df_index}
\title{Create a data frame containing model output index fits from a
model list which is in long format ready for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
\usage{
create_group_df_index(
  model_lst = NULL,
  model_names = NULL,
  survey_type = c("age1", "age2", "edna")
)
}
\arguments{
\item{model_lst}{A list of models, each created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_names}{A vector of model names,the same length as \code{model_lst}}

\item{survey_type}{One of \code{age1}, \code{age2}, or \code{edna}}
}
\value{
A list containing a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create a data frame containing model output index fits from a
model list which is in long format ready for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
