% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-kn-files.R
\name{create_kn_files}
\alias{create_kn_files}
\title{Create the \code{nuisance} and \code{key} posteriors CSV files in the output
mcmc directory of the model}
\usage{
create_kn_files(model)
}
\arguments{
\item{model}{A model object representing the output from the SS model}
}
\value{
Nothing, creates two CSV files
}
\description{
Create the \code{nuisance} and \code{key} posteriors CSV files in the output
mcmc directory of the model
}
