% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-file.R
\name{create_rds_file}
\alias{create_rds_file}
\title{Create an rds file to hold the model's data and outputs.}
\usage{
create_rds_file(model_dir = NULL, verbose = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{model_dir}{Directory name of model to be loaded}

\item{...}{Arguments to pass to [load_extra_mcmc()]}
}
\value{
[base::invisible()]
}
\description{
Create an rds file to hold the model's data and outputs.
}
