% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-file.R
\name{create_rds_file}
\alias{create_rds_file}
\title{Create an rds file to hold the model's data and outputs.}
\usage{
create_rds_file(
  model_dir = NULL,
  ct_levels_lst = set_ct_levels(),
  forecasts_path = file.path(model_dir, "forecasts"),
  ct_levels_path = file.path(model_dir, "catch-levels"),
  retrospectives_path = file.path(model_dir, "retrospectives"),
  default_hr_path = file.path(ct_levels_path, "default-hr"),
  stable_catch_path = file.path(ct_levels_path, "stable-catch"),
  spr_100_path = file.path(ct_levels_path, "spr-100"),
  verbose = TRUE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{model_dir}{Directory name of model to be loaded}

\item{verbose}{Logical. If \code{TRUE}, write more output to the console}

\item{overwrite}{Logical. If \code{TRUE}, overwrite the file if it exists}

\item{...}{Arguments to pass to \code{\link[=load_ss_files]{load_ss_files()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Create an rds file to hold the model's data and outputs.
}
