% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sens-dirs.R
\name{create_sens_dirs}
\alias{create_sens_dirs}
\title{Create standard sensitivity directories}
\usage{
create_sens_dirs(dir_version, sens_dir_name = "03-sensitivity-models")
}
\arguments{
\item{dir_version}{A file path to the current model version of interest. For
example, \verb{/srv/hake/models/2023/01-version}. This directory will contain
subdirectories that store folders for the base model and a separate folder
for sensitivities.}

\item{sens_dir_name}{Name of the sensitivity directory found in the \code{dir_version}
directory}
}
\value{
A string is returned providing the path to the sensitivity directory.
Files are written to the disk inside the following directories:
\itemize{
\item 01-h-prior-mean-low
\item 02-h-fix-high
\item 03-sigma-r-fix-low
\item 04-sigma-r-fix-high
\item 05-m-02-sd
\item 06-m-03-sd
\item 07-m-hamel-prior
\item 08-age-1-survey,
\item 09-comp-weight-harmonic-mean
\item 10-tv-select-phi-extra-low
\item 11-tv-select-phi-low
\item 12-tv-select-phi-high
\item 13-max-sel-age-'5
\item 14-max-sel-age-7
\item 15-max-sel-age-8
\item 18-tv-maturity-182
\item 19-m-at-age
\item 20-m-at-age-fixed
}
}
\description{
The SRG requests that the following sensitivities are run as standard
protocol for the base model on an annual basis.
This function copies and alters the input files from a base model
setting up all of the required sensitivity runs.
}
\author{
Kelli F. Johnson
}
