% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-mcmc-quant.R
\name{extract_mcmc_quant}
\alias{extract_mcmc_quant}
\title{Extract MCMC quantiles from models and return them
in a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}-ready data frame}
\usage{
extract_mcmc_quant(
  model_lst,
  model_names,
  type,
  inc_model_year = FALSE,
  end_yrs = year(Sys.time())
)
}
\arguments{
\item{model_lst}{A list of models, each created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_names}{A vector of model names,the same length as \code{models_lst}}

\item{type}{A name as found in an \code{mcmccalcs} object of a model object,
for example extraction of: \code{base_model$mcmccalcs$smed} would require
\code{type} = "smed"}

\item{inc_model_year}{Logical. If \code{TRUE}, include the \code{model} and \code{year}
columns in the data frame}

\item{end_yrs}{A vector of the end years for each model.
If one value, it will apply to all models}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Extract MCMC quantiles from models and return them
in a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}-ready data frame
}
