% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-sigma-r.R
\name{extract_sigma_r}
\alias{extract_sigma_r}
\title{Extract the \code{sigma_R_info} table from a data frame of recruitment parameters}
\usage{
extract_sigma_r(models = NA, model_names = NA, sigma_r_in = NA)
}
\arguments{
\item{models}{A list of models extracted using \code{\link[r4ss:SS_output]{r4ss::SS_output()}}. If
only one model,it must be a list with one element containing the model}

\item{model_names}{A vector of names, one for each model in \code{models}.
This vector must be the same length as \code{models}}

\item{sigma_r_in}{The input value used for sigma_R, this is also available
in the list output by \code{\link[r4ss:SS_output]{r4ss::SS_output()}} and is used to scale the results}
}
\value{
A data frame is returned with three rows for each model in the
\code{models} list and 11 columns. Rows provide summary statistics for a
given group of recruitments, where the rows are additive in their
inclusiveness, i.e., the data frames used to calculate rows two and three
include all data used in the calculations that led to the results
displayed in the first row. Columns include the following variables:
\itemize{
\item \code{period}: Categories that specify which recruitments were included in the
subset used to calculate the summary
\item \code{N_devs}: The number of recruitment estimates used in the summary
\item \code{SD_of_devs}: The standard deviation (sd) of the recruitment deviations.
\item \code{Var_of_devs}: The variance (var) of the recruitment deviations
\item \code{mean_SE}: The mean of the estimated standard error (se) of the
recruitment deviations, or more precisely, the mean of the \code{Parm_StDev}
column of your data
\item \code{mean_SEsquared}: The mean of the squared estimates of standard error of
the recruitment deviations, i.e., \code{mean(Parm_StDev^2)}
\item \code{sqrt_sum_of_components}: The square root of the sum of the deviation and
the standard error
\item \code{SD_of_devs_over_sigma_R}: The scaled version of \code{SD_of_devs}
\item \code{sqrt_sum_over_sigma_R}: The scaled version of \code{sqrt_sum_of_components}
\item \code{alternative_sigma_R}: The suggested value for a new input sigma_R, which
is just a repeat of the sum of components column
}
}
\description{
Extract the \code{sigma_R_info} table from a data frame of recruitment parameters
}
\details{
Extract the \code{sigma_R_info} table that is available in the object
returned from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}. Typically, this table is determined from
MLE results, but this function also offers the capability to calculate this
info from the table of MCMC results as well.

Based on the non-exported function \code{extract_sigma_R_info()} in \link{r4ss},
but with a wrapper for more than one model
}
