% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-f.R
\name{f}
\alias{f}
\title{Format a number}
\usage{
f(x, digits = 0)
}
\arguments{
\item{x}{The number}

\item{digits}{The number of decimal points to use}
}
\value{
A Character string. A formatted version of the number \code{x}.
}
\description{
Format a number
}
\details{
The formatted number will have comma-separated thousands, millions, etc.
and the exact number of decimal points given by \code{digits}, even if the
last one is a zero.
}
