% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fix-posteriors.R
\name{fix_posteriors}
\alias{fix_posteriors}
\title{Remove multiple header/data rows found in posteriors files to leave just
one}
\usage{
fix_posteriors(fn)
}
\arguments{
\item{fn}{File name for the posterior file}
}
\value{
Nothing. If the file contains only one header/data row (i.e is
of correct format) then nothing happens. If it contains more than one
header row, remove the second header and everything else after that in
the file and overwrite the file
}
\description{
Remove multiple header/data rows found in posteriors files to leave just
one
}
