% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-quantities.R
\name{get_baa}
\alias{get_baa}
\title{Get the median proportions of biomass-at-ages for the final year (MCMC)
for all cohorts}
\usage{
get_baa(model, curr_yr, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{model}{A model as loaded by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{curr_yr}{is the current year. The years will decrease from this to
give the cohorts year values}
}
\value{
A single-column tibble representing the proportion of biomass-at
ages for the terminal year
}
\description{
Get the median proportions of biomass-at-ages for the final year (MCMC)
for all cohorts
}
