% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_posterior_data}
\alias{get_posterior_data}
\title{Get the posterior values for the given regular expressions of
parameter names}
\usage{
get_posterior_data(model, param_regex)
}
\arguments{
\item{model}{The SS model output as loaded by [load_ss_files()]}

\item{param_regex}{A vector of regular expressions used to extract data
for parameter names. If there are no matches, or more than one for any
regular expression, the program will stop}
}
\value{
A list of posterior vectors, one for each of the regular
expressions in `param_regex`
}
\description{
Get the posterior values for the given regular expressions of
parameter names
}
\examples{
get_posterior_data(base_model, "BH_steep")
get_posterior_data(base_model, "e")
get_posterior_data(base_model, "asdfg")
get_posterior_data(base_model, c("NatM", "SR_LN", "SR_BH_steep",
 "Q_extraSD"))
}
