% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_prior_data}
\alias{get_prior_data}
\title{Get the prior and MLE values for the given regular expressions of
parameter names}
\usage{
get_prior_data(model, param_regex, n_points_prior = 3e+06)
}
\arguments{
\item{model}{The SS model output as loaded by [load_ss_files()]}

\item{param_regex}{A vector of regular expressions used to extract data
for parameter names. If there are no matches, or more than one for any
regular expression, the program will stop.}

\item{n_points_prior}{An integer specifying the number of points you want to
be generated from the `r<distribution>()` function, e.g.,
`rnorm(n_points_prior)`.}
}
\value{
A list of prior and MLE data, one for each of the regular
expressions in `param_regex`
}
\description{
Get the prior and MLE values for the given regular expressions of
parameter names
}
\examples{
get_prior_data(base_model, "BH_steep")
get_prior_data(base_model, "e")
get_prior_data(base_model, "asdfg")
get_prior_data(base_model, c("NatM", "SR_LN", "SR_BH_steep", "Q_extraSD"))
}
