% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-prior-data.R
\name{get_prior_data}
\alias{get_prior_data}
\title{Get the prior and MLE values for the given regular expressions of
parameter names}
\usage{
get_prior_data(
  model,
  params_regex = key_posteriors,
  params_titles = key_posteriors_titles,
  n_points_prior = 1e+05,
  ...
)
}
\arguments{
\item{model}{A model object as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{params_regex}{A vector of regular expressions used to extract data
for parameter names. Default value is \link{key_posteriors}}

\item{params_titles}{A list of names for the given elements of
\code{params_regex}. Default value is \link{key_posteriors_titles}}

\item{n_points_prior}{An integer specifying the number of points you want to
be generated from the \verb{r<distribution>()} function, e.g.,
\code{rnorm(n_points_prior)}.}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A list of prior and MLE data, one for each of the regular
expressions in \code{params_regex}
}
\description{
Get the prior and MLE values for the given regular expressions of
parameter names
}
\examples{
\dontrun{
get_prior_data(base_model, "BH_steep")
get_prior_data(base_model, "e")
get_prior_data(base_model, "asdfg")
get_prior_data(base_model, c("NatM", "SR_LN", "SR_BH_steep", "Q_extraSD"))
}
}
