% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-quantities.R
\name{get_rec_ci}
\alias{get_rec_ci}
\title{Get recruitment information for a cohort from two models with
an CI envelope \% difference calculation}
\usage{
get_rec_ci(model1, model2, yr, scale = 1e+06, decimals = 3, perc_decimals = 0)
}
\arguments{
\item{model1}{A model, typically last year's model}

\item{model2}{A model, typically this year's model}

\item{yr}{The cohort to extract}

\item{scale}{Scaling factor}

\item{decimals}{Number of decimal points to round values to}

\item{probs}{The 3-vector of probabilities to be passes to the
\code{\link[=quantile]{quantile()}} function.}

\item{perc.decimals}{Number of decimal points to round percentages to}
}
\value{
A list of length 3: Item 1 is a vector of the \code{model1}
recruitment lower, median, upper, and size of credible interval
(\code{env_diff} column). Item 2 is the same type object for \code{model2},
and item 3 is the ratio of the \code{model2} \code{env_diff} divided by that
of \code{model1} as a percentage.
}
\description{
Get recruitment information for a cohort from two models with
an CI envelope \% difference calculation
}
