% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-rec-ci.R
\name{get_rec_ci}
\alias{get_rec_ci}
\title{Get recruitment information for a cohort from two models with
an CI envelope \% difference calculation}
\usage{
get_rec_ci(model1, model2, yr, digits = 3, perc_digits = 0)
}
\arguments{
\item{model1}{A model object, created with \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model2}{A model object, created with \code{\link[=create_rds_file]{create_rds_file()}}}

\item{yr}{The cohort (year) to extract}

\item{digits}{Number of decimal points to round values to}

\item{perc_digits}{Number of decimal points to round percentages to}

\item{probs}{The 3-vector of probabilities to be passes to the
\code{\link[=quantile]{quantile()}} function.}
}
\value{
A list of length 3: Item 1 is a vector of the \code{model1}
recruitment lower, median, upper, and size of credible interval
(\code{env_diff} column). Item 2 is the same type object for \code{model2},
and item 3 is the ratio of the \code{model2} \code{env_diff} divided by that
of \code{model1} as a percentage.
}
\description{
Get recruitment information for a cohort from two models with
an CI envelope \% difference calculation
}
