% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gotest-goback.R
\name{gotest}
\alias{gotest}
\title{Create and change to a temporary directory where the loading files have
been copied to}
\usage{
gotest(repo_dr = "~/github/pacific-hake/hake", copy_tmpdir = FALSE)
}
\arguments{
\item{copy_tmpdir}{Logical. If \code{TRUE}, copy the temporary path to the
OS clipboard}

\item{doc_dr}{The directory that holds all the RMD files for building the
document. Note: due to \code{tempdir()} issues, we cannot use \code{here()} for this,
it must be a full path}
}
\value{
Nothing
}
\description{
Create and change to a temporary directory where the loading files have
been copied to
}
\details{
Run \code{bookdown::render_book("000-launcher.rmd", output_dir = ".")} once
in the temporary directory to load everything necessary. Once that has
completed you can run any table or figure code in the document for
debugging purposes. You can also see what a plot or table will look like
in the doc by adding it to the blank "006-test.rmd" file and then running
bookdown as above.
}
