% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gotest.R
\name{gotest}
\alias{gotest}
\title{Create and change to a temporary directory and copy a skeleton
version of the document or beamer presentation files for testing}
\usage{
gotest(
  config_fn = "_bookdown.yml",
  figures_dir = NULL,
  knitr_figures_dir = NULL
)
}
\arguments{
\item{config_fn}{The name of the bookdown YAML file. The default is
\verb{_bookdown.yml}}
}
\value{
Nothing, but the global variable \code{goback_dr} is set for the
function \code{\link[=goback]{goback()}} to use to return to the original directory
}
\description{
Create and change to a temporary directory and copy a skeleton
version of the document or beamer presentation files for testing
}
\details{
Extracts the source and destination filenames for copying into the
temporary directory with one of the two functions
\code{\link[=gotest_doc_get_src_dest_filenames]{gotest_doc_get_src_dest_filenames()}} and
\code{\link[=gotest_beamer_get_src_dest_filenames]{gotest_beamer_get_src_dest_filenames()}}
See the \emph{Debugging a figure or table, or any other Rmarkdown code}
section of the README.md file for detailed explanation.
Use the function \code{\link[=goback]{goback()}} to return from the testing directory to the
directory you were in before calling \code{\link[=gotest]{gotest()}}
}
