% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake.R
\docType{package}
\name{hake}
\alias{hake}
\title{hake: A package for creating the Pacific hake stock assessment document}
\description{
Uses the \code{knitr} R package along with \code{lualatex} to generate a PDF
of the stock assessment document. Also included are beamer presentation
templates for the various meetings required.
}
