% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake.R
\docType{_PACKAGE}
\name{hake}
\alias{hake}
\title{hake: A package for creating the Pacific hake stock assessment document}
\description{
Uses the \link{bookdown}, \link{rmarkdown}, and \link{knitr} packages along with
TexLive's \code{lualatex} to generate a PDF of the stock assessment document.
Also included are \link{bookdown} beamer presentation templates for the various
meetings required.
}
\details{
All of these are in alphabetical order, both packages and functions
within them
}
