% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake-pdf.R
\name{hake_pdf}
\alias{hake_pdf}
\title{Custom function to allow the post processing steps to be run before
PDF rendering}
\usage{
hake_pdf(...)
}
\arguments{
\item{...}{Variables from the YAML code (in `000-launcher.rmd``)
passed to \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}} and \code{\link[=post_process]{post_process()}}}
}
\value{
The output from \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}
}
\description{
Custom function to allow the post processing steps to be run before
PDF rendering
}
