% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-theme.R
\name{hake_theme}
\alias{hake_theme}
\title{Theme for the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plots in the document}
\usage{
hake_theme(
  ax_label_color = axis_label_color,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_title_font_face = "plain",
  ax_tick_font_face = "plain",
  ax_tick_length_cm = 0.15,
  margin_def = margin(0, 0, 0, 0)
)
}
\arguments{
\item{ax_label_color}{The color of the text for all tick labels and
axis titles in the project}

\item{ax_title_font_size}{The font size for all the axis titles in the
project}

\item{ax_tick_font_size}{The font size for all the axis tick labels in the
project}

\item{ax_title_font_face}{The font face for all the axis titles in the
project. One of "plain", "bold", or "italics"}

\item{ax_tick_font_face}{The font face for all the axis tick labels in the
project. One of "plain", "bold", or "italics"}

\item{ax_tick_length_cm}{The length of all the tick mark lines in the
project}

\item{margin_def}{A call to the function \code{\link[ggplot2:element]{ggplot2::margin()}}. The format
must be like this: margin(12, 12, 12, 12) where the order of the arguments
is for each side top, right, bottom, left. Defaults to zero margin on all
sides}
}
\description{
Theme for the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plots in the document
}
