% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-heatmap-extract-sample-size.R
\name{heatmap_extract_sample_size}
\alias{heatmap_extract_sample_size}
\title{Extract the sample size data frame, or a data frame of logical values
for presence/absence of sample sizes if \code{ret_mask} is \code{TRUE}}
\usage{
heatmap_extract_sample_size(
  sample_size_df = NULL,
  wa = NULL,
  ret_mask = FALSE,
  ...
)
}
\arguments{
\item{sample_size_df}{A data frame where ages are columns (and start with
the letter 'a'). If the values are zero, the weight-at-age was
extrapolated/interpolated. If there is a value, the weight-at-age is data}

\item{wa}{A weight-at-age data frame as created by \code{\link[=heatmap_extract_wa]{heatmap_extract_wa()}}}

\item{ret_mask}{Logical. If \code{TRUE}, return a data frame the same
dimensions as \code{wa} which contains only the \code{yr} column and \code{TRUE}/\code{FALSE}
values in the table cells}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A data frame containing years (\code{yr}) column and columns for age,
represented as numbers (no text appended to them). If \code{ret_mask} is
\code{TRUE} then each cell is either \code{TRUE}or \code{FALSE} representing presence
or absence of samples for that year/age combination.The dimensions of the
returned data frame will be identical to the dimensions of \code{wa}.
If \code{ret_mask} is \code{FALSE}, the data frame returned will be the same
dimensions as \code{wa} plus two more rows and one more column. The two rows
are one row containing the year and all \code{NA}s (a blank row for the table)
and a row of the column totals, The extra column contains the row sums
Set the first and second elements to \code{NA} so that nothing appears in
row sums column for the first two cells
}
\description{
Extract the sample size data frame, or a data frame of logical values
for presence/absence of sample sizes if \code{ret_mask} is \code{TRUE}
}
