% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-heatmap-set-colors.R
\name{heatmap_set_colors}
\alias{heatmap_set_colors}
\title{Ramp colors given to the length of the number of ages in the data frame \code{w}}
\usage{
heatmap_set_colors(
  wa = NULL,
  cols = c("red", "yellow", "green", "dodgerblue"),
  ...
)
}
\arguments{
\item{wa}{A data frame containing columns named by numbers which are ages}

\item{cols}{A vector of the colors to use to build the color ramp}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A vector of the colors ramped by the function. The vector will
be the number of ages minus 1 found in the \code{wa} data frame. The color
ramp is provided by \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}
}
\description{
Ramp colors given to the length of the number of ages in the data frame \code{w}
}
