% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ct-levels.R
\name{load_ct_levels}
\alias{load_ct_levels}
\title{Load catch levels from SS forecast files.}
\usage{
load_ct_levels(model_path, inc_fi_and_stable_catch = FALSE, ...)
}
\arguments{
\item{model_path}{The model directory name}

\item{inc_fi_and_stable_catch}{Logical. If \code{TRUE}, include the Fishing
intensity = 100\% and the Stable Catch scenarios}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A list of 3-element lists of vectors of 3 catch levels
corresponding to:
\enumerate{
\item vector of values, 1 for each forecast year
\item Pretty name of the catch stream scenario
\item The directory name for the scenario
}

Return object looks the same as the the list returned from
\code{set_ct_levels()} but with the NAs replaced for the catch levels that
needed to be run through algorithms to get the catch values (default-hr,
spr-100, stable-catch)
}
\description{
Load catch levels from SS forecast files.
}
\details{
This is used to retrieve calculated values after running the
\code{\link[=run_ct_levels]{run_ct_levels()}} function for default HR, SPR 100, and stable catch
stream scenarios

Assumes \code{\link[=run_ct_levels]{run_ct_levels()}} function has been run and the forecast
files are populated with 3 forecast years.
}
