% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc.R
\name{load_extra_mcmc}
\alias{load_extra_mcmc}
\title{Create and return a list of output stats to attach to the main model
by looking in the model's path for the report files.}
\usage{
load_extra_mcmc(
  model,
  probs = c(0.025, 0.5, 0.975),
  small = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model}{A model list as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{probs}{The quantile values to use on the MCMC posterior data}

\item{small}{Logical. If \code{TRUE}, only load a small set of things
that are required for the assessment (index fits and catchability
estimates) in order to reduce the memory taken up by loading all the
output into R for the document build.If \code{FALSE}, load everything that
the base model needs, which will make a huge RDS file}

\item{verbose}{Logical. If \code{TRUE}, show progress messages}

\item{...}{Arguments passed to \code{\link[=extract_rep_table]{extract_rep_table()}}}
}
\value{
The extra MCMC list
}
\description{
Create and return a list of output stats to attach to the main model
by looking in the model's path for the report files.
}
