% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc.R
\name{load_extra_mcmc}
\alias{load_extra_mcmc}
\title{Create and return a list of output stats to attach to the main model
by looking in the model's path for the report files.}
\usage{
load_extra_mcmc(model, progress_n = 500, verbose = TRUE, first = 0, ...)
}
\arguments{
\item{model}{A model list as created by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{progress_n}{Report every time this many files are processed. Consider
how many posteriors there are, this should be a fairly large proportion of
that (around 1/8th) or there will be too much output and it will run slow}

\item{verbose}{Logical. If \code{TRUE}, show progress messages}

\item{first}{Load this many of the files. If a non-positive number, load
them all. Used for debugging purposes to cut down the size of the
lists used}

\item{...}{Arguments passed to \code{\link[=extract_rep_table]{extract_rep_table()}}}
}
\value{
The extra MCMC list
}
\description{
Create and return a list of output stats to attach to the main model
by looking in the model's path for the report files.
}
