% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-extra-mcmc-atage.R
\name{load_extra_mcmc_atage}
\alias{load_extra_mcmc_atage}
\title{Load at-age data chunks from a list of report file output,
and calculate the medians at-age by year}
\usage{
load_extra_mcmc_atage(
  reps,
  verbose = TRUE,
  txt = "output",
  scale = 1,
  start_yr = NULL,
  end_yr = NULL,
  ...
)
}
\arguments{
\item{reps}{A list of vectors representing one element per report file}

\item{verbose}{Logical. Show messages}

\item{txt}{Text to write to the screen while processing}

\item{scale}{Value to divide the values by in the output table}

\item{start_yr}{Filter years earlier than this out of the results. If
\code{NULL}, no filtering will occur for the start year}

\item{end_yr}{Filter years greater than this out of the results. If
\code{NULL}, no filtering will occur for the end year}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
A list with one element being the data frame of values by year with
every posterior included (\code{atage}) and the other being a much smaller data
frame containing the median across posteriors by year for the values (\code{med})
}
\description{
Load at-age data chunks from a list of report file output,
and calculate the medians at-age by year
}
