% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}}
\usage{
load_forecasts(
  model_path = NULL,
  first = 0,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  verbose = TRUE,
  is_catch_proj_model = FALSE,
  ...
)
}
\arguments{
\item{first}{Load this many of the files. If a non-positive number, load
them all. Used for debugging purposes to cut down the size of the
lists used}

\item{forecast_yrs}{A vector of forecast years}

\item{verbose}{Logical. If \code{TRUE}, show messages}

\item{is_catch_proj_model}{Logical. If \code{TRUE}, The first forecast year will
be treated as a catch projected year meaning the \code{forecast_yrs} are in
addition to that. All will be loaded as forecasts however}

\item{...}{Absorbs arguments intended for other functions}

\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A list of forecast outputs as read in by \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}}
}
\description{
Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}
}
\details{
If the forecasts directory does not exist or there is a problem
loading the forecasts, return \code{NA}
}
