% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}}
\usage{
load_forecasts(model, first = model$nposts, ...)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{first}{Load this many of the files. If \code{NULL}, load them all. Used
for debugging purposes to cut down the size of the lists used}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A list of forecast outputs as read in by \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}}
}
\description{
Fetch the output from previously-run forecasting using \code{\link[=run_forecasts]{run_forecasts()}}
}
\details{
If the forecasts directory does not exist or there is a problem
loading the forecasts, return \code{NA}
}
