% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-mcmc-vals.R
\name{load_mcmc_vals}
\alias{load_mcmc_vals}
\title{Load some values from the MCMC output and store in a list}
\usage{
load_mcmc_vals(model, assess_yr)
}
\arguments{
\item{model}{A model list as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{assess_yr}{The current assessment year}
}
\value{
A list of values extracted from the MCMC data frame of a model
}
\description{
Load some values from the MCMC output and store in a list
}
\details{
Uses the \link{probs} vector which is included in the package data for
this package
}
