% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-retrospectives.R
\name{load_retrospectives}
\alias{load_retrospectives}
\title{Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt}
\usage{
load_retrospectives(model, base_model_name = "Base model", ...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=load_ss_files]{load_ss_files()}} and \code{\link[=load_extra_mcmc]{load_extra_mcmc()}}}

\item{retro_path}{The path in which the retrospective directories reside}
}
\value{
The list of retrospective outputs
}
\description{
Fetch the retrospectives and return a list of each. If there are no retrospective
directories or there is some other problem, the program will halt
}
