% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load_ss_files}
\alias{load_ss_files}
\title{Load all the SS files for output and input, and return the model object}
\usage{
load_ss_files(
  model_path = NA,
  key_posts = c("NatM", "SR_LN", "SR_BH_steep", "Q_extraSD", "ln.EffN_mult._1",
    "ln.EffN_mult._2"),
  key_posts_fn = "keyposteriors.csv",
  nuisance_posts_fn = "nuisanceposteriors.csv",
  printstats = FALSE,
  ...
)
}
\arguments{
\item{key_posts}{Vector of key posteriors used to create key posteriors file}

\item{key_posts_fn}{Name of the key posteriors file}

\item{nuisance_posts_fn}{Name of the nuisance posteriors file}

\item{printstats}{Print info on each model loaded via \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{model.dir}{Directory the model reesides in}
}
\value{
A model object representing the output from the SS model
}
\description{
Load all the SS files for output and input, and return the model object
}
\details{
If MCMC directory is present, load that and perform calculations for mcmc parameters.
}
