% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ss-files.R
\name{load_ss_files}
\alias{load_ss_files}
\title{Load all the SS files for output and input, and return the model object}
\usage{
load_ss_files(model_path = NA, printstats = FALSE, ...)
}
\arguments{
\item{model_path}{The directory that the model is in}

\item{printstats}{Print info on each model loaded via \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{...}{Passed to \code{calc_mcmc()}}
}
\value{
A model object representing the output from the SS model
}
\description{
Load all the SS files for output and input, and return the model object
}
\details{
If MCMC directory is present, load that and perform calculations
for mcmc parameters.
}
