% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-comps.R
\name{make_age_comp_bubble_plot}
\alias{make_age_comp_bubble_plot}
\title{Make an age composition bubble plot}
\usage{
make_age_comp_bubble_plot(model, subplot = 1, ...)
}
\arguments{
\item{model}{A model object as returned from [load.ss.files()]}

\item{subplot}{1 for fishery, 2 for survey}

\item{...}{Additional parameters passed to [plot_bubbles()]}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Make an age composition bubble plot
}
