% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-makebox.R
\name{makebox}
\alias{makebox}
\title{Make a box plot from a data frame of quantiles}
\usage{
makebox(
  data,
  col = c("gray", "colour"),
  lty = 1,
  labels = c("Year", "Depth (m)"),
  ...
)
}
\arguments{
\item{data}{A data frame of quantiles by year. The function
assumes that the depth data are in meters.}

\item{col}{The default is to use gray for all boxes, but
one can change to "colour" which will use custom colours
per year that are generated by \link{plotcolour}.}

\item{lty}{To change the line type, where the default is
\code{1} and leads to solid lines, where \code{2} would produce dashed
lines, etc. These lines are both for the whiskers and the
median that cuts the center of the box.}

\item{labels}{Character values for the horizontal and
vertical axes labels.}

\item{...}{Items passed to \link[graphics:bxp]{graphics::bxp}. Typically this will
include things such as \code{main}.}
}
\value{
A boxplot is printed to the screen. If you save the object
you will be disappointed because is just a vector of integers
the same length as the number of boxes in the figure; though
the figure is still generated.
}
\description{
Use \link[graphics:bxp]{graphics::bxp}, which is normally called by
\link[graphics:boxplot]{graphics::boxplot} to create a boxplot. Given the data
are provided directly in \code{data} there is no need to use
the wrapper and \link[graphics:bxp]{graphics::bxp} is called directly.
}
\author{
Kelli Faye Johnson
}
