% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{panel.cor}
\alias{panel.cor}
\title{Generate the correlation values as text for the panels}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{The x values for the correlations}

\item{y}{The y values for the correlations}

\item{digits}{The number of decimal places to round to in the panels}

\item{prefix}{A vector of text to paste to the beginning of the
correlation values}

\item{cex.cor}{the font size for the correlation text}
}
\description{
Generate the correlation values as text for the panels
}
