% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-bubbles.R
\name{plot_age_comp_bubbles}
\alias{plot_age_comp_bubbles}
\title{Make an age composition bubble plot}
\usage{
plot_age_comp_bubbles(
  model,
  type = c("fishery", "survey"),
  proportions = TRUE,
  inc_mean_age_line = TRUE,
  ...
)
}
\arguments{
\item{model}{A model object as returned from \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{Either \code{fishery} or \code{survey}}

\item{proportions}{Logical. If \code{TRUE}, If \code{TRUE} plot the input data
proportions-at-age. If \code{FALSE} plot the estimated median numbers-at-age}

\item{inc_mean_age_line}{Logical. If \code{TRUE}, plot the mean age line
over the bubbles}

\item{...}{Additional parameters passed to \code{\link[=plot_bubbles]{plot_bubbles()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make an age composition bubble plot
}
