% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-bubbles-data.R
\name{plot_age_comp_bubbles_data}
\alias{plot_age_comp_bubbles_data}
\title{Create a multi-panel plot of age proportion bubble plots, one for each
sector of the fishery}
\usage{
plot_age_comp_bubbles_data(
  age_data_lst,
  nms_vec,
  years = c(year(now()) - 6, year(now()) - 1)
)
}
\arguments{
\item{age_data_lst}{A list of age proportion data frames}

\item{nms_vec}{A vector of names for the sectors, must be the same length
as \code{age_data-lst}}

\item{years}{A vector of two values re0-presenting the minimum and maximum
year values to plot}

\item{ax_title_font_size}{The font size for the axis titles and strip
titles (sector names)}

\item{ax_tick_font_size}{The foint size for the axis tick labels
(years and ages)}

\item{ax_label_color}{The color for the axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a multi-panel plot of age proportion bubble plots, one for each
sector of the fishery
}
