% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-comp-comparison-bubbles.R
\name{plot_age_comp_comparison_bubbles}
\alias{plot_age_comp_comparison_bubbles}
\title{Create a plot of age proportions as bubbles with fishery and survey values
overlaid in one plot}
\usage{
plot_age_comp_comparison_bubbles(model, alpha = 0.3, remove_yr_labels = NULL)
}
\arguments{
\item{model}{A model object}

\item{alpha}{Transparanecy of the bubbles}

\item{remove_yr_labels}{A vector of years to remove the ,labels for in
case they are overlapping}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of age proportions as bubbles with fishery and survey values
overlaid in one plot
}
