% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-autocor.R
\name{plot_autocor}
\alias{plot_autocor}
\title{Create an autocorrelation plot for a particular parameter}
\usage{
plot_autocor(
  model,
  post_regex,
  lag_max = 20,
  y_lim = c(-1, 1),
  bar_width = 0.5,
  fill = main_fill,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{post_regex}{A regular expression that matches one parameter name
and nothing else}

\item{lag_max}{See \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}

\item{y_lim}{A vector of start and end limits for the y-axis}

\item{bar_width}{The width of the bars}

\item{fill}{The fill color for the bars}

\item{alpha}{The transparency for the bar fill}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create an autocorrelation plot for a particular parameter
}
