% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-comps.R
\name{plot_bubbles}
\alias{plot_bubbles}
\title{Make a bubble plot from the given data}
\usage{
plot_bubbles(
  d,
  clines = c(1980, 1984, 1999, 2010, 2014, 2016),
  mean_age = NULL,
  mean_age_line_color = "red",
  mean_age_line_size = 1.5,
  mean_age_line_type = "solid",
  yrs = NULL,
  by = 5,
  legend.position = "none",
  alpha = 0.3,
  xlim = c(1975, year(Sys.Date())),
  legend.title = "Proportion",
  ...
)
}
\arguments{
\item{d}{a \code{\link[tibble:tibble]{tibble::tibble()}} of the data in long format with column
names \code{Year}, \code{Age}, and \code{Proportion}}

\item{clines}{An optional vector of years to draw cohort lines through}

\item{mean_age}{A two-column tibble with columnn names \code{Year} and \code{Age} where
each row containns a year and \code{Age} represents the mean age for each year}

\item{mean_age_line_color}{The line color for the mean age line}

\item{mean_age_line_size}{The line thickness for the mean age line}

\item{mean_age_line_type}{The line type for the mean age line}

\item{yrs}{A vector of 2, for the years to show on the plot}

\item{by}{How many years between year labels on the x-axis}

\item{legend.position}{See the \code{legend.position} parameter in
\code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{alpha}{See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}

\item{xlim}{Limits for the x-axis}

\item{legend.title}{The legend title text}

\item{...}{Additional parameters passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}} and \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a bubble plot from the given data
}
