% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch-forecast-density.R
\name{plot_catch_forecast_density}
\alias{plot_catch_forecast_density}
\title{Create a plot showing the density of the posterior for the catch forecast
for a given year}
\usage{
plot_catch_forecast_density(
  model,
  yr,
  xlim = c(0, 3000),
  fill_color = ts_single_model_ribbon_fill,
  fill_alpha_tail = 0.3,
  fill_alpha_main = 0.7,
  label_x_offset = 500,
  label_y_offset = 1e-04,
  ...
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
\code{\link[=create_rds_file]{create_rds_file()}}.}

\item{yr}{Forecast year to plot. Must be a column in the MCMC output}

\item{xlim}{A vector of two values. The x-axis minimum and maximum values}

\item{fill_color}{The color used to fill under the density curve}

\item{fill_alpha_tail}{The transparency used for the color of the tails}

\item{fill_alpha_main}{The transparency used for the color of the main part
of the curve (-1SD - 1SD)}

\item{label_x_offset}{An x-axis offset to move the median label to,
Positive means move right}

\item{label_y_offset}{A y-axis offset to move the median label to,
Positive means move up}

\item{...}{Arguments to pass to \code{\link[=theme_axis_text]{theme_axis_text()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot showing the density of the posterior for the catch forecast
for a given year
}
