% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catches.R
\name{plot_catches}
\alias{plot_catches}
\title{Plot catch data as a stacked barplot}
\usage{
plot_catches(ct, mar = c(4, 4, 6, 2) + 0.1, leg.y.loc = 430, leg.cex = 1)
}
\arguments{
\item{ct}{The data frame which is read in from
\code{inst/extdata/data/landings-tac-history.csv}}

\item{mar}{See \code{\link[=par]{par()}}}

\item{leg.y.loc}{Y-position for legend}

\item{leg.cex}{Text size for legend}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot catch data as a stacked barplot
}
