% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cumulative-catch-cohorts.R
\name{plot_cumulative_catch_cohorts}
\alias{plot_cumulative_catch_cohorts}
\title{Creates a plot showing the cumulative catches for certain cohorts
as they age}
\usage{
plot_cumulative_catch_cohorts(
  model,
  cohorts = c(1999, 2010, 2014, 2016, 2020, 2021),
  leg_pos = c(0.7, 0.4),
  leg_ncol = 1
)
}
\arguments{
\item{model}{A model list, as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{cohorts}{A vector of years representing cohorts to how}

\item{leg_pos}{The position of the legend inside the plot. If \code{NULL},
\code{NA}, or "none", the legend will not be shown}

\item{leg_ncol}{The number of columns to show in the legend}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Creates a plot showing the cumulative catches for certain cohorts
as they age
}
