% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_catch.R
\name{plot_cumulative_catches}
\alias{plot_cumulative_catches}
\title{Create a plot of cumulative catches, with one panel for each fishery}
\usage{
plot_cumulative_catches(
  catch_lst,
  names_lst,
  yrs = c((year(now()) - 4), (year(now()) - 1)),
  scale = 1000,
  type = c("default", "proportion", "cumulative"),
  leg_pos = c(0.1, 0.7),
  leg_font_size = 8,
  line_width = ts_linewidth,
  line_gap = ts_linegap,
  point_shape = ts_pointshape,
  point_size = ts_pointsize,
  point_stroke = ts_pointstroke,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color
)
}
\arguments{
\item{catch_lst}{A list of data tables with either three columns: month,
year, and catch or 13 columns: Year, and one column each for the 12 months}

\item{names_lst}{A list of names for the panel titles in the plot. Must be
the same length as the \code{catch_lst} list}

\item{yrs}{A vector of two years representing the maximum and minimum years
to include in the data shown in the plot}

\item{scale}{A number to divide the catch by}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of cumulative catches, with one panel for each fishery
}
