% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{plot_depth}
\alias{plot_depth}
\title{Create a boxplot of depths for a given data frame}
\usage{
plot_depth(
  depth_df,
  yrs,
  title_text = NULL,
  line_col = "grey",
  box_width = 0.75,
  y_lim = NULL,
  y_breaks = NULL,
  reverse_y = TRUE,
  ...
)
}
\arguments{
\item{depth_df}{A data frame containing boxplot input, see files like
\verb{depth-us-*.csv}}

\item{yrs}{A vector of the years to show in the plot}

\item{title_text}{The title to show. If \code{NULL}, do not show the title}

\item{line_col}{The color of the median line and box outline}

\item{box_width}{With of each box}

\item{y_lim}{The vector of two value representing the y-limits for the plot}

\item{y_breaks}{The vector of y-axis values to show on the plot}

\item{reverse_y}{Logical. If \code{TRUE}, place the zero on top of the y-axis
and the depths in reverse order downward, to mirror the water column layout
in the real world}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a boxplot of depths for a given data frame
}
