% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-during-loading.R
\name{plot_during_loading}
\alias{plot_during_loading}
\title{Create plots during the loading phase, and return a list of them to be
attached to the model object for saving in a model RDS file}
\usage{
plot_during_loading(model)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A list of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects
}
\description{
Create plots during the loading phase, and return a list of them to be
attached to the model object for saving in a model RDS file
}
\details{
Uses the \link{probs} vector which is included in the package data for
this package
Primarily used for plots that directly access the \code{mcmc} output because
that is not included in the model object to reduce model object size
}
