% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fore-age-comps.R
\name{plot_fore_age_comps}
\alias{plot_fore_age_comps}
\title{Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)}
\usage{
plot_fore_age_comps(
  model,
  probs = c(0.025, 0.5, 0.975),
  x_lim = c(1, 15),
  x_breaks = seq(x_lim[1], x_lim[2], 2),
  y_lim = c(0, 0.4),
  bar_alpha = 0.3,
  whisker_width = 0.5,
  axis_title_font_size = 14,
  axis_tick_font_size = 12,
  axis_label_color = "black"
)
}
\arguments{
\item{model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{probs}{A vector of three quantiles for the lower CI,
median, and upper CI. The second value must be 0.5}

\item{x_lim}{A vector of 2 representing the minimum and maximum x-axis
values to plot}

\item{x_breaks}{A vector of values to show on the x-axis}

\item{y_lim}{A vector of 2 representing the minimum and maximum y-axis
values to plot}

\item{bar_alpha}{The transparency of the bars from 0 to 1}

\item{whisker_width}{The width of the top and bottom crossbars
(whiskers) on the error bars}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} object
}
\description{
Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)
}
