% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fore-age-comps.R
\name{plot_fore_age_comps}
\alias{plot_fore_age_comps}
\title{Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)}
\usage{
plot_fore_age_comps(
  model,
  x_lim = c(1, 15),
  x_breaks = seq(x_lim[1], x_lim[2], 2),
  y_lim = c(0, 0.55),
  whisker_width = 0.5,
  point_size = 2,
  glow_size = 2.5
)
}
\arguments{
\item{model}{A model object, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{x_lim}{A vector of 2 representing the minimum and maximum x-axis
values to plot}

\item{x_breaks}{A vector of values to show on the x-axis}

\item{y_lim}{A vector of 2 representing the minimum and maximum y-axis
values to plot}

\item{whisker_width}{The width of the top and bottom crossbars
(whiskers) on the error bars}

\item{point_size}{The size of the median points}

\item{glow_size}{The size of the white glow behind the points (must be
larger than \code{point_size} to be seen)}

\item{probs}{A vector of three quantiles for the lower CI,
median, and upper CI. The second value must be 0.5}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} object
}
\description{
Create a two-panel plot showing the forecasted age compositions by
numbers and weight for the first forecast year (\code{endyr} + 1)
}
