% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-management-catch-vs-tac.R
\name{plot_management_catch_vs_tac}
\alias{plot_management_catch_vs_tac}
\title{Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch}
\usage{
plot_management_catch_vs_tac(
  d,
  curr_assess_biomass = NULL,
  connect_vals_linetype = "dashed",
  connect_vals_color = "darkgrey",
  connect_vals_alpha = 0.5,
  connect_vars = FALSE,
  connect_vars_linetype = "dashed",
  connect_vars_alpha = 0.5,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12
)
}
\arguments{
\item{d}{The data as read in using \code{\link[readr:read_delim]{readr::read_csv()}} from the file
"catch-targets.csv"}

\item{curr_assess_biomass}{Current year's assessment-estimated biomass
(or any value). If \code{NULL} it will not be shown}

\item{connect_vals_linetype}{Which linetype to use for connecting the
values with vertical lines}

\item{connect_vals_color}{Which alpha level (0-1) to use for connecting
the values with vertical lines}

\item{connect_vars}{Logical. Connent the TACs and realized catches to
each other}

\item{connect_vars_linetype}{If \code{connect_vars} is TRUE, which line type
to use}

\item{connect_vars_alpha}{If \code{connect_vars} is TRUE, which alpha level
(0-1) to use}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch
}
