% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-management-catch-vs-tac.R
\name{plot_management_catch_vs_tac}
\alias{plot_management_catch_vs_tac}
\title{Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch}
\usage{
plot_management_catch_vs_tac(
  d,
  curr_assess_biomass = NULL,
  line_type = "solid",
  line_width = 1,
  line_alpha = 0.5,
  leg_pos = c(0.65, 0.83),
  leg_ncol = 1,
  leg_font_size = 12,
  point_size = 3,
  ret_tbl = FALSE,
  font_size = 10,
  right_cols_cm = 1.8,
  y_breaks = seq(0, 1000, 100),
  ...
)
}
\arguments{
\item{d}{The data as read in using \code{\link[readr:read_delim]{readr::read_csv()}} from the file
"catch-targets.csv"}

\item{curr_assess_biomass}{Current year's assessment-estimated biomass
(or any value). If \code{NULL} it will not be shown}

\item{line_type}{Which linetype to use for connecting the
values with vertical lines}

\item{line_width}{Width of the connecting lines}

\item{line_alpha}{Which alpha level (0-1) to use for connecting
the values with vertical lines}

\item{leg_pos}{The legend position as x-y coordinates (vector of 2)}

\item{leg_ncol}{The number of columns to split the legend into}

\item{leg_font_size}{The legend font size}

\item{point_size}{The point size}

\item{ret_tbl}{Logical. If \code{TRUE}, return a \code{\link[kableExtra:kbl]{kableExtra::kbl()}} containing
the outputs instead of the \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object}

\item{font_size}{The table data and header font size in points}

\item{right_cols_cm}{The number of centimeters wide to make all of the
rightmost columns (all the value columns)}

\item{y_breaks}{A vector of values to show on the y-axis}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a time series plot of the management-chosen TAC, assessment
estimated TAC, and realized catch
}
