% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_map}
\alias{plot_map}
\title{Create a map of a part of the west coast of North America,
including bathymetry raster and contours}
\usage{
plot_map(
  crs_ll = 4326,
  x_lim = c(-140, -113),
  y_lim = c(33, 58),
  coast_line_color = "black",
  coast_line_thickness = 0.1,
  coast_fill = "grey",
  show_bathy_contours = TRUE,
  show_bathy_raster = TRUE,
  bathy_resolution = 1,
  bathy_alpha = 0.8,
  bathy_contours = c(100, 200, 500, 1000, 1500),
  bathy_contour_color = "black",
  bathy_contour_thickness = 0.25,
  bathy_raster_scale = scale_fill_continuous(type = "gradient")
)
}
\arguments{
\item{crs_ll}{Coordinate Reference System (CRS) number for a
latitude/longitude-based projection. This could be NAD83 or WGS84 or
others. The default is 4326 which is WGS84: See
\href{https://epsg.org/home.html}{Epsg.org} for details. Click \verb{Text search}
tab and look at the \code{code} column for valid crs numbers}

\item{x_lim}{The length-2 vector representing the minimum and maximum
limits of the x-axis in degrees of longitude}

\item{y_lim}{The length-2 vector representing the minimum and maximum
limits of the y-axis in degrees of latitude}

\item{coast_line_color}{The line color for the coastline}

\item{coast_line_thickness}{The line thickness for the coastline}

\item{coast_fill}{The fill color for the land areas}

\item{show_bathy_contours}{Logical. If \code{TRUE} draw the bathymetry contour
lines}

\item{show_bathy_raster}{Logical. If \code{TRUE} draw the bathymetry raster
layer}

\item{bathy_resolution}{The resolution for the bathymetry in minutes.
See \code{\link[marmap:getNOAA.bathy]{marmap::getNOAA.bathy()}}}

\item{bathy_alpha}{The transparency for the bathymetric raster layer}

\item{bathy_contours}{A vector of contours to show contour lines for.
These can be either positive or negative; they will all be made negative
in the function}

\item{bathy_contour_color}{The line color for the bathymetric contours}

\item{bathy_contour_thickness}{The line thickness for the bathymetric
contours}

\item{bathy_raster_scale}{A \verb{ggplot2::scale_*} function describing the
color gradient to use for the bathymetry raster. For example:
\code{scale_fill_gradientn(colors = c("red", "orange", "yellow"))}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a map of a part of the west coast of North America,
including bathymetry raster and contours
}
