% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mcmc-diagnostics-all-params.R
\name{plot_mcmc_diagnostics_all_params}
\alias{plot_mcmc_diagnostics_all_params}
\title{Creates the plot showing the 4 MCMC diagnostics for  all parameters
combined}
\usage{
plot_mcmc_diagnostics_all_params(
  model,
  diag_fill = "royalblue",
  diag_alpha = 0.3,
  diag_outline_color = "black",
  ro_arrow_lengths = rep(20, 4),
  ro_text_nudges = rep(0, 4),
  y_lim_geweke = c(0, 30),
  ...
)
}
\arguments{
\item{model}{A model object as output by \code{\link[=create_rds_file]{create_rds_file()}}.}

\item{diag_fill}{The fill color for the bars in the plots}

\item{diag_alpha}{The transparency for the bars in the plots}

\item{diag_outline_color}{The bar outline color for the bars in the plots}

\item{ro_arrow_lengths}{A vector of 4 values to use for the arrow lengths
in the 4 plots. Note the plot orders in this vector is:
top left, top right bottom left bottom right}

\item{ro_text_nudges}{A vector of 4 for the R0 text left/right nudges.
See \code{\link[=plot_mcmc_histogram]{plot_mcmc_histogram()}}}

\item{y_lim_geweke}{A y limit vector of two values for the Geweke plot
only}

\item{...}{Arguments passed to \code{\link[=calc_mcmc_param_stats]{calc_mcmc_param_stats()}} and
\code{\link[=plot_mcmc_histogram]{plot_mcmc_histogram()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Creates the plot showing the 4 MCMC diagnostics for  all parameters
combined
}
