% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pearson-bubbles.R
\name{plot_pearson_bubbles}
\alias{plot_pearson_bubbles}
\title{Make a bubble plot from the given data}
\usage{
plot_pearson_bubbles(
  model,
  type = 1,
  clines = c(1980, 1984, 1999, 2010, 2014, 2016),
  by = 5,
  legend.position = "none",
  alpha = 0.3,
  xlim = c(1975, year(Sys.Date())),
  ...
)
}
\arguments{
\item{type}{1 = Fishery, any other value = Survey}

\item{clines}{An optional vector of years to draw cohort lines through}

\item{by}{How many years between year labels on the x-axis}

\item{legend.position}{See [ggplot2::theme(legend.position)]}

\item{alpha}{See [ggplot2::geom_point()]}

\item{xlim}{Limits for the x-axis}

\item{...}{Additional parameters passed to [ggplot2::geom_point()],
[ggplot2::geom_segment()] and [ggplot2::theme()]}

\item{d}{a [tibble::tibble()] of the data in long format with column names `Year`, `Age`, and `Proportion`}

\item{yrs}{A vector of 2, for the years to show on the plot}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Make a bubble plot from the given data
}
