% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pearson-bubbles.R
\name{plot_pearson_bubbles}
\alias{plot_pearson_bubbles}
\title{Make a bubble plot from the given data}
\usage{
plot_pearson_bubbles(
  model,
  type = c("fishery", "survey"),
  clines = age_bubble_cohorts,
  tick_prop = 1,
  vjust_x_labels = -1.5,
  remove_yr_labels = NULL,
  x_labs_mod = 5,
  leg_pos = "none",
  alpha = 0.3,
  xlim = c(1975, year(Sys.Date())),
  point_alpha = main_alpha,
  point_fill = main_fill,
  point_color = "black",
  diag_line_color = age_diag_line_color,
  diag_line_width = age_diag_line_width,
  diag_line_type = age_diag_line_type,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{1 = Fishery, any other value = Survey}

\item{clines}{An optional vector of years to draw cohort lines through}

\item{tick_prop}{A value that the length of the major tick marks are
multiplied by. This proportion must be set by trial and error. Make sure
to change \code{vjust_x_labels} so the labels are not overlapping the lines or
are too far away from the lines}

\item{vjust_x_labels}{Adjustment to move the x-axis tick labels and label
up or down. Negative numbers move down}

\item{remove_yr_labels}{A vector of years to remove the ,labels for in
case they are overlapping}

\item{x_labs_mod}{How many years between year labels on the x-axis}

\item{leg_pos}{See the \code{leg_pos} parameter of
\code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{alpha}{See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}

\item{xlim}{Limits for the x-axis}

\item{...}{Additional parameters passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}} and \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Make a bubble plot from the given data
}
