% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-phase.R
\name{plot_phase}
\alias{plot_phase}
\title{Create plot of the relative fishing intensity in year t-1 against relative
spawning biomass in year t, as a historical look at the fishery for the
MCMC given by model}
\usage{
plot_phase(
  model,
  start_yr,
  end_yr,
  x_lim = c(0, 1.5),
  y_lim = c(0, 1.2),
  init_lbl_x_off = 0,
  init_lbl_y_off = 0,
  final_lbl_x_off = 0,
  final_lbl_y_off = 0,
  cross_line_width = 1,
  cross_line_color = "grey",
  title_y_font_size = axis_title_font_size,
  detail_b40_outliers = FALSE,
  detail_fspr_outliers = FALSE
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{start_yr}{The minimum year in the plot}

\item{end_yr}{The maximum year in the plot}

\item{x_lim}{Vector of two values for minimum and maximum x-axis values}

\item{y_lim}{Vector of two values for minimum and maximum y-axis values}

\item{init_lbl_x_off}{Initial year label x offset in units of axis}

\item{init_lbl_y_off}{Initial year label y offset in units of axis}

\item{final_lbl_x_off}{Final year label x offset in units of axis}

\item{final_lbl_y_off}{Final year label y offset in units of axis}

\item{detail_b40_outliers}{Shade the part of the plot less that B40,
plot the missing points at the lower 2.5\% of the depletion distribution,
draw a box around those points and show a label with the description}

\item{detail_fspr_outliers}{Shade the part of the plot greater than
FSPR = 1, plot the missing points at the upper 2.5\% of the FSPR
distribution, draw a box around those points and show a label with the
description}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create plot of the relative fishing intensity in year t-1 against relative
spawning biomass in year t, as a historical look at the fishery for the
MCMC given by model
}
