% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mcmc-running-quants.R
\name{plot_running_quants}
\alias{plot_running_quants}
\title{Create a plot with running quantiles across the posterior for a parameter}
\usage{
plot_running_quants(model, post_regex, rib_alpha = 0.2, y_lim = NULL)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{post_regex}{A regular expression that matches one parameter name
and nothing else}

\item{rib_alpha}{The alpha value for the credible interval ribbon}

\item{y_lim}{The limits for the y-axis. A vector of 2}

\item{probs}{A vector of 3 values for the lower, median, and upper
quantiles}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot with running quantiles across the posterior for a parameter
}
