% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-posteriors.R
\name{plot_selex_posteriors}
\alias{plot_selex_posteriors}
\title{Plot the selectivity parameter estimates from all posteriors
as many lines with an overlaid errorbar for each age}
\usage{
plot_selex_posteriors(
  model = NULL,
  type = c("fishery", "survey"),
  age_range = NULL,
  show_xlab = TRUE,
  post_line_width = 0.1,
  post_line_color = "black",
  post_line_alpha = 0.1,
  post_med_line_width = 1,
  post_med_line_color = "blue3",
  post_med_line_alpha = 1,
  post_line_gap = ts_linegap,
  post_med_point_size = 3,
  unc_line_width = 1,
  unc_line_color = "blue3",
  unc_line_alpha = 1,
  unc_point_size = 3,
  glow = FALSE,
  glow_offset = 0.5,
  glow_color = "white",
  glow_alpha = 1,
  ax_title_font_size = axis_title_font_size,
  ax_tick_font_size = axis_tick_font_size,
  ax_label_color = axis_label_color
)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{One of 1 or 2 where 1 = Fishery and 2 = Acoustic age 2+ survey}

\item{age_range}{A vector of two for the min and max ages to plot. If \code{NULL},
all ages in the data will be plotted}

\item{show_xlab}{Logical. If \code{TRUE}, show the X-axis label. If \code{FALSE},
omit it}

\item{post_line_width}{Thickness of the posterior lines}

\item{post_line_color}{Color of the posterior lines}

\item{post_line_alpha}{Transparency of the posterior lines}

\item{post_med_line_width}{Thickness of the median posterior line}

\item{post_med_line_color}{Color of the median posterior line}

\item{post_med_line_alpha}{Transparency of the median posterior line}

\item{post_med_point_size}{Size of the points on the median posterior line}

\item{unc_line_width}{Thickness of the uncertainty lines}

\item{unc_line_color}{Color of the uncertainty lines and points}

\item{unc_line_alpha}{Transparency for the uncertainty lines and points}

\item{unc_point_size}{Size of the points}

\item{glow}{Logical. If \code{TRUE}, add a white glow around the lines so that
they can more easily be seen in contrast to the background posterior lines}

\item{glow_offset}{The amount to add to the thickness of the lines and
diameter of the points to create the glow effect}

\item{glow_color}{Color of the glow effect}

\item{glow_alpha}{Transparency of the glow effect}

\item{ax_title_font_size}{Size of the font for the X and Y axis labels}

\item{ax_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{ax_label_color}{Color of the font for the X and Y axis tick and
title labels}

\item{probs}{A vector of two quantiles to include for the uncertainty}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the selectivity parameter estimates from all posteriors
as many lines with an overlaid errorbar for each age
}
