% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-squid.R
\name{plot_squid}
\alias{plot_squid}
\title{Make the retrospective recruitment deviations plot (squid plot)}
\usage{
plot_squid(
  model,
  relative = FALSE,
  show_ci = FALSE,
  ci_alpha = 0.2,
  ci_yrs = NULL,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  axis_label_color = "black",
  year_label_font_size = 4
)
}
\arguments{
\item{model}{Model with retrospectives}

\item{relative}{Logical. If \code{TRUE}, plot the deviaitons relative to the
most recent estimate for each cohort}

\item{show_ci}{Logical. If \code{TRUE}, plot the credible interval ribbons
around the median lines}

\item{ci_alpha}{The a transparency value for the credible interval
ribbon fill}

\item{ci_yrs}{A vector of years to include credible intervals for.
If \code{NULL}, all will be shown. Only used if \code{show_ci} is \code{TRUE}}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}

\item{year_label_font_size}{Size of the font for the year labels}
}
\description{
Make the retrospective recruitment deviations plot (squid plot)
}
