% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-squid.R
\name{plot_squid}
\alias{plot_squid}
\title{Make the retrospective recruitment deviations plot (squid plot)}
\usage{
plot_squid(
  model,
  relative = FALSE,
  show_ci = FALSE,
  ci_alpha = 0.2,
  ci_yrs = NULL,
  year_label_font_size = 4,
  y_lim = c(NA, NA),
  x_lim = c(NA, NA),
  surv_point_type = 17,
  reg_point_type = 19,
  cohorts = NULL,
  color_offset = 0
)
}
\arguments{
\item{model}{Model with retrospectives}

\item{relative}{Logical. If \code{TRUE}, plot the deviaitons relative to the
most recent estimate for each cohort}

\item{show_ci}{Logical. If \code{TRUE}, plot the credible interval ribbons
around the median lines}

\item{ci_alpha}{The a transparency value for the credible interval
ribbon fill}

\item{ci_yrs}{A vector of years to include credible intervals for.
If \code{NULL}, all will be shown. Only used if \code{show_ci} is \code{TRUE}}

\item{year_label_font_size}{Size of the font for the year labels}

\item{y_lim}{A vector to specify y limits, default is to span what is being
plotted. If \code{full} then automatically span all the
credible intervals even when not showing them all (useful
for showing sequential plots in a talk). Not tested with \code{relative = TRUE}
and \code{show_ci = TRUE} as we do not currently use those.}

\item{x_lim}{A vector to specify x limits, default is to span what is being
plotted.}

\item{surv_point_type}{The point shape type for ages in age-1 index years}

\item{reg_point_type}{The point shape type for ages in non-age-1-index years}

\item{cohorts}{A vector of years so as to plot only those cohorts if available (useful for
talks). If \code{NULL} (the default) plots all available cohorts.}

\item{color_offset}{A value to offset the colors to compared to the ten
retrospective years. This may be used when comparing model where only five
retrospective years were run with those that have ten.}
}
\description{
Make the retrospective recruitment deviations plot (squid plot)
}
