% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-stock-recruitment.R
\name{plot_stock_recruitment}
\alias{plot_stock_recruitment}
\title{Create a plot of the stock-recruitment relationship with
estimated medians and credible intervals by year overlaid. An expected
distribution of absolute recruitment is also plotted}
\usage{
plot_stock_recruitment(
  model,
  probs = hake::probs,
  xlim = c(0, 1.4),
  ylim = c(0, 7)
)
}
\arguments{
\item{model}{The model output from Stock Synthesis as loaded by
\code{\link[=create_rds_file]{create_rds_file()}}}

\item{probs}{A vector of three probabilities to use in the call to
\code{\link[stats:quantile]{stats::quantile()}}}

\item{xlim}{A vector of two values. The x-axis minimum and maximum values}

\item{ylim}{A vector of two values. The y-axis minimum and maximum values}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of the stock-recruitment relationship with
estimated medians and credible intervals by year overlaid. An expected
distribution of absolute recruitment is also plotted
}
