% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survey-biomass.R
\name{plot_survey_biomass}
\alias{plot_survey_biomass}
\title{Create a plot of survey biomass as points with error   bars}
\usage{
plot_survey_biomass(
  model,
  index = c("age1", "age2"),
  x_lim = c(survey_start_yr, survey_end_yr),
  y_lim = c(0, 3),
  x_labs_mod = 5,
  alpha = 0.3,
  tick_prop = 1,
  vjust_x_labels = -1.5,
  remove_yr_labels = NULL,
  point_size = ts_single_model_pointsize,
  point_color = ts_single_model_pointcolor,
  point_shape = ts_single_model_pointshape,
  point_stroke = ts_single_model_pointstroke,
  line_width = ts_single_model_linewidth,
  line_color = ts_single_model_linecolor,
  line_type = ts_single_model_linetype,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{index}{The index type to plot}

\item{y_lim}{A vector of two for the minimum and maximum values
for the y-axis on the plot}

\item{x_labs_mod}{How many years between year labels on the x-axis}

\item{alpha}{The transparency for the error bars (non-squid). Transparency
for the squid error bars are 1}

\item{tick_prop}{A value that the length of the major tick marks are
multiplied by. This proportion must be set by trial and error. Make sure
to change \code{vjust_x_labels} so the labels are not overlapping the lines or
are too far away from the lines}

\item{vjust_x_labels}{Adjustment to move the x-axis tick labels and label
up or down. Negative numbers move down}

\item{remove_yr_labels}{A vector of years to remove the ,labels for in
case they are overlapping}

\item{point_size}{The size of the points (See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for
more details)}

\item{point_color}{The R color for the points (See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
for more details)}

\item{point_shape}{The R shape number for the points (See
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for more details)}

\item{point_stroke}{The stroke value for the points (See
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for more details)}

\item{line_width}{Width of the error bar lines}

\item{line_color}{R color for the error bar lines}

\item{line_type}{Line type for the error bar lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of survey biomass as points with error   bars
}
\details{
Extra SD is shown for the 2009 "squid year"
}
