% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survey-fit-mcmc.R
\name{plot_survey_fit_mcmc}
\alias{plot_survey_fit_mcmc}
\title{Creates a plot of fits to survey biomass including individual posterior
lines, median posterior line, observed values, and estimated values}
\usage{
plot_survey_fit_mcmc(
  model,
  type = c("age1", "acoustic"),
  n_posts = NULL,
  show_legend = TRUE,
  ylim = c(0, 5),
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  axis_label_color = "black",
  post_line_col = "black",
  post_line_alpha = 0.1,
  post_med_color = "darkblue",
  post_med_line_width = 1,
  post_med_point_size = 1,
  obs_color = "blue",
  obs_line_width = 1,
  obs_point_size = 3,
  extrasd_line_color = "blue",
  extrasd_point_color = "blue",
  extrasd_point_size = 3,
  extrasd_line_width = 1,
  glow = TRUE,
  glow_offset = 0.5,
  glow_col_err = "white",
  glow_col_med = "white",
  glow_alpha_err = 1,
  glow_alpha_med = 1
)
}
\arguments{
\item{model}{A model object, , created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{One of the surveys, either \code{age1} or \code{acoustic}}

\item{n_posts}{The number of posterior lines to plot. The lines are
randomly drawn from all posteriors. if this is larger than the number
of posteriors, all posterior lines will be shown}

\item{show_legend}{Logical. If \code{TRUE}, show the legend}

\item{ylim}{A vector of two values representing the minimum and maximum
values to appear on the y-axis}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}

\item{post_line_col}{The color of the thinner individual posterior lines}

\item{post_line_alpha}{The transparency of the thinner individual
posterior lines}

\item{post_med_color}{The color of the posterior median line and points}

\item{post_med_line_width}{The width of the posterior median line}

\item{post_med_point_size}{The size of the posterior median points}

\item{obs_color}{the color for the uncertainty bars for the observed
values}

\item{obs_line_width}{The width of the error bar lines for observed
values}

\item{obs_point_size}{The size of the points on the error bars for the
observed values}

\item{extrasd_line_color}{The color of the error bars for the
extra SD parameter estimates}

\item{extrasd_point_color}{The color of the median points for the
extra SD parameter estimates}

\item{extrasd_point_size}{The sizeof the median points for the
extra SD parameter estimates}

\item{extrasd_line_width}{The width of the error bar lines for the
extra SD parameter estimates}

\item{glow}{Logical. If \code{TRUE}, add a white glow around the lines so that
they can more easily be seen in contrast to the background posterior lines}

\item{glow_offset}{The amount to add to the lines and points to create the
glow effect}

\item{glow_col_err}{The color of the 'glow' on the error bar lines}

\item{glow_col_med}{The color of the 'glow' on the posterior median line}

\item{glow_alpha_err}{The transparency of the 'glow' on the error bars
and points}

\item{glow_alpha_med}{The transparency of the 'glow' on the posterior
median line}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Creates a plot of fits to survey biomass including individual posterior
lines, median posterior line, observed values, and estimated values
}
\details{
Uses the \link{probs} vector which is included in the package data for
this package
}
