% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survey-fit-mcmc.R
\name{plot_survey_fit_mcmc}
\alias{plot_survey_fit_mcmc}
\title{Creates a plot of fits to survey biomass including individual posterior
lines and the median posterior line}
\usage{
plot_survey_fit_mcmc(
  model,
  type = c("age1", "acoustic"),
  n_posts = 1000,
  probs = c(0.025, 0.5, 0.975),
  show_legend = TRUE,
  leg_ncol = 1,
  leg_font_size = 12,
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  axis_label_color = "black"
)
}
\arguments{
\item{model}{A model object, , created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{One of the surveys, either \code{age1} or \code{acoustic}}

\item{n_posts}{The number of posterior lines to plot. The lines are
randomly drawn from all posteriors. if this is larger than the number
of posteriors, all posterior lines will be shown}

\item{probs}{A vector of three probabilities for the credible interval,
with the middle one being the median (0.5)}

\item{show_legend}{Logical. If \code{TRUE}, show the legend}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_label_color}{Color for the axis labels and tick labels}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Creates a plot of fits to survey biomass including individual posterior
lines and the median posterior line
}
