% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Create a trace plot for a particular parameter}
\usage{
plot_trace(model, post_regex, color = "blue", alpha = 1)
}
\arguments{
\item{model}{A model list, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{post_regex}{A regular expression that matches one parameter name
and nothing else}

\item{color}{The color for the trace lines}

\item{alpha}{The transparency for the trace lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a trace plot for a particular parameter
}
