% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weight-at-age.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Create a plot of weight-at-age lines by age and year}
\usage{
plot_weight_at_age(
  wa,
  ages = NULL,
  bold_ages = NULL,
  cols = c("purple", "darkblue", "yellow", "darkgreen"),
  axis_title_font_size = 14,
  axis_tick_font_size = 11,
  axis_text_color = "black",
  age_label_font_size = 4
)
}
\arguments{
\item{wa}{A data frame of weight-at-age with a \code{yr} column and
numerically-named columns for ages}

\item{ages}{A vector of ages to plot. If \code{NULL}, all ages in the \code{wa}
data frame will be plotted}

\item{bold_ages}{A vector of ages to make bold in the plot. The lines for
these ages will be thicker and the labels for them will be bold as well}

\item{cols}{A vector of colors to base the color ramp on. See
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}

\item{axis_title_font_size}{Size of the font for the X and Y axis labels}

\item{axis_tick_font_size}{Size of the font for the X and Y axis tick labels}

\item{axis_text_color}{The color of the tick labels and axis labels}

\item{age_label_font_size}{The font size for the labels pointing to the age
lines}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of weight-at-age lines by age and year
}
