% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weight-at-age.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Create a plot of weight-at-age lines by age and year}
\usage{
plot_weight_at_age(
  wa,
  ages = NULL,
  bold_ages = NULL,
  cols = c("purple", "darkblue", "yellow", "darkgreen"),
  x_labels_mod = 5,
  age_label_font_size = 4,
  age_label_side = "right",
  ...
)
}
\arguments{
\item{wa}{A data frame of weight-at-age with a \code{yr} column and
numerically-named columns for ages}

\item{ages}{A vector of ages to plot. If \code{NULL}, all ages in the \code{wa}
data frame will be plotted}

\item{bold_ages}{A vector of ages to make bold in the plot. The lines for
these ages will be thicker and the labels for them will be bold as well}

\item{cols}{A vector of colors to base the color ramp on. See
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}

\item{x_labels_mod}{Show every nth year where n is this number}

\item{age_label_font_size}{The font size for the labels pointing to the age
lines}

\item{age_label_side}{Where to place the age labels next to the lines.
One of "none", "both", "left", or "right"}

\item{...}{Arguments passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of weight-at-age lines by age and year
}
