% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\title{Post-process the TEX file output by Bookdown}
\usage{
post_process(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Arguments passed to \code{\link[=post_process_table_of_contents]{post_process_table_of_contents()}},
\code{\link[=post_process_set_latex_placement_options]{post_process_set_latex_placement_options()}}, and \link{post_process_longtables}}
}
\value{
Nothing, overwrite the file \code{fn} with the modified TEX
}
\description{
Post-process the TEX file output by Bookdown
}
\details{
Called by \code{\link[=hake_pdf]{hake_pdf()}} to post-process the LaTeX compiled by the
\link{bookdown} package
}
