% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\title{Post-process the TEX file output by Bookdown}
\usage{
post_process(
  fn,
  section_indent_inch = 0,
  section_num_width_inch = 0.1,
  subsection_indent_inch = 0.2,
  subsection_num_width_inch = 0.1,
  subsubsection_indent_inch = 0.4,
  subsubsection_num_width_inch = 0.1
)
}
\arguments{
\item{fn}{The TEX filename}

\item{section_indent_inch}{Indent value in inches for section headers in
the TOC}

\item{section_num_width_inch}{Width of section numbers in inches in
the TOC}

\item{subsection_indent_inch}{Indent value in inches for subsection headers
in the TOC}

\item{subsection_num_width_inch}{Width of subsection numbers in inches in
the TOC}

\item{subsubsection_indent_inch}{Indent value in inches for subsubsection
headers in the TOC}

\item{subsubsection_num_width_inch}{Width of subsubsection numbers in inches
in the TOC}
}
\value{
Nothing, overwrite the file \code{fn} with the modified TEX
}
\description{
Post-process the TEX file output by Bookdown
}
\details{
Add centering for section names,
}
