% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\title{Post-process the TEX file output by Bookdown}
\usage{
post_process(x, prepub = FALSE, accessibility_tagging = FALSE, ...)
}
\arguments{
\item{accessibility_tagging}{Logical. If \code{TRUE}, include lines for
LaTeX \code{tagpdf} package to make a more web-accessible document. This will
make the compile time much longer and is typically used when done the
document and doing the final compile for government web distribution}

\item{...}{Arguments passed to \code{\link[=post_process_table_of_contents]{post_process_table_of_contents()}}}
}
\value{
Nothing, overwrite the file \code{fn} with the modified TEX
}
\description{
Post-process the TEX file output by Bookdown
}
\details{
Add centering for section names,
}
