% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\title{Post-process the TEX file output by Bookdown}
\usage{
post_process(x, ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{...}{Arguments passed to all the post-processing functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Post-process the TEX file output by Bookdown
}
\details{
Called by \code{\link[=hake_pdf]{hake_pdf()}} to post-process the LaTeX compiled by the
\link{bookdown} package
}
