% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-add-alt-text.R
\name{post_process_add_alt_text}
\alias{post_process_add_alt_text}
\title{Add alternative text to figures in the document via the
\code{pdfmanagement-testphase} LaTeX package}
\usage{
post_process_add_alt_text(
  x,
  accessible_pdf = NULL,
  title_page_image = NULL,
  title_page_image_width_cm = NULL,
  figures_dir = NULL,
  knitr_figures_dir = NULL,
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{accessible_pdf}{Logical. If \code{TRUE}, inject the code needed to include
the `pdfmanagement-testphase`` package and add alternative text}

\item{figures_dir}{The subdirectory of the "doc" directory containing
images that have previously been made such as pictures and logos, and all
other figures made outside the scope of this project}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Add alternative text to figures in the document via the
\code{pdfmanagement-testphase} LaTeX package
}
\details{
Searches the given TEX file for the \verb{includegraphics\{\}},
finds the correct alternative text for those calls by searching through
the project RMD files for the matching \code{(ref:variable-alt)} definitions
and injects the text and the \code{pdfmanagement-testphase} structural
LaTeX lines for web accessibility
}
