% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-longtables.R
\name{post_process_longtables}
\alias{post_process_longtables}
\title{Post-process LaTeX longtables in TEX code}
\usage{
post_process_longtables(x, cont_page_just = c("l", "c", "r"), ...)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{cont_page_just}{The justification for the continue page lines.
Must be one of "l", "c", or "r"}

\item{...}{Absorbs arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Post-process LaTeX longtables in TEX code
}
\details{
Injects some LaTeX code which adds 'Continued on next page ...' and
'... Continued from previous page' to the longtables that split pages.
Removes the caption for the subsequent pages, but leaves the header row.
Assumes that the table TeX was created using the following syntax:
\code{kable(longtable = TRUE) |> kable_styling(latex_options = c("repeat_header"))}
If \code{repeat_header} is not used to create the TeX code an error will be
thrown which will explain that \verb{\\\\endfirsthead} cannot be found.
}
