% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-modify-link-colors.R
\name{post_process_modify_link_colors}
\alias{post_process_modify_link_colors}
\title{Modify reference link colors in the document}
\usage{
post_process_modify_link_colors(
  x,
  underline_links = NULL,
  color_links = NULL,
  toc_underline_link_color = NULL,
  toc_link_text_color = NULL,
  cite_underline_link_color = NULL,
  cite_link_text_color = NULL,
  url_underline_link_color = NULL,
  url_link_text_color = NULL,
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{underline_links}{Logical. If \code{TRUE} make all the links in the
TOC and the section links in the document text be underlined}

\item{color_links}{Logical. If \code{TRUE} make all the links in the
TOC and the section links in the document text be colored}

\item{toc_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color. See LaTeX package \code{xcolor} for allowable
colors}

\item{toc_link_text_color}{If \code{underline_links} is \code{FALSE}, this color
will be the color of the link text (without underlines). See LaTeX package
\code{xcolor} for allowable colors}

\item{cite_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color for citations}

\item{cite_link_text_color}{Text color for citation links. See
\code{toc_link_text_color}}

\item{url_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color for URLs}

\item{url_link_text_color}{Text color for URL links. See
\code{toc_link_text_color}}

\item{...}{Absorb arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Modify reference link colors in the document
}
