% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-table-of-contents.R
\name{post_process_table_of_contents}
\alias{post_process_table_of_contents}
\title{Insert the table of contents into TeX code}
\usage{
post_process_table_of_contents(
  x,
  toc_depth = 2,
  toc_section_indent_inch = 0,
  toc_section_num_width_inch = 0.2,
  toc_subsection_indent_inch = 0.3,
  toc_subsection_num_width_inch = 0.35,
  toc_subsubsection_indent_inch = 0.5,
  toc_subsubsection_num_width_inch = 0.45,
  underline_links = FALSE,
  toc_underline_link_color = "blue",
  toc_link_text_color = "blue",
  cite_underline_link_color = "blue",
  cite_link_text_color = "blue",
  url_underline_link_color = "blue",
  url_link_text_color = "blue",
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{toc_depth}{The depth of the entries in the table of contents:
2 means 1.2, 3 means 1.2.3, 4 means 1.2.3.4, etc}

\item{toc_section_indent_inch}{Indent value in inches for section headers in
the TOC}

\item{toc_section_num_width_inch}{Width of section numbers in inches in
the TOC}

\item{toc_subsection_indent_inch}{Indent value in inches for subsection headers
in the TOC}

\item{toc_subsection_num_width_inch}{Width of subsection numbers in inches in
the TOC}

\item{toc_subsubsection_indent_inch}{Indent value in inches for subsubsection
headers in the TOC}

\item{toc_subsubsection_num_width_inch}{Width of subsubsection numbers in inches
in the TOC}

\item{underline_links}{Logical. If \code{TRUE} make all the ljnks in the
TOC and the section links in the document text be underlined}

\item{toc_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color. See LaTeX package \code{xcolor} for allowable
colors}

\item{toc_link_text_color}{If \code{underline_links} is \code{FALSE}, this color
will be the color of the link text (without underlines). See LaTeX package
\code{xcolor} for allowable colors}

\item{cite_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color for citations}

\item{cite_link_text_color}{Text color for citation links. See
\code{toc_link_text_color}}

\item{url_underline_link_color}{If \code{underline_links} is \code{TRUE}, this
color will be the underline color for URLs}

\item{url_link_text_color}{Text color for URL links. See
\code{toc_link_text_color}}

\item{...}{Absorb arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Insert the table of contents into TeX code
}
