% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process-table-of-contents.R
\name{post_process_table_of_contents}
\alias{post_process_table_of_contents}
\title{Insert the table of contents into TeX code}
\usage{
post_process_table_of_contents(
  x,
  toc_depth = NULL,
  toc_section_indent_inch = NULL,
  toc_section_num_width_inch = NULL,
  toc_subsection_indent_inch = NULL,
  toc_subsection_num_width_inch = NULL,
  toc_subsubsection_indent_inch = NULL,
  toc_subsubsection_num_width_inch = NULL,
  ...
)
}
\arguments{
\item{x}{Tex code, as a vector of lines read in from a TeX file by
\code{\link[=readLines]{readLines()}}}

\item{toc_depth}{The depth of the entries in the table of contents:
2 means 1.2, 3 means 1.2.3, 4 means 1.2.3.4, etc}

\item{toc_section_indent_inch}{Indent value in inches for section headers in
the TOC}

\item{toc_section_num_width_inch}{Width of section numbers in inches in
the TOC}

\item{toc_subsection_indent_inch}{Indent value in inches for subsection headers
in the TOC}

\item{toc_subsection_num_width_inch}{Width of subsection numbers in inches in
the TOC}

\item{toc_subsubsection_indent_inch}{Indent value in inches for subsubsection
headers in the TOC}

\item{toc_subsubsection_num_width_inch}{Width of subsubsection numbers in inches
in the TOC}

\item{...}{Absorb arguments meant for other functions}
}
\value{
The modified Tex code, as a vector
}
\description{
Insert the table of contents into TeX code
}
