% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-beta.R
\name{rbeta_ab}
\alias{rbeta_ab}
\alias{dbeta_ab}
\title{The Beta distribution}
\usage{
rbeta_ab(n, prior, sd, min, max)

dbeta_ab(x, prior, sd, min, max)
}
\arguments{
\item{n}{The number of random values you want in the returned vector.}

\item{prior}{Prior value}

\item{sd}{Standard deviation}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{x}{X value for distribution}
}
\value{
A vector of doubles, where the length of the vector depends on the
length of \code{x} value passed to \code{n}.
\code{dbeta_ab} returns density estimates for each input value and \code{rbeta_ab}
provides random deviates from the Beta distribution.
}
\description{
The Beta distribution
}
\details{
Random generation for the Beta distribution with parameters based on the
mean and standard deviation as well as the ranges as implemented in
CASAL. This formulation is also available on \href{https://en.wikipedia.org/wiki/Beta_distribution#Mean_and_variance}{Wikipedia}.
This function is a direct copy of the code available from \href{https://github.com/chantelwetzel-noaa/XSSS/blob/4dce917ee06d52416b58dd9440796ba1f7357943/R/rbeta_ab_fn.R}{Wetzel and Punt}.
}
