% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run extra models for forecasting, retrospectives, and extra MCMC (one report file per posterior)}
\usage{
run(
  model_dir = NULL,
  run_forecasts = FALSE,
  run_retrospectives = FALSE,
  run_ct_levels = FALSE,
  ...
)
}
\arguments{
\item{model_dir}{The directory the models are located in}

\item{run_forecasts}{Logical. Run forecasting?}

\item{run_retrospectives}{Logical. If \code{TRUE}, run retrospectives?}

\item{run_ct_levels}{Logical. Run catch levels estimation routines?}

\item{...}{Passed to the subroutines}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
Run extra models for forecasting, retrospectives, and extra MCMC (one report file per posterior)
}
\details{
This is a wrapper function for calling \code{\link[=run_ct_levels]{run_ct_levels()}}, \code{\link[=run_forecasts]{run_forecasts()}},
and \code{\link[=run_retrospectives]{run_retrospectives()}} functions.
}
