% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-adnuts.R
\name{run_adnuts}
\alias{run_adnuts}
\title{Run the adnuts MCMC for the model found in the directory given}
\usage{
run_adnuts(
  path,
  num_chains = NULL,
  seed = 42,
  num_samples = 8000,
  num_warmup_samples = 250,
  adapt_delta = 0.95,
  run_extra_mcmc = FALSE,
  hess_step = TRUE,
  fn_exe = ss_executable,
  overwrite = TRUE,
  fn_logfile = model_output_log_fn,
  input_files = ss_input_files
)
}
\arguments{
\item{path}{Directory where the model files reside}

\item{num_chains}{The number of chains to run in parallel. If \code{NULL},
1 less than the number of cores on the machine will be used
1 less than the number of cores on the machine will be used}

\item{seed}{The random seed used to draw the random seeds for each chain}

\item{num_samples}{The number of samples to output}

\item{num_warmup_samples}{The warmup samples (equivalent of burnin)}

\item{adapt_delta}{The target acceptance rate. See \code{\link[adnuts:samplers]{adnuts::sample_admb()}}}

\item{run_extra_mcmc}{If \code{TRUE}, run SS extra mcmc option which outputs
files into the \code{sso} subdirectory. If \code{FALSE}, those files will not be
created and the \code{posteriors.sso} and \code{dervied_posteriors.sso} files
will be in the running directory}

\item{hess_step}{Logical. If \code{TRUE}, use the \code{hess_step} algorithm`}

\item{fn_exe}{The name of the executable which was built using ADMB}

\item{overwrite}{Logical. If \code{TRUE} and in an interactive session,
don't ask user if they want to overwrite if the directory already exists,
just do it. If non-interactive, the output will be overwritten and this
will be ignored}

\item{fn_logfile}{The filename of the logfile}

\item{input_files}{The input files for SS}
}
\value{
Nothing
}
\description{
Run the adnuts MCMC for the model found in the directory given
}
\details{
\code{path} is the directory in which the MLE will be run, a subdirectory of
this called \code{mcmc} is where the MCMC will be run using the \code{NUTS}
algorithm. Inside the \code{mcmc} directory, several temporary subdirectories
will be created, one for each MCMC chain labeled \verb{chain_*}, AkA CPU
number used in the parallel execution. These will disappear once the
run has completed and the output has been merged.
}
