% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels-copy-input-files.R
\name{run_catch_levels_copy_input_files}
\alias{run_catch_levels_copy_input_files}
\title{Copy the necessary files from a models' work directories to the directory
required to run dynamic catch levels runs}
\usage{
run_catch_levels_copy_input_files(model, pth, inp_files_only = FALSE)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{pth}{The path to the directory in which the dynamic catch levels
run will occur. Example:
'/srv/hake/models/2023/01-version/01-base-models/01-base/catch-levels/spr-100'}

\item{inp_files_only}{Logical. If \code{TRUE}, on;ly copy the SS3 inputs files
which are in the package data variable \link{ss_input_files}. If \code{FALSE},
include 'derposts,sso' and 'ss.psv' as well}
}
\value{
\code{\link[base:invisible]{base::invisible()}} If the function returns, all files were copied
successfully. If they were not, an error is thrown
}
\description{
Copy the necessary files from a models' work directories to the directory
required to run dynamic catch levels runs
}
