% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-catch-levels.R
\name{run_catch_levels_default_hr}
\alias{run_catch_levels_default_hr}
\title{Run the model iteratively reducing the difference between the first and second year projections to
find a stable catch within the the given tolerance}
\usage{
run_catch_levels_default_hr(model, default_hr_path, forecast_yrs, ...)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{default_hr_path}{The path in which the default hr forecast resides}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Absorbs arguments intended for other functions}
}
\description{
Run the model iteratively reducing the difference between the first and second year projections to
find a stable catch within the the given tolerance
}
