% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels.R
\name{run_ct_levels}
\alias{run_ct_levels}
\title{A wrapper to run the catch levels determination routines}
\usage{
run_ct_levels(
  model = NULL,
  model_path = NULL,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  run_default_hr = TRUE,
  run_spr_100 = TRUE,
  run_stable_catch = TRUE,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{model_path}{The model directory name}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Passes arguments to \code{\link[=run_ct_levels_default_hr]{run_ct_levels_default_hr()}},
\code{\link[=run_ct_levels_spr_100]{run_ct_levels_spr_100()}}, and \code{\link[=run_ct_levels_stable_catch]{run_ct_levels_stable_catch()}}}
}
\value{
\code{\link[base:invisible]{base::invisible()}}
}
\description{
A wrapper to run the catch levels determination routines
}
