% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels-default-hr.R
\name{run_ct_levels_default_hr}
\alias{run_ct_levels_default_hr}
\title{Run the model iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance}
\usage{
run_ct_levels_default_hr(model, forecast_yrs, ...)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{forecast_yrs}{A vector of forecast years}

\item{...}{Absorbs arguments intended for other functions}
}
\description{
Run the model iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance
}
