% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels-default-hr.R
\name{run_ct_levels_default_hr}
\alias{run_ct_levels_default_hr}
\title{Run the model, iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance}
\usage{
run_ct_levels_default_hr(
  model,
  forecast_yrs = get_assess_yr():(get_assess_yr() + 3),
  ss_exe = NULL,
  keep_files = FALSE,
  ...
)
}
\arguments{
\item{model}{The SS3 model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{forecast_yrs}{A vector of forecast years}

\item{ss_exe}{The name of executable to use or \code{NULL} to use the package
data variable \link{ss_executable}}

\item{keep_files}{Logical. If \code{TRUE}, keep all files in the directory,
if \code{FALSE} delete all files except for the filename contained in the
the \code{forecast_fn} variable. This is 'forecast.ss' by default for SS3}

\item{...}{Absorbs arguments intended for other functions}
}
\value{
Nothing, the 'forecast.ss' file will have the catch numbers at the
end of the file under 'Catch_or_F'
}
\description{
Run the model, iteratively reducing the difference between the first and
second year projections to find a stable catch within the the given
tolerance
}
