% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ct-levels-spr-100.R
\name{run_ct_levels_spr_100}
\alias{run_ct_levels_spr_100}
\title{Run the model iteratively zoning in on a catch value that reduces the
SPR to 1, within the tolerance given (\code{ct_levels_spr_tol})}
\usage{
run_ct_levels_spr_100(
  model,
  forecast_yrs,
  ct_levels_spr_tol,
  ct_levels_catch_tol,
  ct_levels_max_iter,
  ...
)
}
\arguments{
\item{model}{The SS model output as loaded by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{forecast_yrs}{A vector if the years to forecast for}

\item{ct_levels_spr_tol}{The tolerance to be within 1 for the SPR}

\item{ct_levels_catch_tol}{Catch tolerance. If the upper and lower catch
in the algorithm are within this, it is assumed that the SPR is close
enough}

\item{ct_levels_max_iter}{The maximum number of iterations to do}

\item{...}{Not used}
}
\description{
Run the model iteratively zoning in on a catch value that reduces the
SPR to 1, within the tolerance given (\code{ct_levels_spr_tol})
}
