% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-forecasts.R
\name{run_forecasts}
\alias{run_forecasts}
\title{Run forecasting for the model supplied}
\usage{
run_forecasts(model_path, catch_levels_path, ...)
}
\arguments{
\item{model}{The SS model output as loaded by [load_ss_files()]}
}
\value{
[base::invisible()]
}
\description{
Run forecasting for the model supplied
}
\details{
If there is no mcmc component to the model, an error will be given and the program will be stopped
}
