% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save-nuts-fits.R
\name{save_nuts_fits}
\alias{save_nuts_fits}
\title{Read in NUTS run data files ('hake.Rdata' files), fix the parameter names
in the \code{nuts_updated} \code{samples} table, save the \code{nuts_updated} object as an
RDS file and return that object}
\usage{
save_nuts_fits(
  dr = "/srv/hake/models/2024/02-version/05-test-models",
  mdl_nms = c("03-burnin-1", "04-burnin-50", "05-burnin-100", "06-burnin-150",
    "07-burnin-200", "08-burnin-250", "09-burnin-300", "10-burnin-350", "11-burnin-400",
    "12-burnin-450", "13-burnin-500", "14-long-base")
)
}
\arguments{
\item{dr}{The directory in which the model subdirectories reside}

\item{mdl_nms}{A vector of the model directory names}
}
\value{
A list of the fit objects, each of which can be used in a call to
\code{\link[adnuts:launch_shinyadmb]{adnuts::launch_shinyadmb()}}
}
\description{
Read in NUTS run data files ('hake.Rdata' files), fix the parameter names
in the \code{nuts_updated} \code{samples} table, save the \code{nuts_updated} object as an
RDS file and return that object
}
