% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set-ct-levels.R
\name{set_ct_levels}
\alias{set_ct_levels}
\title{Reads in the forecast description data and creates the list containing
the catch levels and extra information used for the assessment document}
\usage{
set_ct_levels(
  fn = here(doc_path, forecast_descriptions_fn),
  inc_fi_and_stable_catch = FALSE
)
}
\arguments{
\item{fn}{The name of the file containing the forecast description data}

\item{inc_fi_and_stable_catch}{Logical. If \code{TRUE}, include the Fishing
intensity = 100\% and the Stable Catch scenarios}
}
\value{
A list of 2 lists:
List 1 is a list of lists of length-3 vectors. Each list represents a catch
level, each vector's three elements are:
1 - A vector of catch levels
2 - The nice name for the catch level scenario
3 - The directory name for the catch level scenario
List 2 is a list of important values and indices referenced in the document.
Be sure to update this each year if forecasts are added and/or removed
}
\description{
Reads in the forecast description data and creates the list containing
the catch levels and extra information used for the assessment document
}
