% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set-dirs.R
\name{set_dirs}
\alias{set_dirs}
\title{Set and return all model directories for the project}
\usage{
set_dirs(
  models_dir = NA,
  last_yr_models_dir = NA,
  base_models_dir = "01-base-models",
  bridge_models_dir = "02-bridging-models",
  sens_models_dir = "03-sensitivity-models",
  request_models_dir = "04-request-models",
  test_models_dir = "05-test-models",
  base_models_dirs = NA,
  bridge_models_dirs = NA,
  sens_models_dirs = NA,
  request_models_dirs = NA,
  test_models_dirs = NA,
  prepend_to_bridge = NA,
  suppress_warnings = FALSE
)
}
\arguments{
\item{models_dir}{Full path in which the SS3 model directories are
located for the current assessment year}

\item{last_yr_models_dir}{Full path in which the SS3 model directories are
located for the the last assessment year}

\item{base_models_dir}{Name of the base models directory}

\item{bridge_models_dir}{Name of the bridging models directory}

\item{sens_models_dir}{Name of the sensitivity models directory}

\item{request_models_dir}{Name of the request models directory}

\item{test_models_dir}{Name of the test models directory}

\item{base_models_dirs}{A vector of subdirectory names in
\code{base_models_dir} that each contain an individual SS3 base model}

\item{bridge_models_dirs}{A vector of subdirectory names in
\code{bridge_models_dir} that each contain an individual SS3 bridge model}

\item{sens_models_dirs}{A vector of subdirectory names in \code{sens_models_dir}
that each contain an individual SS3 sensitivity model}

\item{request_models_dirs}{A vector of subdirectory names in
\code{request_models_dir} that each contain an individual SS3 base model request
model}

\item{test_models_dirs}{A vector of subdirectory names in \code{test_models_dir}
that each contain an individual SS3 base model test model}

\item{prepend_to_bridge}{A vector of logical values, the same length as the
number of groups of bridge models (\code{length(bridge_models_dirs)}) which,
if \code{TRUE} will prepend the \verb{last_yr_base_model_dir``to the beginning of the group of that element number. If }NULL\code{or}NA`, no prepending will
take place}

\item{suppress_warnings}{If \code{TRUE}, warnings about directories not existing
will not be shown}
}
\value{
A list of twelve vectors of full paths, which will have \code{NA}
elements for those which do not exist:
\enumerate{
\item Full path of \code{models_dir}
\item Full path of \code{last_yr_base_model_dir}
\item Full path of \code{base_model_dir}
\item Full path of \code{bridge_models_dir}
\item Full path of \code{sens_models_dir}
\item Full path of \code{request_models_dir}
\item Full path of \code{test_models_dir}
\item A vector of the base model directories
\item A vector of the bridge model directories
\item A vector of the sensitivity model directories
\item A vector of the request model directories
\item A vector of the test model directories
}
}
\description{
Set and return all model directories for the project
}
\details{
Sensitivity groups will have the base model directory prepended,
The bridge model groups can have \code{last_yr_base_model_dir}
prepended if \code{prepend_to_bridge} is set to \code{TRUE} for the model group.
If directories do not exist, they will be assigned \code{NULL} and a warning
issued. If the \code{models_dir} does not exist, an error will be thrown.
}
