% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_df}
\alias{split_df}
\title{Split the posteriors data frame in a list of N data frames based on
\code{from_to}}
\usage{
split_df(df, from_to)
}
\arguments{
\item{df}{A data frame containing rows to split}

\item{from_to}{A data frame with the columns \code{from} and \code{to} which
indicate the rows from and to for each data frame}
}
\value{
A list of data frames of the same length as the number of
rows in \code{from_to}
}
\description{
Split the posteriors data frame in a list of N data frames based on
\code{from_to}
}
