% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stop-quietly.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop the program without issuing an error message}
\usage{
stop_quietly(...)
}
\arguments{
\item{...}{Arguments passed in to be written out as a message}
}
\value{
Nothing
}
\description{
Stop the program without issuing an error message
}
