% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_history_df}
\alias{survey_history_df}
\title{A data frame containing information about the history of all acoustic
surveys done}
\format{
A data frame with 14 rows and 5 variables:
\describe{
\item{year}{Year of survey}
\item{start.date}{Start date of survey}
\item{end.date}{End date of survey}
\item{vessels}{A list of the vessels involved in the survey}
\item{hauls.with.samples}{The number of hauls with samples taken}
}
}
\usage{
survey_history_df
}
\description{
A data frame containing information about the history of all acoustic
surveys done
}
\keyword{datasets}
