% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of total catch}
\usage{
table_catch(
  ct,
  country = c("both", "can", "us"),
  start_yr,
  end_yr,
  inc_foreign_jv = FALSE,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{ct}{Data frame as in the file \code{landing-tac-history.csv}}

\item{country}{One of "both", "can", or "us"}

\item{start_yr}{Start year in table}

\item{end_yr}{End year in table}

\item{inc_foreign_jv}{Logical. If \code{TRUE}, include columns for the U.S.
foreign fleets and Canadian foreign fleet}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a table of total catch
}
