% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-decision.R
\name{table_decision}
\alias{table_decision}
\title{Create the Executive Summary decision tables. \code{rows_to_label} and
\code{row_labels} need changing each year.}
\usage{
table_decision(
  model,
  type = c("biomass", "spr"),
  forecast_inds = seq_along(model$forecasts[[length(model$forecasts)]]),
  rows_to_show = NULL,
  inc_fi_stable_catch = FALSE,
  bold_letters = TRUE,
  digits = 2,
  left_col_cm = 1,
  right_cols_cm = 1,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ret_df = FALSE,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{type}{One of \code{biomass} or \code{spr}}

\item{forecast_inds}{The indices to use in the \code{forecasts} object which
itself is an object of the \code{model} list}

\item{rows_to_show}{A vector of letters of rows to show. If \code{NULL}, show}

\item{inc_fi_stable_catch}{Logical. If \code{TRUE}, include the two rows of
forecasts for fishing intensity equal to 100\% and stable catch for the first
two years}

\item{bold_letters}{Logical. If \code{TRUE}, the letters labelling each row
will be boldface}

\item{digits}{The number of decimal points to show in the table}

\item{left_col_cm}{Number of centimeters wide to make the leftmost column.
This needs to be changed when the font size is changed}

\item{right_cols_cm}{Number of centimeters wide to make the 3 rightmost
columns wide. This needs to be changed when the font size is changed}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{ret_df}{Logical. If \code{TRUE}, return a data frame (tibble()) instead
of the \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create the Executive Summary decision tables. \code{rows_to_label} and
\code{row_labels} need changing each year.
}
