% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-landings-tac.R
\name{table_landings_tac}
\alias{table_landings_tac}
\title{Create a summary table of landings and TACS}
\usage{
table_landings_tac(
  ct,
  start_yr,
  end_yr,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{ct}{Data frame as in the file \code{landing-tac-histroy.csv}}

\item{start_yr}{Start year for the table}

\item{end_yr}{End year for the table. If past data range, the last year
in \code{d} will be used}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Create a summary table of landings and TACS
}
