% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est.R
\name{table_param_est}
\alias{table_param_est}
\title{Returns an [xtable::xtable()] in the proper format for the MCMC posterior
parameter estimates for all the models in the list, one for each column}
\usage{
table_param_est(
  models,
  model_nms,
  end_yr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font_size = 9,
  space_size = 10,
  rec_yrs = NA,
  show_like = TRUE
)
}
\arguments{
\item{models}{A list of models which contain the MCMC output}

\item{model_nms}{A vector of names of the same length as the number of
models in the models list}

\item{end_yr}{The last year to include}

\item{digits}{Number of decimal points for the estimates}

\item{xcaption}{Caption to appear in the calling document}

\item{xlabel}{The label used to reference the table in latex}

\item{font_size}{Size of the font for the table}

\item{space_size}{Size of the vertical spaces for the table}

\item{rec_yrs}{A vector of integers supplying the years for which you
want estimates of recruitment}

\item{show_like}{If `TRUE`, include the negative log-likelihoods (set to
`FALSE` for presentations)}
}
\value{
as [xtable::xtable()]
}
\description{
Returns an [xtable::xtable()] in the proper format for the MCMC posterior
parameter estimates for all the models in the list, one for each column
}
