% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-reference-points.R
\name{table_reference_points}
\alias{table_reference_points}
\title{Creates a table containing reference point estimates}
\usage{
table_reference_points(
  model,
  category_bold = TRUE,
  category_italics = TRUE,
  category_underline = TRUE,
  category_line_above = FALSE,
  category_line_below = FALSE,
  font_size = 10,
  header_font_size = 10,
  header_vert_spacing = 12,
  header_vert_scale = 1.2,
  ...
)
}
\arguments{
\item{model}{A model, created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{category_bold}{Logical. If \code{TRUE}, make the category header lines
boldface}

\item{category_underline}{Logical. If \code{TRUE}, make the category header lines
underlined}

\item{category_line_above}{Logical. If \code{TRUE}, place a horizontal line above
category header lines}

\item{category_line_below}{Logical. If \code{TRUE}, place a horizontal line below
category header lines}

\item{font_size}{The table data and header font size in points}

\item{header_font_size}{The font size for the headers only. If \code{NULL},
the headers will have the same font size as the table cell data}

\item{header_vert_scale}{Scale factor to create the vertical spacing value.
See \code{header_vert_spacing}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{hdr_italics}{Logical. If \code{TRUE}, make the cohort header lines
italicized}

\item{vert_spacing}{The vertical spacing between newlines for this font.
If \code{NULL} this will be calculated as \code{header_font_size * header_vert_scale}}
}
\value{
An \code{\link[knitr:kable]{knitr::kable()}} object
}
\description{
Creates a table containing reference point estimates
}
