% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vec2df.R
\name{vec2df}
\alias{vec2df}
\title{Create a one-row data frame from a vector of values}
\usage{
vec2df(vec, nms = NULL)
}
\arguments{
\item{vec}{A vector}

\item{nms}{Optional column names for the new data frame}
}
\value{
A one row \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create a one-row data frame from a vector of values
}
\details{
If \code{nms} is not given, the names of the vector (if \code{vec} is a named vector)
will be used for the column names
}
