% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age-.R
\name{write_wtatage_file}
\alias{write_wtatage_file}
\title{Write weight-at-age file}
\usage{
write_wtatage_file(
  file = paste0("wtatage_", format(Sys.time(), "\%d-\%b-\%Y_\%H.\%M"), ".ss"),
  data,
  maturity,
  n_fleets = 2
)
}
\arguments{
\item{file}{A filename that you want to save the information as. The default
is to have an extension of \code{.ss} such that the file can be used for
Stock Synthesis. The file path can either be relative or absolute.}

\item{data}{Weight at age matrix.}

\item{maturity}{A vector of maturity at age.}

\item{n_fleets}{An integer specifying the number of fleets that are in the
model. The default is two, e.g., one fishery and one survey. When the
age-1 index is also included in the model, then this value should be
changed to 3, e.g., one fishery, age 2+ survey, and relative age-1 index
of abundance.}
}
\description{
Write weight-at-age file
}
